Move shapes to common_shapes subpackage.
* Dynamically import all shapes defined in common_shapes
* Move all damansceno stuff to damansceno.py
