--- euclid/FreudShape/Tetrahedron.py
@@ -1,12 +1,0 @@
-from __future__ import division
-from numpy import sqrt
-import numpy
-from euclid.FreudShape import ConvexPolyhedron
-
-# Example:
-# from euclid.FreudShape.Tetrahedron import shape
-
-points = [ (1,1,1), (-1,-1,1), (1,-1,-1), (-1,1,-1) ]
-
-shape = ConvexPolyhedron(numpy.array(points))
-
