+++ euclid/MathematicaToPython.py
