+++ euclid/damasceno.py
@@ -6,21 +6,26 @@ import json
 import collections
 
 Shape = collections.namedtuple(
-    'Shape', ['id', 'shape_class', 'ShortName', 'Name', 'AlternativeName', 'vertices', 'Comment'])
+    'Shape', ['id', 'shape_class', 'ShortName', 'Name',
+              'AlternativeName', 'vertices', 'Comment'])
 
 SPHERE = Shape(0, 'sphere', 'SPH', 'Sphere', '', [], '')
+
 
 def reserved(_id):
     return Shape(_id, 'RESERVED', 'RES', 'RESERVED', '', [], '')
 
+
 # 10.1126/science.1220869
-DAMASCENO_SCIENCE_2012="""{"1": {"AlternativeName": null, "Comment": null, "Name": "Acute Golden Rhombohedron", "ShortName": "O19", "vertices": [[-0.947214, -0.58541, -0.425325], [-0.5, -0.309017, 0.425325], [-0.5, 0.309017, -0.425325], [-0.0527864, 0.58541, 0.425325], [0.0527864, -0.58541, -0.425325], [0.5, -0.309017, 0.425325], [0.5, 0.309017, -0.425325], [0.947214, 0.58541, 0.425325]]}, "2": {"AlternativeName": null, "Comment": "edge length=8", "Name": "Square Antiprism", "ShortName": "O08", "vertices": [[-0.5, -0.5, -0.420448208], [-0.5, 0.5, -0.420448208], [0.0, -0.707106781, 0.420448208], [0.0, 0.707106781, 0.420448208], [0.5, -0.5, -0.420448208], [0.5, 0.5, -0.420448208], [-0.707106781, 0.0, 0.420448208], [0.707106781, 0.0, 0.420448208]]}, "3": {"AlternativeName": null, "Comment": "edge length=9", "Name": "Pentagonal Antiprism", "ShortName": "O09", "vertices": [[-0.850650808, 0.0, -0.425325404], [0.850650808, 0.0, 0.425325404], [-0.68819096, -0.5, 0.425325404], [-0.68819096, 0.5, 0.425325404], [-0.262865556, -0.809016994, -0.425325404], [-0.262865556, 0.809016994, -0.425325404], [0.262865556, -0.809016994, 0.425325404], [0.262865556, 0.809016994, 0.425325404], [0.68819096, -0.5, -0.425325404], [0.68819096, 0.5, -0.425325404]]}, "4": {"AlternativeName": null, "Comment": "edge length=10", "Name": "Hexagonal Antiprism", "ShortName": "O10", "vertices": [[-1.0, 0.0, 0.427799839], [-0.5, -0.866025404, 0.427799839], [-0.5, 0.866025404, 0.427799839], [0.0, -1.0, -0.427799839], [0.0, 1.0, -0.427799839], [0.5, -0.866025404, 0.427799839], [0.5, 0.866025404, 0.427799839], [1.0, 0.0, 0.427799839], [-0.866025404, -0.5, -0.427799839], [-0.866025404, 0.5, -0.427799839], [0.866025404, -0.5, -0.427799839], [0.866025404, 0.5, -0.427799839]]}, "5": {"AlternativeName": null, "Comment": "edge length=11", "Name": "Heptagonal Antiprism", "ShortName": "O11", "vertices": [[-1.038260698, -0.5, 0.429236598], [-1.038260698, 0.5, 0.429236598], [1.038260698, -0.5, -0.429236598], [1.038260698, 0.5, -0.429236598], [-1.152382435, 0.0, -0.429236598], [1.152382435, 0.0, 0.429236598], [-0.256429216, -1.123489802, 0.429236598], [-0.256429216, 1.123489802, 0.429236598], [0.256429216, -1.123489802, -0.429236598], [0.256429216, 1.123489802, -0.429236598], [-0.718498696, -0.900968868, -0.429236598], [-0.718498696, 0.900968868, -0.429236598], [0.718498696, -0.900968868, 0.429236598], [0.718498696, 0.900968868, 0.429236598]]}, "6": {"AlternativeName": null, "Comment": "edge length=12", "Name": "Octagonal Antiprism", "ShortName": "O12", "vertices": [[-0.5, -1.207106781, -0.430147785], [-0.5, 1.207106781, -0.430147785], [0.0, -1.306562965, 0.430147785], [0.0, 1.306562965, 0.430147785], [0.5, -1.207106781, -0.430147785], [0.5, 1.207106781, -0.430147785], [-1.207106781, -0.5, -0.430147785], [-1.207106781, 0.5, -0.430147785], [1.207106781, -0.5, -0.430147785], [1.207106781, 0.5, -0.430147785], [-1.306562965, 0.0, 0.430147785], [1.306562965, 0.0, 0.430147785], [-0.923879533, -0.923879533, 0.430147785], [-0.923879533, 0.923879533, 0.430147785], [0.923879533, -0.923879533, 0.430147785], [0.923879533, 0.923879533, 0.430147785]]}, "7": {"AlternativeName": null, "Comment": "edge length=13", "Name": "Nonagonal Antiprism", "ShortName": "O13", "vertices": [[-1.37373871, -0.5, 0.430763044], [-1.37373871, 0.5, 0.430763044], [1.37373871, -0.5, -0.430763044], [1.37373871, 0.5, -0.430763044], [-1.4619022, 0.0, -0.430763044], [-0.7309511, -1.266044443, 0.430763044], [-0.7309511, 1.266044443, 0.430763044], [0.7309511, -1.266044443, -0.430763044], [0.7309511, 1.266044443, -0.430763044], [1.4619022, 0.0, 0.430763044], [-1.119882057, -0.939692621, -0.430763044], [-1.119882057, 0.939692621, -0.430763044], [1.119882057, -0.939692621, 0.430763044], [1.119882057, 0.939692621, 0.430763044], [-0.253856653, -1.439692621, -0.430763044], [-0.253856653, 1.439692621, -0.430763044], [0.253856653, -1.439692621, 0.430763044], [0.253856653, 1.439692621, 0.430763044]]}, "8": {"AlternativeName": null, "Comment": "edge length=14", "Name": "Decagonal Antiprism", "ShortName": "O14", "vertices": [[0.0, -1.618033989, -0.431198502], [0.0, 1.618033989, -0.431198502], [-1.618033989, 0.0, 0.431198502], [1.618033989, 0.0, 0.431198502], [-1.309016994, -0.951056516, 0.431198502], [-1.309016994, 0.951056516, 0.431198502], [-0.5, -1.538841769, 0.431198502], [-0.5, 1.538841769, 0.431198502], [-0.951056516, -1.309016994, -0.431198502], [-0.951056516, 1.309016994, -0.431198502], [0.951056516, -1.309016994, -0.431198502], [0.951056516, 1.309016994, -0.431198502], [-1.538841769, -0.5, -0.431198502], [-1.538841769, 0.5, -0.431198502], [1.538841769, -0.5, -0.431198502], [1.538841769, 0.5, -0.431198502], [0.5, -1.538841769, 0.431198502], [0.5, 1.538841769, 0.431198502], [1.309016994, -0.951056516, 0.431198502], [1.309016994, 0.951056516, 0.431198502]]}, "9": {"AlternativeName": null, "Comment": null, "Name": "Augmented Dodecahedron", "ShortName": "J58", "vertices": [[0.047386905, 0.0, 1.339229524], [0.047386905, 0.0, -1.463290476], [0.225797905, -1.30902, 0.405055524], [0.225797905, 1.30902, 0.405055524], [0.514472905, -0.809017, -1.106470476], [0.514472905, 0.809017, -1.106470476], [1.091826905, -0.809017, 0.405055524], [1.091826905, 0.809017, 0.405055524], [-1.175463095, -0.5, 0.405055524], [-1.175463095, 0.5, 0.405055524], [1.270236905, -0.5, -0.529116476], [1.270236905, 0.5, -0.529116476], [-0.886785095, 0.0, -1.106470476], [-0.419699095, -0.809017, 0.982409524], [-0.419699095, 0.809017, 0.982409524], [0.981558905, 0.0, 0.982409524], [-0.997053095, -0.809017, -0.529116476], [-0.997053095, 0.809017, -0.529116476], [-0.947738095, 0.0, 1.240609524], [-0.131024095, -1.30902, -0.529116476], [-0.131024095, 1.30902, -0.529116476]]}, "10": {"AlternativeName": null, "Comment": null, "Name": "Augmented Hexagonal Prism", "ShortName": "J54", "vertices": [[-1.104797692, -0.060505077, -0.5], [-1.104797692, -0.060505077, 0.5], [-0.604797692, -0.926530077, -0.5], [-0.604797692, -0.926530077, 0.5], [-0.604797692, 0.805519923, -0.5], [-0.604797692, 0.805519923, 0.5], [0.395202308, -0.926530077, -0.5], [0.395202308, -0.926530077, 0.5], [0.395202308, 0.805519923, -0.5], [0.395202308, 0.805519923, 0.5], [0.895202308, -0.060505077, -0.5], [0.895202308, -0.060505077, 0.5], [1.257572308, 0.726060923, 0.0]]}, "11": {"AlternativeName": null, "Comment": null, "Name": "Augmented Pentagonal Prism", "ShortName": "J52", "vertices": [[0.748030818, -0.074557727, -0.5], [0.748030818, -0.074557727, 0.5], [0.160245818, -0.883574727, -0.5], [0.160245818, -0.883574727, 0.5], [0.160245818, 0.734459273, -0.5], [0.160245818, 0.734459273, 0.5], [-0.790811182, -0.574557727, -0.5], [-0.790811182, -0.574557727, 0.5], [-0.790811182, 0.425442273, -0.5], [-0.790811182, 0.425442273, 0.5], [1.026199818, 0.745577273, 0.0]]}, "12": {"AlternativeName": null, "Comment": null, "Name": "Augmented Sphenocorona", "ShortName": "J87", "vertices": [[-0.802793436, -0.741304536, 0.404638136], [-0.734923436, 0.054175364, -0.197521864], [-0.498633436, 0.138255464, 0.770518136], [-0.215103436, 0.845905464, 0.123338136], [-0.198633436, 0.595905464, -0.844761864], [-0.085183636, -0.700254536, -0.290601864], [0.151106564, -0.616184536, 0.677438136], [0.451106564, -0.158534536, -0.937841864], [0.473206564, 0.329465464, 0.632768136], [0.649110564, 0.597815464, -0.314361864], [0.810740564, -0.345244536, -0.023611364]]}, "13": {"AlternativeName": null, "Comment": null, "Name": "Augmented Triangular Prism", "ShortName": "J49", "vertices": [[-0.359802286, -0.623196, -0.5], [-0.359802286, -0.623196, 0.5], [-0.359802286, 0.376804, -0.5], [-0.359802286, 0.376804, 0.5], [0.506222714, -0.123196, -0.5], [0.506222714, -0.123196, 0.5], [0.426763714, 0.739176, 0.0]]}, "14": {"AlternativeName": null, "Comment": null, "Name": "Augmented Tridiminished Icosahedron", "ShortName": "J64", "vertices": [[0.0, 0.0, -1.4685573], [-0.57735, 0.0, -0.6520583], [-0.288675, -0.5, 0.8594637], [-0.288675, 0.5, 0.8594637], [0.288675, -0.5, -0.6520583], [0.288675, 0.5, -0.6520583], [0.57735, 0.0, 0.8594637], [-0.934172, 0.0, 0.2821137], [0.467086, -0.809017, 0.2821137], [0.467086, 0.809017, 0.2821137]]}, "15": {"AlternativeName": null, "Comment": null, "Name": "Augmented Truncated Cube", "ShortName": "J66", "vertices": [[-0.5, 1.20711, -1.480568571], [-0.5, 1.20711, 0.933651429], [-0.5, -1.20711, -1.480568571], [-0.5, -1.20711, 0.933651429], [0.0, -0.707107, 1.640751429], [0.0, 0.707107, 1.640751429], [0.5, 1.20711, -1.480568571], [0.5, 1.20711, 0.933651429], [0.5, -1.20711, -1.480568571], [0.5, -1.20711, 0.933651429], [-0.707107, 0.0, 1.640751429], [0.707107, 0.0, 1.640751429], [1.20711, -0.5, -1.480568571], [1.20711, -0.5, 0.933651429], [1.20711, 0.5, -1.480568571], [1.20711, 0.5, 0.933651429], [1.20711, 1.20711, -0.773458571], [1.20711, 1.20711, 0.226541429], [1.20711, -1.20711, -0.773458571], [1.20711, -1.20711, 0.226541429], [-1.20711, -0.5, -1.480568571], [-1.20711, -0.5, 0.933651429], [-1.20711, 0.5, -1.480568571], [-1.20711, 0.5, 0.933651429], [-1.20711, 1.20711, -0.773458571], [-1.20711, 1.20711, 0.226541429], [-1.20711, -1.20711, -0.773458571], [-1.20711, -1.20711, 0.226541429]]}, "16": {"AlternativeName": null, "Comment": null, "Name": "Augmented Truncated Dodecahedron", "ShortName": "J68", "vertices": [[-3.133096349, -0.186728263, 0.418858126], [-3.115256349, -0.019748363, -0.566939874], [-2.877166349, 0.191938737, 1.308305126], [-2.830456349, 0.629098737, -1.272554874], [-2.801726349, -0.791431263, 1.143105126], [-2.755006349, -0.354271263, -1.437754874], [-2.445226349, 0.971615737, 1.761655126], [-2.387476349, 1.511970737, -1.428454874], [-2.247706349, -1.602879263, 1.329155126], [-2.189956349, -1.062519263, -1.860964874], [-2.002256349, 1.854490737, 1.605755126], [-1.955536349, 2.291650737, -0.975100874], [-1.717456349, 2.503340737, 0.900143126], [-1.699606349, 2.670320737, -0.085655874], [-1.682656349, -2.311129263, 0.905943126], [-1.670886349, 1.249790737, 2.329995126], [-1.635936349, -1.873969263, -1.674914874], [-1.595286349, 1.957130737, -1.845914874], [-1.426736349, -1.932459263, 1.795385126], [-1.351136349, -1.225119263, -2.380524874], [-1.322416349, -2.645649263, 0.035129326], [-1.304566349, -2.478669263, -0.950669874], [-0.925264349, 2.948490737, 0.482683126], [-0.849908349, 0.920206737, 2.796235126], [-0.756471349, 1.794520737, -2.365474874], [-0.652387349, -1.654289263, 2.363725126], [-0.558949349, -0.779969263, -2.797984874], [-0.483593349, -2.808259263, -0.484431874], [-0.295890349, 0.108757337, 2.982285126], [-0.220444349, -0.874612263, 2.817085126], [-0.191424349, 1.086280737, -2.788684874], [-0.115977649, 0.102906237, -2.953884874], [0.071725651, 3.019920737, 0.512829126], [0.147081651, 0.991632737, 2.826375126], [0.240519651, 1.865950737, -2.335324874], [0.344602651, -1.582859263, 2.393875126], [0.438040651, -0.708543263, -2.767834874], [0.513396651, -2.736829263, -0.454285874], [0.892699651, 2.690340737, 0.979067126], [0.910544651, 2.857320737, -0.006731884], [0.939271651, 1.436790737, 2.408915126], [1.014863651, 2.144130737, -1.766994874], [1.183423651, -1.745459263, 1.874315126], [1.224073651, 2.085630737, 1.703305126], [1.259013651, -1.038129263, -2.301594874], [1.270793651, 2.522790737, -0.877545874], [1.287743651, -2.458659263, 0.114053126], [1.305583651, -2.291679263, -0.871745874], [1.543663651, -2.079989263, 1.003495126], [1.590383651, -1.642829263, -1.577354874], [1.778093651, 1.274180737, 1.889355126], [1.835833651, 1.814540737, -1.300754874], [1.975613651, -1.300309263, 1.456855126], [2.033363651, -0.759951263, -1.733254874], [2.145733651, -1.954989263, 0.214898126], [2.163583651, -1.788009263, -0.770900874], [2.343133651, 0.565934737, 1.466155126], [2.389853651, 1.003094737, -1.114704874], [2.418583651, -0.417435263, 1.300945126], [2.465303651, 0.019724807, -1.279904874], [2.577683651, -1.175319263, 0.668251126], [2.606553651, -0.905136263, -0.926804874], [2.703383651, 0.231411737, 0.595338126], [2.721233651, 0.398391737, -0.390460874], [2.862483651, -0.526468263, -0.037359274]]}, "17": {"AlternativeName": null, "Comment": null, "Name": "Augmented Truncated Tetrahedron", "ShortName": "J65", "vertices": [[0.0, -1.0, -0.326598], [0.0, 1.0, -0.326598], [-0.57735, -1.0, 0.489898], [-0.57735, 1.0, 0.489898], [-0.288675, -0.5, 1.306394], [-0.288675, 0.5, 1.306394], [0.288675, -0.5, -1.143096], [0.57735, 0.0, 1.306394], [1.1547, 0.0, 0.489898], [-0.866025, -0.5, -0.326598], [-0.866025, 0.5, -0.326598], [0.866025, -0.5, -0.326598], [0.866025, 0.5, -0.326598], [-0.57735, 0.0, -1.143096], [0.288675, 0.5, -1.143096]]}, "18": {"AlternativeName": null, "Comment": null, "Name": "Biaugmented Pentagonal Prism", "ShortName": "J53", "vertices": [[0.8728575, -0.068344583, -0.5], [0.8728575, -0.068344583, 0.5], [0.2850725, -0.877361583, -0.5], [0.2850725, -0.877361583, 0.5], [0.2850725, 0.740672417, -0.5], [0.2850725, 0.740672417, 0.5], [-0.6659845, -0.568344583, -0.5], [-0.6659845, -0.568344583, 0.5], [-0.6659845, 0.431655417, -0.5], [-0.6659845, 0.431655417, 0.5], [-1.3730935, -0.068344583, 0.0], [1.1510265, 0.751790417, 0.0]]}, "19": {"AlternativeName": null, "Comment": null, "Name": "Biaugmented Triangular Prism", "ShortName": "J50", "vertices": [[-0.226438625, -0.6077965, -0.5], [-0.226438625, -0.6077965, 0.5], [-0.226438625, 0.3922035, -0.5], [-0.226438625, 0.3922035, 0.5], [0.639586375, -0.1077965, -0.5], [0.639586375, -0.1077965, 0.5], [-0.933545625, -0.1077965, 0.0], [0.560127375, 0.7545755, 0.0]]}, "20": {"AlternativeName": null, "Comment": null, "Name": "Biaugmented Truncated Cube", "ShortName": "J67", "vertices": [[-0.5, 1.20711, 1.20711], [-0.5, 1.20711, -1.20711], [-0.5, -1.20711, 1.20711], [-0.5, -1.20711, -1.20711], [0.0, -0.707107, 1.91421], [0.0, -0.707107, -1.91421], [0.0, 0.707107, 1.91421], [0.0, 0.707107, -1.91421], [0.5, 1.20711, 1.20711], [0.5, 1.20711, -1.20711], [0.5, -1.20711, 1.20711], [0.5, -1.20711, -1.20711], [-0.707107, 0.0, 1.91421], [-0.707107, 0.0, -1.91421], [0.707107, 0.0, 1.91421], [0.707107, 0.0, -1.91421], [1.20711, -0.5, 1.20711], [1.20711, -0.5, -1.20711], [1.20711, 0.5, 1.20711], [1.20711, 0.5, -1.20711], [1.20711, 1.20711, -0.5], [1.20711, 1.20711, 0.5], [1.20711, -1.20711, -0.5], [1.20711, -1.20711, 0.5], [-1.20711, -0.5, 1.20711], [-1.20711, -0.5, -1.20711], [-1.20711, 0.5, 1.20711], [-1.20711, 0.5, -1.20711], [-1.20711, 1.20711, -0.5], [-1.20711, 1.20711, 0.5], [-1.20711, -1.20711, -0.5], [-1.20711, -1.20711, 0.5]]}, "21": {"AlternativeName": null, "Comment": null, "Name": "Bigyrate Diminished Rhombicosidodecahedron", "ShortName": "J79", "vertices": [[-2.304394967, 0.39941876, 0.359382255], [-2.287104967, 0.10787476, -0.597019745], [-1.982764967, 0.17542776, 1.279370255], [-1.937504967, -0.58784224, -1.224519745], [-1.854394967, 1.29045576, 0.419002255], [-1.826424967, 0.81872776, -1.128489745], [-1.558994967, 1.54960576, -0.500551745], [-1.532764967, 1.06646576, 1.338990255], [-1.476824967, 0.12300976, -1.755989745], [-1.445074967, -0.47853924, 1.811550255], [-1.389124967, -1.42199424, -1.283429745], [-1.079334967, 1.91085576, 0.299087255], [-1.026824967, 1.01404576, -1.696369745], [-0.896693967, -1.31269424, 1.752630255], [-0.851431967, -2.07596424, -0.751257745], [-0.759390967, 1.74491576, -1.068429745], [-0.757702967, 1.68686576, 1.219080255], [-0.716957967, 0.96318676, 1.908020255], [-0.662758967, 0.00832236, 2.200070255], [-0.643722967, -0.27181124, -2.143369745], [-0.589524967, -1.22667424, -1.851309745], [-0.547090967, -2.00840424, 1.125130255], [-0.529802967, -2.29995424, 0.168733255], [-0.279728967, 2.10616576, -0.268791745], [-0.193722967, 0.61922476, -2.083749745], [-0.114379167, -0.82582924, 2.141160255], [-0.051829567, -1.88064424, -1.319139745], [0.238989033, 1.80180576, -1.067719745], [0.240678033, 1.74374576, 1.219780255], [0.245260033, -0.72512024, -2.078419745], [0.281423033, 1.02007576, 1.908720255], [0.335621033, 0.06520696, 2.200780255], [0.451289033, -1.95152424, 1.125840255], [0.468577033, -2.24306424, 0.169440255], [0.536082033, 2.00289576, 0.300231255], [0.588591033, 1.10608576, -1.695219745], [0.695261033, 0.16591476, -2.018799745], [0.718720033, -1.22065424, 1.753780255], [0.763981033, -1.98392424, -0.750113745], [0.782956033, -1.37908424, -1.546239745], [1.054805033, 1.69852576, -0.498701745], [1.081025033, 1.21538576, 1.340840255], [1.168725033, -0.32961324, 1.813400255], [1.250895033, -1.75620424, 0.557962255], [1.376425033, 1.47453576, 0.421289255], [1.404405033, 1.00280576, -1.126199745], [1.511075033, 0.06263816, -1.449769745], [1.518325033, -1.02533424, 1.185900255], [1.546295033, -1.49705424, -0.361590745], [1.565265033, -0.89222624, -1.157719745], [1.629405033, 0.38123876, 1.281930255], [1.924805033, 0.64038576, 0.362376255], [1.942095033, 0.34884176, -0.594025745], [1.979005033, -0.31447924, 0.654430255], [1.996295033, -0.60602224, -0.301970745]]}, "22": {"AlternativeName": null, "Comment": null, "Name": "Bilunabirotunda", "ShortName": "J91", "vertices": [[-1.068949879, -0.081979664, 0.071567697], [-1.020379879, 0.821387236, 0.497668897], [-0.554082879, 0.043835936, 0.919549897], [-0.550126879, -0.915152764, -0.119827103], [-0.416538879, -0.101788764, -0.686033103], [-0.337937879, 1.359890236, 0.003412297], [-0.035254679, -0.789335764, 0.728153897], [0.035255221, 0.789337236, -0.728157103], [0.337934121, -1.359889764, -0.003409163], [0.416540121, 0.101790236, 0.686032897], [0.550126121, 0.915156236, 0.119827897], [0.554085121, -0.043839464, -0.919552103], [1.020370121, -0.821390764, -0.497667103], [1.068960121, 0.081979836, -0.071567803]]}, "23": {"AlternativeName": null, "Comment": null, "Name": "Cube", "ShortName": "P03", "vertices": [[-0.5, -0.5, -0.5], [-0.5, -0.5, 0.5], [-0.5, 0.5, -0.5], [-0.5, 0.5, 0.5], [0.5, -0.5, -0.5], [0.5, -0.5, 0.5], [0.5, 0.5, -0.5], [0.5, 0.5, 0.5]]}, "24": {"AlternativeName": null, "Comment": null, "Name": "Cuboctahedron", "ShortName": "A01", "vertices": [[-1.0, 0.0, 0.0], [-0.5, -0.5, -0.707107], [-0.5, -0.5, 0.707107], [-0.5, 0.5, -0.707107], [-0.5, 0.5, 0.707107], [0.0, -1.0, 0.0], [0.0, 1.0, 0.0], [0.5, -0.5, -0.707107], [0.5, -0.5, 0.707107], [0.5, 0.5, -0.707107], [0.5, 0.5, 0.707107], [1.0, 0.0, 0.0]]}, "25": {"AlternativeName": null, "Comment": null, "Name": "Deltoidal Hexecontahedron", "ShortName": "C09", "vertices": [[-0.0, 0.0, -2.77775], [-0.0, 0.0, 2.77775], [-0.0, -2.77775, 0.0], [-0.0, 2.77775, 0.0], [-0.0, -2.42408, -1.49817], [-0.0, -2.42408, 1.49817], [-0.0, -0.973249, -2.548], [-0.0, -0.973249, 2.548], [-0.0, 0.973249, -2.548], [-0.0, 0.973249, 2.548], [-0.0, 2.42408, -1.49817], [-0.0, 2.42408, 1.49817], [-2.77775, 0.0, 0.0], [-1.38888, -2.24725, 0.858373], [-1.38888, -2.24725, -0.858373], [-1.38888, 2.24725, 0.858373], [-1.38888, 2.24725, -0.858373], [1.38888, -2.24725, 0.858373], [1.38888, -2.24725, -0.858373], [1.38888, 2.24725, 0.858373], [1.38888, 2.24725, -0.858373], [2.77775, 0.0, 0.0], [-2.548, 0.0, -0.973249], [-2.548, 0.0, 0.973249], [-1.57475, -1.57475, -1.57475], [-1.57475, -1.57475, 1.57475], [-1.57475, 1.57475, -1.57475], [-1.57475, 1.57475, 1.57475], [-2.42408, -1.49817, 0.0], [-2.42408, 1.49817, 0.0], [-0.973249, -2.548, 0.0], [-0.973249, 2.548, 0.0], [-1.49817, 0.0, -2.42408], [-1.49817, 0.0, 2.42408], [-2.24725, 0.858373, -1.38888], [-2.24725, 0.858373, 1.38888], [-2.24725, -0.858373, -1.38888], [-2.24725, -0.858373, 1.38888], [0.858373, -1.38888, -2.24725], [0.858373, -1.38888, 2.24725], [0.858373, 1.38888, -2.24725], [0.858373, 1.38888, 2.24725], [-0.858373, -1.38888, -2.24725], [-0.858373, -1.38888, 2.24725], [-0.858373, 1.38888, -2.24725], [-0.858373, 1.38888, 2.24725], [1.49817, 0.0, -2.42408], [1.49817, 0.0, 2.42408], [2.24725, 0.858373, -1.38888], [2.24725, 0.858373, 1.38888], [2.24725, -0.858373, -1.38888], [2.24725, -0.858373, 1.38888], [0.973249, -2.548, 0.0], [0.973249, 2.548, 0.0], [2.42408, -1.49817, 0.0], [2.42408, 1.49817, 0.0], [1.57475, -1.57475, -1.57475], [1.57475, -1.57475, 1.57475], [1.57475, 1.57475, -1.57475], [1.57475, 1.57475, 1.57475], [2.548, 0.0, -0.973249], [2.548, 0.0, 0.973249]]}, "26": {"AlternativeName": null, "Comment": null, "Name": "Deltoidal Icositetrahedron", "ShortName": "C08", "vertices": [[-1.19448, -1.19448, 0.0], [-1.19448, 0.0, -1.19448], [-1.19448, 0.0, 1.19448], [-1.19448, 1.19448, 0.0], [0.0, -1.19448, -1.19448], [0.0, -1.19448, 1.19448], [0.0, 0.0, -1.68925], [0.0, 0.0, 1.68925], [0.0, 1.19448, -1.19448], [0.0, 1.19448, 1.19448], [0.0, -1.68925, 0.0], [0.0, 1.68925, 0.0], [1.19448, -1.19448, 0.0], [1.19448, 0.0, -1.19448], [1.19448, 0.0, 1.19448], [1.19448, 1.19448, 0.0], [-1.68925, 0.0, 0.0], [1.68925, 0.0, 0.0], [0.92388, 0.92388, 0.92388], [0.92388, 0.92388, -0.92388], [0.92388, -0.92388, 0.92388], [0.92388, -0.92388, -0.92388], [-0.92388, 0.92388, 0.92388], [-0.92388, 0.92388, -0.92388], [-0.92388, -0.92388, 0.92388], [-0.92388, -0.92388, -0.92388]]}, "27": {"AlternativeName": null, "Comment": null, "Name": "Diminished Rhombicosidodecahedron", "ShortName": "J76", "vertices": [[-2.170740182, -0.339993115, -0.364123767], [-2.071220182, 0.646560885, -0.234480767], [-1.970980182, -0.787896115, 0.507356233], [-1.869560182, -0.245683115, -1.313011767], [-1.809950182, 0.808380885, 0.717124233], [-1.807610182, -1.132227115, -0.854534767], [-1.770030182, 0.740870885, -1.183371767], [-1.748000182, -0.078162115, 1.175608233], [-1.607850182, -1.580127115, 0.016945933], [-1.547050182, 1.450602885, -0.515123767], [-1.346580182, -1.418307115, 0.968548233], [-1.285780182, 1.612422885, 0.436481233], [-1.182470182, -0.540989115, -1.976871767], [-1.123590182, -0.708575115, 1.636798233], [-1.120530182, -1.427527115, -1.518391767], [-1.086020182, 1.164522885, 1.307958233], [-1.024070182, 0.277978885, 1.766438233], [-1.021440182, 1.055282885, -1.767101767], [-0.798452182, 1.765022885, -1.098861767], [-0.797305182, -2.152257115, -0.108306967], [-0.658306182, 0.263054885, -2.257511767], [-0.536033182, -1.990437115, 0.843297233], [-0.496122182, -2.057947115, -1.057201767], [-0.375705182, 2.026852885, 0.440871233], [-0.175943182, 1.578952885, 1.312348233], [-0.175235182, -0.842061115, 1.924548233], [-0.111366182, 1.469712885, -1.762711767], [-0.075708782, 0.144491885, 2.054188233], [-0.074521582, 2.121162885, -0.508019767], [0.151052818, -2.285737115, 0.179440233], [0.187896818, -1.634297115, 1.434138233], [0.251765818, 0.677481885, -2.253121767], [0.448459818, 0.948535885, 1.773548233], [0.452235818, -2.191427115, -0.769450767], [0.572652818, 1.893362885, 0.728618233], [0.612564818, 1.825852885, -1.171881767], [0.734836818, -0.427633115, 1.928938233], [0.873836818, 1.987672885, -0.220272767], [0.874982818, -1.929597115, 0.770277233], [1.097969818, -1.219867115, 1.438528233], [1.197059818, 1.262952885, 1.189808233], [1.200119818, 0.543995885, -1.965381767], [1.259009818, 0.376409885, 1.648288233], [1.362309818, -1.777007115, -0.765060767], [1.423109818, 1.253732885, -1.297131767], [1.623579818, -1.615187115, 0.186544233], [1.684379818, 1.415552885, -0.345525767], [1.824529818, -0.086416815, -1.504181767], [1.846569818, -0.905449115, 0.854792233], [1.884139818, 0.967646885, 0.525955233], [1.886479818, -0.972959115, -1.045701767], [1.946089818, 0.081103885, 0.984438233], [2.047509818, 0.623316885, -0.835935767], [2.147749818, -0.811139115, -0.094099167], [2.247279818, 0.175413885, 0.035544333]]}, "28": {"AlternativeName": "Trinagular Dipyramid", "Comment": null, "Name": "Dipyramid 3", "ShortName": "J12", "vertices": [[0.0, 0.0, -0.816496581], [0.0, 0.0, 0.816496581], [-0.288675135, -0.5, 0.0], [-0.288675135, 0.5, 0.0], [0.577350269, 0.0, 0.0]]}, "29": {"AlternativeName": "Pentagonal Dipyramid", "Comment": null, "Name": "Dipyramid 5", "ShortName": "J13", "vertices": [[0.0, 0.0, -0.525731112], [0.0, 0.0, 0.525731112], [0.850650808, 0.0, 0.0], [0.262865556, -0.809016994, 0.0], [0.262865556, 0.809016994, 0.0], [-0.68819096, -0.5, 0.0], [-0.68819096, 0.5, 0.0]]}, "30": {"AlternativeName": null, "Comment": null, "Name": "Disdyakis Dodecahedron", "ShortName": "C10", "vertices": [[0.0, 1.13152, 1.13152], [0.0, -1.13152, -1.13152], [0.0, 1.13152, -1.13152], [0.0, -1.13152, 1.13152], [1.13152, 0.0, 1.13152], [1.13152, 1.13152, 0.0], [0.975287, 0.975287, 0.975287], [-0.975287, -0.975287, -0.975287], [-1.13152, 0.0, -1.13152], [-1.13152, -1.13152, 0.0], [0.975287, 0.975287, -0.975287], [0.975287, -0.975287, 0.975287], [-0.975287, 0.975287, 0.975287], [0.975287, -0.975287, -0.975287], [-0.975287, 0.975287, -0.975287], [-0.975287, -0.975287, 0.975287], [1.13152, 0.0, -1.13152], [1.13152, -1.13152, 0.0], [-1.13152, 0.0, 1.13152], [-1.13152, 1.13152, 0.0], [-1.84505, 0.0, 0.0], [0.0, -1.84505, 0.0], [0.0, 0.0, -1.84505], [1.84505, 0.0, 0.0], [0.0, 1.84505, 0.0], [0.0, 0.0, 1.84505]]}, "31": {"AlternativeName": null, "Comment": null, "Name": "Disdyakis Triacontahedron", "ShortName": "C11", "vertices": [[-0.0, -0.0, -2.72755], [-0.0, -0.0, 2.72755], [-0.0, -2.72755, -0.0], [-0.0, 2.72755, -0.0], [-0.0, -0.991163, -2.5949], [-0.0, 0.991163, -2.5949], [-0.0, -0.991163, 2.5949], [-0.0, 0.991163, 2.5949], [-2.72755, -0.0, -0.0], [2.72755, -0.0, -0.0], [-1.60374, -1.60374, -1.60374], [1.60374, 1.60374, 1.60374], [-2.5949, -0.0, -0.991163], [-0.991163, -2.5949, -0.0], [-0.842858, -1.36377, -2.20663], [-2.20663, -0.842858, -1.36377], [-1.36377, -2.20663, -0.842858], [-0.842858, 1.36377, -2.20663], [-2.20663, -0.842858, 1.36377], [1.36377, -2.20663, -0.842858], [-2.5949, -0.0, 0.991163], [0.991163, -2.5949, -0.0], [-1.55694, -0.0, -2.51918], [-2.51918, -1.55694, -0.0], [-0.0, -2.51918, -1.55694], [-2.20663, 0.842858, -1.36377], [-1.36377, -2.20663, 0.842858], [0.842858, -1.36377, -2.20663], [-2.20663, 0.842858, 1.36377], [1.36377, -2.20663, 0.842858], [0.842858, 1.36377, -2.20663], [-2.51918, 1.55694, -0.0], [-0.0, -2.51918, 1.55694], [1.55694, -0.0, -2.51918], [-1.60374, -1.60374, 1.60374], [-1.60374, 1.60374, -1.60374], [1.60374, -1.60374, -1.60374], [-1.60374, 1.60374, 1.60374], [1.60374, -1.60374, 1.60374], [1.60374, 1.60374, -1.60374], [-0.842858, -1.36377, 2.20663], [-1.36377, 2.20663, -0.842858], [2.20663, -0.842858, -1.36377], [-0.842858, 1.36377, 2.20663], [1.36377, 2.20663, -0.842858], [2.20663, -0.842858, 1.36377], [-1.36377, 2.20663, 0.842858], [0.842858, -1.36377, 2.20663], [2.20663, 0.842858, -1.36377], [1.36377, 2.20663, 0.842858], [0.842858, 1.36377, 2.20663], [2.20663, 0.842858, 1.36377], [-0.991163, 2.5949, -0.0], [2.5949, -0.0, -0.991163], [0.991163, 2.5949, -0.0], [2.5949, -0.0, 0.991163], [-1.55694, -0.0, 2.51918], [-0.0, 2.51918, -1.55694], [2.51918, -1.55694, -0.0], [-0.0, 2.51918, 1.55694], [1.55694, -0.0, 2.51918], [2.51918, 1.55694, -0.0]]}, "32": {"AlternativeName": null, "Comment": null, "Name": "Disphenocingulum", "ShortName": "J90", "vertices": [[-1.171140619, 0.293688206, 0.109311788], [-0.838770619, -0.457648794, 0.679416788], [-0.814439619, -0.537541794, -0.317089213], [-0.666585619, 0.849234206, -0.551588213], [-0.460440619, 0.461075206, 0.792598788], [-0.309880619, 0.018003906, -0.977988213], [-0.117267619, -0.991153794, 0.238052788], [0.010481881, -0.901366794, -0.749679213], [0.044118781, 1.016620206, 0.131698788], [0.113460381, -0.340276794, 0.961324788], [0.297984381, 0.793976206, -0.809566213], [0.530304381, 0.565040206, 0.879826788], [0.656860381, -0.138801794, -0.775936213], [0.803498381, -0.896900794, -0.140495213], [0.887587381, 0.512075206, -0.052664813], [1.034229381, -0.246022794, 0.582776788]]}, "33": {"AlternativeName": null, "Comment": null, "Name": "Dodecahedron", "ShortName": "P05", "vertices": [[-1.37638, 0.0, 0.262866], [1.37638, 0.0, -0.262866], [-0.425325, -1.30902, 0.262866], [-0.425325, 1.30902, 0.262866], [1.11352, -0.809017, 0.262866], [1.11352, 0.809017, 0.262866], [-0.262866, -0.809017, 1.11352], [-0.262866, 0.809017, 1.11352], [-0.688191, -0.5, -1.11352], [-0.688191, 0.5, -1.11352], [0.688191, -0.5, 1.11352], [0.688191, 0.5, 1.11352], [0.850651, 0.0, -1.11352], [-1.11352, -0.809017, -0.262866], [-1.11352, 0.809017, -0.262866], [-0.850651, 0.0, 1.11352], [0.262866, -0.809017, -1.11352], [0.262866, 0.809017, -1.11352], [0.425325, -1.30902, -0.262866], [0.425325, 1.30902, -0.262866]]}, "34": {"AlternativeName": null, "Comment": null, "Name": "Duerers Solid", "ShortName": "O20", "vertices": [[0.0, -1.7769, 0.961343], [0.0, 1.7769, -0.961343], [0.0, -1.09819, 1.69574], [0.0, 1.09819, -1.69574], [-1.53884, -0.888451, -0.961343], [-1.53884, 0.888451, 0.961343], [1.53884, -0.888451, -0.961343], [1.53884, 0.888451, 0.961343], [-0.951057, -0.549093, -1.69574], [-0.951057, 0.549093, 1.69574], [0.951057, -0.549093, -1.69574], [0.951057, 0.549093, 1.69574]]}, "35": {"AlternativeName": null, "Comment": null, "Name": "Elongated Dodecahedron", "ShortName": "O21", "vertices": [[-0.866025, -0.866025, -0.5], [-0.866025, -0.866025, 0.5], [-0.866025, 0.0, -1.0], [-0.866025, 0.0, 1.0], [-0.866025, 0.866025, -0.5], [-0.866025, 0.866025, 0.5], [0.0, -0.866025, -1.0], [0.0, -0.866025, 1.0], [0.0, 0.0, -1.5], [0.0, 0.0, 1.5], [0.0, 0.866025, -1.0], [0.0, 0.866025, 1.0], [0.866025, -0.866025, -0.5], [0.866025, -0.866025, 0.5], [0.866025, 0.0, -1.0], [0.866025, 0.0, 1.0], [0.866025, 0.866025, -0.5], [0.866025, 0.866025, 0.5]]}, "36": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Cupola", "ShortName": "J20", "vertices": [[-4e-08, -1.61803, -0.7051462], [-4e-08, -1.61803, 0.2948538], [-4e-08, 1.61803, -0.7051462], [-4e-08, 1.61803, 0.2948538], [0.85065096, 0.0, 0.8205848], [0.26286596, -0.809017, 0.8205848], [0.26286596, 0.809017, 0.8205848], [-0.95105704, -1.30902, -0.7051462], [-0.95105704, -1.30902, 0.2948538], [-0.95105704, 1.30902, -0.7051462], [-0.95105704, 1.30902, 0.2948538], [0.95105696, -1.30902, -0.7051462], [0.95105696, -1.30902, 0.2948538], [0.95105696, 1.30902, -0.7051462], [0.95105696, 1.30902, 0.2948538], [-0.68819104, -0.5, 0.8205848], [-0.68819104, 0.5, 0.8205848], [-1.53884004, -0.5, -0.7051462], [-1.53884004, -0.5, 0.2948538], [-1.53884004, 0.5, -0.7051462], [-1.53884004, 0.5, 0.2948538], [1.53883996, -0.5, -0.7051462], [1.53883996, -0.5, 0.2948538], [1.53883996, 0.5, -0.7051462], [1.53883996, 0.5, 0.2948538]]}, "37": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Dipyramid", "ShortName": "J16", "vertices": [[-1.67e-07, 0.0, -1.02573], [-1.67e-07, 0.0, 1.02573], [0.850650833, 0.0, -0.5], [0.850650833, 0.0, 0.5], [0.262865833, -0.809017, -0.5], [0.262865833, -0.809017, 0.5], [0.262865833, 0.809017, -0.5], [0.262865833, 0.809017, 0.5], [-0.688191167, -0.5, -0.5], [-0.688191167, -0.5, 0.5], [-0.688191167, 0.5, -0.5], [-0.688191167, 0.5, 0.5]]}, "38": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Gyrobicupola", "ShortName": "J39", "vertices": [[-0.0, -1.61803, -0.5], [-0.0, -1.61803, 0.5], [-0.0, 1.61803, -0.5], [-0.0, 1.61803, 0.5], [-0.688191, -0.5, 1.02573], [-0.688191, 0.5, 1.02573], [0.688191, -0.5, -1.02573], [0.688191, 0.5, -1.02573], [-0.850651, 0.0, -1.02573], [0.850651, 0.0, 1.02573], [-0.262866, -0.809017, -1.02573], [-0.262866, 0.809017, -1.02573], [0.262866, -0.809017, 1.02573], [0.262866, 0.809017, 1.02573], [-0.951057, -1.30902, -0.5], [-0.951057, -1.30902, 0.5], [-0.951057, 1.30902, -0.5], [-0.951057, 1.30902, 0.5], [0.951057, -1.30902, -0.5], [0.951057, -1.30902, 0.5], [0.951057, 1.30902, -0.5], [0.951057, 1.30902, 0.5], [-1.53884, -0.5, -0.5], [-1.53884, -0.5, 0.5], [-1.53884, 0.5, -0.5], [-1.53884, 0.5, 0.5], [1.53884, -0.5, -0.5], [1.53884, -0.5, 0.5], [1.53884, 0.5, -0.5], [1.53884, 0.5, 0.5]]}, "39": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Gyrobirotunda", "ShortName": "J43", "vertices": [[-0.0, -1.61803, -0.5], [-0.0, -1.61803, 0.5], [-0.0, 1.61803, -0.5], [-0.0, 1.61803, 0.5], [-0.425325, -1.30902, -1.35065], [-0.425325, 1.30902, -1.35065], [0.425325, -1.30902, 1.35065], [0.425325, 1.30902, 1.35065], [-0.688191, -0.5, -1.87638], [-0.688191, 0.5, -1.87638], [0.688191, -0.5, 1.87638], [0.688191, 0.5, 1.87638], [-0.850651, 0.0, 1.87638], [0.850651, 0.0, -1.87638], [-1.11352, -0.809017, 1.35065], [-1.11352, 0.809017, 1.35065], [1.11352, -0.809017, -1.35065], [1.11352, 0.809017, -1.35065], [-0.262866, -0.809017, 1.87638], [-0.262866, 0.809017, 1.87638], [0.262866, -0.809017, -1.87638], [0.262866, 0.809017, -1.87638], [-0.951057, -1.30902, -0.5], [-0.951057, -1.30902, 0.5], [-0.951057, 1.30902, -0.5], [-0.951057, 1.30902, 0.5], [0.951057, -1.30902, -0.5], [0.951057, -1.30902, 0.5], [0.951057, 1.30902, -0.5], [0.951057, 1.30902, 0.5], [-1.53884, -0.5, -0.5], [-1.53884, -0.5, 0.5], [-1.53884, 0.5, -0.5], [-1.53884, 0.5, 0.5], [1.53884, -0.5, -0.5], [1.53884, -0.5, 0.5], [1.53884, 0.5, -0.5], [1.53884, 0.5, 0.5], [-1.37638, 0.0, -1.35065], [1.37638, 0.0, 1.35065]]}, "40": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Gyrocupolarotunda", "ShortName": "J41", "vertices": [[2.86e-07, -1.61803, -0.814471429], [2.86e-07, -1.61803, 0.185528571], [2.86e-07, 1.61803, -0.814471429], [2.86e-07, 1.61803, 0.185528571], [0.850651286, 0.0, -1.340201429], [0.262866286, -0.809017, -1.340201429], [0.262866286, 0.809017, -1.340201429], [-0.951056714, -1.30902, -0.814471429], [-0.951056714, -1.30902, 0.185528571], [-0.951056714, 1.30902, -0.814471429], [-0.951056714, 1.30902, 0.185528571], [0.951057286, -1.30902, -0.814471429], [0.951057286, -1.30902, 0.185528571], [0.951057286, 1.30902, -0.814471429], [0.951057286, 1.30902, 0.185528571], [-0.688190714, -0.5, -1.340201429], [-0.688190714, 0.5, -1.340201429], [-1.538839714, -0.5, -0.814471429], [-1.538839714, -0.5, 0.185528571], [-1.538839714, 0.5, -0.814471429], [-1.538839714, 0.5, 0.185528571], [1.538840286, -0.5, -0.814471429], [1.538840286, -0.5, 0.185528571], [1.538840286, 0.5, -0.814471429], [1.538840286, 0.5, 0.185528571], [1.376380286, 0.0, 1.036178571], [0.425325286, -1.30902, 1.036178571], [0.425325286, 1.30902, 1.036178571], [-1.113519714, -0.809017, 1.036178571], [-1.113519714, 0.809017, 1.036178571], [-0.850650714, 0.0, 1.561908571], [-0.262865714, -0.809017, 1.561908571], [-0.262865714, 0.809017, 1.561908571], [0.688191286, -0.5, 1.561908571], [0.688191286, 0.5, 1.561908571]]}, "41": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Orthobicupola", "ShortName": "J38", "vertices": [[-6.7e-08, -1.61803, -0.5], [-6.7e-08, -1.61803, 0.5], [-6.7e-08, 1.61803, -0.5], [-6.7e-08, 1.61803, 0.5], [0.850650933, 0.0, -1.02573], [0.850650933, 0.0, 1.02573], [0.262865933, -0.809017, -1.02573], [0.262865933, -0.809017, 1.02573], [0.262865933, 0.809017, -1.02573], [0.262865933, 0.809017, 1.02573], [-0.951057067, -1.30902, -0.5], [-0.951057067, -1.30902, 0.5], [-0.951057067, 1.30902, -0.5], [-0.951057067, 1.30902, 0.5], [0.951056933, -1.30902, -0.5], [0.951056933, -1.30902, 0.5], [0.951056933, 1.30902, -0.5], [0.951056933, 1.30902, 0.5], [-0.688191067, -0.5, -1.02573], [-0.688191067, -0.5, 1.02573], [-0.688191067, 0.5, -1.02573], [-0.688191067, 0.5, 1.02573], [-1.538840067, -0.5, -0.5], [-1.538840067, -0.5, 0.5], [-1.538840067, 0.5, -0.5], [-1.538840067, 0.5, 0.5], [1.538839933, -0.5, -0.5], [1.538839933, -0.5, 0.5], [1.538839933, 0.5, -0.5], [1.538839933, 0.5, 0.5]]}, "42": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Orthobirotunda", "ShortName": "J42", "vertices": [[5.5e-07, -1.61803, -0.5], [5.5e-07, -1.61803, 0.5], [5.5e-07, 1.61803, -0.5], [5.5e-07, 1.61803, 0.5], [-0.95105645, -1.30902, -0.5], [-0.95105645, -1.30902, 0.5], [-0.95105645, 1.30902, -0.5], [-0.95105645, 1.30902, 0.5], [0.95105755, -1.30902, -0.5], [0.95105755, -1.30902, 0.5], [0.95105755, 1.30902, -0.5], [0.95105755, 1.30902, 0.5], [-1.53883945, -0.5, -0.5], [-1.53883945, -0.5, 0.5], [-1.53883945, 0.5, -0.5], [-1.53883945, 0.5, 0.5], [1.53884055, -0.5, -0.5], [1.53884055, -0.5, 0.5], [1.53884055, 0.5, -0.5], [1.53884055, 0.5, 0.5], [1.37638055, 0.0, -1.35065], [1.37638055, 0.0, 1.35065], [0.42532555, -1.30902, -1.35065], [0.42532555, -1.30902, 1.35065], [0.42532555, 1.30902, -1.35065], [0.42532555, 1.30902, 1.35065], [-1.11351945, -0.809017, -1.35065], [-1.11351945, -0.809017, 1.35065], [-1.11351945, 0.809017, -1.35065], [-1.11351945, 0.809017, 1.35065], [-0.85065045, 0.0, -1.87638], [-0.85065045, 0.0, 1.87638], [-0.26286545, -0.809017, -1.87638], [-0.26286545, -0.809017, 1.87638], [-0.26286545, 0.809017, -1.87638], [-0.26286545, 0.809017, 1.87638], [0.68819155, -0.5, -1.87638], [0.68819155, -0.5, 1.87638], [0.68819155, 0.5, -1.87638], [0.68819155, 0.5, 1.87638]]}, "43": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Orthocupolarotunda", "ShortName": "J40", "vertices": [[3.43e-07, -1.61803, -0.814471429], [3.43e-07, -1.61803, 0.185528571], [3.43e-07, 1.61803, -0.814471429], [3.43e-07, 1.61803, 0.185528571], [0.425325343, -1.30902, 1.036178571], [0.425325343, 1.30902, 1.036178571], [0.688191343, -0.5, -1.340201429], [0.688191343, -0.5, 1.561908571], [0.688191343, 0.5, -1.340201429], [0.688191343, 0.5, 1.561908571], [-0.850650657, 0.0, -1.340201429], [-0.850650657, 0.0, 1.561908571], [-1.113519657, -0.809017, 1.036178571], [-1.113519657, 0.809017, 1.036178571], [-0.262865657, -0.809017, -1.340201429], [-0.262865657, -0.809017, 1.561908571], [-0.262865657, 0.809017, -1.340201429], [-0.262865657, 0.809017, 1.561908571], [-0.951056657, -1.30902, -0.814471429], [-0.951056657, -1.30902, 0.185528571], [-0.951056657, 1.30902, -0.814471429], [-0.951056657, 1.30902, 0.185528571], [0.951057343, -1.30902, -0.814471429], [0.951057343, -1.30902, 0.185528571], [0.951057343, 1.30902, -0.814471429], [0.951057343, 1.30902, 0.185528571], [-1.538839657, -0.5, -0.814471429], [-1.538839657, -0.5, 0.185528571], [-1.538839657, 0.5, -0.814471429], [-1.538839657, 0.5, 0.185528571], [1.538840343, -0.5, -0.814471429], [1.538840343, -0.5, 0.185528571], [1.538840343, 0.5, -0.814471429], [1.538840343, 0.5, 0.185528571], [1.376380343, 0.0, 1.036178571]]}, "44": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Pyramid", "ShortName": "J09", "vertices": [[-1.82e-07, 0.0, 0.932481818], [0.850650818, 0.0, -0.593248182], [0.850650818, 0.0, 0.406751818], [0.262865818, -0.809017, -0.593248182], [0.262865818, -0.809017, 0.406751818], [0.262865818, 0.809017, -0.593248182], [0.262865818, 0.809017, 0.406751818], [-0.688191182, -0.5, -0.593248182], [-0.688191182, -0.5, 0.406751818], [-0.688191182, 0.5, -0.593248182], [-0.688191182, 0.5, 0.406751818]]}, "45": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Rotunda", "ShortName": "J21", "vertices": [[3.67e-07, -1.61803, -1.0378385], [3.67e-07, -1.61803, -0.0378385], [3.67e-07, 1.61803, -1.0378385], [3.67e-07, 1.61803, -0.0378385], [-0.951056633, -1.30902, -1.0378385], [-0.951056633, -1.30902, -0.0378385], [-0.951056633, 1.30902, -1.0378385], [-0.951056633, 1.30902, -0.0378385], [0.951057367, -1.30902, -1.0378385], [0.951057367, -1.30902, -0.0378385], [0.951057367, 1.30902, -1.0378385], [0.951057367, 1.30902, -0.0378385], [-1.538839633, -0.5, -1.0378385], [-1.538839633, -0.5, -0.0378385], [-1.538839633, 0.5, -1.0378385], [-1.538839633, 0.5, -0.0378385], [1.538840367, -0.5, -1.0378385], [1.538840367, -0.5, -0.0378385], [1.538840367, 0.5, -1.0378385], [1.538840367, 0.5, -0.0378385], [1.376380367, 0.0, 0.8128125], [0.425325367, -1.30902, 0.8128125], [0.425325367, 1.30902, 0.8128125], [-1.113519633, -0.809017, 0.8128125], [-1.113519633, 0.809017, 0.8128125], [-0.850650633, 0.0, 1.3385415], [-0.262865633, -0.809017, 1.3385415], [-0.262865633, 0.809017, 1.3385415], [0.688191367, -0.5, 1.3385415], [0.688191367, 0.5, 1.3385415]]}, "46": {"AlternativeName": null, "Comment": null, "Name": "Elongated Square Cupola", "ShortName": "J19", "vertices": [[-0.0, -0.707107, 0.9656856], [-0.0, 0.707107, 0.9656856], [-0.707107, 0.0, 0.9656856], [0.707107, 0.0, 0.9656856], [0.5, 1.20711, -0.7414214], [0.5, 1.20711, 0.2585786], [-0.5, 1.20711, -0.7414214], [-0.5, 1.20711, 0.2585786], [-1.20711, 0.5, -0.7414214], [-1.20711, 0.5, 0.2585786], [-1.20711, -0.5, -0.7414214], [-1.20711, -0.5, 0.2585786], [-0.5, -1.20711, -0.7414214], [-0.5, -1.20711, 0.2585786], [0.5, -1.20711, -0.7414214], [0.5, -1.20711, 0.2585786], [1.20711, -0.5, -0.7414214], [1.20711, -0.5, 0.2585786], [1.20711, 0.5, -0.7414214], [1.20711, 0.5, 0.2585786]]}, "47": {"AlternativeName": null, "Comment": null, "Name": "Elongated Square Dipyramid", "ShortName": "J15", "vertices": [[0.0, 0.0, -1.20711], [0.0, 0.0, 1.20711], [0.0, -0.707107, -0.5], [0.0, -0.707107, 0.5], [0.0, 0.707107, -0.5], [0.0, 0.707107, 0.5], [-0.707107, 0.0, -0.5], [-0.707107, 0.0, 0.5], [0.707107, 0.0, -0.5], [0.707107, 0.0, 0.5]]}, "48": {"AlternativeName": null, "Comment": null, "Name": "Elongated Square Gyrobicupola", "ShortName": "J37", "vertices": [[-0.5, -0.5, -1.20711], [-0.5, 0.5, -1.20711], [-0.0, -0.707107, 1.20711], [-0.0, 0.707107, 1.20711], [0.5, -0.5, -1.20711], [0.5, 0.5, -1.20711], [-0.707107, 0.0, 1.20711], [0.707107, 0.0, 1.20711], [0.5, 1.20711, -0.5], [0.5, 1.20711, 0.5], [-0.5, 1.20711, -0.5], [-0.5, 1.20711, 0.5], [-1.20711, 0.5, -0.5], [-1.20711, 0.5, 0.5], [-1.20711, -0.5, -0.5], [-1.20711, -0.5, 0.5], [-0.5, -1.20711, -0.5], [-0.5, -1.20711, 0.5], [0.5, -1.20711, -0.5], [0.5, -1.20711, 0.5], [1.20711, -0.5, -0.5], [1.20711, -0.5, 0.5], [1.20711, 0.5, -0.5], [1.20711, 0.5, 0.5]]}, "49": {"AlternativeName": null, "Comment": null, "Name": "Elongated Square Pyramid", "ShortName": "J08", "vertices": [[0.0, 0.0, 1.072986667], [0.0, -0.707107, -0.634123333], [0.0, -0.707107, 0.365876667], [0.0, 0.707107, -0.634123333], [0.0, 0.707107, 0.365876667], [-0.707107, 0.0, -0.634123333], [-0.707107, 0.0, 0.365876667], [0.707107, 0.0, -0.634123333], [0.707107, 0.0, 0.365876667]]}, "50": {"AlternativeName": null, "Comment": null, "Name": "Elongated Triangular Cupola", "ShortName": "J18", "vertices": [[0.0, -1.0, -0.7632994], [0.0, -1.0, 0.2367006], [0.0, 1.0, -0.7632994], [0.0, 1.0, 0.2367006], [-0.288675, -0.5, 1.0531976], [-0.288675, 0.5, 1.0531976], [0.57735, 0.0, 1.0531976], [-0.866025, -0.5, -0.7632994], [-0.866025, -0.5, 0.2367006], [-0.866025, 0.5, -0.7632994], [-0.866025, 0.5, 0.2367006], [0.866025, -0.5, -0.7632994], [0.866025, -0.5, 0.2367006], [0.866025, 0.5, -0.7632994], [0.866025, 0.5, 0.2367006]]}, "51": {"AlternativeName": null, "Comment": null, "Name": "Elongated Triangular Dipyramid", "ShortName": "J14", "vertices": [[0.0, 0.0, -1.3165], [0.0, 0.0, 1.3165], [-0.288675, -0.5, -0.5], [-0.288675, -0.5, 0.5], [-0.288675, 0.5, -0.5], [-0.288675, 0.5, 0.5], [0.57735, 0.0, -0.5], [0.57735, 0.0, 0.5]]}, "52": {"AlternativeName": null, "Comment": null, "Name": "Elongated Triangular Gyrobicupola", "ShortName": "J36", "vertices": [[0.0, -1.0, -0.5], [0.0, -1.0, 0.5], [0.0, 1.0, -0.5], [0.0, 1.0, 0.5], [-0.288675, -0.5, 1.3165], [-0.288675, 0.5, 1.3165], [0.288675, -0.5, -1.3165], [0.57735, 0.0, 1.3165], [-0.866025, -0.5, -0.5], [-0.866025, -0.5, 0.5], [-0.866025, 0.5, -0.5], [-0.866025, 0.5, 0.5], [0.866025, -0.5, -0.5], [0.866025, -0.5, 0.5], [0.866025, 0.5, -0.5], [0.866025, 0.5, 0.5], [-0.57735, 0.0, -1.3165], [0.288675, 0.5, -1.3165]]}, "53": {"AlternativeName": null, "Comment": null, "Name": "Elongated Triangular Orthobicupola", "ShortName": "J35", "vertices": [[0.0, -1.0, -0.5], [0.0, -1.0, 0.5], [0.0, 1.0, -0.5], [0.0, 1.0, 0.5], [-0.288675, -0.5, -1.3165], [-0.288675, -0.5, 1.3165], [-0.288675, 0.5, -1.3165], [-0.288675, 0.5, 1.3165], [0.57735, 0.0, -1.3165], [0.57735, 0.0, 1.3165], [-0.866025, -0.5, -0.5], [-0.866025, -0.5, 0.5], [-0.866025, 0.5, -0.5], [-0.866025, 0.5, 0.5], [0.866025, -0.5, -0.5], [0.866025, -0.5, 0.5], [0.866025, 0.5, -0.5], [0.866025, 0.5, 0.5]]}, "54": {"AlternativeName": null, "Comment": null, "Name": "Elongated Triangular Pyramid", "ShortName": "J07", "vertices": [[0.0, 0.0, 1.128428571], [-0.288675, -0.5, -0.688071429], [-0.288675, -0.5, 0.311928571], [-0.288675, 0.5, -0.688071429], [-0.288675, 0.5, 0.311928571], [0.57735, 0.0, -0.688071429], [0.57735, 0.0, 0.311928571]]}, "55": {"AlternativeName": null, "Comment": null, "Name": "Truncated Icosidodecahedron", "ShortName": "A11", "vertices": [[-1.0, -1.30902, -3.42705], [-1.0, -1.30902, 3.42705], [-1.0, 1.30902, -3.42705], [-1.0, 1.30902, 3.42705], [-0.5, -0.5, -3.73607], [-0.5, -0.5, 3.73607], [-0.5, 0.5, -3.73607], [-0.5, 0.5, 3.73607], [-0.5, -3.73607, -0.5], [-0.5, -3.73607, 0.5], [-0.5, -2.11803, -3.11803], [-0.5, -2.11803, 3.11803], [-0.5, 3.73607, -0.5], [-0.5, 3.73607, 0.5], [-0.5, 2.11803, -3.11803], [-0.5, 2.11803, 3.11803], [0.5, -0.5, -3.73607], [0.5, -0.5, 3.73607], [0.5, 0.5, -3.73607], [0.5, 0.5, 3.73607], [0.5, -3.73607, -0.5], [0.5, -3.73607, 0.5], [0.5, -2.11803, -3.11803], [0.5, -2.11803, 3.11803], [0.5, 3.73607, -0.5], [0.5, 3.73607, 0.5], [0.5, 2.11803, -3.11803], [0.5, 2.11803, 3.11803], [1.0, -1.30902, -3.42705], [1.0, -1.30902, 3.42705], [1.0, 1.30902, -3.42705], [1.0, 1.30902, 3.42705], [-3.42705, -1.0, -1.30902], [-3.42705, -1.0, 1.30902], [-3.42705, 1.0, -1.30902], [-3.42705, 1.0, 1.30902], [-2.92705, -1.80902, -1.61803], [-2.92705, -1.80902, 1.61803], [-2.92705, 1.80902, -1.61803], [-2.92705, 1.80902, 1.61803], [-1.80902, -1.61803, -2.92705], [-1.80902, -1.61803, 2.92705], [-1.80902, 1.61803, -2.92705], [-1.80902, 1.61803, 2.92705], [-1.30902, -3.42705, -1.0], [-1.30902, -3.42705, 1.0], [-1.30902, -2.42705, -2.61803], [-1.30902, -2.42705, 2.61803], [-1.30902, 2.42705, -2.61803], [-1.30902, 2.42705, 2.61803], [-1.30902, 3.42705, -1.0], [-1.30902, 3.42705, 1.0], [-2.61803, -1.30902, -2.42705], [-2.61803, -1.30902, 2.42705], [-2.61803, 1.30902, -2.42705], [-2.61803, 1.30902, 2.42705], [-3.73607, -0.5, -0.5], [-3.73607, -0.5, 0.5], [-3.73607, 0.5, -0.5], [-3.73607, 0.5, 0.5], [-1.61803, -2.92705, -1.80902], [-1.61803, -2.92705, 1.80902], [-1.61803, 2.92705, -1.80902], [-1.61803, 2.92705, 1.80902], [-3.11803, -0.5, -2.11803], [-3.11803, -0.5, 2.11803], [-3.11803, 0.5, -2.11803], [-3.11803, 0.5, 2.11803], [-2.11803, -3.11803, -0.5], [-2.11803, -3.11803, 0.5], [-2.11803, 3.11803, -0.5], [-2.11803, 3.11803, 0.5], [-2.42705, -2.61803, -1.30902], [-2.42705, -2.61803, 1.30902], [-2.42705, 2.61803, -1.30902], [-2.42705, 2.61803, 1.30902], [1.61803, -2.92705, -1.80902], [1.61803, -2.92705, 1.80902], [1.61803, 2.92705, -1.80902], [1.61803, 2.92705, 1.80902], [2.42705, -2.61803, -1.30902], [2.42705, -2.61803, 1.30902], [2.42705, 2.61803, -1.30902], [2.42705, 2.61803, 1.30902], [3.73607, -0.5, -0.5], [3.73607, -0.5, 0.5], [3.73607, 0.5, -0.5], [3.73607, 0.5, 0.5], [2.11803, -3.11803, -0.5], [2.11803, -3.11803, 0.5], [2.11803, 3.11803, -0.5], [2.11803, 3.11803, 0.5], [1.30902, -3.42705, -1.0], [1.30902, -3.42705, 1.0], [1.30902, -2.42705, -2.61803], [1.30902, -2.42705, 2.61803], [1.30902, 2.42705, -2.61803], [1.30902, 2.42705, 2.61803], [1.30902, 3.42705, -1.0], [1.30902, 3.42705, 1.0], [2.61803, -1.30902, -2.42705], [2.61803, -1.30902, 2.42705], [2.61803, 1.30902, -2.42705], [2.61803, 1.30902, 2.42705], [3.11803, -0.5, -2.11803], [3.11803, -0.5, 2.11803], [3.11803, 0.5, -2.11803], [3.11803, 0.5, 2.11803], [1.80902, -1.61803, -2.92705], [1.80902, -1.61803, 2.92705], [1.80902, 1.61803, -2.92705], [1.80902, 1.61803, 2.92705], [2.92705, -1.80902, -1.61803], [2.92705, -1.80902, 1.61803], [2.92705, 1.80902, -1.61803], [2.92705, 1.80902, 1.61803], [3.42705, -1.0, -1.30902], [3.42705, -1.0, 1.30902], [3.42705, 1.0, -1.30902], [3.42705, 1.0, 1.30902]]}, "56": {"AlternativeName": null, "Comment": null, "Name": "Truncated Cuboctahedron", "ShortName": "A10", "vertices": [[-0.5, 1.20711, -1.91421], [-0.5, 1.20711, 1.91421], [-0.5, -1.20711, -1.91421], [-0.5, -1.20711, 1.91421], [-0.5, -1.91421, 1.20711], [-0.5, -1.91421, -1.20711], [-0.5, 1.91421, 1.20711], [-0.5, 1.91421, -1.20711], [0.5, 1.20711, -1.91421], [0.5, 1.20711, 1.91421], [0.5, -1.20711, -1.91421], [0.5, -1.20711, 1.91421], [0.5, -1.91421, 1.20711], [0.5, -1.91421, -1.20711], [0.5, 1.91421, 1.20711], [0.5, 1.91421, -1.20711], [1.20711, -0.5, -1.91421], [1.20711, -0.5, 1.91421], [1.20711, 0.5, -1.91421], [1.20711, 0.5, 1.91421], [1.20711, -1.91421, -0.5], [1.20711, -1.91421, 0.5], [1.20711, 1.91421, -0.5], [1.20711, 1.91421, 0.5], [-1.20711, -0.5, -1.91421], [-1.20711, -0.5, 1.91421], [-1.20711, 0.5, -1.91421], [-1.20711, 0.5, 1.91421], [-1.20711, -1.91421, -0.5], [-1.20711, -1.91421, 0.5], [-1.20711, 1.91421, -0.5], [-1.20711, 1.91421, 0.5], [-1.91421, -0.5, 1.20711], [-1.91421, -0.5, -1.20711], [-1.91421, 0.5, 1.20711], [-1.91421, 0.5, -1.20711], [-1.91421, 1.20711, -0.5], [-1.91421, 1.20711, 0.5], [-1.91421, -1.20711, -0.5], [-1.91421, -1.20711, 0.5], [1.91421, -0.5, 1.20711], [1.91421, -0.5, -1.20711], [1.91421, 0.5, 1.20711], [1.91421, 0.5, -1.20711], [1.91421, 1.20711, -0.5], [1.91421, 1.20711, 0.5], [1.91421, -1.20711, -0.5], [1.91421, -1.20711, 0.5]]}, "57": {"AlternativeName": null, "Comment": null, "Name": "Gyrate Bidiminished Rhombicosidodecahedron", "ShortName": "J82", "vertices": [[-2.03400905, -0.332349036, 0.010840126], [-1.90637905, 0.580113964, -0.377903174], [-1.89745905, 0.039699764, 0.928952826], [-1.76981905, 0.952161964, 0.540209826], [-1.71681905, -0.741266036, -0.844833174], [-1.68827905, -1.270655036, 0.003068226], [-1.58918905, 0.171196964, -1.233572174], [-1.46732905, -0.668664036, 1.488607826], [-1.41958905, 0.744307964, 1.453517826], [-1.35412905, 1.118204964, -1.014672174], [-1.33804905, -1.478505036, 0.916377826], [-1.13316905, 1.720194964, 0.470863826], [-1.02978905, -1.347005036, -1.246152174], [-1.00124905, -1.876395036, -0.398250174], [-0.98945905, 0.035944964, 2.013177826], [-0.87626405, 1.822814964, -0.490107174], [-0.82326705, 0.129387964, -1.875152174], [-0.78294305, 1.512344964, 1.384177826], [-0.65101505, -2.084245036, 0.515059826], [-0.58821005, 1.076394964, -1.656252174], [-0.47753805, -0.808914036, -1.882922174], [-0.23532605, -1.918205036, -1.039822174], [-0.23069505, 2.050434964, 0.747402826], [-0.11034305, 1.781004964, -1.131682174], [-0.08698205, 0.366183964, 2.289707826], [0.02621525, 2.153054964, -0.213568174], [0.04065315, 1.278644964, 1.900967826], [0.09882795, 0.470656964, -2.057562174], [0.31692195, -1.380105036, -1.676592174], [0.33135995, -2.254515036, 0.437941826], [0.44455795, -0.467645036, -2.065342174], [0.58826995, -2.151895036, -0.523029174], [0.59290095, 1.816744964, 1.264197826], [0.87203195, 1.610734964, -1.208802174], [0.89539295, 0.195914964, 2.212597826], [1.00130695, 0.800895964, -1.781032174], [1.00859095, 1.982784964, -0.290686174], [1.02302795, 1.108374964, 1.823847826], [1.14052095, -1.613805036, -1.159802174], [1.23384095, -1.924275036, 0.714480826], [1.34704095, -0.137406036, -1.788802174], [1.35882095, 1.774934964, 0.622623826], [1.49075095, -1.821655036, -0.246491174], [1.58243095, -0.409827036, 1.811277826], [1.71170095, -1.219665036, 1.239047826], [1.77716095, -0.845769036, -1.229142174], [1.78895095, 1.066564964, 1.182277826], [2.12739095, -1.053625036, -0.315837174], [2.13468095, 0.128265964, 1.174507826], [2.26395095, -0.681575036, 0.602276826]]}, "58": {"AlternativeName": null, "Comment": null, "Name": "Gyrate Rhombicosidodecahedron", "ShortName": "J72", "vertices": [[-2.180729817, 0.476838883, -0.055880833], [-2.105409817, -0.300487117, -0.680454933], [-2.049319817, 0.138432883, 0.875899067], [-1.927459817, -1.119310117, -0.134682933], [-1.892789817, -0.848038117, 0.827197067], [-1.801859817, 1.078529883, -0.759038933], [-1.767189817, 1.349799883, 0.202841067], [-1.726549817, 0.301201883, -1.383609933], [-1.635789817, 1.011389883, 1.134620067], [-1.570019817, -0.685269117, -1.432309933], [-1.457829817, 0.192570883, 1.680390067], [-1.392059817, -1.504090117, -0.886542933], [-1.301299817, -0.793900117, 1.631690067], [-1.260649817, -1.842490117, 0.045237067], [-1.225979817, -1.571230117, 1.007120067], [-1.057439817, 1.713679883, -0.964992933], [-1.022769817, 1.984939883, -0.003113083], [-0.935572817, 0.455936883, -1.975569933], [-0.810144817, 1.437389883, 1.504540067], [-0.779041817, -0.530535117, -2.024279933], [-0.632187817, 0.618573883, 2.050310067], [-0.522039817, 1.328889883, -1.716849933], [-0.491102817, -1.855410117, -1.141199933], [-0.431278817, 2.039079883, 0.801380067], [-0.378915817, -0.977571117, 1.971510067], [-0.359693817, -2.193820117, -0.209418933], [-0.303600817, -1.754900117, 1.346930067], [-0.231794817, 2.139679883, -0.595074933], [-0.112236817, -1.253720117, -1.844359933], [-0.034616117, 0.104613883, -2.230229933], [0.034616483, -0.104614117, 2.230230067], [0.167767183, 1.646339883, 1.499160067], [0.231795183, -2.139680117, 0.595075067], [0.303601183, 1.754899883, -1.346929933], [0.345724183, 0.827521883, 2.044930067], [0.359694183, 2.193819883, 0.209419067], [0.378916183, 0.977570883, -1.971509933], [0.431279183, -2.039080117, -0.801379933], [0.522040183, -1.328890117, 1.716850067], [0.632188183, -0.618574117, -2.050309933], [0.810145183, -1.437390117, -1.504539933], [0.935573183, -0.455937117, 1.975570067], [0.958740183, 1.801079883, 0.907199067], [1.022770183, -1.984940117, 0.003113207], [1.057440183, -1.713680117, 0.964993067], [1.225980183, 1.571229883, -1.007119933], [1.246680183, 0.476198883, 1.790280067], [1.260650183, 1.842489883, -0.045236933], [1.301300183, 0.793899883, -1.631689933], [1.457830183, -0.192571117, -1.680389933], [1.625550183, 1.077889883, 1.087120067], [1.635790183, -1.011390117, -1.134619933], [1.726550183, -0.301202117, 1.383610067], [1.767190183, -1.349800117, -0.202840933], [1.801860183, -1.078530117, 0.759039067], [1.892790183, 0.848037883, -0.827196933], [1.927450183, 1.119309883, 0.134683067], [2.049320183, -0.138433117, -0.875898933], [2.105410183, 0.300486883, 0.680455067], [2.180730183, -0.476839117, 0.055880967]]}, "59": {"AlternativeName": null, "Comment": null, "Name": "Gyrobifastigium", "ShortName": "J26", "vertices": [[-0.5, -0.5, 0.0], [-0.5, 0.0, 0.866025], [-0.5, 0.5, 0.0], [0.0, -0.5, -0.866025], [0.0, 0.5, -0.866025], [0.5, -0.5, 0.0], [0.5, 0.0, 0.866025], [0.5, 0.5, 0.0]]}, "60": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Bicupola", "ShortName": "J46", "vertices": [[-1.618030033, -3.3e-08, 0.431199], [-1.538840033, -0.500000033, -0.431199], [-1.538840033, 0.499999967, -0.431199], [-1.309020033, -0.951057033, 0.431199], [-1.309020033, 0.951056967, 0.431199], [-0.951057033, -1.309020033, -0.431199], [-0.951057033, 1.309019967, -0.431199], [-0.809017033, 0.262865967, 0.95693], [-0.688191033, -0.500000033, -0.95693], [-0.688191033, 0.499999967, -0.95693], [-0.500000033, -1.538840033, 0.431199], [-0.500000033, -0.688191033, 0.95693], [-0.500000033, 1.538839967, 0.431199], [-3.3e-08, -1.618030033, -0.431199], [-3.3e-08, 0.850650967, 0.95693], [-3.3e-08, 1.618029967, -0.431199], [0.262865967, -0.809017033, -0.95693], [0.262865967, 0.809016967, -0.95693], [0.499999967, -1.538840033, 0.431199], [0.499999967, -0.688191033, 0.95693], [0.499999967, 1.538839967, 0.431199], [0.809016967, 0.262865967, 0.95693], [0.850650967, -3.3e-08, -0.95693], [0.951056967, -1.309020033, -0.431199], [0.951056967, 1.309019967, -0.431199], [1.309019967, -0.951057033, 0.431199], [1.309019967, 0.951056967, 0.431199], [1.538839967, -0.500000033, -0.431199], [1.538839967, 0.499999967, -0.431199], [1.618029967, -3.3e-08, 0.431199]]}, "61": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Birotunda", "ShortName": "J48", "vertices": [[0.0, -1.618033989, -0.431198502], [0.0, 1.618033989, -0.431198502], [0.0, -1.37638192, 1.28184931], [0.0, 0.850650808, 1.807580422], [-1.618033989, -0.0, 0.431198502], [-0.5, -1.538841769, 0.431198502], [1.618033989, -0.0, 0.431198502], [-0.951056516, -1.309016994, -0.431198502], [-0.951056516, 1.309016994, -0.431198502], [0.951056516, -1.309016994, -0.431198502], [0.951056516, 1.309016994, -0.431198502], [-0.5, 1.538841769, 0.431198502], [0.5, -1.538841769, 0.431198502], [0.5, 1.538841769, 0.431198502], [-1.309016994, -0.951056516, 0.431198502], [-1.309016994, 0.951056516, 0.431198502], [1.309016994, -0.951056516, 0.431198502], [1.309016994, 0.951056516, 0.431198502], [-1.538841769, -0.5, -0.431198502], [-1.538841769, 0.5, -0.431198502], [1.538841769, -0.5, -0.431198502], [1.538841769, 0.5, -0.431198502], [1.37638192, -0.0, -1.28184931], [0.425325404, -1.309016994, -1.28184931], [0.425325404, 1.309016994, -1.28184931], [-1.113516364, -0.809016994, -1.28184931], [-1.113516364, 0.809016994, -1.28184931], [1.309016994, -0.425325404, 1.28184931], [0.809016994, 1.113516364, 1.28184931], [-0.850650808, -0.0, -1.807580422], [-0.262865556, -0.809016994, -1.807580422], [-0.262865556, 0.809016994, -1.807580422], [0.68819096, -0.5, -1.807580422], [0.68819096, 0.5, -1.807580422], [-0.809016994, 0.262865556, 1.807580422], [-0.5, -0.68819096, 1.807580422], [-1.309016994, -0.425325404, 1.28184931], [-0.809016994, 1.113516364, 1.28184931], [0.5, -0.68819096, 1.807580422], [0.809016994, 0.262865556, 1.807580422]]}, "62": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Cupola", "ShortName": "J24", "vertices": [[-4e-08, -1.61803, 0.2398126], [-4e-08, 1.61803, 0.2398126], [-1.61803004, 0.0, -0.6225844], [0.85065096, 0.0, 0.7655436], [0.26286596, -0.809017, 0.7655436], [0.26286596, 0.809017, 0.7655436], [1.61802996, 0.0, -0.6225844], [-0.95105704, -1.30902, 0.2398126], [-0.95105704, 1.30902, 0.2398126], [0.95105696, -1.30902, 0.2398126], [0.95105696, 1.30902, 0.2398126], [-0.68819104, -0.5, 0.7655436], [-0.68819104, 0.5, 0.7655436], [-0.50000004, -1.53884, -0.6225844], [-0.50000004, 1.53884, -0.6225844], [0.49999996, -1.53884, -0.6225844], [0.49999996, 1.53884, -0.6225844], [-1.30902004, -0.951057, -0.6225844], [-1.30902004, 0.951057, -0.6225844], [1.30901996, -0.951057, -0.6225844], [1.30901996, 0.951057, -0.6225844], [-1.53884004, -0.5, 0.2398126], [-1.53884004, 0.5, 0.2398126], [1.53883996, -0.5, 0.2398126], [1.53883996, 0.5, 0.2398126]]}, "63": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Cupolarotunda", "ShortName": "J47", "vertices": [[-2.9e-08, -1.618030314, -0.735841857], [-2.9e-08, 1.618029686, -0.735841857], [-2.9e-08, -1.376380314, 0.977207143], [-2.9e-08, 0.850650686, 1.502937143], [-1.618030029, -3.14e-07, 0.126556143], [0.850650971, -3.14e-07, -1.261572857], [-0.500000029, -1.538840314, 0.126556143], [0.262865971, -0.809017314, -1.261572857], [0.262865971, 0.809016686, -1.261572857], [1.618029971, -3.14e-07, 0.126556143], [-0.951057029, -1.309020314, -0.735841857], [-0.951057029, 1.309019686, -0.735841857], [0.951056971, -1.309020314, -0.735841857], [0.951056971, 1.309019686, -0.735841857], [-0.688191029, -0.500000314, -1.261572857], [-0.688191029, 0.499999686, -1.261572857], [-0.500000029, 1.538839686, 0.126556143], [0.499999971, -1.538840314, 0.126556143], [0.499999971, 1.538839686, 0.126556143], [-1.309020029, -0.951057314, 0.126556143], [-1.309020029, 0.951056686, 0.126556143], [1.309019971, -0.951057314, 0.126556143], [1.309019971, 0.951056686, 0.126556143], [-1.538840029, -0.500000314, -0.735841857], [-1.538840029, 0.499999686, -0.735841857], [1.538839971, -0.500000314, -0.735841857], [1.538839971, 0.499999686, -0.735841857], [1.309019971, -0.425325314, 0.977207143], [0.809016971, 1.113519686, 0.977207143], [-0.809017029, 0.262865686, 1.502937143], [-0.500000029, -0.688191314, 1.502937143], [-1.309020029, -0.425325314, 0.977207143], [-0.809017029, 1.113519686, 0.977207143], [0.499999971, -0.688191314, 1.502937143], [0.809016971, 0.262865686, 1.502937143]]}, "64": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Pyramid", "ShortName": "J11", "vertices": [[0.0, 0.0, 0.864597273], [-0.850651, 0.0, -0.511784727], [0.850651, 0.0, 0.338865273], [-0.262866, -0.809017, -0.511784727], [-0.262866, 0.809017, -0.511784727], [0.262866, -0.809017, 0.338865273], [0.262866, 0.809017, 0.338865273], [-0.688191, -0.5, 0.338865273], [-0.688191, 0.5, 0.338865273], [0.688191, -0.5, -0.511784727], [0.688191, 0.5, -0.511784727]]}, "65": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Rotunda", "ShortName": "J25", "vertices": [[3.67e-07, -1.61803, -0.083706167], [3.67e-07, 1.61803, -0.083706167], [-1.618029633, 0.0, -0.946103167], [1.618030367, 0.0, -0.946103167], [-0.951056633, -1.30902, -0.083706167], [-0.951056633, 1.30902, -0.083706167], [0.951057367, -1.30902, -0.083706167], [0.951057367, 1.30902, -0.083706167], [-0.499999633, -1.53884, -0.946103167], [-0.499999633, 1.53884, -0.946103167], [0.500000367, -1.53884, -0.946103167], [0.500000367, 1.53884, -0.946103167], [-1.309019633, -0.951057, -0.946103167], [-1.309019633, 0.951057, -0.946103167], [1.309020367, -0.951057, -0.946103167], [1.309020367, 0.951057, -0.946103167], [-1.538839633, -0.5, -0.083706167], [-1.538839633, 0.5, -0.083706167], [1.538840367, -0.5, -0.083706167], [1.538840367, 0.5, -0.083706167], [1.376380367, 0.0, 0.766944833], [0.425325367, -1.30902, 0.766944833], [0.425325367, 1.30902, 0.766944833], [-1.113519633, -0.809017, 0.766944833], [-1.113519633, 0.809017, 0.766944833], [-0.850650633, 0.0, 1.292673833], [-0.262865633, -0.809017, 1.292673833], [-0.262865633, 0.809017, 1.292673833], [0.688191367, -0.5, 1.292673833], [0.688191367, 0.5, 1.292673833]]}, "66": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Square Bicupola", "ShortName": "J45", "vertices": [[0.0, -0.707107, -1.13725], [0.0, 0.707107, -1.13725], [0.0, -1.30656, 0.430148], [0.0, 1.30656, 0.430148], [-0.707107, 0.0, -1.13725], [0.707107, 0.0, -1.13725], [-0.653281, 0.270598, 1.13725], [0.653281, -0.270598, 1.13725], [1.20711, 0.5, -0.430148], [-1.30656, 0.0, 0.430148], [1.30656, 0.0, 0.430148], [-0.92388, -0.92388, 0.430148], [-0.92388, 0.92388, 0.430148], [0.92388, -0.92388, 0.430148], [0.92388, 0.92388, 0.430148], [-0.5, -1.20711, -0.430148], [0.5, 1.20711, -0.430148], [0.5, -1.20711, -0.430148], [-0.5, 1.20711, -0.430148], [1.20711, -0.5, -0.430148], [-1.20711, 0.5, -0.430148], [-1.20711, -0.5, -0.430148], [-0.270598, -0.653281, 1.13725], [0.270598, 0.653281, 1.13725]]}, "67": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Square Cupola", "ShortName": "J23", "vertices": [[-0.0, -0.707107, 0.909804], [-0.0, 0.707107, 0.909804], [-0.0, -1.30656, -0.657599], [-0.0, 1.30656, -0.657599], [-0.707107, 0.0, 0.909804], [0.707107, 0.0, 0.909804], [-1.20711, -0.5, 0.202697], [-1.30656, 0.0, -0.657599], [1.30656, 0.0, -0.657599], [-0.92388, -0.92388, -0.657599], [-0.92388, 0.92388, -0.657599], [0.92388, -0.92388, -0.657599], [0.92388, 0.92388, -0.657599], [-0.5, -1.20711, 0.202697], [0.5, 1.20711, 0.202697], [0.5, -1.20711, 0.202697], [-0.5, 1.20711, 0.202697], [1.20711, -0.5, 0.202697], [-1.20711, 0.5, 0.202697], [1.20711, 0.5, 0.202697]]}, "68": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Square Dipyramid", "ShortName": "J17", "vertices": [[-0.5, -0.5, -0.420448], [-0.5, 0.5, -0.420448], [0.0, 0.0, -1.12755], [0.0, 0.0, 1.12755], [0.0, -0.707107, 0.420448], [0.0, 0.707107, 0.420448], [0.5, -0.5, -0.420448], [0.5, 0.5, -0.420448], [-0.707107, 0.0, 0.420448], [0.707107, 0.0, 0.420448]]}, "69": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Square Pyramid", "ShortName": "J10", "vertices": [[-0.5, -0.5, -0.545731333], [-0.5, 0.5, -0.545731333], [0.0, 0.0, 1.002266667], [0.0, -0.707107, 0.295164667], [0.0, 0.707107, 0.295164667], [0.5, -0.5, -0.545731333], [0.5, 0.5, -0.545731333], [-0.707107, 0.0, 0.295164667], [0.707107, 0.0, 0.295164667]]}, "70": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Triangular Bicupola", "ShortName": "J44", "vertices": [[-1.0, 0.0, 0.4278], [-0.5, -0.866025, 0.4278], [-0.5, 0.866025, 0.4278], [-0.5, -0.288675, 1.2443], [0.0, -1.0, -0.4278], [0.0, 1.0, -0.4278], [0.0, 0.57735, 1.2443], [0.5, -0.288675, 1.2443], [0.5, -0.866025, 0.4278], [0.5, 0.866025, 0.4278], [1.0, 0.0, 0.4278], [-0.57735, 0.0, -1.2443], [0.288675, -0.5, -1.2443], [0.288675, 0.5, -1.2443], [-0.866025, -0.5, -0.4278], [-0.866025, 0.5, -0.4278], [0.866025, -0.5, -0.4278], [0.866025, 0.5, -0.4278]]}, "71": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Triangular Cupola", "ShortName": "J22", "vertices": [[-1.0, 0.0, -0.6766594], [-0.5, -0.866025, -0.6766594], [-0.5, 0.866025, -0.6766594], [0.0, -1.0, 0.1789406], [0.0, 1.0, 0.1789406], [0.5, -0.866025, -0.6766594], [0.5, 0.866025, -0.6766594], [1.0, 0.0, -0.6766594], [-0.288675, -0.5, 0.9954376], [-0.288675, 0.5, 0.9954376], [0.57735, 0.0, 0.9954376], [-0.866025, -0.5, 0.1789406], [-0.866025, 0.5, 0.1789406], [0.866025, -0.5, 0.1789406], [0.866025, 0.5, 0.1789406]]}, "72": {"AlternativeName": null, "Comment": null, "Name": "Hebesphenomegacorona", "ShortName": "J89", "vertices": [[-0.871243429, 0.139987014, -0.012209536], [-0.733789429, -0.845001986, -0.116605036], [-0.661500429, -0.423077986, 0.787138964], [-0.603870429, -0.236300986, -0.899294036], [-0.467798429, 0.557870014, 0.801796964], [-0.180802429, 0.856974014, -0.108242036], [0.086570571, 0.480685014, -0.995326036], [0.110277571, -0.890736986, -0.650886036], [0.152619571, -0.798265986, 0.343928964], [0.246344571, -0.096561286, 1.050203964], [0.492445571, 0.775233014, 0.626638964], [0.786968571, 0.734226014, -0.328124036], [0.800718571, -0.173749986, -0.746918036], [0.843059571, -0.081279886, 0.247896964]]}, "73": {"AlternativeName": null, "Comment": null, "Name": "Icosahedron", "ShortName": "P04", "vertices": [[0.0, 0.0, -0.951057], [0.0, 0.0, 0.951057], [-0.850651, 0.0, -0.425325], [0.850651, 0.0, 0.425325], [0.688191, -0.5, -0.425325], [0.688191, 0.5, -0.425325], [-0.688191, -0.5, 0.425325], [-0.688191, 0.5, 0.425325], [-0.262866, -0.809017, -0.425325], [-0.262866, 0.809017, -0.425325], [0.262866, -0.809017, 0.425325], [0.262866, 0.809017, 0.425325]]}, "74": {"AlternativeName": null, "Comment": null, "Name": "Icosidodecahedron", "ShortName": "A02", "vertices": [[-0.0, -1.61803, 0.0], [-0.0, 1.61803, 0.0], [0.262866, -0.809017, -1.37638], [0.262866, 0.809017, -1.37638], [0.425325, -1.30902, 0.850651], [0.425325, 1.30902, 0.850651], [0.688191, -0.5, 1.37638], [0.688191, 0.5, 1.37638], [1.11352, -0.809017, -0.850651], [1.11352, 0.809017, -0.850651], [-1.37638, 0.0, -0.850651], [-0.688191, -0.5, -1.37638], [-0.688191, 0.5, -1.37638], [1.37638, 0.0, 0.850651], [0.951057, -1.30902, 0.0], [0.951057, 1.30902, 0.0], [0.850651, 0.0, -1.37638], [-0.951057, -1.30902, 0.0], [-0.951057, 1.30902, 0.0], [-1.53884, -0.5, 0.0], [-1.53884, 0.5, 0.0], [1.53884, -0.5, 0.0], [1.53884, 0.5, 0.0], [-0.850651, 0.0, 1.37638], [-1.11352, -0.809017, 0.850651], [-1.11352, 0.809017, 0.850651], [-0.425325, -1.30902, -0.850651], [-0.425325, 1.30902, -0.850651], [-0.262866, -0.809017, 1.37638], [-0.262866, 0.809017, 1.37638]]}, "75": {"AlternativeName": null, "Comment": null, "Name": "Metabiaugmented Dodecahedron", "ShortName": "J60", "vertices": [[0.008638727, 0.063383182, 1.356026818], [0.008638727, 0.063383182, -1.446493182], [0.187049727, -1.245636818, 0.421852818], [0.187049727, 1.372403182, 0.421852818], [0.475724727, -0.745633818, -1.089673182], [0.475724727, 0.872400182, -1.089673182], [1.053078727, -0.745633818, 0.421852818], [1.053078727, 0.872400182, 0.421852818], [-1.214211273, -0.436616818, 0.421852818], [-1.214211273, 0.563383182, 0.421852818], [1.231488727, -0.436616818, -0.512319182], [1.231488727, 0.563383182, -0.512319182], [-0.925533273, 0.063383182, -1.089673182], [-0.458447273, -0.745633818, 0.999206818], [-0.458447273, 0.872400182, 0.999206818], [0.942810727, 0.063383182, 0.999206818], [-1.035801273, -0.745633818, -0.512319182], [-1.035801273, 0.872400182, -0.512319182], [-0.986486273, 0.063383182, 1.257406818], [-0.169772273, -1.245636818, -0.512319182], [-0.169772273, 1.372403182, -0.512319182], [0.813711727, -1.331046818, -0.352743182]]}, "76": {"AlternativeName": null, "Comment": null, "Name": "Metabiaugmented Hexagonal Prism", "ShortName": "J56", "vertices": [[-1.0, -0.112366571, -0.5], [-1.0, -0.112366571, 0.5], [-0.5, -0.978391571, -0.5], [-0.5, -0.978391571, 0.5], [-0.5, 0.753658429, -0.5], [-0.5, 0.753658429, 0.5], [0.5, -0.978391571, -0.5], [0.5, -0.978391571, 0.5], [0.5, 0.753658429, -0.5], [0.5, 0.753658429, 0.5], [1.0, -0.112366571, -0.5], [1.0, -0.112366571, 0.5], [-1.36237, 0.674199429, 0.0], [1.36237, 0.674199429, 0.0]]}, "77": {"AlternativeName": null, "Comment": null, "Name": "Metabiaugmented Truncated Dodecahedron", "ShortName": "J70", "vertices": [[-2.952571157, -0.2730359, -0.305566317], [-2.943501157, -0.3283119, 0.692863683], [-2.712911157, 0.6964551, -0.254068317], [-2.703851157, 0.6411791, 0.744360683], [-2.663851157, -1.1596999, -0.666779317], [-2.649181157, -1.2491299, 0.948714683], [-2.540081157, 0.1826311, -1.094377317], [-2.516341157, 0.0379168, 1.519552683], [-2.476341157, -1.7629599, 0.108408483], [-2.418931157, 1.4839301, 0.287644683], [-2.251361157, -0.7040289, -1.455587317], [-2.222021157, -0.8829059, 1.775402683], [-1.966441157, 0.1387231, -1.912307317], [-1.957041157, -1.6248499, -1.199737317], [-1.933301157, -1.7695699, 1.414182683], [-1.928031157, -0.0954302, 2.317112683], [-1.770411157, 2.2442701, 0.323848683], [-1.769531157, -2.2281099, -0.424549317], [-1.760461157, -2.2833899, 0.573880683], [-1.211111157, 0.5815011, -2.395437317], [-1.195891157, -2.2720199, -1.242477317], [-1.172151157, -2.4167399, 1.371442683], [-1.163621157, 0.2920721, 2.832412683], [-1.015071157, 2.6870501, -0.159284317], [-1.006011157, 2.6317701, 0.839145683], [-0.562589157, 1.3418401, -2.359237317], [-0.515109157, 1.0524101, 2.868612683], [-0.441435157, 2.6431401, -0.977217317], [-0.417695157, 2.4984201, 1.636712683], [-0.273870157, 0.4551791, -2.720447317], [-0.268601157, 2.1293101, -1.817517317], [-0.258655157, -2.3983399, -1.567487317], [-0.230188157, 1.8951601, 2.411892683], [-0.229310157, -2.5772199, 1.663502683], [-0.220785157, 0.1315881, 3.124462683], [0.376103843, 1.6782401, -2.434677317], [0.487275843, -0.1919909, -2.763187317], [0.495801843, 2.5168201, -1.302227317], [0.496678843, -1.9555699, -2.050617317], [0.525145843, 2.3379401, 1.928762683], [0.535091843, -2.1897199, 2.178792683], [0.540360843, -0.5155829, 3.081722683], [0.664822843, 0.7915781, -2.795887317], [0.684185843, -2.5588299, -1.275437317], [0.707925843, -2.7035399, 1.338492683], [0.781599843, -1.1128099, -2.507337317], [0.829080843, -1.4022399, 2.720512683], [1.140508843, 2.0657401, -1.919377317], [1.272498843, -2.6921699, -0.477870317], [1.281568843, -2.7474499, 0.520559683], [1.430118843, -0.3524749, -2.471137317], [1.438638843, 2.3563301, -1.010167317], [1.462378843, 2.2116201, 1.603752683], [1.477598843, -0.6419039, 2.756712683], [1.607658843, 0.6310941, -2.503837317], [1.901648843, 1.4185701, -1.962127317], [2.026958843, 2.2229901, -0.212604317], [2.036018843, 2.1677101, 0.785825683], [2.036898843, -2.3046699, 0.037427063], [2.194518843, 0.0350269, -1.955837317], [2.199788843, 1.7091601, -1.052907317], [2.223528843, 1.5644501, 1.561012683], [2.232928843, -0.1991259, 2.273582683], [2.488508843, 0.8225021, -1.414127317], [2.517848843, 0.6436251, 1.816862683], [2.685418843, -1.5443299, 0.073631683], [2.782828843, -0.0983201, -1.158277317], [2.806568843, -0.2430349, 1.455652683], [2.970338843, -0.7015819, -0.383085317], [2.979408843, -0.7568579, 0.615344683]]}, "78": {"AlternativeName": null, "Comment": null, "Name": "Metabidiminished Icosahedron", "ShortName": "J62", "vertices": [[-0.5, 0.0, -0.709017], [-0.5, 0.0, 0.909017], [0.0, -0.809017, -0.4], [0.0, 0.809017, -0.4], [0.5, 0.0, -0.709017], [0.5, 0.0, 0.909017], [-0.809017, -0.5, 0.1], [-0.809017, 0.5, 0.1], [0.809017, -0.5, 0.1], [0.809017, 0.5, 0.1]]}, "79": {"AlternativeName": null, "Comment": null, "Name": "Metabidiminished Rhombicosidodecahedron", "ShortName": "J81", "vertices": [[-2.034672474, -0.381270867, 0.423803728], [-1.942632474, -0.485832867, -0.566446272], [-1.941272474, 0.609723133, 0.327843728], [-1.849232474, 0.505160133, -0.662406272], [-1.757862474, -1.261133867, 0.037516308], [-1.642312474, -0.644739867, 1.305071728], [-1.513342474, 1.333326133, -0.213709272], [-1.491192474, 0.958722133, 1.149811728], [-1.401342474, -0.918487867, -1.287428272], [-1.365502474, -1.524603867, 0.918788728], [-1.306422474, 0.183421133, 1.753771728], [-1.250222474, 0.684973133, -1.442698272], [-1.216572474, -1.693783867, -0.683469272], [-1.063262474, 1.682316133, 0.608255728], [-0.973412474, -0.194889867, -1.828988272], [-0.914332474, 1.513136133, -0.993998272], [-0.670903474, 1.418856133, 1.489531728], [-0.617568474, -1.513973867, -1.463758272], [-0.581731474, -2.120093867, 0.742458728], [-0.489688474, -2.224653867, -0.247791272], [-0.486137474, 0.643551133, 2.093491728], [-0.373047474, 1.080476133, -1.714988272], [-0.189637474, -0.790376867, -2.005318272], [-0.186086474, 2.077826133, 0.335966728], [-0.096238474, 0.200616133, -2.101278272], [-0.094044474, 1.973266133, -0.654283272], [0.109320226, -2.044843867, -1.028088272], [0.206269526, 1.814356133, 1.217241728], [0.409633526, -2.203743867, 0.843438728], [0.447240526, 1.540606133, -1.375268272], [0.501676526, -2.308303867, -0.146811272], [0.505227526, 0.559895133, 2.194471728], [0.801727526, -0.874032867, -1.904338272], [0.805278526, 1.994166133, 0.436946728], [0.895126526, 0.116960133, -2.000298272], [0.897320526, 1.889606133, -0.553303272], [0.933158526, 1.283496133, 1.652921728], [0.986493526, -1.649333867, -1.300368272], [1.229917526, -1.743613867, 1.183161728], [1.231017526, 0.945121133, -1.551598272], [1.288997526, -0.035592527, 2.018141728], [1.378847526, -1.912803867, -0.419101272], [1.532167526, 1.463306133, 0.872624728], [1.565807526, -0.915455867, 1.631851728], [1.681097526, 1.294116133, -0.729633272], [1.716937526, 0.688006133, 1.476591728], [1.828927526, -1.563803867, 0.402864728], [2.073447526, 1.030656133, 0.151638728], [2.164817526, -0.735642867, 0.851560728], [2.258217526, 0.255351133, 0.755601728]]}, "80": {"AlternativeName": null, "Comment": null, "Name": "Metabigyrate Rhombicosidodecahedron", "ShortName": "J74", "vertices": [[-2.15713033, -0.114704717, 0.565411907], [-2.12641033, 0.680289283, -0.040427493], [-2.10096033, -0.721244717, -0.227654093], [-2.07025033, 0.073748983, -0.833493093], [-1.84552033, -1.058149717, 0.678568907], [-1.76510033, 1.023170283, -0.907538093], [-1.71309033, 0.135164283, 1.425869907], [-1.66339033, 1.421490283, 0.445604907], [-1.56604033, -1.452779717, -0.650401093], [-1.51634033, -0.166451717, -1.630670093], [-1.40794033, 1.084590283, 1.351829907], [-1.40148033, -0.808281717, 1.539029907], [-1.31059033, -1.789679717, 0.255822907], [-1.30207033, 1.764370283, -0.421506093], [-1.21119033, 0.782970283, -1.704710093], [-1.20472033, -1.109899717, -1.517510093], [-0.93845133, -0.067079217, 2.025059907], [-0.88729333, 1.826350283, 0.929092907], [-0.75668433, -2.029889717, -0.541353093], [-0.67626433, 0.051435983, -2.127460093], [-0.63330133, 0.882342283, 1.951019907], [-0.59212033, -1.385389717, 1.648079907], [-0.53595133, -1.991929717, 0.855011907], [-0.52597533, 2.169230283, 0.061981607], [-0.46199833, 1.982260283, -0.918298093], [-0.40582933, 1.375720283, -1.711360093], [-0.39536533, -1.686999717, -1.408460093], [-0.36464833, -0.892009717, -2.014300093], [-0.12909533, -0.644185717, 2.134109907], [-0.11265233, 1.624110283, 1.528279907], [0.01795747, -2.232129717, 0.057836307], [0.12909467, 0.644186283, -2.134110093], [0.36464767, 0.892010283, 2.014299907], [0.40582867, -1.375719717, 1.711359907], [0.46199767, -1.982259717, 0.918297907], [0.47197367, 2.178900283, 0.125266907], [0.53595067, 1.991930283, -0.855012093], [0.59211967, 1.385390283, -1.648080093], [0.60258367, -1.677339717, -1.345180093], [0.63330067, -0.882341717, -1.951020093], [0.67626367, -0.051435417, 2.127459907], [0.72742167, 1.842000283, 1.031489907], [0.85803067, -2.014239717, -0.438955093], [0.93845067, 0.067079783, -2.025060093], [1.20471967, 1.109900283, 1.517509907], [1.21118967, -0.782969717, 1.704709907], [1.30206967, -1.764369717, 0.421505907], [1.31058967, 1.789680283, -0.255823093], [1.40147967, 0.808282283, -1.539030093], [1.47005967, -1.397439717, -0.933904093], [1.50076967, -0.602440717, -1.539740093], [1.51633967, 0.166452283, 1.630669907], [1.56603967, 1.452780283, 0.650400907], [1.76509967, -1.023169717, 0.907537907], [1.84551967, 1.058150283, -0.678569093], [1.91409967, -1.147569717, -0.073443193], [1.96379967, 0.138761283, -1.053710093], [2.07024967, -0.073748417, 0.833493907], [2.10095967, 0.721245283, 0.227653907], [2.21924967, -0.198144717, -0.147488093]]}, "81": {"AlternativeName": null, "Comment": null, "Name": "Metagyrate Diminished Rhombicosidodecahedron", "ShortName": "J78", "vertices": [[-2.156463838, 0.479212042, -0.283451002], [-2.118633838, -0.371178958, -0.808241002], [-2.106893838, -0.043710058, 0.567486998], [-2.069063838, -0.894100958, 0.042697398], [-1.869223838, 0.495055042, -1.241186002], [-1.739463838, -0.873972958, 0.986603998], [-1.703353838, 1.357881042, -0.133026002], [-1.623153838, 0.511779042, 1.243823998], [-1.604313838, -0.868466958, -1.506946002], [-1.524123838, -1.714568958, -0.130095502], [-1.416123838, 1.373731042, -1.090756002], [-1.373753838, 1.378011042, 0.810880998], [-1.354913838, -0.002233758, -1.939886002], [-1.255733838, -0.318483958, 1.662933998], [-1.236883838, -1.698728958, -1.087826002], [-1.194523838, -1.694448958, 0.813810998], [-0.932383838, 1.929221042, -0.414421002], [-0.802622838, 0.560189042, 1.813363998], [-0.680204838, -2.191738958, 0.115108998], [-0.621771838, 1.419491042, -1.696486002], [-0.602783838, 1.949341042, 0.529484998], [-0.583942838, 0.569098042, -2.221276002], [-0.553219838, 1.426421042, 1.380423998], [-0.411813838, -0.795644958, 1.908143998], [-0.392972838, -2.175888958, -0.842621002], [-0.373984838, -1.646038958, 1.383353998], [-0.138032838, 1.974981042, -1.020155002], [0.041292262, 0.083028642, 2.058563998], [0.140330162, -2.143328958, 0.684651998], [0.376283162, 1.477691042, -1.718856002], [0.395270162, 2.007551042, 0.507117998], [0.414112162, 0.627299042, -2.243646002], [0.444834162, 1.484621042, 1.358053998], [0.471460162, -1.215348958, 1.699153998], [0.605081162, -2.117688958, -0.864988002], [0.682502162, 2.023391042, -0.450612002], [0.812263162, 0.654361042, 1.777173998], [0.924567162, -0.336671958, 1.849573998], [0.934681162, -2.097558958, 0.078917998], [0.985776162, -1.712628958, 1.000453998], [1.196816162, 1.526101042, -1.149316002], [1.239186162, 1.530381042, 0.752321998], [1.258026162, 0.150139042, -1.998446002], [1.376056162, -1.546358958, -1.146386002], [1.526416162, 1.546231042, -0.205408002], [1.606616162, 0.700122042, 1.171443998], [1.625456162, -0.680124958, -1.579326002], [1.705656162, -1.526228958, -0.202477002], [1.718916162, -0.290909958, 1.243843998], [1.741766162, 0.705628042, -1.322106002], [1.756746162, -1.141298958, 0.719052998], [2.071366162, 0.725756042, -0.378200002], [2.109196162, -0.124634958, -0.902990002], [2.120926162, 0.202834042, 0.472737998], [2.158756162, -0.647556958, -0.052051802]]}, "82": {"AlternativeName": null, "Comment": null, "Name": "Obtuse Golden Rhombohedron", "ShortName": "O18", "vertices": [[-0.947214, 0.0854102, 0.262866], [-0.5, -0.809017, 0.262866], [-0.5, 0.809017, -0.262866], [-0.0527864, -0.0854102, -0.262866], [0.0527864, 0.0854102, 0.262866], [0.5, -0.809017, 0.262866], [0.5, 0.809017, -0.262866], [0.947214, -0.0854102, -0.262866]]}, "83": {"AlternativeName": null, "Comment": null, "Name": "Octahedron", "ShortName": "P02", "vertices": [[-0.707107, 0.0, 0.0], [0.0, 0.707107, 0.0], [0.0, 0.0, -0.707107], [0.0, 0.0, 0.707107], [0.0, -0.707107, 0.0], [0.707107, 0.0, 0.0]]}, "84": {"AlternativeName": null, "Comment": null, "Name": "Parabiaugmented Dodecahedron", "ShortName": "J59", "vertices": [[0.0, 0.0, 1.40126], [0.0, 0.0, -1.40126], [0.178411, -1.30902, 0.467086], [0.178411, 1.30902, 0.467086], [0.467086, -0.809017, -1.04444], [0.467086, 0.809017, -1.04444], [1.04444, -0.809017, 0.467086], [1.04444, 0.809017, 0.467086], [-1.22285, -0.5, 0.467086], [-1.22285, 0.5, 0.467086], [1.22285, -0.5, -0.467086], [1.22285, 0.5, -0.467086], [-0.934172, 0.0, -1.04444], [-0.467086, -0.809017, 1.04444], [-0.467086, 0.809017, 1.04444], [0.934172, 0.0, 1.04444], [-1.04444, -0.809017, -0.467086], [-1.04444, 0.809017, -0.467086], [0.995125, 0.0, -1.30264], [-0.995125, 0.0, 1.30264], [-0.178411, -1.30902, -0.467086], [-0.178411, 1.30902, -0.467086]]}, "85": {"AlternativeName": null, "Comment": null, "Name": "Parabiaugmented Hexagonal Prism", "ShortName": "J55", "vertices": [[-1.0, 0.0, -0.5], [-1.0, 0.0, 0.5], [-0.5, -0.866025, -0.5], [-0.5, -0.866025, 0.5], [-0.5, 0.866025, -0.5], [-0.5, 0.866025, 0.5], [0.5, -0.866025, -0.5], [0.5, -0.866025, 0.5], [0.5, 0.866025, -0.5], [0.5, 0.866025, 0.5], [1.0, 0.0, -0.5], [1.0, 0.0, 0.5], [-1.36237, -0.786566, 0.0], [1.36237, 0.786566, 0.0]]}, "86": {"AlternativeName": null, "Comment": null, "Name": "Parabiaugmented Truncated Dodecahedron", "ShortName": "J69", "vertices": [[-2.95761, 0.246541, -0.096963], [-2.81512, 0.316165, 0.890381], [-2.8109, -0.335955, -0.896447], [-2.72848, 1.12885, 0.314162], [-2.43785, -0.153677, 1.68845], [-2.43103, -1.20883, -1.20269], [-2.34439, -0.39614, -1.77891], [-2.21104, 1.97397, 0.179892], [-1.96992, -0.983522, 1.99242], [-1.96309, -2.03867, -0.89873], [-1.85913, -1.723, -1.84188], [-1.77249, -0.910309, -2.4181], [-1.74079, -0.101208, 2.40354], [-1.73628, 0.0889763, -2.40729], [-1.60293, 2.45909, -0.448487], [-1.59005, -1.8564, 1.68617], [-1.58583, -2.50852, -0.100657], [-1.46044, 2.52871, 0.538858], [-1.44334, -2.43889, 0.886688], [-1.21883, 0.934094, -2.54156], [-1.13642, 2.3989, -1.33095], [-1.11932, -2.5687, -0.983124], [-1.01536, -2.25303, -1.92628], [-0.990189, 0.453532, 2.76251], [-0.989708, 1.81641, -2.13044], [-0.875179, -0.938065, -2.85862], [-0.838961, 0.0612202, -2.84781], [-0.763375, 2.58118, 1.25395], [-0.746275, -2.38642, 1.60178], [-0.472746, 1.29865, 2.62824], [-0.407243, -1.76791, -2.55466], [-0.386111, 2.11134, 2.05202], [-0.239103, 2.37115, -1.77147], [-0.222004, -2.59646, -1.42364], [-0.00481035, 0.468806, 2.9322], [0.00481035, -0.468806, -2.9322], [0.222004, 2.59646, 1.42364], [0.239103, -2.37115, 1.77147], [0.386111, -2.11134, -2.05202], [0.407243, 1.76791, 2.55466], [0.472746, -1.29865, -2.62824], [0.746275, 2.38642, -1.60178], [0.763375, -2.58118, -1.25395], [0.838961, -0.0612202, 2.84781], [0.875179, 0.938065, 2.85862], [0.989708, -1.81641, 2.13044], [0.990189, -0.453532, -2.76251], [1.01536, 2.25303, 1.92628], [1.11932, 2.5687, 0.983124], [1.13642, -2.3989, 1.33095], [1.21883, -0.934094, 2.54156], [1.44334, 2.43889, -0.886688], [1.46044, -2.52871, -0.538858], [1.58583, 2.50852, 0.100657], [1.59005, 1.8564, -1.68617], [1.60293, -2.45909, 0.448487], [1.73628, -0.0889763, 2.40729], [1.74079, 0.101208, -2.40354], [1.77249, 0.910309, 2.4181], [1.85913, 1.723, 1.84188], [1.96309, 2.03867, 0.89873], [1.96992, 0.983522, -1.99242], [2.21104, -1.97397, -0.179892], [2.34439, 0.39614, 1.77891], [2.43103, 1.20883, 1.20269], [2.43785, 0.153677, -1.68845], [2.72848, -1.12885, -0.314162], [2.8109, 0.335955, 0.896447], [2.81512, -0.316165, -0.890381], [2.95761, -0.246541, 0.096963]]}, "87": {"AlternativeName": null, "Comment": null, "Name": "Parabidiminished Rhombicosidodecahedron", "ShortName": "J80", "vertices": [[-2.19149, -0.351484, -0.244698], [-2.10378, -0.155834, 0.732046], [-2.04429, 0.501655, -0.745182], [-1.90236, 0.818222, 0.835221], [-1.86559, 1.22457, -0.0777559], [-1.79587, -0.866028, -1.00544], [-1.64867, -0.0128901, -1.50592], [-1.56623, -0.353811, 1.55171], [-1.44725, 0.961166, -1.40275], [-1.36482, 0.620245, 1.65488], [-1.26855, 1.68408, -0.735321], [-1.21761, 1.47338, 1.1544], [-1.18084, 1.87973, 0.241422], [-1.06803, -1.50293, -1.2596], [-0.829841, -0.122523, -2.0694], [-0.784177, -0.869794, 1.90121], [-0.628423, 0.851534, -1.96623], [-0.470986, -1.04342, -1.91716], [-0.458277, 0.706262, 2.06815], [-0.339292, 2.02124, -0.886307], [-0.31107, 1.5594, 1.56766], [-0.28597, -2.01891, -0.9101], [-0.251577, 2.21689, 0.0904368], [-0.0994221, -0.214633, 2.22039], [-0.0563312, -1.5067, 1.64705], [0.0563312, 1.5067, -1.64705], [0.0994221, 0.214633, -2.22039], [0.251577, -2.21689, -0.0904368], [0.28597, 2.01891, 0.9101], [0.31107, -1.5594, -1.56766], [0.339292, -2.02124, 0.886307], [0.458277, -0.706262, -2.06815], [0.470986, 1.04342, 1.91716], [0.628423, -0.851534, 1.96623], [0.784177, 0.869794, -1.90121], [0.829841, 0.122523, 2.0694], [1.06803, 1.50293, 1.2596], [1.18084, -1.87973, -0.241422], [1.21761, -1.47338, -1.1544], [1.26855, -1.68408, 0.735321], [1.36482, -0.620245, -1.65488], [1.44725, -0.961166, 1.40275], [1.56623, 0.353811, -1.55171], [1.64867, 0.0128901, 1.50592], [1.79587, 0.866028, 1.00544], [1.86559, -1.22457, 0.0777559], [1.90236, -0.818222, -0.835221], [2.04429, -0.501655, 0.745182], [2.10378, 0.155834, -0.732046], [2.19149, 0.351484, 0.244698]]}, "88": {"AlternativeName": null, "Comment": null, "Name": "Parabigyrate Rhombicosidodecahedron", "ShortName": "J73", "vertices": [[-2.19876998, 0.279547983, 0.270815], [-2.13820998, -0.168840017, -0.620971], [-2.10445998, -0.714504017, 0.216352], [-1.91264998, 1.143279983, -0.144039], [-1.85802998, 0.260375983, 1.21078], [-1.85209998, 0.694891983, -1.03583], [-1.76371998, -0.733677017, 1.15632], [-1.69949998, -0.913518017, -1.12395], [-1.66573998, -1.459180017, -0.286624], [-1.39508998, 1.657919983, 0.539531], [-1.36132998, 1.112259983, 1.37685], [-1.35538998, 1.546779983, -0.869749], [-1.24615998, -0.219033017, 1.83989], [-1.23654998, 0.484029983, -1.79519], [-1.14223998, -0.510022017, -1.84966], [-1.11441998, -1.490200017, 1.23427], [-1.05386998, -1.938590017, 0.342484], [-0.90550698, -1.407700017, -1.478], [-0.87175098, -1.953370017, -0.640681], [-0.83782798, 2.061419983, -0.186179], [-0.74945498, 0.632849983, 2.00596], [-0.73984798, 1.335909983, -1.62912], [-0.64178598, 1.779909983, 1.1858], [-0.59685798, -0.975560017, 1.91784], [-0.30113398, 0.591235983, -2.13209], [-0.20682398, -0.402816017, -2.18656], [-0.15814398, -1.720240017, 1.41486], [-0.09758938, -2.168620017, 0.523079], [-0.08452808, 2.183399983, 0.460087], [-0.02991078, 1.300499983, 1.8149], [0.02991082, -1.300500017, -1.8149], [0.08452802, -2.183400017, -0.460087], [0.09758932, 2.168619983, -0.523079], [0.15814402, 1.720239983, -1.41486], [0.20682402, 0.402815983, 2.18656], [0.30113402, -0.591236017, 2.13209], [0.59685802, 0.975559983, -1.91784], [0.64178602, -1.779910017, -1.1858], [0.73984802, -1.335910017, 1.62912], [0.74945502, -0.632850017, -2.00596], [0.83782802, -2.061420017, 0.186179], [0.87175102, 1.953369983, 0.640681], [0.90550602, 1.407699983, 1.478], [1.05387002, 1.938589983, -0.342484], [1.11442002, 1.490199983, -1.23427], [1.14224002, 0.510021983, 1.84966], [1.23655002, -0.484030017, 1.79519], [1.24616002, 0.219032983, -1.83989], [1.35539002, -1.546780017, 0.869749], [1.36133002, -1.112260017, -1.37685], [1.39509002, -1.657920017, -0.539531], [1.66574002, 1.459179983, 0.286624], [1.69950002, 0.913517983, 1.12395], [1.76372002, 0.733676983, -1.15632], [1.85210002, -0.694892017, 1.03583], [1.85803002, -0.260376017, -1.21078], [1.91265002, -1.143280017, 0.144039], [2.10446002, 0.714504983, -0.216352], [2.13821002, 0.168839983, 0.620971], [2.19877002, -0.279548017, -0.270815]]}, "89": {"AlternativeName": null, "Comment": null, "Name": "Paragyrate Diminished Rhombicosidodecahedron", "ShortName": "J77", "vertices": [[-2.147782129, -0.396877776, -0.435706127], [-2.049702129, 0.598175224, -0.451537127], [-2.002372129, -0.706814776, 0.503865873], [-1.843672129, 0.903214224, 0.478249873], [-1.814422129, 0.096686224, 1.068720873], [-1.790842129, -0.446900776, -1.368499127], [-1.761592129, -1.253427776, -0.778023127], [-1.692762129, 0.548151224, -1.384329127], [-1.616192129, -1.563367776, 0.161548873], [-1.504812129, 1.351652224, -0.819470127], [-1.410162129, -1.258327776, 1.091340873], [-1.298792129, 1.656692224, 0.110316873], [-1.295902129, 0.915579224, 1.314780873], [-1.222212129, -0.454826776, 1.656190873], [-1.067902129, -0.837777776, -1.938209127], [-1.038652129, -1.644307776, -1.347739127], [-0.909206129, 0.772250224, -1.963819127], [-0.803381129, -2.145797776, 0.172523873], [-0.751017129, 1.669052224, 0.946852873], [-0.721258129, 1.575752224, -1.398969127], [-0.703691129, 0.364066224, 1.902260873], [-0.597356129, -1.840757776, 1.102310873], [-0.523022129, -0.084301076, -2.306139127], [-0.446446129, -2.195817776, -0.760265127], [-0.387902129, 2.069312224, 0.105460873], [-0.293249129, -0.540665776, 2.016270873], [-0.030967029, 2.019292224, -0.827328127], [0.001682591, 1.184872224, -1.968679127], [0.092935171, -1.397217776, 1.673950873], [0.125584871, -2.231637776, 0.532598873], [0.177948871, 1.583222224, 1.306930873], [0.207197871, 0.776689224, 1.897400873], [0.387866871, 0.328322224, -2.310999127], [0.482519871, -2.281657776, -0.400190127], [0.541063871, 1.983472224, 0.465535873], [0.617639871, -0.128041776, 2.011410873], [0.691973871, 1.628412224, -1.397039127], [0.815875871, -1.788097776, 1.104240873], [0.897998871, 1.933452224, -0.467253127], [1.003823871, -0.984593776, 1.669090873], [1.133277871, 1.431962224, 1.053010873], [1.162527871, 0.625435224, 1.643480873], [1.316827871, 0.242484224, -1.950919127], [1.393407871, -1.869037776, -0.405046127], [1.504777871, 1.045982224, -1.386069127], [1.599437871, -1.563997776, 0.524741873], [1.710807871, 1.351022224, -0.456278127], [1.787377871, -0.760494776, 1.089600873], [1.856217871, 1.041082224, 0.483293873], [1.885467871, 0.234558224, 1.073760873], [1.909037871, -0.309028776, -1.363449127], [1.938287871, -1.115557776, -0.772979127], [2.096987871, 0.494471224, -0.798595127], [2.144317871, -0.810517776, 0.156808873], [2.242397871, 0.184534224, 0.140976873]]}, "90": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Cupola", "ShortName": "J05", "vertices": [[-6.7e-08, -1.61803, -0.175243667], [-6.7e-08, 1.61803, -0.175243667], [0.850650933, 0.0, 0.350487333], [0.262865933, -0.809017, 0.350487333], [0.262865933, 0.809017, 0.350487333], [-0.951057067, -1.30902, -0.175243667], [-0.951057067, 1.30902, -0.175243667], [0.951056933, -1.30902, -0.175243667], [0.951056933, 1.30902, -0.175243667], [-0.688191067, -0.5, 0.350487333], [-0.688191067, 0.5, 0.350487333], [-1.538840067, -0.5, -0.175243667], [-1.538840067, 0.5, -0.175243667], [1.538839933, -0.5, -0.175243667], [1.538839933, 0.5, -0.175243667]]}, "91": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Gyrobicupola", "ShortName": "J31", "vertices": [[0.0, -1.61803, 0.0], [0.0, 1.61803, 0.0], [-0.688191, -0.5, 0.525731], [-0.688191, 0.5, 0.525731], [0.688191, -0.5, -0.525731], [0.688191, 0.5, -0.525731], [-0.850651, 0.0, -0.525731], [0.850651, 0.0, 0.525731], [-0.262866, -0.809017, -0.525731], [-0.262866, 0.809017, -0.525731], [0.262866, -0.809017, 0.525731], [0.262866, 0.809017, 0.525731], [-0.951057, -1.30902, 0.0], [-0.951057, 1.30902, 0.0], [0.951057, -1.30902, 0.0], [0.951057, 1.30902, 0.0], [-1.53884, -0.5, 0.0], [-1.53884, 0.5, 0.0], [1.53884, -0.5, 0.0], [1.53884, 0.5, 0.0]]}, "92": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Gyrocupolarotunda", "ShortName": "J33", "vertices": [[4e-07, -1.61803, -0.34026], [4e-07, 1.61803, -0.34026], [0.8506514, 0.0, -0.865991], [0.2628664, -0.809017, -0.865991], [0.2628664, 0.809017, -0.865991], [-0.9510566, -1.30902, -0.34026], [-0.9510566, 1.30902, -0.34026], [0.9510574, -1.30902, -0.34026], [0.9510574, 1.30902, -0.34026], [-0.6881906, -0.5, -0.865991], [-0.6881906, 0.5, -0.865991], [-1.5388396, -0.5, -0.34026], [-1.5388396, 0.5, -0.34026], [1.5388404, -0.5, -0.34026], [1.5388404, 0.5, -0.34026], [1.3763804, 0.0, 0.510391], [0.4253254, -1.30902, 0.510391], [0.4253254, 1.30902, 0.510391], [-1.1135196, -0.809017, 0.510391], [-1.1135196, 0.809017, 0.510391], [-0.8506506, 0.0, 1.03612], [-0.2628656, -0.809017, 1.03612], [-0.2628656, 0.809017, 1.03612], [0.6881914, -0.5, 1.03612], [0.6881914, 0.5, 1.03612]]}, "93": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Hexecontahedron", "ShortName": "C13", "vertices": [[0.0, 0.0, -3.6322], [0.0, 0.0, 3.6322], [-2.42147, 0.0, -2.70728], [2.42147, 0.0, 2.70728], [-3.74093, 0.0, -0.714455], [3.74093, 0.0, 0.714455], [-2.31202, 0.0, 3.02648], [2.31202, 0.0, -3.02648], [-2.97388, 1.2495, -1.66963], [2.97388, 1.2495, 1.66963], [-3.20333, -1.64692, -0.468228], [3.20333, -1.64692, 0.468228], [-2.89122, -0.64303, -2.10244], [2.89122, -0.64303, 2.10244], [-3.33708, 1.41527, 0.232068], [3.33708, 1.41527, -0.232068], [-3.47888, -0.374822, 0.974575], [3.47888, -0.374822, -0.974575], [-0.175393, 3.59762, 0.468228], [0.175393, 3.59762, -0.468228], [-2.70728, -2.09705, -1.21073], [-2.70728, 2.09705, -1.21073], [2.70728, -2.09705, 1.21073], [2.70728, 2.09705, 1.21073], [-2.46401, 0.973951, -2.48455], [2.46401, 0.973951, 2.48455], [-3.24272, 0.374822, 1.59285], [3.24272, 0.374822, -1.59285], [-3.02794, -1.95071, 0.468228], [3.02794, -1.95071, -0.468228], [-2.89419, 2.18236, -0.232068], [2.89419, 2.18236, 0.232068], [-3.16974, -1.29605, 1.21073], [-3.16974, 1.29605, 1.21073], [3.16974, -1.29605, -1.21073], [3.16974, 1.29605, -1.21073], [-2.54667, -1.58042, -2.05174], [2.54667, -1.58042, 2.05174], [-0.753368, 0.64303, 3.49455], [0.753368, 0.64303, -3.49455], [-0.442882, -3.59762, -0.232068], [0.442882, -3.59762, 0.232068], [-0.40484, -3.20021, 1.66963], [0.40484, -3.20021, -1.66963], [-1.10352, -1.2495, 3.22706], [1.10352, -1.2495, -3.22706], [-0.462459, -3.3931, -1.21073], [-0.462459, 3.3931, -1.21073], [0.462459, -3.3931, 1.21073], [0.462459, 3.3931, 1.21073], [-0.530344, 1.58042, 3.22706], [0.530344, 1.58042, -3.22706], [-0.388538, -2.62087, 2.48455], [0.388538, -2.62087, -2.48455], [-2.64202, -1.41527, 2.05174], [2.64202, -1.41527, -2.05174], [-0.0953548, 2.99569, -2.05174], [0.0953548, 2.99569, 2.05174], [-2.06405, -2.82539, -0.974575], [2.06405, -2.82539, 0.974575], [-2.07547, 1.64692, 2.48455], [2.07547, 1.64692, -2.48455], [-0.180196, -0.973951, 3.49455], [0.180196, -0.973951, -3.49455], [-0.888732, 2.82539, 2.10244], [0.888732, 2.82539, -2.10244], [-2.56904, 1.95071, 1.66963], [2.56904, 1.95071, -1.66963], [-1.21073, -2.09705, 2.70728], [-1.21073, 2.09705, 2.70728], [1.21073, -2.09705, -2.70728], [1.21073, 2.09705, -2.70728], [-1.29676, -2.99569, -1.59285], [1.29676, -2.99569, 1.59285], [-1.94597, 2.62087, -1.59285], [1.94597, 2.62087, 1.59285], [-1.41484, 3.20021, -0.974575], [1.41484, 3.20021, 0.974575], [-1.63386, -0.330921, -3.22706], [1.63386, -0.330921, 3.22706], [-2.00249, -2.18236, 2.10244], [2.00249, -2.18236, -2.10244], [-0.933564, 0.330921, -3.49455], [0.933564, 0.330921, 3.49455], [-1.87047, -3.23974, 0.714455], [-1.87047, 3.23974, 0.714455], [1.87047, -3.23974, -0.714455], [1.87047, 3.23974, -0.714455], [-1.15601, -2.00227, -3.02648], [-1.15601, 2.00227, -3.02648], [1.15601, -2.00227, 3.02648], [1.15601, 2.00227, 3.02648]]}, "94": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Icositetrahedron", "ShortName": "C12", "vertices": [[0.0, 0.0, -2.294], [0.0, 0.0, 2.294], [0.0, -2.294, 0.0], [0.0, 2.294, 0.0], [-1.72488, -1.24722, -0.368676], [-1.72488, 1.24722, 0.368676], [-1.72488, -0.368676, 1.24722], [-1.72488, 0.368676, -1.24722], [1.72488, -1.24722, 0.368676], [1.72488, 1.24722, -0.368676], [1.72488, -0.368676, -1.24722], [1.72488, 0.368676, 1.24722], [-2.294, 0.0, 0.0], [2.294, 0.0, 0.0], [-1.24722, -1.72488, 0.368676], [-1.24722, 1.72488, -0.368676], [-1.24722, -1.24722, -1.24722], [-1.24722, -1.24722, 1.24722], [-1.24722, 1.24722, -1.24722], [-1.24722, 1.24722, 1.24722], [-1.24722, -0.368676, -1.72488], [-1.24722, 0.368676, 1.72488], [1.24722, -1.72488, -0.368676], [1.24722, 1.72488, 0.368676], [1.24722, -1.24722, -1.24722], [1.24722, -1.24722, 1.24722], [1.24722, 1.24722, -1.24722], [1.24722, 1.24722, 1.24722], [1.24722, -0.368676, 1.72488], [1.24722, 0.368676, -1.72488], [-0.368676, -1.72488, -1.24722], [-0.368676, 1.72488, 1.24722], [-0.368676, -1.24722, 1.72488], [-0.368676, 1.24722, -1.72488], [0.368676, -1.72488, 1.24722], [0.368676, 1.72488, -1.24722], [0.368676, -1.24722, -1.72488], [0.368676, 1.24722, 1.72488]]}, "95": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Orthobicupola", "ShortName": "J30", "vertices": [[-1e-07, -1.61803, 0.0], [-1e-07, 1.61803, 0.0], [0.8506509, 0.0, -0.525731], [0.8506509, 0.0, 0.525731], [0.2628659, -0.809017, -0.525731], [0.2628659, -0.809017, 0.525731], [0.2628659, 0.809017, -0.525731], [0.2628659, 0.809017, 0.525731], [-0.9510571, -1.30902, 0.0], [-0.9510571, 1.30902, 0.0], [0.9510569, -1.30902, 0.0], [0.9510569, 1.30902, 0.0], [-0.6881911, -0.5, -0.525731], [-0.6881911, -0.5, 0.525731], [-0.6881911, 0.5, -0.525731], [-0.6881911, 0.5, 0.525731], [-1.5388401, -0.5, 0.0], [-1.5388401, 0.5, 0.0], [1.5388399, -0.5, 0.0], [1.5388399, 0.5, 0.0]]}, "96": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Orthobirotunda", "ShortName": "J34", "vertices": [[7.33e-07, -1.61803, 0.0], [7.33e-07, 1.61803, 0.0], [-0.951056267, -1.30902, 0.0], [-0.951056267, 1.30902, 0.0], [0.951057733, -1.30902, 0.0], [0.951057733, 1.30902, 0.0], [-1.538839267, -0.5, 0.0], [-1.538839267, 0.5, 0.0], [1.538840733, -0.5, 0.0], [1.538840733, 0.5, 0.0], [1.376380733, 0.0, -0.850651], [1.376380733, 0.0, 0.850651], [0.425325733, -1.30902, -0.850651], [0.425325733, -1.30902, 0.850651], [0.425325733, 1.30902, -0.850651], [0.425325733, 1.30902, 0.850651], [-1.113519267, -0.809017, -0.850651], [-1.113519267, -0.809017, 0.850651], [-1.113519267, 0.809017, -0.850651], [-1.113519267, 0.809017, 0.850651], [-0.850650267, 0.0, -1.37638], [-0.850650267, 0.0, 1.37638], [-0.262865267, -0.809017, -1.37638], [-0.262865267, -0.809017, 1.37638], [-0.262865267, 0.809017, -1.37638], [-0.262865267, 0.809017, 1.37638], [0.688191733, -0.5, -1.37638], [0.688191733, -0.5, 1.37638], [0.688191733, 0.5, -1.37638], [0.688191733, 0.5, 1.37638]]}, "97": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Orthocupolarotunda", "ShortName": "J32", "vertices": [[4.8e-07, -1.61803, -0.34026], [4.8e-07, 1.61803, -0.34026], [0.42532548, -1.30902, 0.510391], [0.42532548, 1.30902, 0.510391], [0.68819148, -0.5, -0.865991], [0.68819148, -0.5, 1.03612], [0.68819148, 0.5, -0.865991], [0.68819148, 0.5, 1.03612], [-0.85065052, 0.0, -0.865991], [-0.85065052, 0.0, 1.03612], [-1.11351952, -0.809017, 0.510391], [-1.11351952, 0.809017, 0.510391], [-0.26286552, -0.809017, -0.865991], [-0.26286552, -0.809017, 1.03612], [-0.26286552, 0.809017, -0.865991], [-0.26286552, 0.809017, 1.03612], [-0.95105652, -1.30902, -0.34026], [-0.95105652, 1.30902, -0.34026], [0.95105748, -1.30902, -0.34026], [0.95105748, 1.30902, -0.34026], [-1.53883952, -0.5, -0.34026], [-1.53883952, 0.5, -0.34026], [1.53884048, -0.5, -0.34026], [1.53884048, 0.5, -0.34026], [1.37638048, 0.0, 0.510391]]}, "98": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Rotunda", "ShortName": "J06", "vertices": [[5.5e-07, -1.61803, -0.55675775], [5.5e-07, 1.61803, -0.55675775], [-0.95105645, -1.30902, -0.55675775], [-0.95105645, 1.30902, -0.55675775], [0.95105755, -1.30902, -0.55675775], [0.95105755, 1.30902, -0.55675775], [-1.53883945, -0.5, -0.55675775], [-1.53883945, 0.5, -0.55675775], [1.53884055, -0.5, -0.55675775], [1.53884055, 0.5, -0.55675775], [1.37638055, 0.0, 0.29389325], [0.42532555, -1.30902, 0.29389325], [0.42532555, 1.30902, 0.29389325], [-1.11351945, -0.809017, 0.29389325], [-1.11351945, 0.809017, 0.29389325], [-0.85065045, 0.0, 0.81962225], [-0.26286545, -0.809017, 0.81962225], [-0.26286545, 0.809017, 0.81962225], [0.68819155, -0.5, 0.81962225], [0.68819155, 0.5, 0.81962225]]}, "99": {"AlternativeName": null, "Comment": null, "Name": "Pentakis Dodecahedron", "ShortName": "C06", "vertices": [[-0.958957, 0.0, 1.25529], [1.37638, 0.0, 0.688191], [-0.0, 0.0, 1.53884], [-1.11352, -0.809017, 0.688191], [-1.11352, 0.809017, 0.688191], [-0.0, 0.0, -1.53884], [0.425325, -1.30902, 0.688191], [0.425325, 1.30902, 0.688191], [-0.296334, -0.912023, 1.25529], [-0.296334, 0.912023, 1.25529], [0.775813, 0.563661, 1.25529], [0.775813, -0.563661, 1.25529], [0.958957, 0.0, -1.25529], [0.296334, -0.912023, -1.25529], [0.296334, 0.912023, -1.25529], [-0.775813, 0.563661, -1.25529], [-0.775813, -0.563661, -1.25529], [1.25529, -0.912023, 0.296334], [-0.479479, -1.47568, 0.296334], [1.25529, 0.912023, 0.296334], [-0.479479, 1.47568, 0.296334], [-1.55163, 0.0, 0.296334], [1.55163, 0.0, -0.296334], [0.479479, -1.47568, -0.296334], [0.479479, 1.47568, -0.296334], [-1.25529, -0.912023, -0.296334], [-1.25529, 0.912023, -0.296334], [1.11352, -0.809017, -0.688191], [-1.37638, 0.0, -0.688191], [1.11352, 0.809017, -0.688191], [-0.425325, -1.30902, -0.688191], [-0.425325, 1.30902, -0.688191]]}, "100": {"AlternativeName": null, "Comment": "edge length=1", "Name": "Triangular Prism", "ShortName": "O01", "vertices": [[-0.288675, -0.5, -0.5], [-0.288675, -0.5, 0.5], [-0.288675, 0.5, -0.5], [-0.288675, 0.5, 0.5], [0.57735, 0.0, -0.5], [0.57735, 0.0, 0.5]]}, "101": {"AlternativeName": null, "Comment": "edge length=2", "Name": "Pentagonal Prism", "ShortName": "O02", "vertices": [[0.850650808, 0.0, -0.5], [0.850650808, 0.0, 0.5], [-0.68819096, -0.5, -0.5], [-0.68819096, -0.5, 0.5], [-0.68819096, 0.5, -0.5], [-0.68819096, 0.5, 0.5], [0.262865556, -0.809016994, -0.5], [0.262865556, -0.809016994, 0.5], [0.262865556, 0.809016994, -0.5], [0.262865556, 0.809016994, 0.5]]}, "102": {"AlternativeName": null, "Comment": "edge length=3", "Name": "Hexagonal Prism", "ShortName": "O03", "vertices": [[-1.0, 0.0, -0.5], [-1.0, 0.0, 0.5], [-0.5, -0.866025404, -0.5], [-0.5, -0.866025404, 0.5], [-0.5, 0.866025404, -0.5], [-0.5, 0.866025404, 0.5], [0.5, -0.866025404, -0.5], [0.5, -0.866025404, 0.5], [0.5, 0.866025404, -0.5], [0.5, 0.866025404, 0.5], [1.0, 0.0, -0.5], [1.0, 0.0, 0.5]]}, "103": {"AlternativeName": null, "Comment": "edge length=4", "Name": "Heptagonal Prism", "ShortName": "O04", "vertices": [[-1.038260698, -0.5, -0.5], [-1.038260698, -0.5, 0.5], [-1.038260698, 0.5, -0.5], [-1.038260698, 0.5, 0.5], [1.152382435, 0.0, -0.5], [1.152382435, 0.0, 0.5], [-0.256429216, -1.123489802, -0.5], [-0.256429216, -1.123489802, 0.5], [-0.256429216, 1.123489802, -0.5], [-0.256429216, 1.123489802, 0.5], [0.718498696, -0.900968868, -0.5], [0.718498696, -0.900968868, 0.5], [0.718498696, 0.900968868, -0.5], [0.718498696, 0.900968868, 0.5]]}, "104": {"AlternativeName": null, "Comment": "edge length=5", "Name": "Octagonal Prism", "ShortName": "O05", "vertices": [[-0.0, -1.306562965, -0.5], [-0.0, -1.306562965, 0.5], [-0.0, 1.306562965, -0.5], [-0.0, 1.306562965, 0.5], [-1.306562965, 0.0, -0.5], [-1.306562965, 0.0, 0.5], [1.306562965, 0.0, -0.5], [1.306562965, 0.0, 0.5], [-0.923879533, -0.923879533, -0.5], [-0.923879533, -0.923879533, 0.5], [-0.923879533, 0.923879533, -0.5], [-0.923879533, 0.923879533, 0.5], [0.923879533, -0.923879533, -0.5], [0.923879533, -0.923879533, 0.5], [0.923879533, 0.923879533, -0.5], [0.923879533, 0.923879533, 0.5]]}, "105": {"AlternativeName": null, "Comment": "edge length=6", "Name": "Nonagonal Prism", "ShortName": "O06", "vertices": [[-1.37373871, -0.5, -0.5], [-1.37373871, -0.5, 0.5], [-1.37373871, 0.5, -0.5], [-1.37373871, 0.5, 0.5], [-0.7309511, -1.266044443, -0.5], [-0.7309511, -1.266044443, 0.5], [-0.7309511, 1.266044443, -0.5], [-0.7309511, 1.266044443, 0.5], [1.4619022, 0.0, -0.5], [1.4619022, 0.0, 0.5], [1.119882057, -0.939692621, -0.5], [1.119882057, -0.939692621, 0.5], [1.119882057, 0.939692621, -0.5], [1.119882057, 0.939692621, 0.5], [0.253856653, -1.439692621, -0.5], [0.253856653, -1.439692621, 0.5], [0.253856653, 1.439692621, -0.5], [0.253856653, 1.439692621, 0.5]]}, "106": {"AlternativeName": null, "Comment": "edge length=7", "Name": "Decagonal Prism", "ShortName": "O07", "vertices": [[-1.618033989, 0.0, -0.5], [-1.618033989, 0.0, 0.5], [1.618033989, 0.0, -0.5], [1.618033989, 0.0, 0.5], [-1.309016994, -0.951056516, -0.5], [-1.309016994, -0.951056516, 0.5], [-1.309016994, 0.951056516, -0.5], [-1.309016994, 0.951056516, 0.5], [-0.5, -1.538841769, -0.5], [-0.5, -1.538841769, 0.5], [-0.5, 1.538841769, -0.5], [-0.5, 1.538841769, 0.5], [0.5, -1.538841769, -0.5], [0.5, -1.538841769, 0.5], [0.5, 1.538841769, -0.5], [0.5, 1.538841769, 0.5], [1.309016994, -0.951056516, -0.5], [1.309016994, -0.951056516, 0.5], [1.309016994, 0.951056516, -0.5], [1.309016994, 0.951056516, 0.5]]}, "107": {"AlternativeName": "Square Pyramid", "Comment": null, "Name": "Pyramid 4", "ShortName": "J01", "vertices": [[0.0, 0.0, 0.565685425], [0.0, -0.707106781, -0.141421356], [0.0, 0.707106781, -0.141421356], [-0.707106781, 0.0, -0.141421356], [0.707106781, 0.0, -0.141421356]]}, "108": {"AlternativeName": null, "Comment": null, "Name": "Pyramid 5", "ShortName": "J02", "vertices": [[0.0, 0.0, 0.43810926], [0.850650808, 0.0, -0.087621852], [0.262865556, -0.809016994, -0.087621852], [0.262865556, 0.809016994, -0.087621852], [-0.68819096, -0.5, -0.087621852], [-0.68819096, 0.5, -0.087621852]]}, "109": {"AlternativeName": null, "Comment": null, "Name": "Rhombic Dodecahedron", "ShortName": "C01", "vertices": [[-0.816497, -0.816497, 0.0], [-0.816497, 0.0, -0.57735], [-0.816497, 0.0, 0.57735], [-0.816497, 0.816497, 0.0], [0.0, -0.816497, -0.57735], [0.0, -0.816497, 0.57735], [0.0, 0.0, -1.1547], [0.0, 0.0, 1.1547], [0.0, 0.816497, -0.57735], [0.0, 0.816497, 0.57735], [0.816497, -0.816497, 0.0], [0.816497, 0.0, -0.57735], [0.816497, 0.0, 0.57735], [0.816497, 0.816497, 0.0]]}, "110": {"AlternativeName": null, "Comment": null, "Name": "Rhombic Enneacontahedron", "ShortName": "O17", "vertices": [[-2.41202, 0.0, -0.460654738], [-2.41202, 0.0, 0.539345262], [-2.2847, -0.934172, -0.127321738], [-2.2847, -0.934172, 0.872678262], [-2.2847, 0.934172, -0.127321738], [-2.2847, 0.934172, 0.872678262], [-2.15738, 0.0, 1.206011262], [-2.07869, -0.57735, -1.206013738], [-2.07869, 0.57735, -1.206013738], [-1.95137, -1.51152, -0.872673738], [-1.95137, -1.51152, 0.127322262], [-1.95137, 1.51152, -0.872673738], [-1.95137, 1.51152, 0.127322262], [-1.74536, 0.0, -1.951363738], [-1.61803, -0.934172, -1.618033738], [-1.61803, -0.934172, 1.618036262], [-1.61803, 0.934172, -1.618033738], [-1.61803, 0.934172, 1.618036262], [-1.53934, -1.51152, 1.206011262], [-1.53934, 1.51152, 1.206011262], [-1.49071, 0.0, 1.951366262], [-1.20601, -2.08887, -0.539344738], [-1.20601, -2.08887, 0.460655562], [-1.20601, 2.08887, -0.539344738], [-1.20601, 2.08887, 0.460655562], [-1.07869, -1.86834, -1.206013738], [-1.07869, 1.86834, -1.206013738], [-0.872678, -1.51152, 1.951366262], [-0.872678, -0.356822, -2.284703738], [-0.872678, 0.356822, -2.284703738], [-0.872678, 1.51152, 1.951366262], [-0.745356, -1.29099, -1.951363738], [-0.745356, -0.57735, 2.284696262], [-0.745356, 0.57735, 2.284696262], [-0.745356, 1.29099, -1.951363738], [-0.539345, -2.08887, 1.206011262], [-0.539345, 2.08887, 1.206011262], [-0.333333, -2.44569, -0.872673738], [-0.333333, -2.44569, 0.127322262], [-0.333333, 2.44569, -0.872673738], [-0.333333, 2.44569, 0.127322262], [-0.127322, -0.934172, 2.284696262], [-0.127322, 0.934172, 2.284696262], [-0.0, -1.86834, -1.618033738], [-0.0, -1.86834, 1.618036262], [-0.0, 0.0, -2.618033738], [-0.0, 0.0, 2.618036262], [-0.0, 1.86834, -1.618033738], [-0.0, 1.86834, 1.618036262], [0.127322, -0.934172, -2.284703738], [0.127322, 0.934172, -2.284703738], [0.333333, -2.44569, -0.127321738], [0.333333, -2.44569, 0.872678262], [0.333333, 2.44569, -0.127321738], [0.333333, 2.44569, 0.872678262], [0.539345, -2.08887, -1.206013738], [0.539345, 2.08887, -1.206013738], [0.745356, -1.29099, 1.951366262], [0.745356, -0.57735, -2.284703738], [0.745356, 0.57735, -2.284703738], [0.745356, 1.29099, 1.951366262], [0.872678, -1.51152, -1.951363738], [0.872678, -0.356822, 2.284696262], [0.872678, 0.356822, 2.284696262], [0.872678, 1.51152, -1.951363738], [1.07869, -1.86834, 1.206011262], [1.07869, 1.86834, 1.206011262], [1.20601, -2.08887, -0.460654738], [1.20601, -2.08887, 0.539345262], [1.20601, 2.08887, -0.460654738], [1.20601, 2.08887, 0.539345262], [1.49071, 0.0, -1.951363738], [1.53934, -1.51152, -1.206013738], [1.53934, 1.51152, -1.206013738], [1.61803, -0.934172, -1.618033738], [1.61803, -0.934172, 1.618036262], [1.61803, 0.934172, -1.618033738], [1.61803, 0.934172, 1.618036262], [1.74536, 0.0, 1.951366262], [1.95137, -1.51152, -0.127321738], [1.95137, -1.51152, 0.872678262], [1.95137, 1.51152, -0.127321738], [1.95137, 1.51152, 0.872678262], [2.07869, -0.57735, 1.206011262], [2.07869, 0.57735, 1.206011262], [2.15738, 0.0, -1.206013738], [2.2847, -0.934172, -0.872673738], [2.2847, -0.934172, 0.127322262], [2.2847, 0.934172, -0.872673738], [2.2847, 0.934172, 0.127322262], [2.41202, 0.0, -0.539344738], [2.41202, 0.0, 0.460655562]]}, "111": {"AlternativeName": null, "Comment": null, "Name": "Rhombic Icosahedron", "ShortName": "O16", "vertices": [[-1.44721, 0.0, 0.223607], [-1.17082, -0.850651, -0.223607], [-1.17082, 0.850651, -0.223607], [-0.894427, 0.0, -0.67082], [-0.723607, -0.525731, 0.67082], [-0.723607, 0.525731, 0.67082], [-0.447214, -1.37638, 0.223607], [-0.447214, 1.37638, 0.223607], [-0.276393, -0.850651, -0.67082], [-0.276393, 0.850651, -0.67082], [0.0, 0.0, -1.11803], [0.0, 0.0, 1.11803], [0.276393, -0.850651, 0.67082], [0.276393, 0.850651, 0.67082], [0.447214, -1.37638, -0.223607], [0.447214, 1.37638, -0.223607], [0.723607, -0.525731, -0.67082], [0.723607, 0.525731, -0.67082], [0.894427, 0.0, 0.67082], [1.17082, -0.850651, 0.223607], [1.17082, 0.850651, 0.223607], [1.44721, 0.0, -0.223607]]}, "112": {"AlternativeName": null, "Comment": null, "Name": "Rhombic Triacontahedron", "ShortName": "C02", "vertices": [[-0.0, 0.0, -1.61803], [-0.0, 0.0, 1.61803], [0.276393, -0.850651, 1.17082], [0.276393, 0.850651, 1.17082], [0.894427, 0.0, 1.17082], [1.17082, -0.850651, 0.723607], [1.17082, -0.850651, -0.276393], [1.17082, 0.850651, 0.723607], [1.17082, 0.850651, -0.276393], [-0.894427, 0.0, -1.17082], [-0.447214, -1.37638, 0.723607], [-0.447214, -1.37638, -0.276393], [-0.447214, 1.37638, 0.723607], [-0.447214, 1.37638, -0.276393], [0.447214, -1.37638, 0.276393], [0.447214, -1.37638, -0.723607], [0.447214, 1.37638, 0.276393], [0.447214, 1.37638, -0.723607], [-1.44721, 0.0, 0.723607], [-1.44721, 0.0, -0.276393], [-0.723607, -0.525731, 1.17082], [-0.723607, 0.525731, 1.17082], [0.723607, -0.525731, -1.17082], [0.723607, 0.525731, -1.17082], [1.44721, 0.0, 0.276393], [1.44721, 0.0, -0.723607], [-1.17082, -0.850651, 0.276393], [-1.17082, -0.850651, -0.723607], [-1.17082, 0.850651, 0.276393], [-1.17082, 0.850651, -0.723607], [-0.276393, -0.850651, -1.17082], [-0.276393, 0.850651, -1.17082]]}, "113": {"AlternativeName": null, "Comment": null, "Name": "Rhombicosidodecahedron", "ShortName": "A09", "vertices": [[-0.5, -0.5, -2.11803], [-0.5, -0.5, 2.11803], [-0.5, 0.5, -2.11803], [-0.5, 0.5, 2.11803], [-0.5, -2.11803, -0.5], [-0.5, -2.11803, 0.5], [-0.5, 2.11803, -0.5], [-0.5, 2.11803, 0.5], [-0.0, -1.30902, -1.80902], [-0.0, -1.30902, 1.80902], [-0.0, 1.30902, -1.80902], [-0.0, 1.30902, 1.80902], [0.5, -0.5, -2.11803], [0.5, -0.5, 2.11803], [0.5, 0.5, -2.11803], [0.5, 0.5, 2.11803], [0.5, -2.11803, -0.5], [0.5, -2.11803, 0.5], [0.5, 2.11803, -0.5], [0.5, 2.11803, 0.5], [-1.80902, 0.0, -1.30902], [-1.80902, 0.0, 1.30902], [-0.809017, -1.61803, -1.30902], [-0.809017, -1.61803, 1.30902], [-0.809017, 1.61803, -1.30902], [-0.809017, 1.61803, 1.30902], [-1.61803, -1.30902, -0.809017], [-1.61803, -1.30902, 0.809017], [-1.61803, 1.30902, -0.809017], [-1.61803, 1.30902, 0.809017], [-2.11803, -0.5, -0.5], [-2.11803, -0.5, 0.5], [-2.11803, 0.5, -0.5], [-2.11803, 0.5, 0.5], [-1.30902, -1.80902, 0.0], [-1.30902, -0.809017, -1.61803], [-1.30902, -0.809017, 1.61803], [-1.30902, 0.809017, -1.61803], [-1.30902, 0.809017, 1.61803], [-1.30902, 1.80902, 0.0], [0.809017, -1.61803, -1.30902], [0.809017, -1.61803, 1.30902], [0.809017, 1.61803, -1.30902], [0.809017, 1.61803, 1.30902], [1.61803, -1.30902, -0.809017], [1.61803, -1.30902, 0.809017], [1.61803, 1.30902, -0.809017], [1.61803, 1.30902, 0.809017], [2.11803, -0.5, -0.5], [2.11803, -0.5, 0.5], [2.11803, 0.5, -0.5], [2.11803, 0.5, 0.5], [1.30902, -1.80902, 0.0], [1.30902, -0.809017, -1.61803], [1.30902, -0.809017, 1.61803], [1.30902, 0.809017, -1.61803], [1.30902, 0.809017, 1.61803], [1.30902, 1.80902, 0.0], [1.80902, 0.0, -1.30902], [1.80902, 0.0, 1.30902]]}, "114": {"AlternativeName": null, "Comment": null, "Name": "Rhombicuboctahedron", "ShortName": "A08", "vertices": [[-0.5, -0.5, -1.20711], [-0.5, -0.5, 1.20711], [-0.5, 0.5, -1.20711], [-0.5, 0.5, 1.20711], [-0.5, -1.20711, -0.5], [-0.5, -1.20711, 0.5], [-0.5, 1.20711, -0.5], [-0.5, 1.20711, 0.5], [0.5, -0.5, -1.20711], [0.5, -0.5, 1.20711], [0.5, 0.5, -1.20711], [0.5, 0.5, 1.20711], [0.5, -1.20711, -0.5], [0.5, -1.20711, 0.5], [0.5, 1.20711, -0.5], [0.5, 1.20711, 0.5], [-1.20711, -0.5, -0.5], [-1.20711, -0.5, 0.5], [-1.20711, 0.5, -0.5], [-1.20711, 0.5, 0.5], [1.20711, -0.5, -0.5], [1.20711, -0.5, 0.5], [1.20711, 0.5, -0.5], [1.20711, 0.5, 0.5]]}, "115": {"AlternativeName": null, "Comment": null, "Name": "Triakis Octahedron", "ShortName": "C05", "vertices": [[-0.5, -0.5, -0.5], [-0.5, -0.5, 0.5], [-0.5, 0.5, -0.5], [-0.5, 0.5, 0.5], [0.0, 0.0, -1.20711], [0.0, 0.0, 1.20711], [0.0, -1.20711, 0.0], [0.0, 1.20711, 0.0], [0.5, -0.5, -0.5], [0.5, -0.5, 0.5], [0.5, 0.5, -0.5], [0.5, 0.5, 0.5], [-1.20711, 0.0, 0.0], [1.20711, 0.0, 0.0]]}, "116": {"AlternativeName": null, "Comment": null, "Name": "Snub Cuboctahedron", "ShortName": "A12", "vertices": [[-1.14261, -0.337754, -0.621226], [-1.14261, 0.337754, 0.621226], [-1.14261, -0.621226, 0.337754], [-1.14261, 0.621226, -0.337754], [1.14261, -0.337754, 0.621226], [1.14261, 0.337754, -0.621226], [1.14261, -0.621226, -0.337754], [1.14261, 0.621226, 0.337754], [-0.337754, -1.14261, 0.621226], [-0.337754, 1.14261, -0.621226], [-0.337754, -0.621226, -1.14261], [-0.337754, 0.621226, 1.14261], [0.337754, -1.14261, -0.621226], [0.337754, 1.14261, 0.621226], [0.337754, -0.621226, 1.14261], [0.337754, 0.621226, -1.14261], [-0.621226, -1.14261, -0.337754], [-0.621226, 1.14261, 0.337754], [-0.621226, -0.337754, 1.14261], [-0.621226, 0.337754, -1.14261], [0.621226, -1.14261, 0.337754], [0.621226, 1.14261, -0.337754], [0.621226, -0.337754, -1.14261], [0.621226, 0.337754, 1.14261]]}, "117": {"AlternativeName": null, "Comment": null, "Name": "Snub Disphenoid", "ShortName": "J84", "vertices": [[-0.5, 0.0, -0.78393025], [0.0, -0.5, 0.78392975], [0.0, 0.5, 0.78392975], [0.0, -0.644584, -0.20556125], [0.0, 0.644584, -0.20556125], [0.5, 0.0, -0.78393025], [-0.644584, 0.0, 0.20556175], [0.644584, 0.0, 0.20556175]]}, "118": {"AlternativeName": null, "Comment": null, "Name": "Snub Icosidodecahedron", "ShortName": "A13", "vertices": [[-2.05022, -0.64303, 0.175393], [2.05022, -0.64303, -0.175393], [-1.64507, 0.64303, 1.23608], [1.64507, 0.64303, -1.23608], [-2.09275, 0.330921, 0.398127], [2.09275, 0.330921, -0.398127], [-1.33296, 1.64692, -0.398127], [1.33296, 1.64692, 0.398127], [-1.82527, -0.330921, 1.09842], [1.82527, -0.330921, -1.09842], [-0.626047, 1.74619, -1.09842], [0.626047, 1.74619, 1.09842], [-1.06222, 1.45402, 1.18539], [1.06222, 1.45402, -1.18539], [-1.93214, 0.84755, -0.442882], [1.93214, 0.84755, 0.442882], [-1.14487, -0.84755, 1.6182], [1.14487, -0.84755, -1.6182], [-1.58199, -1.45402, -0.175393], [1.58199, -1.45402, 0.175393], [-1.05741, 0.374822, -1.84093], [1.05741, 0.374822, 1.84093], [-0.439138, -0.374822, -2.07709], [0.439138, -0.374822, 2.07709], [-1.56241, -1.2495, 0.803274], [1.56241, -1.2495, -0.803274], [-1.86331, -0.728335, -0.803274], [1.86331, -0.728335, 0.803274], [-1.70007, 1.2495, 0.442882], [1.70007, 1.2495, -0.442882], [-0.728114, -1.64692, 1.18539], [0.728114, -1.64692, -1.18539], [-0.265655, -1.74619, -1.23608], [0.265655, -1.74619, 1.23608], [-0.759791, -1.97784, -0.398127], [0.759791, -1.97784, 0.398127], [-1.19922, -1.41527, -1.09842], [1.19922, -1.41527, 1.09842], [-1.79033, 0.192894, -1.18539], [1.79033, 0.192894, 1.18539], [-1.30644, -0.567715, -1.6182], [1.30644, -0.567715, 1.6182], [-0.853311, 0.728335, 1.84093], [0.853311, 0.728335, -1.84093], [-1.37941, 1.10316, -1.23608], [1.37941, 1.10316, 1.23608], [-0.105036, 0.567715, -2.07709], [0.105036, 0.567715, 2.07709], [-0.468228, 2.09705, -0.175393], [0.468228, 2.09705, 0.175393], [-0.300897, 1.97784, 0.803274], [0.300897, 1.97784, -0.803274], [-0.161563, 1.41527, 1.6182], [0.161563, 1.41527, -1.6182], [-0.544174, -0.192894, 2.07709], [0.544174, -0.192894, -2.07709], [-0.232068, -2.09705, 0.442882], [0.232068, -2.09705, -0.442882], [-0.204101, -1.10316, 1.84093], [0.204101, -1.10316, -1.84093]]}, "119": {"AlternativeName": null, "Comment": null, "Name": "Snub Square Antiprism", "ShortName": "J85", "vertices": [[-1.093170494, 0.431564919, -0.353605944], [-1.005700494, -0.562189081, -0.422882944], [-0.859633494, -0.107437081, 0.455675056], [-0.719877494, 0.882747919, 0.457000056], [-0.363678494, 0.031303219, -0.908244944], [-0.309622494, -0.905009081, 0.207952056], [-0.197089494, 0.873472919, -0.395411944], [-0.149505494, -0.936855081, -0.778631944], [-0.074030594, 0.285999919, 0.933206056], [0.254707506, 1.106599919, 0.465710056], [0.475980506, -0.511572081, 0.685483056], [0.580741506, -0.254465081, -0.745745944], [0.614673506, -1.052490081, -0.144077944], [0.747331506, 0.587704919, -0.232912944], [0.900494506, 0.392447919, 0.735804056], [1.198379506, -0.261824081, 0.040684156]]}, "120": {"AlternativeName": null, "Comment": null, "Name": "Spenocorona", "ShortName": "J86", "vertices": [[0.0, -0.5, -0.6630428], [0.0, 0.5, -0.6630428], [-0.852727, 0.5, -0.1406858], [-0.5, 0.0, 0.6502572], [-0.852727, -0.5, -0.1406858], [0.852727, 0.5, -0.1406858], [0.5, 0.0, 0.6502572], [0.852727, -0.5, -0.1406858], [0.0, 0.789428, 0.2941572], [0.0, -0.789428, 0.2941572]]}, "121": {"AlternativeName": null, "Comment": null, "Name": "Sphenomegacorona", "ShortName": "J88", "vertices": [[-0.708649899, -0.296833667, -0.152242192], [-0.649328899, -0.105651667, 0.827517808], [-0.619595899, 0.646067333, 0.168705808], [-0.612442898, -1.049476667, 0.499124808], [-0.165930899, 0.360258333, -0.675392192], [-0.002052116, -0.990752667, -0.290796192], [0.165479102, 0.470573333, 0.762723808], [0.225204102, -0.506004667, 0.553818808], [0.278442101, 1.017633333, -0.066799092], [0.540667102, -0.333660667, -0.813946192], [0.767923102, 0.151087333, 0.030669008], [0.780285102, 0.636760333, -0.843384192]]}, "122": {"AlternativeName": null, "Comment": null, "Name": "Square Cupola", "ShortName": "J04", "vertices": [[0, -0.707107, 0.707107], [0, 0.707107, 0.707107], [-0.707107, 0, 0.707107], [0.707107, 0, 0.707107], [0.5, 1.20711, 0], [-0.5, 1.20711, 0], [-1.20711, 0.5, 0], [-1.20711, -0.5, 0], [-0.5, -1.20711, 0], [0.5, -1.20711, 0], [1.20711, -0.5, 0], [1.20711, 0.5, 0]]}, "123": {"AlternativeName": null, "Comment": null, "Name": "Square Gyrobicupola", "ShortName": "J29", "vertices": [[-0.5, -0.5, -0.707107], [-0.5, 0.5, -0.707107], [0.0, -0.707107, 0.707107], [0.0, 0.707107, 0.707107], [0.5, -0.5, -0.707107], [0.5, 0.5, -0.707107], [-0.707107, 0.0, 0.707107], [0.707107, 0.0, 0.707107], [0.5, 1.20711, 0.0], [-0.5, 1.20711, 0.0], [-1.20711, 0.5, 0.0], [-1.20711, -0.5, 0.0], [-0.5, -1.20711, 0.0], [0.5, -1.20711, 0.0], [1.20711, -0.5, 0.0], [1.20711, 0.5, 0.0]]}, "124": {"AlternativeName": null, "Comment": null, "Name": "Square Orthobicupola", "ShortName": "J28", "vertices": [[0.0, -0.707107, -0.707107], [0.0, -0.707107, 0.707107], [0.0, 0.707107, -0.707107], [0.0, 0.707107, 0.707107], [-0.707107, 0.0, -0.707107], [-0.707107, 0.0, 0.707107], [0.707107, 0.0, -0.707107], [0.707107, 0.0, 0.707107], [0.5, 1.20711, 0.0], [-0.5, 1.20711, 0.0], [-1.20711, 0.5, 0.0], [-1.20711, -0.5, 0.0], [-0.5, -1.20711, 0.0], [0.5, -1.20711, 0.0], [1.20711, -0.5, 0.0], [1.20711, 0.5, 0.0]]}, "125": {"AlternativeName": null, "Comment": null, "Name": "Squashed Dodecahedron", "ShortName": "O15", "vertices": [[-1.732050357, 3.57e-07, -3.57e-07], [-0.866025357, -0.866024643, -0.866025357], [-0.866025357, 0.866025357, -0.866025357], [-0.866025357, 0.866025357, 0.866024643], [-0.577350357, -1.154699643, 1.154699643], [-0.288675357, -1.443379643, -0.288675357], [-0.288675357, 0.288675357, 1.443379643], [-3.57e-07, 3.57e-07, -1.732050357], [-3.57e-07, 1.732050357, -3.57e-07], [0.866024643, -0.866024643, 0.866024643], [0.866024643, 0.866025357, -0.866025357], [1.154699643, -1.154699643, -0.577350357], [1.154699643, 0.577350357, 1.154699643], [1.443379643, 0.288675357, -0.288675357]]}, "126": {"AlternativeName": null, "Comment": null, "Name": "Tetrahedron", "ShortName": "P01", "vertices": [[0.0, 0.0, 0.612372], [-0.288675, -0.5, -0.204124], [-0.288675, 0.5, -0.204124], [0.57735, 0.0, -0.204124]]}, "127": {"AlternativeName": null, "Comment": null, "Name": "Tetrakis Hexahedron", "ShortName": "C04", "vertices": [[-0.942809, 0.0, -0.666667], [-0.942809, 0.0, 0.666667], [-0.707107, -0.707107, 0.0], [-0.707107, 0.707107, 0.0], [0.0, -0.942809, -0.666667], [0.0, -0.942809, 0.666667], [0.0, 0.0, -1.0], [0.0, 0.0, 1.0], [0.0, 0.942809, -0.666667], [0.0, 0.942809, 0.666667], [0.707107, -0.707107, 0.0], [0.707107, 0.707107, 0.0], [0.942809, 0.0, -0.666667], [0.942809, 0.0, 0.666667]]}, "128": {"AlternativeName": null, "Comment": null, "Name": "Triakis Icosahedron", "ShortName": "C07", "vertices": [[-0.0, 0.0, 1.63925], [0.262866, -0.809017, 1.11352], [0.262866, 0.809017, 1.11352], [0.850651, 0.0, 1.11352], [-0.0, 0.0, -1.63925], [-0.262866, -0.809017, -1.11352], [-0.262866, 0.809017, -1.11352], [0.688191, -0.5, -1.11352], [0.688191, 0.5, -1.11352], [-0.850651, 0.0, -1.11352], [-0.688191, -0.5, 1.11352], [-0.688191, 0.5, 1.11352], [-1.37638, 0.0, -0.262866], [1.11352, -0.809017, -0.262866], [1.11352, 0.809017, -0.262866], [-0.425325, -1.30902, -0.262866], [-0.425325, 1.30902, -0.262866], [1.37638, 0.0, 0.262866], [0.425325, -1.30902, 0.262866], [0.425325, 1.30902, 0.262866], [-1.11352, -0.809017, 0.262866], [-1.11352, 0.809017, 0.262866], [1.18617, -0.861803, 0.733094], [1.18617, 0.861803, 0.733094], [-1.46619, 0.0, 0.733094], [-0.453077, -1.39443, 0.733094], [-0.453077, 1.39443, 0.733094], [1.46619, 0.0, -0.733094], [-1.18617, -0.861803, -0.733094], [-1.18617, 0.861803, -0.733094], [0.453077, -1.39443, -0.733094], [0.453077, 1.39443, -0.733094]]}, "129": {"AlternativeName": null, "Comment": null, "Name": "Triakis Tetrahedron", "ShortName": "C03", "vertices": [[0.0, 0.0, -1.020620125], [0.0, 0.0, 0.612371875], [-0.96225, 0.0, 0.340206875], [-0.288675, -0.5, -0.204124125], [-0.288675, 0.5, -0.204124125], [0.481125, -0.833333, 0.340206875], [0.481125, 0.833333, 0.340206875], [0.57735, 0.0, -0.204124125]]}, "130": {"AlternativeName": null, "Comment": null, "Name": "Triangular Cupola", "ShortName": "J03", "vertices": [[0.0, -1.0, -0.272165667], [0.0, 1.0, -0.272165667], [-0.288675, -0.5, 0.544331333], [-0.288675, 0.5, 0.544331333], [0.57735, 0.0, 0.544331333], [-0.866025, -0.5, -0.272165667], [-0.866025, 0.5, -0.272165667], [0.866025, -0.5, -0.272165667], [0.866025, 0.5, -0.272165667]]}, "131": {"AlternativeName": null, "Comment": null, "Name": "Triangular Hebesphenorotunda", "ShortName": "J92", "vertices": [[-0.960178539, 0.715213333, -0.038936944], [-0.818773539, 0.161289333, -0.859408644], [-0.761159539, 0.332418333, 0.863204356], [-0.547987539, 1.123904333, -0.853225644], [-0.478350539, -0.775429667, -0.777738644], [-0.420736539, -0.604300667, 0.944875356], [-0.279332539, -1.158225667, 0.124403356], [-0.181614539, 1.336044333, 0.052737256], [-0.026950639, 0.121738333, 1.508613356], [0.017403761, 0.953251333, 0.954879356], [0.047182961, 0.439778333, -1.274816644], [0.343251461, -1.328455667, -0.639410644], [0.387606461, -0.496940667, -1.193146644], [0.568220461, -0.562393667, 1.087023356], [0.639988461, 0.783023333, 0.191065356], [0.709625461, -1.116315667, 0.266553356], [0.781393461, 0.229098333, -0.629407644], [0.980411461, -0.153697667, 0.272736356]]}, "132": {"AlternativeName": null, "Comment": null, "Name": "Triangular Orthobicupola", "ShortName": "J27", "vertices": [[0.0, -1.0, 0.0], [0.0, 1.0, 0.0], [-0.288675, -0.5, -0.816497], [-0.288675, -0.5, 0.816497], [-0.288675, 0.5, -0.816497], [-0.288675, 0.5, 0.816497], [0.57735, 0.0, -0.816497], [0.57735, 0.0, 0.816497], [-0.866025, -0.5, 0.0], [-0.866025, 0.5, 0.0], [0.866025, -0.5, 0.0], [0.866025, 0.5, 0.0]]}, "133": {"AlternativeName": null, "Comment": null, "Name": "Triaugmented Dodecahedron", "ShortName": "J61", "vertices": [[-0.026740043, 0.0, 1.371363478], [-0.026740043, 0.0, -1.431156522], [0.151670957, -1.30902, 0.437189478], [0.151670957, 1.30902, 0.437189478], [0.440345957, -0.809017, -1.074336522], [0.440345957, 0.809017, -1.074336522], [1.017699957, -0.809017, 0.437189478], [1.017699957, 0.809017, 0.437189478], [-1.249590043, -0.5, 0.437189478], [-1.249590043, 0.5, 0.437189478], [1.196109957, -0.5, -0.496982522], [1.196109957, 0.5, -0.496982522], [-0.960912043, 0.0, -1.074336522], [-0.493826043, -0.809017, 1.014543478], [-0.493826043, 0.809017, 1.014543478], [0.907431957, 0.0, 1.014543478], [-1.071180043, -0.809017, -0.496982522], [-1.071180043, 0.809017, -0.496982522], [-1.021865043, 0.0, 1.272743478], [-0.205151043, -1.30902, -0.496982522], [-0.205151043, 1.30902, -0.496982522], [0.778332957, 1.39443, -0.337406522], [0.778332957, -1.39443, -0.337406522]]}, "134": {"AlternativeName": null, "Comment": null, "Name": "Triaugmented Hexagonal Prism", "ShortName": "J57", "vertices": [[-1.0, -1.33e-07, -0.5], [-1.0, -1.33e-07, 0.5], [-0.5, -0.866025133, -0.5], [-0.5, -0.866025133, 0.5], [-0.5, 0.866024867, -0.5], [-0.5, 0.866024867, 0.5], [0.0, -1.573130133, 0.0], [0.5, -0.866025133, -0.5], [0.5, -0.866025133, 0.5], [0.5, 0.866024867, -0.5], [0.5, 0.866024867, 0.5], [1.0, -1.33e-07, -0.5], [1.0, -1.33e-07, 0.5], [-1.36237, 0.786565867, 0.0], [1.36237, 0.786565867, 0.0]]}, "135": {"AlternativeName": null, "Comment": null, "Name": "Triaugmented Triangular Prism", "ShortName": "J51", "vertices": [[-0.288675, -0.5, -0.5], [-0.288675, -0.5, 0.5], [-0.288675, 0.5, -0.5], [-0.288675, 0.5, 0.5], [0.57735, 0.0, -0.5], [0.57735, 0.0, 0.5], [-0.995782, 0.0, 0.0], [0.497891, -0.862372, 0.0], [0.497891, 0.862372, 0.0]]}, "136": {"AlternativeName": null, "Comment": null, "Name": "Triaugmented Truncated Dodecahedron", "ShortName": "J71", "vertices": [[-2.829466198, 0.210868048, -0.449141693], [-2.721906198, 0.967199048, 0.196144307], [-2.691866198, -0.771698952, -0.574157693], [-2.556136198, 1.114481048, -0.778961693], [-2.410266198, 1.208401048, 1.115224307], [-2.408266198, 1.896771048, 0.389863307], [-2.361666198, -1.605198952, -0.131152693], [-2.242496198, 2.044051048, -0.585242693], [-2.195906198, -1.457908952, -1.106255693], [-2.013596198, 0.842345048, 1.957034307], [-1.976296198, 1.593991048, -1.437635693], [-1.964986198, -1.971248952, 0.710661307], [-1.740266198, 1.745961048, 1.627214307], [-1.738266198, 2.434321048, 0.901857307], [-1.683396198, 0.008850038, 2.400044307], [-1.653356198, -1.730048952, 1.629744307], [-1.545796198, -0.973716952, 2.275024307], [-1.531016198, -1.585668952, -1.842195693], [-1.470056198, 2.672631048, -0.675896693], [-1.311416198, 1.466231048, -2.173575693], [-1.203856198, 2.222561048, -1.528295693], [-1.158416198, 2.913831048, 0.243181307], [-1.157396198, -2.416258952, 1.097634307], [-0.967824198, 2.374531048, 1.536564307], [-0.953109198, -1.776348952, 2.342134307], [-0.951175198, -1.106158952, -2.500875693], [-0.875795198, -0.436161952, 2.787024307], [-0.815451198, 0.780019048, -2.705675693], [-0.677853198, -0.202547952, -2.830695693], [-0.620976198, -1.939658952, -2.057865693], [-0.533854198, 2.760121048, -1.016295693], [-0.457147198, -2.462568952, 1.810034307], [-0.387981198, 2.854041048, 0.877885307], [-0.283111198, -1.238798952, 2.854134307], [-0.247349198, -2.770248952, 0.881968307], [-0.222219198, 3.001321048, -0.097220693], [0.008694472, 2.487981048, 1.719694307], [0.094591702, 0.426031048, -2.921345693], [0.100723602, -0.322711952, 2.970154307], [0.186620802, -2.384668952, -1.670895693], [0.417534802, -2.898008952, 0.146027307], [0.442664802, 2.873571048, -0.833161693], [0.519371802, -2.349118952, 1.993174307], [0.583296802, -2.750728952, -0.829077693], [0.626931802, -1.592788952, 2.638464307], [0.729169802, -2.656798952, 1.065104307], [0.816290802, 2.042971048, 2.106674307], [0.873167802, 0.305866048, 2.879504307], [1.010766802, -0.676699952, 2.754484307], [1.071113802, 0.539481048, -2.738215693], [1.146493802, 1.209481048, 2.549684307], [1.163143802, -2.271218952, -1.487755693], [1.352703802, 2.519581048, -1.048835693], [1.399163802, -2.119248952, 1.577104307], [1.506723802, -1.362918952, 2.222384307], [1.726333802, 1.688981048, 1.891004307], [1.741113802, 1.077031048, -2.226215693], [-0.056852698, -0.980022952, -2.930145693], [1.848673802, 1.833371048, -1.580935693], [1.878703802, 0.094468448, -2.351235693], [1.935583802, -1.642638952, -1.578405693], [0.715591802, -0.351444952, -3.020805693], [2.160303802, 2.074571048, -0.661854693], [2.171613802, -1.490668952, 1.486444307], [2.208903802, -0.739026952, -1.908225693], [2.391213802, 1.561231048, 1.155064307], [0.273346802, -1.813518952, -2.487145693], [2.556983802, 1.708511048, 0.179959307], [2.605583802, -1.105078952, -1.066415693], [1.523183802, -0.796456952, -2.633825693], [2.751453802, -1.011158952, 0.827769307], [2.887183802, 0.875017048, 0.622965307], [1.249863802, -1.700068952, -2.304005693], [2.917213802, -0.863880952, -0.147336693], [3.024773802, -0.107549952, 0.497948307]]}, "137": {"AlternativeName": null, "Comment": null, "Name": "Tridiminished Icosahedron", "ShortName": "J63", "vertices": [[-0.57735, 0.0, -0.815231333], [-0.288675, -0.5, 0.696290667], [-0.288675, 0.5, 0.696290667], [0.288675, -0.5, -0.815231333], [0.288675, 0.5, -0.815231333], [0.57735, 0.0, 0.696290667], [0.467086, -0.809017, 0.118940667], [0.467086, 0.809017, 0.118940667], [-0.934172, 0.0, 0.118940667]]}, "138": {"AlternativeName": null, "Comment": null, "Name": "Tridiminished Rhombicosidodecahedron", "ShortName": "J83", "vertices": [[-2.079718862, 0.421930022, 0.428667444], [-2.056218862, 0.538149022, -0.564278556], [-2.033028862, -0.557217978, 0.626379444], [-1.995018862, -0.369170978, -0.980239556], [-1.980688862, -1.046144978, -0.244372556], [-1.603218862, 1.293775022, 0.541985444], [-1.588888862, 0.616801022, 1.277854444], [-1.579728862, 1.409995022, -0.450960556], [-1.542208862, -0.362346978, 1.475564444], [-1.419508862, -0.965401978, -1.539965556], [-1.405168862, -1.642374978, -0.804094556], [-0.785559862, 1.725305022, 0.923051444], [-0.771227862, 1.048325022, 1.658914444], [-0.747548862, 1.913345022, -0.683568556], [-0.695694862, -0.535965978, 1.978824444], [-0.549496862, -1.022801978, -2.029645556], [-0.526307862, -2.118164978, -0.838989556], [-0.256723862, 2.108215022, 0.165617444], [-0.219201862, 0.335879022, 2.092144444], [0.002528728, -1.735254978, -1.596425556], [0.060954338, 1.551685022, 1.426314444], [0.122458138, 1.855945022, -1.173255556], [0.183170138, -1.011758978, 1.943924444], [0.282685138, -0.519447978, -2.262255556], [0.320207138, -2.291784978, -0.335730556], [0.589790138, 1.934605022, 0.668875444], [0.612980138, 0.839233022, 1.859534444], [0.613283138, 2.050815022, -0.324069556], [0.659662138, -0.139913978, 2.057244444], [0.697976138, 1.259715022, -1.732975556], [0.758688138, -1.607984978, 1.384204444], [0.759178138, 0.352398022, -2.148935556], [0.811032138, -2.096914978, 0.513454444], [0.834711138, -1.231894978, -1.829035556], [0.849043138, -1.908874978, -1.093165556], [1.468651138, 1.458805022, 0.633980444], [1.482991138, 0.781833022, 1.369844444], [1.492151138, 1.575025022, -0.358965556], [1.529671138, -0.197314978, 1.567564444], [1.544491138, 1.086095022, -1.229715556], [1.590871138, -1.104634978, 1.151594444], [1.605691138, 0.178778022, -1.645675556], [1.643211138, -1.593564978, 0.280846444], [1.652371138, -0.800369978, -1.447965556], [1.666711138, -1.477344978, -0.712099556]]}, "139": {"AlternativeName": null, "Comment": null, "Name": "Trigyrate Rhombicosidodecahedron", "ShortName": "J75", "vertices": [[-2.189130133, -0.439999098, -0.013079717], [-2.113110133, 0.336730902, -0.638307917], [-2.089600133, 0.178000902, 0.766773083], [-2.013580133, 0.954731902, 0.141544083], [-1.813430133, -1.189070098, 0.532579083], [-1.797080133, -1.240750098, -0.465950917], [-1.713910133, -0.571070098, 1.312430083], [-1.695260133, 1.236569902, -0.763573917], [-1.674080133, 0.016025502, -1.477589917], [-1.536510133, 0.377194902, 1.575730083], [-1.478760133, -0.958912098, -1.371069917], [-1.413510133, 1.633969902, 0.564089083], [-1.256230133, 0.915864902, -1.602859917], [-1.118660133, 1.277029902, 1.450460083], [-1.095190133, 1.915809902, -0.341028917], [-1.079550133, -1.866710098, 0.579671083], [-1.063200133, -1.918390098, -0.418858917], [-0.918512133, -0.866768098, 1.841500083], [-0.864176133, 0.115112902, -2.055729917], [-0.744877133, -1.636560098, -1.323979917], [-0.741119133, 0.081497002, 2.104800083], [-0.668857133, -0.859825098, -1.949199917], [-0.548585133, 1.938449902, 0.963084083], [-0.526460133, -1.667520098, 1.388630083], [-0.384815133, 1.396899902, -1.699019917], [-0.323269133, 0.981335902, 1.979530083], [-0.285287133, 2.014899902, -0.919165917], [-0.267805133, -2.214090098, 0.110209083], [-0.230264133, 2.220289902, 0.057966183], [-0.007237073, -0.596146098, 2.151890083], [0.007236797, 0.596145902, -2.151889917], [0.246808867, 1.642749902, 1.492150083], [0.247249867, -1.758070098, -1.354299917], [0.285286867, -2.014900098, 0.919166083], [0.323268867, -0.981336098, -1.979529917], [0.384814867, -1.396900098, 1.699020083], [0.542096867, -2.115000098, -0.467927917], [0.607311867, 1.275389902, -1.729339917], [0.668856867, 0.859823902, 1.949200083], [0.706839867, 1.893389902, -0.949491917], [0.741118867, -0.081497298, -2.104799917], [0.761862867, 2.098779902, 0.027639883], [0.864175867, -0.115113098, 2.055730083], [1.056709867, 1.741839902, 0.914015083], [1.095189867, -1.915810098, 0.341029083], [1.179149867, -1.590370098, -1.032669917], [1.255159867, -0.813642098, -1.657899917], [1.256229867, -0.915865098, 1.602860083], [1.341189867, 0.597742902, -1.682249917], [1.478759867, 0.958911902, 1.371070083], [1.502229867, 1.597689902, -0.420424917], [1.674079867, -0.016025698, 1.477590083], [1.695259867, -1.236570098, 0.763574083], [1.732239867, -1.391180098, -0.223709917], [1.797079867, 1.240749902, 0.465951083], [1.855239867, -0.134402098, -1.235349917], [1.894279867, 0.796936902, -0.873294917], [2.113109867, -0.336731098, 0.638308083], [2.150089867, -0.491339098, -0.348975917], [2.189129867, 0.439998902, 0.013079983]]}, "140": {"AlternativeName": null, "Comment": null, "Name": "Truncated Cube", "ShortName": "A05", "vertices": [[-0.5, 1.20711, 1.20711], [-0.5, 1.20711, -1.20711], [-0.5, -1.20711, 1.20711], [-0.5, -1.20711, -1.20711], [0.5, 1.20711, 1.20711], [0.5, 1.20711, -1.20711], [0.5, -1.20711, 1.20711], [0.5, -1.20711, -1.20711], [1.20711, -0.5, 1.20711], [1.20711, -0.5, -1.20711], [1.20711, 0.5, 1.20711], [1.20711, 0.5, -1.20711], [1.20711, 1.20711, -0.5], [1.20711, 1.20711, 0.5], [1.20711, -1.20711, -0.5], [1.20711, -1.20711, 0.5], [-1.20711, -0.5, 1.20711], [-1.20711, -0.5, -1.20711], [-1.20711, 0.5, 1.20711], [-1.20711, 0.5, -1.20711], [-1.20711, 1.20711, -0.5], [-1.20711, 1.20711, 0.5], [-1.20711, -1.20711, -0.5], [-1.20711, -1.20711, 0.5]]}, "141": {"AlternativeName": null, "Comment": null, "Name": "Truncated Dodecahedron", "ShortName": "A07", "vertices": [[-0.0, -1.61803, 2.4899], [-0.0, -1.61803, -2.4899], [-0.0, 1.61803, 2.4899], [-0.0, 1.61803, -2.4899], [0.425325, -2.92705, 0.262866], [0.425325, 2.92705, 0.262866], [0.688191, -2.11803, 1.96417], [0.688191, 2.11803, 1.96417], [-2.75276, 0.0, -1.11352], [-2.06457, -2.11803, 0.262866], [-2.06457, 2.11803, 0.262866], [-1.37638, -2.61803, -0.262866], [-1.37638, 2.61803, -0.262866], [-0.688191, -2.11803, -1.96417], [-0.688191, 2.11803, -1.96417], [1.37638, -2.61803, 0.262866], [1.37638, 2.61803, 0.262866], [2.75276, 0.0, 1.11352], [1.80171, -1.30902, -1.96417], [1.80171, 1.30902, -1.96417], [2.06457, -2.11803, -0.262866], [2.06457, 2.11803, -0.262866], [2.22703, 0.0, 1.96417], [2.22703, -1.61803, -1.11352], [2.22703, 1.61803, -1.11352], [-2.65236, -1.30902, 0.262866], [-2.65236, 1.30902, 0.262866], [2.65236, -1.30902, -0.262866], [2.65236, 1.30902, -0.262866], [2.91522, -0.5, 0.262866], [2.91522, 0.5, 0.262866], [-2.91522, -0.5, -0.262866], [-2.91522, 0.5, -0.262866], [0.951057, -1.30902, 2.4899], [0.951057, -1.30902, -2.4899], [0.951057, 1.30902, 2.4899], [0.951057, 1.30902, -2.4899], [0.850651, -2.61803, 1.11352], [0.850651, 2.61803, 1.11352], [-0.951057, -1.30902, 2.4899], [-0.951057, -1.30902, -2.4899], [-0.951057, 1.30902, 2.4899], [-0.951057, 1.30902, -2.4899], [-1.53884, -0.5, 2.4899], [-1.53884, -0.5, -2.4899], [-1.53884, 0.5, 2.4899], [-1.53884, 0.5, -2.4899], [1.53884, -0.5, 2.4899], [1.53884, -0.5, -2.4899], [1.53884, 0.5, 2.4899], [1.53884, 0.5, -2.4899], [-2.22703, 0.0, -1.96417], [-2.22703, -1.61803, 1.11352], [-2.22703, 1.61803, 1.11352], [-0.850651, -2.61803, -1.11352], [-0.850651, 2.61803, -1.11352], [-1.80171, -1.30902, 1.96417], [-1.80171, 1.30902, 1.96417], [-0.425325, -2.92705, -0.262866], [-0.425325, 2.92705, -0.262866]]}, "142": {"AlternativeName": null, "Comment": null, "Name": "Truncated Icosahedron", "ShortName": "A06", "vertices": [[-0.16246, -2.11803, 1.27598], [-0.16246, 2.11803, 1.27598], [0.16246, -2.11803, -1.27598], [0.16246, 2.11803, -1.27598], [-0.262866, -0.809017, -2.32744], [-0.262866, -2.42705, -0.425325], [-0.262866, 0.809017, -2.32744], [-0.262866, 2.42705, -0.425325], [0.262866, -0.809017, 2.32744], [0.262866, -2.42705, 0.425325], [0.262866, 0.809017, 2.32744], [0.262866, 2.42705, 0.425325], [0.688191, -0.5, -2.32744], [0.688191, 0.5, -2.32744], [1.21392, -2.11803, 0.425325], [1.21392, 2.11803, 0.425325], [-2.06457, -0.5, 1.27598], [-2.06457, 0.5, 1.27598], [-1.37638, -1.0, 1.80171], [-1.37638, 1.0, 1.80171], [-1.37638, -1.61803, -1.27598], [-1.37638, 1.61803, -1.27598], [-0.688191, -0.5, 2.32744], [-0.688191, 0.5, 2.32744], [1.37638, -1.0, -1.80171], [1.37638, 1.0, -1.80171], [1.37638, -1.61803, 1.27598], [1.37638, 1.61803, 1.27598], [-1.7013, 0.0, -1.80171], [1.7013, 0.0, 1.80171], [-1.21392, -2.11803, -0.425325], [-1.21392, 2.11803, -0.425325], [-1.96417, -0.809017, -1.27598], [-1.96417, 0.809017, -1.27598], [2.06457, -0.5, -1.27598], [2.06457, 0.5, -1.27598], [2.22703, -1.0, -0.425325], [2.22703, 1.0, -0.425325], [2.38949, -0.5, 0.425325], [2.38949, 0.5, 0.425325], [-1.11352, -1.80902, 1.27598], [-1.11352, 1.80902, 1.27598], [1.11352, -1.80902, -1.27598], [1.11352, 1.80902, -1.27598], [-2.38949, -0.5, -0.425325], [-2.38949, 0.5, -0.425325], [-1.63925, -1.80902, 0.425325], [-1.63925, 1.80902, 0.425325], [1.63925, -1.80902, -0.425325], [1.63925, 1.80902, -0.425325], [1.96417, -0.809017, 1.27598], [1.96417, 0.809017, 1.27598], [0.850651, 0.0, 2.32744], [-2.22703, -1.0, 0.425325], [-2.22703, 1.0, 0.425325], [-0.850651, 0.0, -2.32744], [-0.525731, -1.61803, -1.80171], [-0.525731, 1.61803, -1.80171], [0.525731, -1.61803, 1.80171], [0.525731, 1.61803, 1.80171]]}, "143": {"AlternativeName": null, "Comment": null, "Name": "Truncated Octahedron", "ShortName": "A04", "vertices": [[-1.5, -0.5, 0.0], [-1.5, 0.5, 0.0], [-1.0, -1.0, -0.707107], [-1.0, -1.0, 0.707107], [-1.0, 1.0, -0.707107], [-1.0, 1.0, 0.707107], [-0.5, -1.5, 0.0], [-0.5, -0.5, -1.41421], [-0.5, -0.5, 1.41421], [-0.5, 0.5, -1.41421], [-0.5, 0.5, 1.41421], [-0.5, 1.5, 0.0], [0.5, -1.5, 0.0], [0.5, -0.5, -1.41421], [0.5, -0.5, 1.41421], [0.5, 0.5, -1.41421], [0.5, 0.5, 1.41421], [0.5, 1.5, 0.0], [1.0, -1.0, -0.707107], [1.0, -1.0, 0.707107], [1.0, 1.0, -0.707107], [1.0, 1.0, 0.707107], [1.5, -0.5, 0.0], [1.5, 0.5, 0.0]]}, "144": {"AlternativeName": null, "Comment": null, "Name": "Truncated Tetrahedron", "ShortName": "A03", "vertices": [[0, -1, -0.612372], [0, 1, -0.612372], [-0.57735, -1, 0.204124], [-0.57735, 1, 0.204124], [-0.288675, -0.5, 1.02062], [-0.288675, 0.5, 1.02062], [0.57735, 0, 1.02062], [1.1547, 0, 0.204124], [-0.866025, -0.5, -0.612372], [-0.866025, 0.5, -0.612372], [0.866025, -0.5, -0.612372], [0.866025, 0.5, -0.612372]]}}"""
+DAMASCENO_SCIENCE_2012 = """{"1": {"AlternativeName": null, "Comment": null, "Name": "Acute Golden Rhombohedron", "ShortName": "O19", "vertices": [[-0.947214, -0.58541, -0.425325], [-0.5, -0.309017, 0.425325], [-0.5, 0.309017, -0.425325], [-0.0527864, 0.58541, 0.425325], [0.0527864, -0.58541, -0.425325], [0.5, -0.309017, 0.425325], [0.5, 0.309017, -0.425325], [0.947214, 0.58541, 0.425325]]}, "2": {"AlternativeName": null, "Comment": "edge length=8", "Name": "Square Antiprism", "ShortName": "O08", "vertices": [[-0.5, -0.5, -0.420448208], [-0.5, 0.5, -0.420448208], [0.0, -0.707106781, 0.420448208], [0.0, 0.707106781, 0.420448208], [0.5, -0.5, -0.420448208], [0.5, 0.5, -0.420448208], [-0.707106781, 0.0, 0.420448208], [0.707106781, 0.0, 0.420448208]]}, "3": {"AlternativeName": null, "Comment": "edge length=9", "Name": "Pentagonal Antiprism", "ShortName": "O09", "vertices": [[-0.850650808, 0.0, -0.425325404], [0.850650808, 0.0, 0.425325404], [-0.68819096, -0.5, 0.425325404], [-0.68819096, 0.5, 0.425325404], [-0.262865556, -0.809016994, -0.425325404], [-0.262865556, 0.809016994, -0.425325404], [0.262865556, -0.809016994, 0.425325404], [0.262865556, 0.809016994, 0.425325404], [0.68819096, -0.5, -0.425325404], [0.68819096, 0.5, -0.425325404]]}, "4": {"AlternativeName": null, "Comment": "edge length=10", "Name": "Hexagonal Antiprism", "ShortName": "O10", "vertices": [[-1.0, 0.0, 0.427799839], [-0.5, -0.866025404, 0.427799839], [-0.5, 0.866025404, 0.427799839], [0.0, -1.0, -0.427799839], [0.0, 1.0, -0.427799839], [0.5, -0.866025404, 0.427799839], [0.5, 0.866025404, 0.427799839], [1.0, 0.0, 0.427799839], [-0.866025404, -0.5, -0.427799839], [-0.866025404, 0.5, -0.427799839], [0.866025404, -0.5, -0.427799839], [0.866025404, 0.5, -0.427799839]]}, "5": {"AlternativeName": null, "Comment": "edge length=11", "Name": "Heptagonal Antiprism", "ShortName": "O11", "vertices": [[-1.038260698, -0.5, 0.429236598], [-1.038260698, 0.5, 0.429236598], [1.038260698, -0.5, -0.429236598], [1.038260698, 0.5, -0.429236598], [-1.152382435, 0.0, -0.429236598], [1.152382435, 0.0, 0.429236598], [-0.256429216, -1.123489802, 0.429236598], [-0.256429216, 1.123489802, 0.429236598], [0.256429216, -1.123489802, -0.429236598], [0.256429216, 1.123489802, -0.429236598], [-0.718498696, -0.900968868, -0.429236598], [-0.718498696, 0.900968868, -0.429236598], [0.718498696, -0.900968868, 0.429236598], [0.718498696, 0.900968868, 0.429236598]]}, "6": {"AlternativeName": null, "Comment": "edge length=12", "Name": "Octagonal Antiprism", "ShortName": "O12", "vertices": [[-0.5, -1.207106781, -0.430147785], [-0.5, 1.207106781, -0.430147785], [0.0, -1.306562965, 0.430147785], [0.0, 1.306562965, 0.430147785], [0.5, -1.207106781, -0.430147785], [0.5, 1.207106781, -0.430147785], [-1.207106781, -0.5, -0.430147785], [-1.207106781, 0.5, -0.430147785], [1.207106781, -0.5, -0.430147785], [1.207106781, 0.5, -0.430147785], [-1.306562965, 0.0, 0.430147785], [1.306562965, 0.0, 0.430147785], [-0.923879533, -0.923879533, 0.430147785], [-0.923879533, 0.923879533, 0.430147785], [0.923879533, -0.923879533, 0.430147785], [0.923879533, 0.923879533, 0.430147785]]}, "7": {"AlternativeName": null, "Comment": "edge length=13", "Name": "Nonagonal Antiprism", "ShortName": "O13", "vertices": [[-1.37373871, -0.5, 0.430763044], [-1.37373871, 0.5, 0.430763044], [1.37373871, -0.5, -0.430763044], [1.37373871, 0.5, -0.430763044], [-1.4619022, 0.0, -0.430763044], [-0.7309511, -1.266044443, 0.430763044], [-0.7309511, 1.266044443, 0.430763044], [0.7309511, -1.266044443, -0.430763044], [0.7309511, 1.266044443, -0.430763044], [1.4619022, 0.0, 0.430763044], [-1.119882057, -0.939692621, -0.430763044], [-1.119882057, 0.939692621, -0.430763044], [1.119882057, -0.939692621, 0.430763044], [1.119882057, 0.939692621, 0.430763044], [-0.253856653, -1.439692621, -0.430763044], [-0.253856653, 1.439692621, -0.430763044], [0.253856653, -1.439692621, 0.430763044], [0.253856653, 1.439692621, 0.430763044]]}, "8": {"AlternativeName": null, "Comment": "edge length=14", "Name": "Decagonal Antiprism", "ShortName": "O14", "vertices": [[0.0, -1.618033989, -0.431198502], [0.0, 1.618033989, -0.431198502], [-1.618033989, 0.0, 0.431198502], [1.618033989, 0.0, 0.431198502], [-1.309016994, -0.951056516, 0.431198502], [-1.309016994, 0.951056516, 0.431198502], [-0.5, -1.538841769, 0.431198502], [-0.5, 1.538841769, 0.431198502], [-0.951056516, -1.309016994, -0.431198502], [-0.951056516, 1.309016994, -0.431198502], [0.951056516, -1.309016994, -0.431198502], [0.951056516, 1.309016994, -0.431198502], [-1.538841769, -0.5, -0.431198502], [-1.538841769, 0.5, -0.431198502], [1.538841769, -0.5, -0.431198502], [1.538841769, 0.5, -0.431198502], [0.5, -1.538841769, 0.431198502], [0.5, 1.538841769, 0.431198502], [1.309016994, -0.951056516, 0.431198502], [1.309016994, 0.951056516, 0.431198502]]}, "9": {"AlternativeName": null, "Comment": null, "Name": "Augmented Dodecahedron", "ShortName": "J58", "vertices": [[0.047386905, 0.0, 1.339229524], [0.047386905, 0.0, -1.463290476], [0.225797905, -1.30902, 0.405055524], [0.225797905, 1.30902, 0.405055524], [0.514472905, -0.809017, -1.106470476], [0.514472905, 0.809017, -1.106470476], [1.091826905, -0.809017, 0.405055524], [1.091826905, 0.809017, 0.405055524], [-1.175463095, -0.5, 0.405055524], [-1.175463095, 0.5, 0.405055524], [1.270236905, -0.5, -0.529116476], [1.270236905, 0.5, -0.529116476], [-0.886785095, 0.0, -1.106470476], [-0.419699095, -0.809017, 0.982409524], [-0.419699095, 0.809017, 0.982409524], [0.981558905, 0.0, 0.982409524], [-0.997053095, -0.809017, -0.529116476], [-0.997053095, 0.809017, -0.529116476], [-0.947738095, 0.0, 1.240609524], [-0.131024095, -1.30902, -0.529116476], [-0.131024095, 1.30902, -0.529116476]]}, "10": {"AlternativeName": null, "Comment": null, "Name": "Augmented Hexagonal Prism", "ShortName": "J54", "vertices": [[-1.104797692, -0.060505077, -0.5], [-1.104797692, -0.060505077, 0.5], [-0.604797692, -0.926530077, -0.5], [-0.604797692, -0.926530077, 0.5], [-0.604797692, 0.805519923, -0.5], [-0.604797692, 0.805519923, 0.5], [0.395202308, -0.926530077, -0.5], [0.395202308, -0.926530077, 0.5], [0.395202308, 0.805519923, -0.5], [0.395202308, 0.805519923, 0.5], [0.895202308, -0.060505077, -0.5], [0.895202308, -0.060505077, 0.5], [1.257572308, 0.726060923, 0.0]]}, "11": {"AlternativeName": null, "Comment": null, "Name": "Augmented Pentagonal Prism", "ShortName": "J52", "vertices": [[0.748030818, -0.074557727, -0.5], [0.748030818, -0.074557727, 0.5], [0.160245818, -0.883574727, -0.5], [0.160245818, -0.883574727, 0.5], [0.160245818, 0.734459273, -0.5], [0.160245818, 0.734459273, 0.5], [-0.790811182, -0.574557727, -0.5], [-0.790811182, -0.574557727, 0.5], [-0.790811182, 0.425442273, -0.5], [-0.790811182, 0.425442273, 0.5], [1.026199818, 0.745577273, 0.0]]}, "12": {"AlternativeName": null, "Comment": null, "Name": "Augmented Sphenocorona", "ShortName": "J87", "vertices": [[-0.802793436, -0.741304536, 0.404638136], [-0.734923436, 0.054175364, -0.197521864], [-0.498633436, 0.138255464, 0.770518136], [-0.215103436, 0.845905464, 0.123338136], [-0.198633436, 0.595905464, -0.844761864], [-0.085183636, -0.700254536, -0.290601864], [0.151106564, -0.616184536, 0.677438136], [0.451106564, -0.158534536, -0.937841864], [0.473206564, 0.329465464, 0.632768136], [0.649110564, 0.597815464, -0.314361864], [0.810740564, -0.345244536, -0.023611364]]}, "13": {"AlternativeName": null, "Comment": null, "Name": "Augmented Triangular Prism", "ShortName": "J49", "vertices": [[-0.359802286, -0.623196, -0.5], [-0.359802286, -0.623196, 0.5], [-0.359802286, 0.376804, -0.5], [-0.359802286, 0.376804, 0.5], [0.506222714, -0.123196, -0.5], [0.506222714, -0.123196, 0.5], [0.426763714, 0.739176, 0.0]]}, "14": {"AlternativeName": null, "Comment": null, "Name": "Augmented Tridiminished Icosahedron", "ShortName": "J64", "vertices": [[0.0, 0.0, -1.4685573], [-0.57735, 0.0, -0.6520583], [-0.288675, -0.5, 0.8594637], [-0.288675, 0.5, 0.8594637], [0.288675, -0.5, -0.6520583], [0.288675, 0.5, -0.6520583], [0.57735, 0.0, 0.8594637], [-0.934172, 0.0, 0.2821137], [0.467086, -0.809017, 0.2821137], [0.467086, 0.809017, 0.2821137]]}, "15": {"AlternativeName": null, "Comment": null, "Name": "Augmented Truncated Cube", "ShortName": "J66", "vertices": [[-0.5, 1.20711, -1.480568571], [-0.5, 1.20711, 0.933651429], [-0.5, -1.20711, -1.480568571], [-0.5, -1.20711, 0.933651429], [0.0, -0.707107, 1.640751429], [0.0, 0.707107, 1.640751429], [0.5, 1.20711, -1.480568571], [0.5, 1.20711, 0.933651429], [0.5, -1.20711, -1.480568571], [0.5, -1.20711, 0.933651429], [-0.707107, 0.0, 1.640751429], [0.707107, 0.0, 1.640751429], [1.20711, -0.5, -1.480568571], [1.20711, -0.5, 0.933651429], [1.20711, 0.5, -1.480568571], [1.20711, 0.5, 0.933651429], [1.20711, 1.20711, -0.773458571], [1.20711, 1.20711, 0.226541429], [1.20711, -1.20711, -0.773458571], [1.20711, -1.20711, 0.226541429], [-1.20711, -0.5, -1.480568571], [-1.20711, -0.5, 0.933651429], [-1.20711, 0.5, -1.480568571], [-1.20711, 0.5, 0.933651429], [-1.20711, 1.20711, -0.773458571], [-1.20711, 1.20711, 0.226541429], [-1.20711, -1.20711, -0.773458571], [-1.20711, -1.20711, 0.226541429]]}, "16": {"AlternativeName": null, "Comment": null, "Name": "Augmented Truncated Dodecahedron", "ShortName": "J68", "vertices": [[-3.133096349, -0.186728263, 0.418858126], [-3.115256349, -0.019748363, -0.566939874], [-2.877166349, 0.191938737, 1.308305126], [-2.830456349, 0.629098737, -1.272554874], [-2.801726349, -0.791431263, 1.143105126], [-2.755006349, -0.354271263, -1.437754874], [-2.445226349, 0.971615737, 1.761655126], [-2.387476349, 1.511970737, -1.428454874], [-2.247706349, -1.602879263, 1.329155126], [-2.189956349, -1.062519263, -1.860964874], [-2.002256349, 1.854490737, 1.605755126], [-1.955536349, 2.291650737, -0.975100874], [-1.717456349, 2.503340737, 0.900143126], [-1.699606349, 2.670320737, -0.085655874], [-1.682656349, -2.311129263, 0.905943126], [-1.670886349, 1.249790737, 2.329995126], [-1.635936349, -1.873969263, -1.674914874], [-1.595286349, 1.957130737, -1.845914874], [-1.426736349, -1.932459263, 1.795385126], [-1.351136349, -1.225119263, -2.380524874], [-1.322416349, -2.645649263, 0.035129326], [-1.304566349, -2.478669263, -0.950669874], [-0.925264349, 2.948490737, 0.482683126], [-0.849908349, 0.920206737, 2.796235126], [-0.756471349, 1.794520737, -2.365474874], [-0.652387349, -1.654289263, 2.363725126], [-0.558949349, -0.779969263, -2.797984874], [-0.483593349, -2.808259263, -0.484431874], [-0.295890349, 0.108757337, 2.982285126], [-0.220444349, -0.874612263, 2.817085126], [-0.191424349, 1.086280737, -2.788684874], [-0.115977649, 0.102906237, -2.953884874], [0.071725651, 3.019920737, 0.512829126], [0.147081651, 0.991632737, 2.826375126], [0.240519651, 1.865950737, -2.335324874], [0.344602651, -1.582859263, 2.393875126], [0.438040651, -0.708543263, -2.767834874], [0.513396651, -2.736829263, -0.454285874], [0.892699651, 2.690340737, 0.979067126], [0.910544651, 2.857320737, -0.006731884], [0.939271651, 1.436790737, 2.408915126], [1.014863651, 2.144130737, -1.766994874], [1.183423651, -1.745459263, 1.874315126], [1.224073651, 2.085630737, 1.703305126], [1.259013651, -1.038129263, -2.301594874], [1.270793651, 2.522790737, -0.877545874], [1.287743651, -2.458659263, 0.114053126], [1.305583651, -2.291679263, -0.871745874], [1.543663651, -2.079989263, 1.003495126], [1.590383651, -1.642829263, -1.577354874], [1.778093651, 1.274180737, 1.889355126], [1.835833651, 1.814540737, -1.300754874], [1.975613651, -1.300309263, 1.456855126], [2.033363651, -0.759951263, -1.733254874], [2.145733651, -1.954989263, 0.214898126], [2.163583651, -1.788009263, -0.770900874], [2.343133651, 0.565934737, 1.466155126], [2.389853651, 1.003094737, -1.114704874], [2.418583651, -0.417435263, 1.300945126], [2.465303651, 0.019724807, -1.279904874], [2.577683651, -1.175319263, 0.668251126], [2.606553651, -0.905136263, -0.926804874], [2.703383651, 0.231411737, 0.595338126], [2.721233651, 0.398391737, -0.390460874], [2.862483651, -0.526468263, -0.037359274]]}, "17": {"AlternativeName": null, "Comment": null, "Name": "Augmented Truncated Tetrahedron", "ShortName": "J65", "vertices": [[0.0, -1.0, -0.326598], [0.0, 1.0, -0.326598], [-0.57735, -1.0, 0.489898], [-0.57735, 1.0, 0.489898], [-0.288675, -0.5, 1.306394], [-0.288675, 0.5, 1.306394], [0.288675, -0.5, -1.143096], [0.57735, 0.0, 1.306394], [1.1547, 0.0, 0.489898], [-0.866025, -0.5, -0.326598], [-0.866025, 0.5, -0.326598], [0.866025, -0.5, -0.326598], [0.866025, 0.5, -0.326598], [-0.57735, 0.0, -1.143096], [0.288675, 0.5, -1.143096]]}, "18": {"AlternativeName": null, "Comment": null, "Name": "Biaugmented Pentagonal Prism", "ShortName": "J53", "vertices": [[0.8728575, -0.068344583, -0.5], [0.8728575, -0.068344583, 0.5], [0.2850725, -0.877361583, -0.5], [0.2850725, -0.877361583, 0.5], [0.2850725, 0.740672417, -0.5], [0.2850725, 0.740672417, 0.5], [-0.6659845, -0.568344583, -0.5], [-0.6659845, -0.568344583, 0.5], [-0.6659845, 0.431655417, -0.5], [-0.6659845, 0.431655417, 0.5], [-1.3730935, -0.068344583, 0.0], [1.1510265, 0.751790417, 0.0]]}, "19": {"AlternativeName": null, "Comment": null, "Name": "Biaugmented Triangular Prism", "ShortName": "J50", "vertices": [[-0.226438625, -0.6077965, -0.5], [-0.226438625, -0.6077965, 0.5], [-0.226438625, 0.3922035, -0.5], [-0.226438625, 0.3922035, 0.5], [0.639586375, -0.1077965, -0.5], [0.639586375, -0.1077965, 0.5], [-0.933545625, -0.1077965, 0.0], [0.560127375, 0.7545755, 0.0]]}, "20": {"AlternativeName": null, "Comment": null, "Name": "Biaugmented Truncated Cube", "ShortName": "J67", "vertices": [[-0.5, 1.20711, 1.20711], [-0.5, 1.20711, -1.20711], [-0.5, -1.20711, 1.20711], [-0.5, -1.20711, -1.20711], [0.0, -0.707107, 1.91421], [0.0, -0.707107, -1.91421], [0.0, 0.707107, 1.91421], [0.0, 0.707107, -1.91421], [0.5, 1.20711, 1.20711], [0.5, 1.20711, -1.20711], [0.5, -1.20711, 1.20711], [0.5, -1.20711, -1.20711], [-0.707107, 0.0, 1.91421], [-0.707107, 0.0, -1.91421], [0.707107, 0.0, 1.91421], [0.707107, 0.0, -1.91421], [1.20711, -0.5, 1.20711], [1.20711, -0.5, -1.20711], [1.20711, 0.5, 1.20711], [1.20711, 0.5, -1.20711], [1.20711, 1.20711, -0.5], [1.20711, 1.20711, 0.5], [1.20711, -1.20711, -0.5], [1.20711, -1.20711, 0.5], [-1.20711, -0.5, 1.20711], [-1.20711, -0.5, -1.20711], [-1.20711, 0.5, 1.20711], [-1.20711, 0.5, -1.20711], [-1.20711, 1.20711, -0.5], [-1.20711, 1.20711, 0.5], [-1.20711, -1.20711, -0.5], [-1.20711, -1.20711, 0.5]]}, "21": {"AlternativeName": null, "Comment": null, "Name": "Bigyrate Diminished Rhombicosidodecahedron", "ShortName": "J79", "vertices": [[-2.304394967, 0.39941876, 0.359382255], [-2.287104967, 0.10787476, -0.597019745], [-1.982764967, 0.17542776, 1.279370255], [-1.937504967, -0.58784224, -1.224519745], [-1.854394967, 1.29045576, 0.419002255], [-1.826424967, 0.81872776, -1.128489745], [-1.558994967, 1.54960576, -0.500551745], [-1.532764967, 1.06646576, 1.338990255], [-1.476824967, 0.12300976, -1.755989745], [-1.445074967, -0.47853924, 1.811550255], [-1.389124967, -1.42199424, -1.283429745], [-1.079334967, 1.91085576, 0.299087255], [-1.026824967, 1.01404576, -1.696369745], [-0.896693967, -1.31269424, 1.752630255], [-0.851431967, -2.07596424, -0.751257745], [-0.759390967, 1.74491576, -1.068429745], [-0.757702967, 1.68686576, 1.219080255], [-0.716957967, 0.96318676, 1.908020255], [-0.662758967, 0.00832236, 2.200070255], [-0.643722967, -0.27181124, -2.143369745], [-0.589524967, -1.22667424, -1.851309745], [-0.547090967, -2.00840424, 1.125130255], [-0.529802967, -2.29995424, 0.168733255], [-0.279728967, 2.10616576, -0.268791745], [-0.193722967, 0.61922476, -2.083749745], [-0.114379167, -0.82582924, 2.141160255], [-0.051829567, -1.88064424, -1.319139745], [0.238989033, 1.80180576, -1.067719745], [0.240678033, 1.74374576, 1.219780255], [0.245260033, -0.72512024, -2.078419745], [0.281423033, 1.02007576, 1.908720255], [0.335621033, 0.06520696, 2.200780255], [0.451289033, -1.95152424, 1.125840255], [0.468577033, -2.24306424, 0.169440255], [0.536082033, 2.00289576, 0.300231255], [0.588591033, 1.10608576, -1.695219745], [0.695261033, 0.16591476, -2.018799745], [0.718720033, -1.22065424, 1.753780255], [0.763981033, -1.98392424, -0.750113745], [0.782956033, -1.37908424, -1.546239745], [1.054805033, 1.69852576, -0.498701745], [1.081025033, 1.21538576, 1.340840255], [1.168725033, -0.32961324, 1.813400255], [1.250895033, -1.75620424, 0.557962255], [1.376425033, 1.47453576, 0.421289255], [1.404405033, 1.00280576, -1.126199745], [1.511075033, 0.06263816, -1.449769745], [1.518325033, -1.02533424, 1.185900255], [1.546295033, -1.49705424, -0.361590745], [1.565265033, -0.89222624, -1.157719745], [1.629405033, 0.38123876, 1.281930255], [1.924805033, 0.64038576, 0.362376255], [1.942095033, 0.34884176, -0.594025745], [1.979005033, -0.31447924, 0.654430255], [1.996295033, -0.60602224, -0.301970745]]}, "22": {"AlternativeName": null, "Comment": null, "Name": "Bilunabirotunda", "ShortName": "J91", "vertices": [[-1.068949879, -0.081979664, 0.071567697], [-1.020379879, 0.821387236, 0.497668897], [-0.554082879, 0.043835936, 0.919549897], [-0.550126879, -0.915152764, -0.119827103], [-0.416538879, -0.101788764, -0.686033103], [-0.337937879, 1.359890236, 0.003412297], [-0.035254679, -0.789335764, 0.728153897], [0.035255221, 0.789337236, -0.728157103], [0.337934121, -1.359889764, -0.003409163], [0.416540121, 0.101790236, 0.686032897], [0.550126121, 0.915156236, 0.119827897], [0.554085121, -0.043839464, -0.919552103], [1.020370121, -0.821390764, -0.497667103], [1.068960121, 0.081979836, -0.071567803]]}, "23": {"AlternativeName": null, "Comment": null, "Name": "Cube", "ShortName": "P03", "vertices": [[-0.5, -0.5, -0.5], [-0.5, -0.5, 0.5], [-0.5, 0.5, -0.5], [-0.5, 0.5, 0.5], [0.5, -0.5, -0.5], [0.5, -0.5, 0.5], [0.5, 0.5, -0.5], [0.5, 0.5, 0.5]]}, "24": {"AlternativeName": null, "Comment": null, "Name": "Cuboctahedron", "ShortName": "A01", "vertices": [[-1.0, 0.0, 0.0], [-0.5, -0.5, -0.707107], [-0.5, -0.5, 0.707107], [-0.5, 0.5, -0.707107], [-0.5, 0.5, 0.707107], [0.0, -1.0, 0.0], [0.0, 1.0, 0.0], [0.5, -0.5, -0.707107], [0.5, -0.5, 0.707107], [0.5, 0.5, -0.707107], [0.5, 0.5, 0.707107], [1.0, 0.0, 0.0]]}, "25": {"AlternativeName": null, "Comment": null, "Name": "Deltoidal Hexecontahedron", "ShortName": "C09", "vertices": [[-0.0, 0.0, -2.77775], [-0.0, 0.0, 2.77775], [-0.0, -2.77775, 0.0], [-0.0, 2.77775, 0.0], [-0.0, -2.42408, -1.49817], [-0.0, -2.42408, 1.49817], [-0.0, -0.973249, -2.548], [-0.0, -0.973249, 2.548], [-0.0, 0.973249, -2.548], [-0.0, 0.973249, 2.548], [-0.0, 2.42408, -1.49817], [-0.0, 2.42408, 1.49817], [-2.77775, 0.0, 0.0], [-1.38888, -2.24725, 0.858373], [-1.38888, -2.24725, -0.858373], [-1.38888, 2.24725, 0.858373], [-1.38888, 2.24725, -0.858373], [1.38888, -2.24725, 0.858373], [1.38888, -2.24725, -0.858373], [1.38888, 2.24725, 0.858373], [1.38888, 2.24725, -0.858373], [2.77775, 0.0, 0.0], [-2.548, 0.0, -0.973249], [-2.548, 0.0, 0.973249], [-1.57475, -1.57475, -1.57475], [-1.57475, -1.57475, 1.57475], [-1.57475, 1.57475, -1.57475], [-1.57475, 1.57475, 1.57475], [-2.42408, -1.49817, 0.0], [-2.42408, 1.49817, 0.0], [-0.973249, -2.548, 0.0], [-0.973249, 2.548, 0.0], [-1.49817, 0.0, -2.42408], [-1.49817, 0.0, 2.42408], [-2.24725, 0.858373, -1.38888], [-2.24725, 0.858373, 1.38888], [-2.24725, -0.858373, -1.38888], [-2.24725, -0.858373, 1.38888], [0.858373, -1.38888, -2.24725], [0.858373, -1.38888, 2.24725], [0.858373, 1.38888, -2.24725], [0.858373, 1.38888, 2.24725], [-0.858373, -1.38888, -2.24725], [-0.858373, -1.38888, 2.24725], [-0.858373, 1.38888, -2.24725], [-0.858373, 1.38888, 2.24725], [1.49817, 0.0, -2.42408], [1.49817, 0.0, 2.42408], [2.24725, 0.858373, -1.38888], [2.24725, 0.858373, 1.38888], [2.24725, -0.858373, -1.38888], [2.24725, -0.858373, 1.38888], [0.973249, -2.548, 0.0], [0.973249, 2.548, 0.0], [2.42408, -1.49817, 0.0], [2.42408, 1.49817, 0.0], [1.57475, -1.57475, -1.57475], [1.57475, -1.57475, 1.57475], [1.57475, 1.57475, -1.57475], [1.57475, 1.57475, 1.57475], [2.548, 0.0, -0.973249], [2.548, 0.0, 0.973249]]}, "26": {"AlternativeName": null, "Comment": null, "Name": "Deltoidal Icositetrahedron", "ShortName": "C08", "vertices": [[-1.19448, -1.19448, 0.0], [-1.19448, 0.0, -1.19448], [-1.19448, 0.0, 1.19448], [-1.19448, 1.19448, 0.0], [0.0, -1.19448, -1.19448], [0.0, -1.19448, 1.19448], [0.0, 0.0, -1.68925], [0.0, 0.0, 1.68925], [0.0, 1.19448, -1.19448], [0.0, 1.19448, 1.19448], [0.0, -1.68925, 0.0], [0.0, 1.68925, 0.0], [1.19448, -1.19448, 0.0], [1.19448, 0.0, -1.19448], [1.19448, 0.0, 1.19448], [1.19448, 1.19448, 0.0], [-1.68925, 0.0, 0.0], [1.68925, 0.0, 0.0], [0.92388, 0.92388, 0.92388], [0.92388, 0.92388, -0.92388], [0.92388, -0.92388, 0.92388], [0.92388, -0.92388, -0.92388], [-0.92388, 0.92388, 0.92388], [-0.92388, 0.92388, -0.92388], [-0.92388, -0.92388, 0.92388], [-0.92388, -0.92388, -0.92388]]}, "27": {"AlternativeName": null, "Comment": null, "Name": "Diminished Rhombicosidodecahedron", "ShortName": "J76", "vertices": [[-2.170740182, -0.339993115, -0.364123767], [-2.071220182, 0.646560885, -0.234480767], [-1.970980182, -0.787896115, 0.507356233], [-1.869560182, -0.245683115, -1.313011767], [-1.809950182, 0.808380885, 0.717124233], [-1.807610182, -1.132227115, -0.854534767], [-1.770030182, 0.740870885, -1.183371767], [-1.748000182, -0.078162115, 1.175608233], [-1.607850182, -1.580127115, 0.016945933], [-1.547050182, 1.450602885, -0.515123767], [-1.346580182, -1.418307115, 0.968548233], [-1.285780182, 1.612422885, 0.436481233], [-1.182470182, -0.540989115, -1.976871767], [-1.123590182, -0.708575115, 1.636798233], [-1.120530182, -1.427527115, -1.518391767], [-1.086020182, 1.164522885, 1.307958233], [-1.024070182, 0.277978885, 1.766438233], [-1.021440182, 1.055282885, -1.767101767], [-0.798452182, 1.765022885, -1.098861767], [-0.797305182, -2.152257115, -0.108306967], [-0.658306182, 0.263054885, -2.257511767], [-0.536033182, -1.990437115, 0.843297233], [-0.496122182, -2.057947115, -1.057201767], [-0.375705182, 2.026852885, 0.440871233], [-0.175943182, 1.578952885, 1.312348233], [-0.175235182, -0.842061115, 1.924548233], [-0.111366182, 1.469712885, -1.762711767], [-0.075708782, 0.144491885, 2.054188233], [-0.074521582, 2.121162885, -0.508019767], [0.151052818, -2.285737115, 0.179440233], [0.187896818, -1.634297115, 1.434138233], [0.251765818, 0.677481885, -2.253121767], [0.448459818, 0.948535885, 1.773548233], [0.452235818, -2.191427115, -0.769450767], [0.572652818, 1.893362885, 0.728618233], [0.612564818, 1.825852885, -1.171881767], [0.734836818, -0.427633115, 1.928938233], [0.873836818, 1.987672885, -0.220272767], [0.874982818, -1.929597115, 0.770277233], [1.097969818, -1.219867115, 1.438528233], [1.197059818, 1.262952885, 1.189808233], [1.200119818, 0.543995885, -1.965381767], [1.259009818, 0.376409885, 1.648288233], [1.362309818, -1.777007115, -0.765060767], [1.423109818, 1.253732885, -1.297131767], [1.623579818, -1.615187115, 0.186544233], [1.684379818, 1.415552885, -0.345525767], [1.824529818, -0.086416815, -1.504181767], [1.846569818, -0.905449115, 0.854792233], [1.884139818, 0.967646885, 0.525955233], [1.886479818, -0.972959115, -1.045701767], [1.946089818, 0.081103885, 0.984438233], [2.047509818, 0.623316885, -0.835935767], [2.147749818, -0.811139115, -0.094099167], [2.247279818, 0.175413885, 0.035544333]]}, "28": {"AlternativeName": "Trinagular Dipyramid", "Comment": null, "Name": "Dipyramid 3", "ShortName": "J12", "vertices": [[0.0, 0.0, -0.816496581], [0.0, 0.0, 0.816496581], [-0.288675135, -0.5, 0.0], [-0.288675135, 0.5, 0.0], [0.577350269, 0.0, 0.0]]}, "29": {"AlternativeName": "Pentagonal Dipyramid", "Comment": null, "Name": "Dipyramid 5", "ShortName": "J13", "vertices": [[0.0, 0.0, -0.525731112], [0.0, 0.0, 0.525731112], [0.850650808, 0.0, 0.0], [0.262865556, -0.809016994, 0.0], [0.262865556, 0.809016994, 0.0], [-0.68819096, -0.5, 0.0], [-0.68819096, 0.5, 0.0]]}, "30": {"AlternativeName": null, "Comment": null, "Name": "Disdyakis Dodecahedron", "ShortName": "C10", "vertices": [[0.0, 1.13152, 1.13152], [0.0, -1.13152, -1.13152], [0.0, 1.13152, -1.13152], [0.0, -1.13152, 1.13152], [1.13152, 0.0, 1.13152], [1.13152, 1.13152, 0.0], [0.975287, 0.975287, 0.975287], [-0.975287, -0.975287, -0.975287], [-1.13152, 0.0, -1.13152], [-1.13152, -1.13152, 0.0], [0.975287, 0.975287, -0.975287], [0.975287, -0.975287, 0.975287], [-0.975287, 0.975287, 0.975287], [0.975287, -0.975287, -0.975287], [-0.975287, 0.975287, -0.975287], [-0.975287, -0.975287, 0.975287], [1.13152, 0.0, -1.13152], [1.13152, -1.13152, 0.0], [-1.13152, 0.0, 1.13152], [-1.13152, 1.13152, 0.0], [-1.84505, 0.0, 0.0], [0.0, -1.84505, 0.0], [0.0, 0.0, -1.84505], [1.84505, 0.0, 0.0], [0.0, 1.84505, 0.0], [0.0, 0.0, 1.84505]]}, "31": {"AlternativeName": null, "Comment": null, "Name": "Disdyakis Triacontahedron", "ShortName": "C11", "vertices": [[-0.0, -0.0, -2.72755], [-0.0, -0.0, 2.72755], [-0.0, -2.72755, -0.0], [-0.0, 2.72755, -0.0], [-0.0, -0.991163, -2.5949], [-0.0, 0.991163, -2.5949], [-0.0, -0.991163, 2.5949], [-0.0, 0.991163, 2.5949], [-2.72755, -0.0, -0.0], [2.72755, -0.0, -0.0], [-1.60374, -1.60374, -1.60374], [1.60374, 1.60374, 1.60374], [-2.5949, -0.0, -0.991163], [-0.991163, -2.5949, -0.0], [-0.842858, -1.36377, -2.20663], [-2.20663, -0.842858, -1.36377], [-1.36377, -2.20663, -0.842858], [-0.842858, 1.36377, -2.20663], [-2.20663, -0.842858, 1.36377], [1.36377, -2.20663, -0.842858], [-2.5949, -0.0, 0.991163], [0.991163, -2.5949, -0.0], [-1.55694, -0.0, -2.51918], [-2.51918, -1.55694, -0.0], [-0.0, -2.51918, -1.55694], [-2.20663, 0.842858, -1.36377], [-1.36377, -2.20663, 0.842858], [0.842858, -1.36377, -2.20663], [-2.20663, 0.842858, 1.36377], [1.36377, -2.20663, 0.842858], [0.842858, 1.36377, -2.20663], [-2.51918, 1.55694, -0.0], [-0.0, -2.51918, 1.55694], [1.55694, -0.0, -2.51918], [-1.60374, -1.60374, 1.60374], [-1.60374, 1.60374, -1.60374], [1.60374, -1.60374, -1.60374], [-1.60374, 1.60374, 1.60374], [1.60374, -1.60374, 1.60374], [1.60374, 1.60374, -1.60374], [-0.842858, -1.36377, 2.20663], [-1.36377, 2.20663, -0.842858], [2.20663, -0.842858, -1.36377], [-0.842858, 1.36377, 2.20663], [1.36377, 2.20663, -0.842858], [2.20663, -0.842858, 1.36377], [-1.36377, 2.20663, 0.842858], [0.842858, -1.36377, 2.20663], [2.20663, 0.842858, -1.36377], [1.36377, 2.20663, 0.842858], [0.842858, 1.36377, 2.20663], [2.20663, 0.842858, 1.36377], [-0.991163, 2.5949, -0.0], [2.5949, -0.0, -0.991163], [0.991163, 2.5949, -0.0], [2.5949, -0.0, 0.991163], [-1.55694, -0.0, 2.51918], [-0.0, 2.51918, -1.55694], [2.51918, -1.55694, -0.0], [-0.0, 2.51918, 1.55694], [1.55694, -0.0, 2.51918], [2.51918, 1.55694, -0.0]]}, "32": {"AlternativeName": null, "Comment": null, "Name": "Disphenocingulum", "ShortName": "J90", "vertices": [[-1.171140619, 0.293688206, 0.109311788], [-0.838770619, -0.457648794, 0.679416788], [-0.814439619, -0.537541794, -0.317089213], [-0.666585619, 0.849234206, -0.551588213], [-0.460440619, 0.461075206, 0.792598788], [-0.309880619, 0.018003906, -0.977988213], [-0.117267619, -0.991153794, 0.238052788], [0.010481881, -0.901366794, -0.749679213], [0.044118781, 1.016620206, 0.131698788], [0.113460381, -0.340276794, 0.961324788], [0.297984381, 0.793976206, -0.809566213], [0.530304381, 0.565040206, 0.879826788], [0.656860381, -0.138801794, -0.775936213], [0.803498381, -0.896900794, -0.140495213], [0.887587381, 0.512075206, -0.052664813], [1.034229381, -0.246022794, 0.582776788]]}, "33": {"AlternativeName": null, "Comment": null, "Name": "Dodecahedron", "ShortName": "P05", "vertices": [[-1.37638, 0.0, 0.262866], [1.37638, 0.0, -0.262866], [-0.425325, -1.30902, 0.262866], [-0.425325, 1.30902, 0.262866], [1.11352, -0.809017, 0.262866], [1.11352, 0.809017, 0.262866], [-0.262866, -0.809017, 1.11352], [-0.262866, 0.809017, 1.11352], [-0.688191, -0.5, -1.11352], [-0.688191, 0.5, -1.11352], [0.688191, -0.5, 1.11352], [0.688191, 0.5, 1.11352], [0.850651, 0.0, -1.11352], [-1.11352, -0.809017, -0.262866], [-1.11352, 0.809017, -0.262866], [-0.850651, 0.0, 1.11352], [0.262866, -0.809017, -1.11352], [0.262866, 0.809017, -1.11352], [0.425325, -1.30902, -0.262866], [0.425325, 1.30902, -0.262866]]}, "34": {"AlternativeName": null, "Comment": null, "Name": "Duerers Solid", "ShortName": "O20", "vertices": [[0.0, -1.7769, 0.961343], [0.0, 1.7769, -0.961343], [0.0, -1.09819, 1.69574], [0.0, 1.09819, -1.69574], [-1.53884, -0.888451, -0.961343], [-1.53884, 0.888451, 0.961343], [1.53884, -0.888451, -0.961343], [1.53884, 0.888451, 0.961343], [-0.951057, -0.549093, -1.69574], [-0.951057, 0.549093, 1.69574], [0.951057, -0.549093, -1.69574], [0.951057, 0.549093, 1.69574]]}, "35": {"AlternativeName": null, "Comment": null, "Name": "Elongated Dodecahedron", "ShortName": "O21", "vertices": [[-0.866025, -0.866025, -0.5], [-0.866025, -0.866025, 0.5], [-0.866025, 0.0, -1.0], [-0.866025, 0.0, 1.0], [-0.866025, 0.866025, -0.5], [-0.866025, 0.866025, 0.5], [0.0, -0.866025, -1.0], [0.0, -0.866025, 1.0], [0.0, 0.0, -1.5], [0.0, 0.0, 1.5], [0.0, 0.866025, -1.0], [0.0, 0.866025, 1.0], [0.866025, -0.866025, -0.5], [0.866025, -0.866025, 0.5], [0.866025, 0.0, -1.0], [0.866025, 0.0, 1.0], [0.866025, 0.866025, -0.5], [0.866025, 0.866025, 0.5]]}, "36": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Cupola", "ShortName": "J20", "vertices": [[-4e-08, -1.61803, -0.7051462], [-4e-08, -1.61803, 0.2948538], [-4e-08, 1.61803, -0.7051462], [-4e-08, 1.61803, 0.2948538], [0.85065096, 0.0, 0.8205848], [0.26286596, -0.809017, 0.8205848], [0.26286596, 0.809017, 0.8205848], [-0.95105704, -1.30902, -0.7051462], [-0.95105704, -1.30902, 0.2948538], [-0.95105704, 1.30902, -0.7051462], [-0.95105704, 1.30902, 0.2948538], [0.95105696, -1.30902, -0.7051462], [0.95105696, -1.30902, 0.2948538], [0.95105696, 1.30902, -0.7051462], [0.95105696, 1.30902, 0.2948538], [-0.68819104, -0.5, 0.8205848], [-0.68819104, 0.5, 0.8205848], [-1.53884004, -0.5, -0.7051462], [-1.53884004, -0.5, 0.2948538], [-1.53884004, 0.5, -0.7051462], [-1.53884004, 0.5, 0.2948538], [1.53883996, -0.5, -0.7051462], [1.53883996, -0.5, 0.2948538], [1.53883996, 0.5, -0.7051462], [1.53883996, 0.5, 0.2948538]]}, "37": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Dipyramid", "ShortName": "J16", "vertices": [[-1.67e-07, 0.0, -1.02573], [-1.67e-07, 0.0, 1.02573], [0.850650833, 0.0, -0.5], [0.850650833, 0.0, 0.5], [0.262865833, -0.809017, -0.5], [0.262865833, -0.809017, 0.5], [0.262865833, 0.809017, -0.5], [0.262865833, 0.809017, 0.5], [-0.688191167, -0.5, -0.5], [-0.688191167, -0.5, 0.5], [-0.688191167, 0.5, -0.5], [-0.688191167, 0.5, 0.5]]}, "38": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Gyrobicupola", "ShortName": "J39", "vertices": [[-0.0, -1.61803, -0.5], [-0.0, -1.61803, 0.5], [-0.0, 1.61803, -0.5], [-0.0, 1.61803, 0.5], [-0.688191, -0.5, 1.02573], [-0.688191, 0.5, 1.02573], [0.688191, -0.5, -1.02573], [0.688191, 0.5, -1.02573], [-0.850651, 0.0, -1.02573], [0.850651, 0.0, 1.02573], [-0.262866, -0.809017, -1.02573], [-0.262866, 0.809017, -1.02573], [0.262866, -0.809017, 1.02573], [0.262866, 0.809017, 1.02573], [-0.951057, -1.30902, -0.5], [-0.951057, -1.30902, 0.5], [-0.951057, 1.30902, -0.5], [-0.951057, 1.30902, 0.5], [0.951057, -1.30902, -0.5], [0.951057, -1.30902, 0.5], [0.951057, 1.30902, -0.5], [0.951057, 1.30902, 0.5], [-1.53884, -0.5, -0.5], [-1.53884, -0.5, 0.5], [-1.53884, 0.5, -0.5], [-1.53884, 0.5, 0.5], [1.53884, -0.5, -0.5], [1.53884, -0.5, 0.5], [1.53884, 0.5, -0.5], [1.53884, 0.5, 0.5]]}, "39": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Gyrobirotunda", "ShortName": "J43", "vertices": [[-0.0, -1.61803, -0.5], [-0.0, -1.61803, 0.5], [-0.0, 1.61803, -0.5], [-0.0, 1.61803, 0.5], [-0.425325, -1.30902, -1.35065], [-0.425325, 1.30902, -1.35065], [0.425325, -1.30902, 1.35065], [0.425325, 1.30902, 1.35065], [-0.688191, -0.5, -1.87638], [-0.688191, 0.5, -1.87638], [0.688191, -0.5, 1.87638], [0.688191, 0.5, 1.87638], [-0.850651, 0.0, 1.87638], [0.850651, 0.0, -1.87638], [-1.11352, -0.809017, 1.35065], [-1.11352, 0.809017, 1.35065], [1.11352, -0.809017, -1.35065], [1.11352, 0.809017, -1.35065], [-0.262866, -0.809017, 1.87638], [-0.262866, 0.809017, 1.87638], [0.262866, -0.809017, -1.87638], [0.262866, 0.809017, -1.87638], [-0.951057, -1.30902, -0.5], [-0.951057, -1.30902, 0.5], [-0.951057, 1.30902, -0.5], [-0.951057, 1.30902, 0.5], [0.951057, -1.30902, -0.5], [0.951057, -1.30902, 0.5], [0.951057, 1.30902, -0.5], [0.951057, 1.30902, 0.5], [-1.53884, -0.5, -0.5], [-1.53884, -0.5, 0.5], [-1.53884, 0.5, -0.5], [-1.53884, 0.5, 0.5], [1.53884, -0.5, -0.5], [1.53884, -0.5, 0.5], [1.53884, 0.5, -0.5], [1.53884, 0.5, 0.5], [-1.37638, 0.0, -1.35065], [1.37638, 0.0, 1.35065]]}, "40": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Gyrocupolarotunda", "ShortName": "J41", "vertices": [[2.86e-07, -1.61803, -0.814471429], [2.86e-07, -1.61803, 0.185528571], [2.86e-07, 1.61803, -0.814471429], [2.86e-07, 1.61803, 0.185528571], [0.850651286, 0.0, -1.340201429], [0.262866286, -0.809017, -1.340201429], [0.262866286, 0.809017, -1.340201429], [-0.951056714, -1.30902, -0.814471429], [-0.951056714, -1.30902, 0.185528571], [-0.951056714, 1.30902, -0.814471429], [-0.951056714, 1.30902, 0.185528571], [0.951057286, -1.30902, -0.814471429], [0.951057286, -1.30902, 0.185528571], [0.951057286, 1.30902, -0.814471429], [0.951057286, 1.30902, 0.185528571], [-0.688190714, -0.5, -1.340201429], [-0.688190714, 0.5, -1.340201429], [-1.538839714, -0.5, -0.814471429], [-1.538839714, -0.5, 0.185528571], [-1.538839714, 0.5, -0.814471429], [-1.538839714, 0.5, 0.185528571], [1.538840286, -0.5, -0.814471429], [1.538840286, -0.5, 0.185528571], [1.538840286, 0.5, -0.814471429], [1.538840286, 0.5, 0.185528571], [1.376380286, 0.0, 1.036178571], [0.425325286, -1.30902, 1.036178571], [0.425325286, 1.30902, 1.036178571], [-1.113519714, -0.809017, 1.036178571], [-1.113519714, 0.809017, 1.036178571], [-0.850650714, 0.0, 1.561908571], [-0.262865714, -0.809017, 1.561908571], [-0.262865714, 0.809017, 1.561908571], [0.688191286, -0.5, 1.561908571], [0.688191286, 0.5, 1.561908571]]}, "41": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Orthobicupola", "ShortName": "J38", "vertices": [[-6.7e-08, -1.61803, -0.5], [-6.7e-08, -1.61803, 0.5], [-6.7e-08, 1.61803, -0.5], [-6.7e-08, 1.61803, 0.5], [0.850650933, 0.0, -1.02573], [0.850650933, 0.0, 1.02573], [0.262865933, -0.809017, -1.02573], [0.262865933, -0.809017, 1.02573], [0.262865933, 0.809017, -1.02573], [0.262865933, 0.809017, 1.02573], [-0.951057067, -1.30902, -0.5], [-0.951057067, -1.30902, 0.5], [-0.951057067, 1.30902, -0.5], [-0.951057067, 1.30902, 0.5], [0.951056933, -1.30902, -0.5], [0.951056933, -1.30902, 0.5], [0.951056933, 1.30902, -0.5], [0.951056933, 1.30902, 0.5], [-0.688191067, -0.5, -1.02573], [-0.688191067, -0.5, 1.02573], [-0.688191067, 0.5, -1.02573], [-0.688191067, 0.5, 1.02573], [-1.538840067, -0.5, -0.5], [-1.538840067, -0.5, 0.5], [-1.538840067, 0.5, -0.5], [-1.538840067, 0.5, 0.5], [1.538839933, -0.5, -0.5], [1.538839933, -0.5, 0.5], [1.538839933, 0.5, -0.5], [1.538839933, 0.5, 0.5]]}, "42": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Orthobirotunda", "ShortName": "J42", "vertices": [[5.5e-07, -1.61803, -0.5], [5.5e-07, -1.61803, 0.5], [5.5e-07, 1.61803, -0.5], [5.5e-07, 1.61803, 0.5], [-0.95105645, -1.30902, -0.5], [-0.95105645, -1.30902, 0.5], [-0.95105645, 1.30902, -0.5], [-0.95105645, 1.30902, 0.5], [0.95105755, -1.30902, -0.5], [0.95105755, -1.30902, 0.5], [0.95105755, 1.30902, -0.5], [0.95105755, 1.30902, 0.5], [-1.53883945, -0.5, -0.5], [-1.53883945, -0.5, 0.5], [-1.53883945, 0.5, -0.5], [-1.53883945, 0.5, 0.5], [1.53884055, -0.5, -0.5], [1.53884055, -0.5, 0.5], [1.53884055, 0.5, -0.5], [1.53884055, 0.5, 0.5], [1.37638055, 0.0, -1.35065], [1.37638055, 0.0, 1.35065], [0.42532555, -1.30902, -1.35065], [0.42532555, -1.30902, 1.35065], [0.42532555, 1.30902, -1.35065], [0.42532555, 1.30902, 1.35065], [-1.11351945, -0.809017, -1.35065], [-1.11351945, -0.809017, 1.35065], [-1.11351945, 0.809017, -1.35065], [-1.11351945, 0.809017, 1.35065], [-0.85065045, 0.0, -1.87638], [-0.85065045, 0.0, 1.87638], [-0.26286545, -0.809017, -1.87638], [-0.26286545, -0.809017, 1.87638], [-0.26286545, 0.809017, -1.87638], [-0.26286545, 0.809017, 1.87638], [0.68819155, -0.5, -1.87638], [0.68819155, -0.5, 1.87638], [0.68819155, 0.5, -1.87638], [0.68819155, 0.5, 1.87638]]}, "43": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Orthocupolarotunda", "ShortName": "J40", "vertices": [[3.43e-07, -1.61803, -0.814471429], [3.43e-07, -1.61803, 0.185528571], [3.43e-07, 1.61803, -0.814471429], [3.43e-07, 1.61803, 0.185528571], [0.425325343, -1.30902, 1.036178571], [0.425325343, 1.30902, 1.036178571], [0.688191343, -0.5, -1.340201429], [0.688191343, -0.5, 1.561908571], [0.688191343, 0.5, -1.340201429], [0.688191343, 0.5, 1.561908571], [-0.850650657, 0.0, -1.340201429], [-0.850650657, 0.0, 1.561908571], [-1.113519657, -0.809017, 1.036178571], [-1.113519657, 0.809017, 1.036178571], [-0.262865657, -0.809017, -1.340201429], [-0.262865657, -0.809017, 1.561908571], [-0.262865657, 0.809017, -1.340201429], [-0.262865657, 0.809017, 1.561908571], [-0.951056657, -1.30902, -0.814471429], [-0.951056657, -1.30902, 0.185528571], [-0.951056657, 1.30902, -0.814471429], [-0.951056657, 1.30902, 0.185528571], [0.951057343, -1.30902, -0.814471429], [0.951057343, -1.30902, 0.185528571], [0.951057343, 1.30902, -0.814471429], [0.951057343, 1.30902, 0.185528571], [-1.538839657, -0.5, -0.814471429], [-1.538839657, -0.5, 0.185528571], [-1.538839657, 0.5, -0.814471429], [-1.538839657, 0.5, 0.185528571], [1.538840343, -0.5, -0.814471429], [1.538840343, -0.5, 0.185528571], [1.538840343, 0.5, -0.814471429], [1.538840343, 0.5, 0.185528571], [1.376380343, 0.0, 1.036178571]]}, "44": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Pyramid", "ShortName": "J09", "vertices": [[-1.82e-07, 0.0, 0.932481818], [0.850650818, 0.0, -0.593248182], [0.850650818, 0.0, 0.406751818], [0.262865818, -0.809017, -0.593248182], [0.262865818, -0.809017, 0.406751818], [0.262865818, 0.809017, -0.593248182], [0.262865818, 0.809017, 0.406751818], [-0.688191182, -0.5, -0.593248182], [-0.688191182, -0.5, 0.406751818], [-0.688191182, 0.5, -0.593248182], [-0.688191182, 0.5, 0.406751818]]}, "45": {"AlternativeName": null, "Comment": null, "Name": "Elongated Pentagonal Rotunda", "ShortName": "J21", "vertices": [[3.67e-07, -1.61803, -1.0378385], [3.67e-07, -1.61803, -0.0378385], [3.67e-07, 1.61803, -1.0378385], [3.67e-07, 1.61803, -0.0378385], [-0.951056633, -1.30902, -1.0378385], [-0.951056633, -1.30902, -0.0378385], [-0.951056633, 1.30902, -1.0378385], [-0.951056633, 1.30902, -0.0378385], [0.951057367, -1.30902, -1.0378385], [0.951057367, -1.30902, -0.0378385], [0.951057367, 1.30902, -1.0378385], [0.951057367, 1.30902, -0.0378385], [-1.538839633, -0.5, -1.0378385], [-1.538839633, -0.5, -0.0378385], [-1.538839633, 0.5, -1.0378385], [-1.538839633, 0.5, -0.0378385], [1.538840367, -0.5, -1.0378385], [1.538840367, -0.5, -0.0378385], [1.538840367, 0.5, -1.0378385], [1.538840367, 0.5, -0.0378385], [1.376380367, 0.0, 0.8128125], [0.425325367, -1.30902, 0.8128125], [0.425325367, 1.30902, 0.8128125], [-1.113519633, -0.809017, 0.8128125], [-1.113519633, 0.809017, 0.8128125], [-0.850650633, 0.0, 1.3385415], [-0.262865633, -0.809017, 1.3385415], [-0.262865633, 0.809017, 1.3385415], [0.688191367, -0.5, 1.3385415], [0.688191367, 0.5, 1.3385415]]}, "46": {"AlternativeName": null, "Comment": null, "Name": "Elongated Square Cupola", "ShortName": "J19", "vertices": [[-0.0, -0.707107, 0.9656856], [-0.0, 0.707107, 0.9656856], [-0.707107, 0.0, 0.9656856], [0.707107, 0.0, 0.9656856], [0.5, 1.20711, -0.7414214], [0.5, 1.20711, 0.2585786], [-0.5, 1.20711, -0.7414214], [-0.5, 1.20711, 0.2585786], [-1.20711, 0.5, -0.7414214], [-1.20711, 0.5, 0.2585786], [-1.20711, -0.5, -0.7414214], [-1.20711, -0.5, 0.2585786], [-0.5, -1.20711, -0.7414214], [-0.5, -1.20711, 0.2585786], [0.5, -1.20711, -0.7414214], [0.5, -1.20711, 0.2585786], [1.20711, -0.5, -0.7414214], [1.20711, -0.5, 0.2585786], [1.20711, 0.5, -0.7414214], [1.20711, 0.5, 0.2585786]]}, "47": {"AlternativeName": null, "Comment": null, "Name": "Elongated Square Dipyramid", "ShortName": "J15", "vertices": [[0.0, 0.0, -1.20711], [0.0, 0.0, 1.20711], [0.0, -0.707107, -0.5], [0.0, -0.707107, 0.5], [0.0, 0.707107, -0.5], [0.0, 0.707107, 0.5], [-0.707107, 0.0, -0.5], [-0.707107, 0.0, 0.5], [0.707107, 0.0, -0.5], [0.707107, 0.0, 0.5]]}, "48": {"AlternativeName": null, "Comment": null, "Name": "Elongated Square Gyrobicupola", "ShortName": "J37", "vertices": [[-0.5, -0.5, -1.20711], [-0.5, 0.5, -1.20711], [-0.0, -0.707107, 1.20711], [-0.0, 0.707107, 1.20711], [0.5, -0.5, -1.20711], [0.5, 0.5, -1.20711], [-0.707107, 0.0, 1.20711], [0.707107, 0.0, 1.20711], [0.5, 1.20711, -0.5], [0.5, 1.20711, 0.5], [-0.5, 1.20711, -0.5], [-0.5, 1.20711, 0.5], [-1.20711, 0.5, -0.5], [-1.20711, 0.5, 0.5], [-1.20711, -0.5, -0.5], [-1.20711, -0.5, 0.5], [-0.5, -1.20711, -0.5], [-0.5, -1.20711, 0.5], [0.5, -1.20711, -0.5], [0.5, -1.20711, 0.5], [1.20711, -0.5, -0.5], [1.20711, -0.5, 0.5], [1.20711, 0.5, -0.5], [1.20711, 0.5, 0.5]]}, "49": {"AlternativeName": null, "Comment": null, "Name": "Elongated Square Pyramid", "ShortName": "J08", "vertices": [[0.0, 0.0, 1.072986667], [0.0, -0.707107, -0.634123333], [0.0, -0.707107, 0.365876667], [0.0, 0.707107, -0.634123333], [0.0, 0.707107, 0.365876667], [-0.707107, 0.0, -0.634123333], [-0.707107, 0.0, 0.365876667], [0.707107, 0.0, -0.634123333], [0.707107, 0.0, 0.365876667]]}, "50": {"AlternativeName": null, "Comment": null, "Name": "Elongated Triangular Cupola", "ShortName": "J18", "vertices": [[0.0, -1.0, -0.7632994], [0.0, -1.0, 0.2367006], [0.0, 1.0, -0.7632994], [0.0, 1.0, 0.2367006], [-0.288675, -0.5, 1.0531976], [-0.288675, 0.5, 1.0531976], [0.57735, 0.0, 1.0531976], [-0.866025, -0.5, -0.7632994], [-0.866025, -0.5, 0.2367006], [-0.866025, 0.5, -0.7632994], [-0.866025, 0.5, 0.2367006], [0.866025, -0.5, -0.7632994], [0.866025, -0.5, 0.2367006], [0.866025, 0.5, -0.7632994], [0.866025, 0.5, 0.2367006]]}, "51": {"AlternativeName": null, "Comment": null, "Name": "Elongated Triangular Dipyramid", "ShortName": "J14", "vertices": [[0.0, 0.0, -1.3165], [0.0, 0.0, 1.3165], [-0.288675, -0.5, -0.5], [-0.288675, -0.5, 0.5], [-0.288675, 0.5, -0.5], [-0.288675, 0.5, 0.5], [0.57735, 0.0, -0.5], [0.57735, 0.0, 0.5]]}, "52": {"AlternativeName": null, "Comment": null, "Name": "Elongated Triangular Gyrobicupola", "ShortName": "J36", "vertices": [[0.0, -1.0, -0.5], [0.0, -1.0, 0.5], [0.0, 1.0, -0.5], [0.0, 1.0, 0.5], [-0.288675, -0.5, 1.3165], [-0.288675, 0.5, 1.3165], [0.288675, -0.5, -1.3165], [0.57735, 0.0, 1.3165], [-0.866025, -0.5, -0.5], [-0.866025, -0.5, 0.5], [-0.866025, 0.5, -0.5], [-0.866025, 0.5, 0.5], [0.866025, -0.5, -0.5], [0.866025, -0.5, 0.5], [0.866025, 0.5, -0.5], [0.866025, 0.5, 0.5], [-0.57735, 0.0, -1.3165], [0.288675, 0.5, -1.3165]]}, "53": {"AlternativeName": null, "Comment": null, "Name": "Elongated Triangular Orthobicupola", "ShortName": "J35", "vertices": [[0.0, -1.0, -0.5], [0.0, -1.0, 0.5], [0.0, 1.0, -0.5], [0.0, 1.0, 0.5], [-0.288675, -0.5, -1.3165], [-0.288675, -0.5, 1.3165], [-0.288675, 0.5, -1.3165], [-0.288675, 0.5, 1.3165], [0.57735, 0.0, -1.3165], [0.57735, 0.0, 1.3165], [-0.866025, -0.5, -0.5], [-0.866025, -0.5, 0.5], [-0.866025, 0.5, -0.5], [-0.866025, 0.5, 0.5], [0.866025, -0.5, -0.5], [0.866025, -0.5, 0.5], [0.866025, 0.5, -0.5], [0.866025, 0.5, 0.5]]}, "54": {"AlternativeName": null, "Comment": null, "Name": "Elongated Triangular Pyramid", "ShortName": "J07", "vertices": [[0.0, 0.0, 1.128428571], [-0.288675, -0.5, -0.688071429], [-0.288675, -0.5, 0.311928571], [-0.288675, 0.5, -0.688071429], [-0.288675, 0.5, 0.311928571], [0.57735, 0.0, -0.688071429], [0.57735, 0.0, 0.311928571]]}, "55": {"AlternativeName": null, "Comment": null, "Name": "Truncated Icosidodecahedron", "ShortName": "A11", "vertices": [[-1.0, -1.30902, -3.42705], [-1.0, -1.30902, 3.42705], [-1.0, 1.30902, -3.42705], [-1.0, 1.30902, 3.42705], [-0.5, -0.5, -3.73607], [-0.5, -0.5, 3.73607], [-0.5, 0.5, -3.73607], [-0.5, 0.5, 3.73607], [-0.5, -3.73607, -0.5], [-0.5, -3.73607, 0.5], [-0.5, -2.11803, -3.11803], [-0.5, -2.11803, 3.11803], [-0.5, 3.73607, -0.5], [-0.5, 3.73607, 0.5], [-0.5, 2.11803, -3.11803], [-0.5, 2.11803, 3.11803], [0.5, -0.5, -3.73607], [0.5, -0.5, 3.73607], [0.5, 0.5, -3.73607], [0.5, 0.5, 3.73607], [0.5, -3.73607, -0.5], [0.5, -3.73607, 0.5], [0.5, -2.11803, -3.11803], [0.5, -2.11803, 3.11803], [0.5, 3.73607, -0.5], [0.5, 3.73607, 0.5], [0.5, 2.11803, -3.11803], [0.5, 2.11803, 3.11803], [1.0, -1.30902, -3.42705], [1.0, -1.30902, 3.42705], [1.0, 1.30902, -3.42705], [1.0, 1.30902, 3.42705], [-3.42705, -1.0, -1.30902], [-3.42705, -1.0, 1.30902], [-3.42705, 1.0, -1.30902], [-3.42705, 1.0, 1.30902], [-2.92705, -1.80902, -1.61803], [-2.92705, -1.80902, 1.61803], [-2.92705, 1.80902, -1.61803], [-2.92705, 1.80902, 1.61803], [-1.80902, -1.61803, -2.92705], [-1.80902, -1.61803, 2.92705], [-1.80902, 1.61803, -2.92705], [-1.80902, 1.61803, 2.92705], [-1.30902, -3.42705, -1.0], [-1.30902, -3.42705, 1.0], [-1.30902, -2.42705, -2.61803], [-1.30902, -2.42705, 2.61803], [-1.30902, 2.42705, -2.61803], [-1.30902, 2.42705, 2.61803], [-1.30902, 3.42705, -1.0], [-1.30902, 3.42705, 1.0], [-2.61803, -1.30902, -2.42705], [-2.61803, -1.30902, 2.42705], [-2.61803, 1.30902, -2.42705], [-2.61803, 1.30902, 2.42705], [-3.73607, -0.5, -0.5], [-3.73607, -0.5, 0.5], [-3.73607, 0.5, -0.5], [-3.73607, 0.5, 0.5], [-1.61803, -2.92705, -1.80902], [-1.61803, -2.92705, 1.80902], [-1.61803, 2.92705, -1.80902], [-1.61803, 2.92705, 1.80902], [-3.11803, -0.5, -2.11803], [-3.11803, -0.5, 2.11803], [-3.11803, 0.5, -2.11803], [-3.11803, 0.5, 2.11803], [-2.11803, -3.11803, -0.5], [-2.11803, -3.11803, 0.5], [-2.11803, 3.11803, -0.5], [-2.11803, 3.11803, 0.5], [-2.42705, -2.61803, -1.30902], [-2.42705, -2.61803, 1.30902], [-2.42705, 2.61803, -1.30902], [-2.42705, 2.61803, 1.30902], [1.61803, -2.92705, -1.80902], [1.61803, -2.92705, 1.80902], [1.61803, 2.92705, -1.80902], [1.61803, 2.92705, 1.80902], [2.42705, -2.61803, -1.30902], [2.42705, -2.61803, 1.30902], [2.42705, 2.61803, -1.30902], [2.42705, 2.61803, 1.30902], [3.73607, -0.5, -0.5], [3.73607, -0.5, 0.5], [3.73607, 0.5, -0.5], [3.73607, 0.5, 0.5], [2.11803, -3.11803, -0.5], [2.11803, -3.11803, 0.5], [2.11803, 3.11803, -0.5], [2.11803, 3.11803, 0.5], [1.30902, -3.42705, -1.0], [1.30902, -3.42705, 1.0], [1.30902, -2.42705, -2.61803], [1.30902, -2.42705, 2.61803], [1.30902, 2.42705, -2.61803], [1.30902, 2.42705, 2.61803], [1.30902, 3.42705, -1.0], [1.30902, 3.42705, 1.0], [2.61803, -1.30902, -2.42705], [2.61803, -1.30902, 2.42705], [2.61803, 1.30902, -2.42705], [2.61803, 1.30902, 2.42705], [3.11803, -0.5, -2.11803], [3.11803, -0.5, 2.11803], [3.11803, 0.5, -2.11803], [3.11803, 0.5, 2.11803], [1.80902, -1.61803, -2.92705], [1.80902, -1.61803, 2.92705], [1.80902, 1.61803, -2.92705], [1.80902, 1.61803, 2.92705], [2.92705, -1.80902, -1.61803], [2.92705, -1.80902, 1.61803], [2.92705, 1.80902, -1.61803], [2.92705, 1.80902, 1.61803], [3.42705, -1.0, -1.30902], [3.42705, -1.0, 1.30902], [3.42705, 1.0, -1.30902], [3.42705, 1.0, 1.30902]]}, "56": {"AlternativeName": null, "Comment": null, "Name": "Truncated Cuboctahedron", "ShortName": "A10", "vertices": [[-0.5, 1.20711, -1.91421], [-0.5, 1.20711, 1.91421], [-0.5, -1.20711, -1.91421], [-0.5, -1.20711, 1.91421], [-0.5, -1.91421, 1.20711], [-0.5, -1.91421, -1.20711], [-0.5, 1.91421, 1.20711], [-0.5, 1.91421, -1.20711], [0.5, 1.20711, -1.91421], [0.5, 1.20711, 1.91421], [0.5, -1.20711, -1.91421], [0.5, -1.20711, 1.91421], [0.5, -1.91421, 1.20711], [0.5, -1.91421, -1.20711], [0.5, 1.91421, 1.20711], [0.5, 1.91421, -1.20711], [1.20711, -0.5, -1.91421], [1.20711, -0.5, 1.91421], [1.20711, 0.5, -1.91421], [1.20711, 0.5, 1.91421], [1.20711, -1.91421, -0.5], [1.20711, -1.91421, 0.5], [1.20711, 1.91421, -0.5], [1.20711, 1.91421, 0.5], [-1.20711, -0.5, -1.91421], [-1.20711, -0.5, 1.91421], [-1.20711, 0.5, -1.91421], [-1.20711, 0.5, 1.91421], [-1.20711, -1.91421, -0.5], [-1.20711, -1.91421, 0.5], [-1.20711, 1.91421, -0.5], [-1.20711, 1.91421, 0.5], [-1.91421, -0.5, 1.20711], [-1.91421, -0.5, -1.20711], [-1.91421, 0.5, 1.20711], [-1.91421, 0.5, -1.20711], [-1.91421, 1.20711, -0.5], [-1.91421, 1.20711, 0.5], [-1.91421, -1.20711, -0.5], [-1.91421, -1.20711, 0.5], [1.91421, -0.5, 1.20711], [1.91421, -0.5, -1.20711], [1.91421, 0.5, 1.20711], [1.91421, 0.5, -1.20711], [1.91421, 1.20711, -0.5], [1.91421, 1.20711, 0.5], [1.91421, -1.20711, -0.5], [1.91421, -1.20711, 0.5]]}, "57": {"AlternativeName": null, "Comment": null, "Name": "Gyrate Bidiminished Rhombicosidodecahedron", "ShortName": "J82", "vertices": [[-2.03400905, -0.332349036, 0.010840126], [-1.90637905, 0.580113964, -0.377903174], [-1.89745905, 0.039699764, 0.928952826], [-1.76981905, 0.952161964, 0.540209826], [-1.71681905, -0.741266036, -0.844833174], [-1.68827905, -1.270655036, 0.003068226], [-1.58918905, 0.171196964, -1.233572174], [-1.46732905, -0.668664036, 1.488607826], [-1.41958905, 0.744307964, 1.453517826], [-1.35412905, 1.118204964, -1.014672174], [-1.33804905, -1.478505036, 0.916377826], [-1.13316905, 1.720194964, 0.470863826], [-1.02978905, -1.347005036, -1.246152174], [-1.00124905, -1.876395036, -0.398250174], [-0.98945905, 0.035944964, 2.013177826], [-0.87626405, 1.822814964, -0.490107174], [-0.82326705, 0.129387964, -1.875152174], [-0.78294305, 1.512344964, 1.384177826], [-0.65101505, -2.084245036, 0.515059826], [-0.58821005, 1.076394964, -1.656252174], [-0.47753805, -0.808914036, -1.882922174], [-0.23532605, -1.918205036, -1.039822174], [-0.23069505, 2.050434964, 0.747402826], [-0.11034305, 1.781004964, -1.131682174], [-0.08698205, 0.366183964, 2.289707826], [0.02621525, 2.153054964, -0.213568174], [0.04065315, 1.278644964, 1.900967826], [0.09882795, 0.470656964, -2.057562174], [0.31692195, -1.380105036, -1.676592174], [0.33135995, -2.254515036, 0.437941826], [0.44455795, -0.467645036, -2.065342174], [0.58826995, -2.151895036, -0.523029174], [0.59290095, 1.816744964, 1.264197826], [0.87203195, 1.610734964, -1.208802174], [0.89539295, 0.195914964, 2.212597826], [1.00130695, 0.800895964, -1.781032174], [1.00859095, 1.982784964, -0.290686174], [1.02302795, 1.108374964, 1.823847826], [1.14052095, -1.613805036, -1.159802174], [1.23384095, -1.924275036, 0.714480826], [1.34704095, -0.137406036, -1.788802174], [1.35882095, 1.774934964, 0.622623826], [1.49075095, -1.821655036, -0.246491174], [1.58243095, -0.409827036, 1.811277826], [1.71170095, -1.219665036, 1.239047826], [1.77716095, -0.845769036, -1.229142174], [1.78895095, 1.066564964, 1.182277826], [2.12739095, -1.053625036, -0.315837174], [2.13468095, 0.128265964, 1.174507826], [2.26395095, -0.681575036, 0.602276826]]}, "58": {"AlternativeName": null, "Comment": null, "Name": "Gyrate Rhombicosidodecahedron", "ShortName": "J72", "vertices": [[-2.180729817, 0.476838883, -0.055880833], [-2.105409817, -0.300487117, -0.680454933], [-2.049319817, 0.138432883, 0.875899067], [-1.927459817, -1.119310117, -0.134682933], [-1.892789817, -0.848038117, 0.827197067], [-1.801859817, 1.078529883, -0.759038933], [-1.767189817, 1.349799883, 0.202841067], [-1.726549817, 0.301201883, -1.383609933], [-1.635789817, 1.011389883, 1.134620067], [-1.570019817, -0.685269117, -1.432309933], [-1.457829817, 0.192570883, 1.680390067], [-1.392059817, -1.504090117, -0.886542933], [-1.301299817, -0.793900117, 1.631690067], [-1.260649817, -1.842490117, 0.045237067], [-1.225979817, -1.571230117, 1.007120067], [-1.057439817, 1.713679883, -0.964992933], [-1.022769817, 1.984939883, -0.003113083], [-0.935572817, 0.455936883, -1.975569933], [-0.810144817, 1.437389883, 1.504540067], [-0.779041817, -0.530535117, -2.024279933], [-0.632187817, 0.618573883, 2.050310067], [-0.522039817, 1.328889883, -1.716849933], [-0.491102817, -1.855410117, -1.141199933], [-0.431278817, 2.039079883, 0.801380067], [-0.378915817, -0.977571117, 1.971510067], [-0.359693817, -2.193820117, -0.209418933], [-0.303600817, -1.754900117, 1.346930067], [-0.231794817, 2.139679883, -0.595074933], [-0.112236817, -1.253720117, -1.844359933], [-0.034616117, 0.104613883, -2.230229933], [0.034616483, -0.104614117, 2.230230067], [0.167767183, 1.646339883, 1.499160067], [0.231795183, -2.139680117, 0.595075067], [0.303601183, 1.754899883, -1.346929933], [0.345724183, 0.827521883, 2.044930067], [0.359694183, 2.193819883, 0.209419067], [0.378916183, 0.977570883, -1.971509933], [0.431279183, -2.039080117, -0.801379933], [0.522040183, -1.328890117, 1.716850067], [0.632188183, -0.618574117, -2.050309933], [0.810145183, -1.437390117, -1.504539933], [0.935573183, -0.455937117, 1.975570067], [0.958740183, 1.801079883, 0.907199067], [1.022770183, -1.984940117, 0.003113207], [1.057440183, -1.713680117, 0.964993067], [1.225980183, 1.571229883, -1.007119933], [1.246680183, 0.476198883, 1.790280067], [1.260650183, 1.842489883, -0.045236933], [1.301300183, 0.793899883, -1.631689933], [1.457830183, -0.192571117, -1.680389933], [1.625550183, 1.077889883, 1.087120067], [1.635790183, -1.011390117, -1.134619933], [1.726550183, -0.301202117, 1.383610067], [1.767190183, -1.349800117, -0.202840933], [1.801860183, -1.078530117, 0.759039067], [1.892790183, 0.848037883, -0.827196933], [1.927450183, 1.119309883, 0.134683067], [2.049320183, -0.138433117, -0.875898933], [2.105410183, 0.300486883, 0.680455067], [2.180730183, -0.476839117, 0.055880967]]}, "59": {"AlternativeName": null, "Comment": null, "Name": "Gyrobifastigium", "ShortName": "J26", "vertices": [[-0.5, -0.5, 0.0], [-0.5, 0.0, 0.866025], [-0.5, 0.5, 0.0], [0.0, -0.5, -0.866025], [0.0, 0.5, -0.866025], [0.5, -0.5, 0.0], [0.5, 0.0, 0.866025], [0.5, 0.5, 0.0]]}, "60": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Bicupola", "ShortName": "J46", "vertices": [[-1.618030033, -3.3e-08, 0.431199], [-1.538840033, -0.500000033, -0.431199], [-1.538840033, 0.499999967, -0.431199], [-1.309020033, -0.951057033, 0.431199], [-1.309020033, 0.951056967, 0.431199], [-0.951057033, -1.309020033, -0.431199], [-0.951057033, 1.309019967, -0.431199], [-0.809017033, 0.262865967, 0.95693], [-0.688191033, -0.500000033, -0.95693], [-0.688191033, 0.499999967, -0.95693], [-0.500000033, -1.538840033, 0.431199], [-0.500000033, -0.688191033, 0.95693], [-0.500000033, 1.538839967, 0.431199], [-3.3e-08, -1.618030033, -0.431199], [-3.3e-08, 0.850650967, 0.95693], [-3.3e-08, 1.618029967, -0.431199], [0.262865967, -0.809017033, -0.95693], [0.262865967, 0.809016967, -0.95693], [0.499999967, -1.538840033, 0.431199], [0.499999967, -0.688191033, 0.95693], [0.499999967, 1.538839967, 0.431199], [0.809016967, 0.262865967, 0.95693], [0.850650967, -3.3e-08, -0.95693], [0.951056967, -1.309020033, -0.431199], [0.951056967, 1.309019967, -0.431199], [1.309019967, -0.951057033, 0.431199], [1.309019967, 0.951056967, 0.431199], [1.538839967, -0.500000033, -0.431199], [1.538839967, 0.499999967, -0.431199], [1.618029967, -3.3e-08, 0.431199]]}, "61": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Birotunda", "ShortName": "J48", "vertices": [[0.0, -1.618033989, -0.431198502], [0.0, 1.618033989, -0.431198502], [0.0, -1.37638192, 1.28184931], [0.0, 0.850650808, 1.807580422], [-1.618033989, -0.0, 0.431198502], [-0.5, -1.538841769, 0.431198502], [1.618033989, -0.0, 0.431198502], [-0.951056516, -1.309016994, -0.431198502], [-0.951056516, 1.309016994, -0.431198502], [0.951056516, -1.309016994, -0.431198502], [0.951056516, 1.309016994, -0.431198502], [-0.5, 1.538841769, 0.431198502], [0.5, -1.538841769, 0.431198502], [0.5, 1.538841769, 0.431198502], [-1.309016994, -0.951056516, 0.431198502], [-1.309016994, 0.951056516, 0.431198502], [1.309016994, -0.951056516, 0.431198502], [1.309016994, 0.951056516, 0.431198502], [-1.538841769, -0.5, -0.431198502], [-1.538841769, 0.5, -0.431198502], [1.538841769, -0.5, -0.431198502], [1.538841769, 0.5, -0.431198502], [1.37638192, -0.0, -1.28184931], [0.425325404, -1.309016994, -1.28184931], [0.425325404, 1.309016994, -1.28184931], [-1.113516364, -0.809016994, -1.28184931], [-1.113516364, 0.809016994, -1.28184931], [1.309016994, -0.425325404, 1.28184931], [0.809016994, 1.113516364, 1.28184931], [-0.850650808, -0.0, -1.807580422], [-0.262865556, -0.809016994, -1.807580422], [-0.262865556, 0.809016994, -1.807580422], [0.68819096, -0.5, -1.807580422], [0.68819096, 0.5, -1.807580422], [-0.809016994, 0.262865556, 1.807580422], [-0.5, -0.68819096, 1.807580422], [-1.309016994, -0.425325404, 1.28184931], [-0.809016994, 1.113516364, 1.28184931], [0.5, -0.68819096, 1.807580422], [0.809016994, 0.262865556, 1.807580422]]}, "62": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Cupola", "ShortName": "J24", "vertices": [[-4e-08, -1.61803, 0.2398126], [-4e-08, 1.61803, 0.2398126], [-1.61803004, 0.0, -0.6225844], [0.85065096, 0.0, 0.7655436], [0.26286596, -0.809017, 0.7655436], [0.26286596, 0.809017, 0.7655436], [1.61802996, 0.0, -0.6225844], [-0.95105704, -1.30902, 0.2398126], [-0.95105704, 1.30902, 0.2398126], [0.95105696, -1.30902, 0.2398126], [0.95105696, 1.30902, 0.2398126], [-0.68819104, -0.5, 0.7655436], [-0.68819104, 0.5, 0.7655436], [-0.50000004, -1.53884, -0.6225844], [-0.50000004, 1.53884, -0.6225844], [0.49999996, -1.53884, -0.6225844], [0.49999996, 1.53884, -0.6225844], [-1.30902004, -0.951057, -0.6225844], [-1.30902004, 0.951057, -0.6225844], [1.30901996, -0.951057, -0.6225844], [1.30901996, 0.951057, -0.6225844], [-1.53884004, -0.5, 0.2398126], [-1.53884004, 0.5, 0.2398126], [1.53883996, -0.5, 0.2398126], [1.53883996, 0.5, 0.2398126]]}, "63": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Cupolarotunda", "ShortName": "J47", "vertices": [[-2.9e-08, -1.618030314, -0.735841857], [-2.9e-08, 1.618029686, -0.735841857], [-2.9e-08, -1.376380314, 0.977207143], [-2.9e-08, 0.850650686, 1.502937143], [-1.618030029, -3.14e-07, 0.126556143], [0.850650971, -3.14e-07, -1.261572857], [-0.500000029, -1.538840314, 0.126556143], [0.262865971, -0.809017314, -1.261572857], [0.262865971, 0.809016686, -1.261572857], [1.618029971, -3.14e-07, 0.126556143], [-0.951057029, -1.309020314, -0.735841857], [-0.951057029, 1.309019686, -0.735841857], [0.951056971, -1.309020314, -0.735841857], [0.951056971, 1.309019686, -0.735841857], [-0.688191029, -0.500000314, -1.261572857], [-0.688191029, 0.499999686, -1.261572857], [-0.500000029, 1.538839686, 0.126556143], [0.499999971, -1.538840314, 0.126556143], [0.499999971, 1.538839686, 0.126556143], [-1.309020029, -0.951057314, 0.126556143], [-1.309020029, 0.951056686, 0.126556143], [1.309019971, -0.951057314, 0.126556143], [1.309019971, 0.951056686, 0.126556143], [-1.538840029, -0.500000314, -0.735841857], [-1.538840029, 0.499999686, -0.735841857], [1.538839971, -0.500000314, -0.735841857], [1.538839971, 0.499999686, -0.735841857], [1.309019971, -0.425325314, 0.977207143], [0.809016971, 1.113519686, 0.977207143], [-0.809017029, 0.262865686, 1.502937143], [-0.500000029, -0.688191314, 1.502937143], [-1.309020029, -0.425325314, 0.977207143], [-0.809017029, 1.113519686, 0.977207143], [0.499999971, -0.688191314, 1.502937143], [0.809016971, 0.262865686, 1.502937143]]}, "64": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Pyramid", "ShortName": "J11", "vertices": [[0.0, 0.0, 0.864597273], [-0.850651, 0.0, -0.511784727], [0.850651, 0.0, 0.338865273], [-0.262866, -0.809017, -0.511784727], [-0.262866, 0.809017, -0.511784727], [0.262866, -0.809017, 0.338865273], [0.262866, 0.809017, 0.338865273], [-0.688191, -0.5, 0.338865273], [-0.688191, 0.5, 0.338865273], [0.688191, -0.5, -0.511784727], [0.688191, 0.5, -0.511784727]]}, "65": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Pentagonal Rotunda", "ShortName": "J25", "vertices": [[3.67e-07, -1.61803, -0.083706167], [3.67e-07, 1.61803, -0.083706167], [-1.618029633, 0.0, -0.946103167], [1.618030367, 0.0, -0.946103167], [-0.951056633, -1.30902, -0.083706167], [-0.951056633, 1.30902, -0.083706167], [0.951057367, -1.30902, -0.083706167], [0.951057367, 1.30902, -0.083706167], [-0.499999633, -1.53884, -0.946103167], [-0.499999633, 1.53884, -0.946103167], [0.500000367, -1.53884, -0.946103167], [0.500000367, 1.53884, -0.946103167], [-1.309019633, -0.951057, -0.946103167], [-1.309019633, 0.951057, -0.946103167], [1.309020367, -0.951057, -0.946103167], [1.309020367, 0.951057, -0.946103167], [-1.538839633, -0.5, -0.083706167], [-1.538839633, 0.5, -0.083706167], [1.538840367, -0.5, -0.083706167], [1.538840367, 0.5, -0.083706167], [1.376380367, 0.0, 0.766944833], [0.425325367, -1.30902, 0.766944833], [0.425325367, 1.30902, 0.766944833], [-1.113519633, -0.809017, 0.766944833], [-1.113519633, 0.809017, 0.766944833], [-0.850650633, 0.0, 1.292673833], [-0.262865633, -0.809017, 1.292673833], [-0.262865633, 0.809017, 1.292673833], [0.688191367, -0.5, 1.292673833], [0.688191367, 0.5, 1.292673833]]}, "66": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Square Bicupola", "ShortName": "J45", "vertices": [[0.0, -0.707107, -1.13725], [0.0, 0.707107, -1.13725], [0.0, -1.30656, 0.430148], [0.0, 1.30656, 0.430148], [-0.707107, 0.0, -1.13725], [0.707107, 0.0, -1.13725], [-0.653281, 0.270598, 1.13725], [0.653281, -0.270598, 1.13725], [1.20711, 0.5, -0.430148], [-1.30656, 0.0, 0.430148], [1.30656, 0.0, 0.430148], [-0.92388, -0.92388, 0.430148], [-0.92388, 0.92388, 0.430148], [0.92388, -0.92388, 0.430148], [0.92388, 0.92388, 0.430148], [-0.5, -1.20711, -0.430148], [0.5, 1.20711, -0.430148], [0.5, -1.20711, -0.430148], [-0.5, 1.20711, -0.430148], [1.20711, -0.5, -0.430148], [-1.20711, 0.5, -0.430148], [-1.20711, -0.5, -0.430148], [-0.270598, -0.653281, 1.13725], [0.270598, 0.653281, 1.13725]]}, "67": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Square Cupola", "ShortName": "J23", "vertices": [[-0.0, -0.707107, 0.909804], [-0.0, 0.707107, 0.909804], [-0.0, -1.30656, -0.657599], [-0.0, 1.30656, -0.657599], [-0.707107, 0.0, 0.909804], [0.707107, 0.0, 0.909804], [-1.20711, -0.5, 0.202697], [-1.30656, 0.0, -0.657599], [1.30656, 0.0, -0.657599], [-0.92388, -0.92388, -0.657599], [-0.92388, 0.92388, -0.657599], [0.92388, -0.92388, -0.657599], [0.92388, 0.92388, -0.657599], [-0.5, -1.20711, 0.202697], [0.5, 1.20711, 0.202697], [0.5, -1.20711, 0.202697], [-0.5, 1.20711, 0.202697], [1.20711, -0.5, 0.202697], [-1.20711, 0.5, 0.202697], [1.20711, 0.5, 0.202697]]}, "68": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Square Dipyramid", "ShortName": "J17", "vertices": [[-0.5, -0.5, -0.420448], [-0.5, 0.5, -0.420448], [0.0, 0.0, -1.12755], [0.0, 0.0, 1.12755], [0.0, -0.707107, 0.420448], [0.0, 0.707107, 0.420448], [0.5, -0.5, -0.420448], [0.5, 0.5, -0.420448], [-0.707107, 0.0, 0.420448], [0.707107, 0.0, 0.420448]]}, "69": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Square Pyramid", "ShortName": "J10", "vertices": [[-0.5, -0.5, -0.545731333], [-0.5, 0.5, -0.545731333], [0.0, 0.0, 1.002266667], [0.0, -0.707107, 0.295164667], [0.0, 0.707107, 0.295164667], [0.5, -0.5, -0.545731333], [0.5, 0.5, -0.545731333], [-0.707107, 0.0, 0.295164667], [0.707107, 0.0, 0.295164667]]}, "70": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Triangular Bicupola", "ShortName": "J44", "vertices": [[-1.0, 0.0, 0.4278], [-0.5, -0.866025, 0.4278], [-0.5, 0.866025, 0.4278], [-0.5, -0.288675, 1.2443], [0.0, -1.0, -0.4278], [0.0, 1.0, -0.4278], [0.0, 0.57735, 1.2443], [0.5, -0.288675, 1.2443], [0.5, -0.866025, 0.4278], [0.5, 0.866025, 0.4278], [1.0, 0.0, 0.4278], [-0.57735, 0.0, -1.2443], [0.288675, -0.5, -1.2443], [0.288675, 0.5, -1.2443], [-0.866025, -0.5, -0.4278], [-0.866025, 0.5, -0.4278], [0.866025, -0.5, -0.4278], [0.866025, 0.5, -0.4278]]}, "71": {"AlternativeName": null, "Comment": null, "Name": "Gyroelongated Triangular Cupola", "ShortName": "J22", "vertices": [[-1.0, 0.0, -0.6766594], [-0.5, -0.866025, -0.6766594], [-0.5, 0.866025, -0.6766594], [0.0, -1.0, 0.1789406], [0.0, 1.0, 0.1789406], [0.5, -0.866025, -0.6766594], [0.5, 0.866025, -0.6766594], [1.0, 0.0, -0.6766594], [-0.288675, -0.5, 0.9954376], [-0.288675, 0.5, 0.9954376], [0.57735, 0.0, 0.9954376], [-0.866025, -0.5, 0.1789406], [-0.866025, 0.5, 0.1789406], [0.866025, -0.5, 0.1789406], [0.866025, 0.5, 0.1789406]]}, "72": {"AlternativeName": null, "Comment": null, "Name": "Hebesphenomegacorona", "ShortName": "J89", "vertices": [[-0.871243429, 0.139987014, -0.012209536], [-0.733789429, -0.845001986, -0.116605036], [-0.661500429, -0.423077986, 0.787138964], [-0.603870429, -0.236300986, -0.899294036], [-0.467798429, 0.557870014, 0.801796964], [-0.180802429, 0.856974014, -0.108242036], [0.086570571, 0.480685014, -0.995326036], [0.110277571, -0.890736986, -0.650886036], [0.152619571, -0.798265986, 0.343928964], [0.246344571, -0.096561286, 1.050203964], [0.492445571, 0.775233014, 0.626638964], [0.786968571, 0.734226014, -0.328124036], [0.800718571, -0.173749986, -0.746918036], [0.843059571, -0.081279886, 0.247896964]]}, "73": {"AlternativeName": null, "Comment": null, "Name": "Icosahedron", "ShortName": "P04", "vertices": [[0.0, 0.0, -0.951057], [0.0, 0.0, 0.951057], [-0.850651, 0.0, -0.425325], [0.850651, 0.0, 0.425325], [0.688191, -0.5, -0.425325], [0.688191, 0.5, -0.425325], [-0.688191, -0.5, 0.425325], [-0.688191, 0.5, 0.425325], [-0.262866, -0.809017, -0.425325], [-0.262866, 0.809017, -0.425325], [0.262866, -0.809017, 0.425325], [0.262866, 0.809017, 0.425325]]}, "74": {"AlternativeName": null, "Comment": null, "Name": "Icosidodecahedron", "ShortName": "A02", "vertices": [[-0.0, -1.61803, 0.0], [-0.0, 1.61803, 0.0], [0.262866, -0.809017, -1.37638], [0.262866, 0.809017, -1.37638], [0.425325, -1.30902, 0.850651], [0.425325, 1.30902, 0.850651], [0.688191, -0.5, 1.37638], [0.688191, 0.5, 1.37638], [1.11352, -0.809017, -0.850651], [1.11352, 0.809017, -0.850651], [-1.37638, 0.0, -0.850651], [-0.688191, -0.5, -1.37638], [-0.688191, 0.5, -1.37638], [1.37638, 0.0, 0.850651], [0.951057, -1.30902, 0.0], [0.951057, 1.30902, 0.0], [0.850651, 0.0, -1.37638], [-0.951057, -1.30902, 0.0], [-0.951057, 1.30902, 0.0], [-1.53884, -0.5, 0.0], [-1.53884, 0.5, 0.0], [1.53884, -0.5, 0.0], [1.53884, 0.5, 0.0], [-0.850651, 0.0, 1.37638], [-1.11352, -0.809017, 0.850651], [-1.11352, 0.809017, 0.850651], [-0.425325, -1.30902, -0.850651], [-0.425325, 1.30902, -0.850651], [-0.262866, -0.809017, 1.37638], [-0.262866, 0.809017, 1.37638]]}, "75": {"AlternativeName": null, "Comment": null, "Name": "Metabiaugmented Dodecahedron", "ShortName": "J60", "vertices": [[0.008638727, 0.063383182, 1.356026818], [0.008638727, 0.063383182, -1.446493182], [0.187049727, -1.245636818, 0.421852818], [0.187049727, 1.372403182, 0.421852818], [0.475724727, -0.745633818, -1.089673182], [0.475724727, 0.872400182, -1.089673182], [1.053078727, -0.745633818, 0.421852818], [1.053078727, 0.872400182, 0.421852818], [-1.214211273, -0.436616818, 0.421852818], [-1.214211273, 0.563383182, 0.421852818], [1.231488727, -0.436616818, -0.512319182], [1.231488727, 0.563383182, -0.512319182], [-0.925533273, 0.063383182, -1.089673182], [-0.458447273, -0.745633818, 0.999206818], [-0.458447273, 0.872400182, 0.999206818], [0.942810727, 0.063383182, 0.999206818], [-1.035801273, -0.745633818, -0.512319182], [-1.035801273, 0.872400182, -0.512319182], [-0.986486273, 0.063383182, 1.257406818], [-0.169772273, -1.245636818, -0.512319182], [-0.169772273, 1.372403182, -0.512319182], [0.813711727, -1.331046818, -0.352743182]]}, "76": {"AlternativeName": null, "Comment": null, "Name": "Metabiaugmented Hexagonal Prism", "ShortName": "J56", "vertices": [[-1.0, -0.112366571, -0.5], [-1.0, -0.112366571, 0.5], [-0.5, -0.978391571, -0.5], [-0.5, -0.978391571, 0.5], [-0.5, 0.753658429, -0.5], [-0.5, 0.753658429, 0.5], [0.5, -0.978391571, -0.5], [0.5, -0.978391571, 0.5], [0.5, 0.753658429, -0.5], [0.5, 0.753658429, 0.5], [1.0, -0.112366571, -0.5], [1.0, -0.112366571, 0.5], [-1.36237, 0.674199429, 0.0], [1.36237, 0.674199429, 0.0]]}, "77": {"AlternativeName": null, "Comment": null, "Name": "Metabiaugmented Truncated Dodecahedron", "ShortName": "J70", "vertices": [[-2.952571157, -0.2730359, -0.305566317], [-2.943501157, -0.3283119, 0.692863683], [-2.712911157, 0.6964551, -0.254068317], [-2.703851157, 0.6411791, 0.744360683], [-2.663851157, -1.1596999, -0.666779317], [-2.649181157, -1.2491299, 0.948714683], [-2.540081157, 0.1826311, -1.094377317], [-2.516341157, 0.0379168, 1.519552683], [-2.476341157, -1.7629599, 0.108408483], [-2.418931157, 1.4839301, 0.287644683], [-2.251361157, -0.7040289, -1.455587317], [-2.222021157, -0.8829059, 1.775402683], [-1.966441157, 0.1387231, -1.912307317], [-1.957041157, -1.6248499, -1.199737317], [-1.933301157, -1.7695699, 1.414182683], [-1.928031157, -0.0954302, 2.317112683], [-1.770411157, 2.2442701, 0.323848683], [-1.769531157, -2.2281099, -0.424549317], [-1.760461157, -2.2833899, 0.573880683], [-1.211111157, 0.5815011, -2.395437317], [-1.195891157, -2.2720199, -1.242477317], [-1.172151157, -2.4167399, 1.371442683], [-1.163621157, 0.2920721, 2.832412683], [-1.015071157, 2.6870501, -0.159284317], [-1.006011157, 2.6317701, 0.839145683], [-0.562589157, 1.3418401, -2.359237317], [-0.515109157, 1.0524101, 2.868612683], [-0.441435157, 2.6431401, -0.977217317], [-0.417695157, 2.4984201, 1.636712683], [-0.273870157, 0.4551791, -2.720447317], [-0.268601157, 2.1293101, -1.817517317], [-0.258655157, -2.3983399, -1.567487317], [-0.230188157, 1.8951601, 2.411892683], [-0.229310157, -2.5772199, 1.663502683], [-0.220785157, 0.1315881, 3.124462683], [0.376103843, 1.6782401, -2.434677317], [0.487275843, -0.1919909, -2.763187317], [0.495801843, 2.5168201, -1.302227317], [0.496678843, -1.9555699, -2.050617317], [0.525145843, 2.3379401, 1.928762683], [0.535091843, -2.1897199, 2.178792683], [0.540360843, -0.5155829, 3.081722683], [0.664822843, 0.7915781, -2.795887317], [0.684185843, -2.5588299, -1.275437317], [0.707925843, -2.7035399, 1.338492683], [0.781599843, -1.1128099, -2.507337317], [0.829080843, -1.4022399, 2.720512683], [1.140508843, 2.0657401, -1.919377317], [1.272498843, -2.6921699, -0.477870317], [1.281568843, -2.7474499, 0.520559683], [1.430118843, -0.3524749, -2.471137317], [1.438638843, 2.3563301, -1.010167317], [1.462378843, 2.2116201, 1.603752683], [1.477598843, -0.6419039, 2.756712683], [1.607658843, 0.6310941, -2.503837317], [1.901648843, 1.4185701, -1.962127317], [2.026958843, 2.2229901, -0.212604317], [2.036018843, 2.1677101, 0.785825683], [2.036898843, -2.3046699, 0.037427063], [2.194518843, 0.0350269, -1.955837317], [2.199788843, 1.7091601, -1.052907317], [2.223528843, 1.5644501, 1.561012683], [2.232928843, -0.1991259, 2.273582683], [2.488508843, 0.8225021, -1.414127317], [2.517848843, 0.6436251, 1.816862683], [2.685418843, -1.5443299, 0.073631683], [2.782828843, -0.0983201, -1.158277317], [2.806568843, -0.2430349, 1.455652683], [2.970338843, -0.7015819, -0.383085317], [2.979408843, -0.7568579, 0.615344683]]}, "78": {"AlternativeName": null, "Comment": null, "Name": "Metabidiminished Icosahedron", "ShortName": "J62", "vertices": [[-0.5, 0.0, -0.709017], [-0.5, 0.0, 0.909017], [0.0, -0.809017, -0.4], [0.0, 0.809017, -0.4], [0.5, 0.0, -0.709017], [0.5, 0.0, 0.909017], [-0.809017, -0.5, 0.1], [-0.809017, 0.5, 0.1], [0.809017, -0.5, 0.1], [0.809017, 0.5, 0.1]]}, "79": {"AlternativeName": null, "Comment": null, "Name": "Metabidiminished Rhombicosidodecahedron", "ShortName": "J81", "vertices": [[-2.034672474, -0.381270867, 0.423803728], [-1.942632474, -0.485832867, -0.566446272], [-1.941272474, 0.609723133, 0.327843728], [-1.849232474, 0.505160133, -0.662406272], [-1.757862474, -1.261133867, 0.037516308], [-1.642312474, -0.644739867, 1.305071728], [-1.513342474, 1.333326133, -0.213709272], [-1.491192474, 0.958722133, 1.149811728], [-1.401342474, -0.918487867, -1.287428272], [-1.365502474, -1.524603867, 0.918788728], [-1.306422474, 0.183421133, 1.753771728], [-1.250222474, 0.684973133, -1.442698272], [-1.216572474, -1.693783867, -0.683469272], [-1.063262474, 1.682316133, 0.608255728], [-0.973412474, -0.194889867, -1.828988272], [-0.914332474, 1.513136133, -0.993998272], [-0.670903474, 1.418856133, 1.489531728], [-0.617568474, -1.513973867, -1.463758272], [-0.581731474, -2.120093867, 0.742458728], [-0.489688474, -2.224653867, -0.247791272], [-0.486137474, 0.643551133, 2.093491728], [-0.373047474, 1.080476133, -1.714988272], [-0.189637474, -0.790376867, -2.005318272], [-0.186086474, 2.077826133, 0.335966728], [-0.096238474, 0.200616133, -2.101278272], [-0.094044474, 1.973266133, -0.654283272], [0.109320226, -2.044843867, -1.028088272], [0.206269526, 1.814356133, 1.217241728], [0.409633526, -2.203743867, 0.843438728], [0.447240526, 1.540606133, -1.375268272], [0.501676526, -2.308303867, -0.146811272], [0.505227526, 0.559895133, 2.194471728], [0.801727526, -0.874032867, -1.904338272], [0.805278526, 1.994166133, 0.436946728], [0.895126526, 0.116960133, -2.000298272], [0.897320526, 1.889606133, -0.553303272], [0.933158526, 1.283496133, 1.652921728], [0.986493526, -1.649333867, -1.300368272], [1.229917526, -1.743613867, 1.183161728], [1.231017526, 0.945121133, -1.551598272], [1.288997526, -0.035592527, 2.018141728], [1.378847526, -1.912803867, -0.419101272], [1.532167526, 1.463306133, 0.872624728], [1.565807526, -0.915455867, 1.631851728], [1.681097526, 1.294116133, -0.729633272], [1.716937526, 0.688006133, 1.476591728], [1.828927526, -1.563803867, 0.402864728], [2.073447526, 1.030656133, 0.151638728], [2.164817526, -0.735642867, 0.851560728], [2.258217526, 0.255351133, 0.755601728]]}, "80": {"AlternativeName": null, "Comment": null, "Name": "Metabigyrate Rhombicosidodecahedron", "ShortName": "J74", "vertices": [[-2.15713033, -0.114704717, 0.565411907], [-2.12641033, 0.680289283, -0.040427493], [-2.10096033, -0.721244717, -0.227654093], [-2.07025033, 0.073748983, -0.833493093], [-1.84552033, -1.058149717, 0.678568907], [-1.76510033, 1.023170283, -0.907538093], [-1.71309033, 0.135164283, 1.425869907], [-1.66339033, 1.421490283, 0.445604907], [-1.56604033, -1.452779717, -0.650401093], [-1.51634033, -0.166451717, -1.630670093], [-1.40794033, 1.084590283, 1.351829907], [-1.40148033, -0.808281717, 1.539029907], [-1.31059033, -1.789679717, 0.255822907], [-1.30207033, 1.764370283, -0.421506093], [-1.21119033, 0.782970283, -1.704710093], [-1.20472033, -1.109899717, -1.517510093], [-0.93845133, -0.067079217, 2.025059907], [-0.88729333, 1.826350283, 0.929092907], [-0.75668433, -2.029889717, -0.541353093], [-0.67626433, 0.051435983, -2.127460093], [-0.63330133, 0.882342283, 1.951019907], [-0.59212033, -1.385389717, 1.648079907], [-0.53595133, -1.991929717, 0.855011907], [-0.52597533, 2.169230283, 0.061981607], [-0.46199833, 1.982260283, -0.918298093], [-0.40582933, 1.375720283, -1.711360093], [-0.39536533, -1.686999717, -1.408460093], [-0.36464833, -0.892009717, -2.014300093], [-0.12909533, -0.644185717, 2.134109907], [-0.11265233, 1.624110283, 1.528279907], [0.01795747, -2.232129717, 0.057836307], [0.12909467, 0.644186283, -2.134110093], [0.36464767, 0.892010283, 2.014299907], [0.40582867, -1.375719717, 1.711359907], [0.46199767, -1.982259717, 0.918297907], [0.47197367, 2.178900283, 0.125266907], [0.53595067, 1.991930283, -0.855012093], [0.59211967, 1.385390283, -1.648080093], [0.60258367, -1.677339717, -1.345180093], [0.63330067, -0.882341717, -1.951020093], [0.67626367, -0.051435417, 2.127459907], [0.72742167, 1.842000283, 1.031489907], [0.85803067, -2.014239717, -0.438955093], [0.93845067, 0.067079783, -2.025060093], [1.20471967, 1.109900283, 1.517509907], [1.21118967, -0.782969717, 1.704709907], [1.30206967, -1.764369717, 0.421505907], [1.31058967, 1.789680283, -0.255823093], [1.40147967, 0.808282283, -1.539030093], [1.47005967, -1.397439717, -0.933904093], [1.50076967, -0.602440717, -1.539740093], [1.51633967, 0.166452283, 1.630669907], [1.56603967, 1.452780283, 0.650400907], [1.76509967, -1.023169717, 0.907537907], [1.84551967, 1.058150283, -0.678569093], [1.91409967, -1.147569717, -0.073443193], [1.96379967, 0.138761283, -1.053710093], [2.07024967, -0.073748417, 0.833493907], [2.10095967, 0.721245283, 0.227653907], [2.21924967, -0.198144717, -0.147488093]]}, "81": {"AlternativeName": null, "Comment": null, "Name": "Metagyrate Diminished Rhombicosidodecahedron", "ShortName": "J78", "vertices": [[-2.156463838, 0.479212042, -0.283451002], [-2.118633838, -0.371178958, -0.808241002], [-2.106893838, -0.043710058, 0.567486998], [-2.069063838, -0.894100958, 0.042697398], [-1.869223838, 0.495055042, -1.241186002], [-1.739463838, -0.873972958, 0.986603998], [-1.703353838, 1.357881042, -0.133026002], [-1.623153838, 0.511779042, 1.243823998], [-1.604313838, -0.868466958, -1.506946002], [-1.524123838, -1.714568958, -0.130095502], [-1.416123838, 1.373731042, -1.090756002], [-1.373753838, 1.378011042, 0.810880998], [-1.354913838, -0.002233758, -1.939886002], [-1.255733838, -0.318483958, 1.662933998], [-1.236883838, -1.698728958, -1.087826002], [-1.194523838, -1.694448958, 0.813810998], [-0.932383838, 1.929221042, -0.414421002], [-0.802622838, 0.560189042, 1.813363998], [-0.680204838, -2.191738958, 0.115108998], [-0.621771838, 1.419491042, -1.696486002], [-0.602783838, 1.949341042, 0.529484998], [-0.583942838, 0.569098042, -2.221276002], [-0.553219838, 1.426421042, 1.380423998], [-0.411813838, -0.795644958, 1.908143998], [-0.392972838, -2.175888958, -0.842621002], [-0.373984838, -1.646038958, 1.383353998], [-0.138032838, 1.974981042, -1.020155002], [0.041292262, 0.083028642, 2.058563998], [0.140330162, -2.143328958, 0.684651998], [0.376283162, 1.477691042, -1.718856002], [0.395270162, 2.007551042, 0.507117998], [0.414112162, 0.627299042, -2.243646002], [0.444834162, 1.484621042, 1.358053998], [0.471460162, -1.215348958, 1.699153998], [0.605081162, -2.117688958, -0.864988002], [0.682502162, 2.023391042, -0.450612002], [0.812263162, 0.654361042, 1.777173998], [0.924567162, -0.336671958, 1.849573998], [0.934681162, -2.097558958, 0.078917998], [0.985776162, -1.712628958, 1.000453998], [1.196816162, 1.526101042, -1.149316002], [1.239186162, 1.530381042, 0.752321998], [1.258026162, 0.150139042, -1.998446002], [1.376056162, -1.546358958, -1.146386002], [1.526416162, 1.546231042, -0.205408002], [1.606616162, 0.700122042, 1.171443998], [1.625456162, -0.680124958, -1.579326002], [1.705656162, -1.526228958, -0.202477002], [1.718916162, -0.290909958, 1.243843998], [1.741766162, 0.705628042, -1.322106002], [1.756746162, -1.141298958, 0.719052998], [2.071366162, 0.725756042, -0.378200002], [2.109196162, -0.124634958, -0.902990002], [2.120926162, 0.202834042, 0.472737998], [2.158756162, -0.647556958, -0.052051802]]}, "82": {"AlternativeName": null, "Comment": null, "Name": "Obtuse Golden Rhombohedron", "ShortName": "O18", "vertices": [[-0.947214, 0.0854102, 0.262866], [-0.5, -0.809017, 0.262866], [-0.5, 0.809017, -0.262866], [-0.0527864, -0.0854102, -0.262866], [0.0527864, 0.0854102, 0.262866], [0.5, -0.809017, 0.262866], [0.5, 0.809017, -0.262866], [0.947214, -0.0854102, -0.262866]]}, "83": {"AlternativeName": null, "Comment": null, "Name": "Octahedron", "ShortName": "P02", "vertices": [[-0.707107, 0.0, 0.0], [0.0, 0.707107, 0.0], [0.0, 0.0, -0.707107], [0.0, 0.0, 0.707107], [0.0, -0.707107, 0.0], [0.707107, 0.0, 0.0]]}, "84": {"AlternativeName": null, "Comment": null, "Name": "Parabiaugmented Dodecahedron", "ShortName": "J59", "vertices": [[0.0, 0.0, 1.40126], [0.0, 0.0, -1.40126], [0.178411, -1.30902, 0.467086], [0.178411, 1.30902, 0.467086], [0.467086, -0.809017, -1.04444], [0.467086, 0.809017, -1.04444], [1.04444, -0.809017, 0.467086], [1.04444, 0.809017, 0.467086], [-1.22285, -0.5, 0.467086], [-1.22285, 0.5, 0.467086], [1.22285, -0.5, -0.467086], [1.22285, 0.5, -0.467086], [-0.934172, 0.0, -1.04444], [-0.467086, -0.809017, 1.04444], [-0.467086, 0.809017, 1.04444], [0.934172, 0.0, 1.04444], [-1.04444, -0.809017, -0.467086], [-1.04444, 0.809017, -0.467086], [0.995125, 0.0, -1.30264], [-0.995125, 0.0, 1.30264], [-0.178411, -1.30902, -0.467086], [-0.178411, 1.30902, -0.467086]]}, "85": {"AlternativeName": null, "Comment": null, "Name": "Parabiaugmented Hexagonal Prism", "ShortName": "J55", "vertices": [[-1.0, 0.0, -0.5], [-1.0, 0.0, 0.5], [-0.5, -0.866025, -0.5], [-0.5, -0.866025, 0.5], [-0.5, 0.866025, -0.5], [-0.5, 0.866025, 0.5], [0.5, -0.866025, -0.5], [0.5, -0.866025, 0.5], [0.5, 0.866025, -0.5], [0.5, 0.866025, 0.5], [1.0, 0.0, -0.5], [1.0, 0.0, 0.5], [-1.36237, -0.786566, 0.0], [1.36237, 0.786566, 0.0]]}, "86": {"AlternativeName": null, "Comment": null, "Name": "Parabiaugmented Truncated Dodecahedron", "ShortName": "J69", "vertices": [[-2.95761, 0.246541, -0.096963], [-2.81512, 0.316165, 0.890381], [-2.8109, -0.335955, -0.896447], [-2.72848, 1.12885, 0.314162], [-2.43785, -0.153677, 1.68845], [-2.43103, -1.20883, -1.20269], [-2.34439, -0.39614, -1.77891], [-2.21104, 1.97397, 0.179892], [-1.96992, -0.983522, 1.99242], [-1.96309, -2.03867, -0.89873], [-1.85913, -1.723, -1.84188], [-1.77249, -0.910309, -2.4181], [-1.74079, -0.101208, 2.40354], [-1.73628, 0.0889763, -2.40729], [-1.60293, 2.45909, -0.448487], [-1.59005, -1.8564, 1.68617], [-1.58583, -2.50852, -0.100657], [-1.46044, 2.52871, 0.538858], [-1.44334, -2.43889, 0.886688], [-1.21883, 0.934094, -2.54156], [-1.13642, 2.3989, -1.33095], [-1.11932, -2.5687, -0.983124], [-1.01536, -2.25303, -1.92628], [-0.990189, 0.453532, 2.76251], [-0.989708, 1.81641, -2.13044], [-0.875179, -0.938065, -2.85862], [-0.838961, 0.0612202, -2.84781], [-0.763375, 2.58118, 1.25395], [-0.746275, -2.38642, 1.60178], [-0.472746, 1.29865, 2.62824], [-0.407243, -1.76791, -2.55466], [-0.386111, 2.11134, 2.05202], [-0.239103, 2.37115, -1.77147], [-0.222004, -2.59646, -1.42364], [-0.00481035, 0.468806, 2.9322], [0.00481035, -0.468806, -2.9322], [0.222004, 2.59646, 1.42364], [0.239103, -2.37115, 1.77147], [0.386111, -2.11134, -2.05202], [0.407243, 1.76791, 2.55466], [0.472746, -1.29865, -2.62824], [0.746275, 2.38642, -1.60178], [0.763375, -2.58118, -1.25395], [0.838961, -0.0612202, 2.84781], [0.875179, 0.938065, 2.85862], [0.989708, -1.81641, 2.13044], [0.990189, -0.453532, -2.76251], [1.01536, 2.25303, 1.92628], [1.11932, 2.5687, 0.983124], [1.13642, -2.3989, 1.33095], [1.21883, -0.934094, 2.54156], [1.44334, 2.43889, -0.886688], [1.46044, -2.52871, -0.538858], [1.58583, 2.50852, 0.100657], [1.59005, 1.8564, -1.68617], [1.60293, -2.45909, 0.448487], [1.73628, -0.0889763, 2.40729], [1.74079, 0.101208, -2.40354], [1.77249, 0.910309, 2.4181], [1.85913, 1.723, 1.84188], [1.96309, 2.03867, 0.89873], [1.96992, 0.983522, -1.99242], [2.21104, -1.97397, -0.179892], [2.34439, 0.39614, 1.77891], [2.43103, 1.20883, 1.20269], [2.43785, 0.153677, -1.68845], [2.72848, -1.12885, -0.314162], [2.8109, 0.335955, 0.896447], [2.81512, -0.316165, -0.890381], [2.95761, -0.246541, 0.096963]]}, "87": {"AlternativeName": null, "Comment": null, "Name": "Parabidiminished Rhombicosidodecahedron", "ShortName": "J80", "vertices": [[-2.19149, -0.351484, -0.244698], [-2.10378, -0.155834, 0.732046], [-2.04429, 0.501655, -0.745182], [-1.90236, 0.818222, 0.835221], [-1.86559, 1.22457, -0.0777559], [-1.79587, -0.866028, -1.00544], [-1.64867, -0.0128901, -1.50592], [-1.56623, -0.353811, 1.55171], [-1.44725, 0.961166, -1.40275], [-1.36482, 0.620245, 1.65488], [-1.26855, 1.68408, -0.735321], [-1.21761, 1.47338, 1.1544], [-1.18084, 1.87973, 0.241422], [-1.06803, -1.50293, -1.2596], [-0.829841, -0.122523, -2.0694], [-0.784177, -0.869794, 1.90121], [-0.628423, 0.851534, -1.96623], [-0.470986, -1.04342, -1.91716], [-0.458277, 0.706262, 2.06815], [-0.339292, 2.02124, -0.886307], [-0.31107, 1.5594, 1.56766], [-0.28597, -2.01891, -0.9101], [-0.251577, 2.21689, 0.0904368], [-0.0994221, -0.214633, 2.22039], [-0.0563312, -1.5067, 1.64705], [0.0563312, 1.5067, -1.64705], [0.0994221, 0.214633, -2.22039], [0.251577, -2.21689, -0.0904368], [0.28597, 2.01891, 0.9101], [0.31107, -1.5594, -1.56766], [0.339292, -2.02124, 0.886307], [0.458277, -0.706262, -2.06815], [0.470986, 1.04342, 1.91716], [0.628423, -0.851534, 1.96623], [0.784177, 0.869794, -1.90121], [0.829841, 0.122523, 2.0694], [1.06803, 1.50293, 1.2596], [1.18084, -1.87973, -0.241422], [1.21761, -1.47338, -1.1544], [1.26855, -1.68408, 0.735321], [1.36482, -0.620245, -1.65488], [1.44725, -0.961166, 1.40275], [1.56623, 0.353811, -1.55171], [1.64867, 0.0128901, 1.50592], [1.79587, 0.866028, 1.00544], [1.86559, -1.22457, 0.0777559], [1.90236, -0.818222, -0.835221], [2.04429, -0.501655, 0.745182], [2.10378, 0.155834, -0.732046], [2.19149, 0.351484, 0.244698]]}, "88": {"AlternativeName": null, "Comment": null, "Name": "Parabigyrate Rhombicosidodecahedron", "ShortName": "J73", "vertices": [[-2.19876998, 0.279547983, 0.270815], [-2.13820998, -0.168840017, -0.620971], [-2.10445998, -0.714504017, 0.216352], [-1.91264998, 1.143279983, -0.144039], [-1.85802998, 0.260375983, 1.21078], [-1.85209998, 0.694891983, -1.03583], [-1.76371998, -0.733677017, 1.15632], [-1.69949998, -0.913518017, -1.12395], [-1.66573998, -1.459180017, -0.286624], [-1.39508998, 1.657919983, 0.539531], [-1.36132998, 1.112259983, 1.37685], [-1.35538998, 1.546779983, -0.869749], [-1.24615998, -0.219033017, 1.83989], [-1.23654998, 0.484029983, -1.79519], [-1.14223998, -0.510022017, -1.84966], [-1.11441998, -1.490200017, 1.23427], [-1.05386998, -1.938590017, 0.342484], [-0.90550698, -1.407700017, -1.478], [-0.87175098, -1.953370017, -0.640681], [-0.83782798, 2.061419983, -0.186179], [-0.74945498, 0.632849983, 2.00596], [-0.73984798, 1.335909983, -1.62912], [-0.64178598, 1.779909983, 1.1858], [-0.59685798, -0.975560017, 1.91784], [-0.30113398, 0.591235983, -2.13209], [-0.20682398, -0.402816017, -2.18656], [-0.15814398, -1.720240017, 1.41486], [-0.09758938, -2.168620017, 0.523079], [-0.08452808, 2.183399983, 0.460087], [-0.02991078, 1.300499983, 1.8149], [0.02991082, -1.300500017, -1.8149], [0.08452802, -2.183400017, -0.460087], [0.09758932, 2.168619983, -0.523079], [0.15814402, 1.720239983, -1.41486], [0.20682402, 0.402815983, 2.18656], [0.30113402, -0.591236017, 2.13209], [0.59685802, 0.975559983, -1.91784], [0.64178602, -1.779910017, -1.1858], [0.73984802, -1.335910017, 1.62912], [0.74945502, -0.632850017, -2.00596], [0.83782802, -2.061420017, 0.186179], [0.87175102, 1.953369983, 0.640681], [0.90550602, 1.407699983, 1.478], [1.05387002, 1.938589983, -0.342484], [1.11442002, 1.490199983, -1.23427], [1.14224002, 0.510021983, 1.84966], [1.23655002, -0.484030017, 1.79519], [1.24616002, 0.219032983, -1.83989], [1.35539002, -1.546780017, 0.869749], [1.36133002, -1.112260017, -1.37685], [1.39509002, -1.657920017, -0.539531], [1.66574002, 1.459179983, 0.286624], [1.69950002, 0.913517983, 1.12395], [1.76372002, 0.733676983, -1.15632], [1.85210002, -0.694892017, 1.03583], [1.85803002, -0.260376017, -1.21078], [1.91265002, -1.143280017, 0.144039], [2.10446002, 0.714504983, -0.216352], [2.13821002, 0.168839983, 0.620971], [2.19877002, -0.279548017, -0.270815]]}, "89": {"AlternativeName": null, "Comment": null, "Name": "Paragyrate Diminished Rhombicosidodecahedron", "ShortName": "J77", "vertices": [[-2.147782129, -0.396877776, -0.435706127], [-2.049702129, 0.598175224, -0.451537127], [-2.002372129, -0.706814776, 0.503865873], [-1.843672129, 0.903214224, 0.478249873], [-1.814422129, 0.096686224, 1.068720873], [-1.790842129, -0.446900776, -1.368499127], [-1.761592129, -1.253427776, -0.778023127], [-1.692762129, 0.548151224, -1.384329127], [-1.616192129, -1.563367776, 0.161548873], [-1.504812129, 1.351652224, -0.819470127], [-1.410162129, -1.258327776, 1.091340873], [-1.298792129, 1.656692224, 0.110316873], [-1.295902129, 0.915579224, 1.314780873], [-1.222212129, -0.454826776, 1.656190873], [-1.067902129, -0.837777776, -1.938209127], [-1.038652129, -1.644307776, -1.347739127], [-0.909206129, 0.772250224, -1.963819127], [-0.803381129, -2.145797776, 0.172523873], [-0.751017129, 1.669052224, 0.946852873], [-0.721258129, 1.575752224, -1.398969127], [-0.703691129, 0.364066224, 1.902260873], [-0.597356129, -1.840757776, 1.102310873], [-0.523022129, -0.084301076, -2.306139127], [-0.446446129, -2.195817776, -0.760265127], [-0.387902129, 2.069312224, 0.105460873], [-0.293249129, -0.540665776, 2.016270873], [-0.030967029, 2.019292224, -0.827328127], [0.001682591, 1.184872224, -1.968679127], [0.092935171, -1.397217776, 1.673950873], [0.125584871, -2.231637776, 0.532598873], [0.177948871, 1.583222224, 1.306930873], [0.207197871, 0.776689224, 1.897400873], [0.387866871, 0.328322224, -2.310999127], [0.482519871, -2.281657776, -0.400190127], [0.541063871, 1.983472224, 0.465535873], [0.617639871, -0.128041776, 2.011410873], [0.691973871, 1.628412224, -1.397039127], [0.815875871, -1.788097776, 1.104240873], [0.897998871, 1.933452224, -0.467253127], [1.003823871, -0.984593776, 1.669090873], [1.133277871, 1.431962224, 1.053010873], [1.162527871, 0.625435224, 1.643480873], [1.316827871, 0.242484224, -1.950919127], [1.393407871, -1.869037776, -0.405046127], [1.504777871, 1.045982224, -1.386069127], [1.599437871, -1.563997776, 0.524741873], [1.710807871, 1.351022224, -0.456278127], [1.787377871, -0.760494776, 1.089600873], [1.856217871, 1.041082224, 0.483293873], [1.885467871, 0.234558224, 1.073760873], [1.909037871, -0.309028776, -1.363449127], [1.938287871, -1.115557776, -0.772979127], [2.096987871, 0.494471224, -0.798595127], [2.144317871, -0.810517776, 0.156808873], [2.242397871, 0.184534224, 0.140976873]]}, "90": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Cupola", "ShortName": "J05", "vertices": [[-6.7e-08, -1.61803, -0.175243667], [-6.7e-08, 1.61803, -0.175243667], [0.850650933, 0.0, 0.350487333], [0.262865933, -0.809017, 0.350487333], [0.262865933, 0.809017, 0.350487333], [-0.951057067, -1.30902, -0.175243667], [-0.951057067, 1.30902, -0.175243667], [0.951056933, -1.30902, -0.175243667], [0.951056933, 1.30902, -0.175243667], [-0.688191067, -0.5, 0.350487333], [-0.688191067, 0.5, 0.350487333], [-1.538840067, -0.5, -0.175243667], [-1.538840067, 0.5, -0.175243667], [1.538839933, -0.5, -0.175243667], [1.538839933, 0.5, -0.175243667]]}, "91": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Gyrobicupola", "ShortName": "J31", "vertices": [[0.0, -1.61803, 0.0], [0.0, 1.61803, 0.0], [-0.688191, -0.5, 0.525731], [-0.688191, 0.5, 0.525731], [0.688191, -0.5, -0.525731], [0.688191, 0.5, -0.525731], [-0.850651, 0.0, -0.525731], [0.850651, 0.0, 0.525731], [-0.262866, -0.809017, -0.525731], [-0.262866, 0.809017, -0.525731], [0.262866, -0.809017, 0.525731], [0.262866, 0.809017, 0.525731], [-0.951057, -1.30902, 0.0], [-0.951057, 1.30902, 0.0], [0.951057, -1.30902, 0.0], [0.951057, 1.30902, 0.0], [-1.53884, -0.5, 0.0], [-1.53884, 0.5, 0.0], [1.53884, -0.5, 0.0], [1.53884, 0.5, 0.0]]}, "92": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Gyrocupolarotunda", "ShortName": "J33", "vertices": [[4e-07, -1.61803, -0.34026], [4e-07, 1.61803, -0.34026], [0.8506514, 0.0, -0.865991], [0.2628664, -0.809017, -0.865991], [0.2628664, 0.809017, -0.865991], [-0.9510566, -1.30902, -0.34026], [-0.9510566, 1.30902, -0.34026], [0.9510574, -1.30902, -0.34026], [0.9510574, 1.30902, -0.34026], [-0.6881906, -0.5, -0.865991], [-0.6881906, 0.5, -0.865991], [-1.5388396, -0.5, -0.34026], [-1.5388396, 0.5, -0.34026], [1.5388404, -0.5, -0.34026], [1.5388404, 0.5, -0.34026], [1.3763804, 0.0, 0.510391], [0.4253254, -1.30902, 0.510391], [0.4253254, 1.30902, 0.510391], [-1.1135196, -0.809017, 0.510391], [-1.1135196, 0.809017, 0.510391], [-0.8506506, 0.0, 1.03612], [-0.2628656, -0.809017, 1.03612], [-0.2628656, 0.809017, 1.03612], [0.6881914, -0.5, 1.03612], [0.6881914, 0.5, 1.03612]]}, "93": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Hexecontahedron", "ShortName": "C13", "vertices": [[0.0, 0.0, -3.6322], [0.0, 0.0, 3.6322], [-2.42147, 0.0, -2.70728], [2.42147, 0.0, 2.70728], [-3.74093, 0.0, -0.714455], [3.74093, 0.0, 0.714455], [-2.31202, 0.0, 3.02648], [2.31202, 0.0, -3.02648], [-2.97388, 1.2495, -1.66963], [2.97388, 1.2495, 1.66963], [-3.20333, -1.64692, -0.468228], [3.20333, -1.64692, 0.468228], [-2.89122, -0.64303, -2.10244], [2.89122, -0.64303, 2.10244], [-3.33708, 1.41527, 0.232068], [3.33708, 1.41527, -0.232068], [-3.47888, -0.374822, 0.974575], [3.47888, -0.374822, -0.974575], [-0.175393, 3.59762, 0.468228], [0.175393, 3.59762, -0.468228], [-2.70728, -2.09705, -1.21073], [-2.70728, 2.09705, -1.21073], [2.70728, -2.09705, 1.21073], [2.70728, 2.09705, 1.21073], [-2.46401, 0.973951, -2.48455], [2.46401, 0.973951, 2.48455], [-3.24272, 0.374822, 1.59285], [3.24272, 0.374822, -1.59285], [-3.02794, -1.95071, 0.468228], [3.02794, -1.95071, -0.468228], [-2.89419, 2.18236, -0.232068], [2.89419, 2.18236, 0.232068], [-3.16974, -1.29605, 1.21073], [-3.16974, 1.29605, 1.21073], [3.16974, -1.29605, -1.21073], [3.16974, 1.29605, -1.21073], [-2.54667, -1.58042, -2.05174], [2.54667, -1.58042, 2.05174], [-0.753368, 0.64303, 3.49455], [0.753368, 0.64303, -3.49455], [-0.442882, -3.59762, -0.232068], [0.442882, -3.59762, 0.232068], [-0.40484, -3.20021, 1.66963], [0.40484, -3.20021, -1.66963], [-1.10352, -1.2495, 3.22706], [1.10352, -1.2495, -3.22706], [-0.462459, -3.3931, -1.21073], [-0.462459, 3.3931, -1.21073], [0.462459, -3.3931, 1.21073], [0.462459, 3.3931, 1.21073], [-0.530344, 1.58042, 3.22706], [0.530344, 1.58042, -3.22706], [-0.388538, -2.62087, 2.48455], [0.388538, -2.62087, -2.48455], [-2.64202, -1.41527, 2.05174], [2.64202, -1.41527, -2.05174], [-0.0953548, 2.99569, -2.05174], [0.0953548, 2.99569, 2.05174], [-2.06405, -2.82539, -0.974575], [2.06405, -2.82539, 0.974575], [-2.07547, 1.64692, 2.48455], [2.07547, 1.64692, -2.48455], [-0.180196, -0.973951, 3.49455], [0.180196, -0.973951, -3.49455], [-0.888732, 2.82539, 2.10244], [0.888732, 2.82539, -2.10244], [-2.56904, 1.95071, 1.66963], [2.56904, 1.95071, -1.66963], [-1.21073, -2.09705, 2.70728], [-1.21073, 2.09705, 2.70728], [1.21073, -2.09705, -2.70728], [1.21073, 2.09705, -2.70728], [-1.29676, -2.99569, -1.59285], [1.29676, -2.99569, 1.59285], [-1.94597, 2.62087, -1.59285], [1.94597, 2.62087, 1.59285], [-1.41484, 3.20021, -0.974575], [1.41484, 3.20021, 0.974575], [-1.63386, -0.330921, -3.22706], [1.63386, -0.330921, 3.22706], [-2.00249, -2.18236, 2.10244], [2.00249, -2.18236, -2.10244], [-0.933564, 0.330921, -3.49455], [0.933564, 0.330921, 3.49455], [-1.87047, -3.23974, 0.714455], [-1.87047, 3.23974, 0.714455], [1.87047, -3.23974, -0.714455], [1.87047, 3.23974, -0.714455], [-1.15601, -2.00227, -3.02648], [-1.15601, 2.00227, -3.02648], [1.15601, -2.00227, 3.02648], [1.15601, 2.00227, 3.02648]]}, "94": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Icositetrahedron", "ShortName": "C12", "vertices": [[0.0, 0.0, -2.294], [0.0, 0.0, 2.294], [0.0, -2.294, 0.0], [0.0, 2.294, 0.0], [-1.72488, -1.24722, -0.368676], [-1.72488, 1.24722, 0.368676], [-1.72488, -0.368676, 1.24722], [-1.72488, 0.368676, -1.24722], [1.72488, -1.24722, 0.368676], [1.72488, 1.24722, -0.368676], [1.72488, -0.368676, -1.24722], [1.72488, 0.368676, 1.24722], [-2.294, 0.0, 0.0], [2.294, 0.0, 0.0], [-1.24722, -1.72488, 0.368676], [-1.24722, 1.72488, -0.368676], [-1.24722, -1.24722, -1.24722], [-1.24722, -1.24722, 1.24722], [-1.24722, 1.24722, -1.24722], [-1.24722, 1.24722, 1.24722], [-1.24722, -0.368676, -1.72488], [-1.24722, 0.368676, 1.72488], [1.24722, -1.72488, -0.368676], [1.24722, 1.72488, 0.368676], [1.24722, -1.24722, -1.24722], [1.24722, -1.24722, 1.24722], [1.24722, 1.24722, -1.24722], [1.24722, 1.24722, 1.24722], [1.24722, -0.368676, 1.72488], [1.24722, 0.368676, -1.72488], [-0.368676, -1.72488, -1.24722], [-0.368676, 1.72488, 1.24722], [-0.368676, -1.24722, 1.72488], [-0.368676, 1.24722, -1.72488], [0.368676, -1.72488, 1.24722], [0.368676, 1.72488, -1.24722], [0.368676, -1.24722, -1.72488], [0.368676, 1.24722, 1.72488]]}, "95": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Orthobicupola", "ShortName": "J30", "vertices": [[-1e-07, -1.61803, 0.0], [-1e-07, 1.61803, 0.0], [0.8506509, 0.0, -0.525731], [0.8506509, 0.0, 0.525731], [0.2628659, -0.809017, -0.525731], [0.2628659, -0.809017, 0.525731], [0.2628659, 0.809017, -0.525731], [0.2628659, 0.809017, 0.525731], [-0.9510571, -1.30902, 0.0], [-0.9510571, 1.30902, 0.0], [0.9510569, -1.30902, 0.0], [0.9510569, 1.30902, 0.0], [-0.6881911, -0.5, -0.525731], [-0.6881911, -0.5, 0.525731], [-0.6881911, 0.5, -0.525731], [-0.6881911, 0.5, 0.525731], [-1.5388401, -0.5, 0.0], [-1.5388401, 0.5, 0.0], [1.5388399, -0.5, 0.0], [1.5388399, 0.5, 0.0]]}, "96": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Orthobirotunda", "ShortName": "J34", "vertices": [[7.33e-07, -1.61803, 0.0], [7.33e-07, 1.61803, 0.0], [-0.951056267, -1.30902, 0.0], [-0.951056267, 1.30902, 0.0], [0.951057733, -1.30902, 0.0], [0.951057733, 1.30902, 0.0], [-1.538839267, -0.5, 0.0], [-1.538839267, 0.5, 0.0], [1.538840733, -0.5, 0.0], [1.538840733, 0.5, 0.0], [1.376380733, 0.0, -0.850651], [1.376380733, 0.0, 0.850651], [0.425325733, -1.30902, -0.850651], [0.425325733, -1.30902, 0.850651], [0.425325733, 1.30902, -0.850651], [0.425325733, 1.30902, 0.850651], [-1.113519267, -0.809017, -0.850651], [-1.113519267, -0.809017, 0.850651], [-1.113519267, 0.809017, -0.850651], [-1.113519267, 0.809017, 0.850651], [-0.850650267, 0.0, -1.37638], [-0.850650267, 0.0, 1.37638], [-0.262865267, -0.809017, -1.37638], [-0.262865267, -0.809017, 1.37638], [-0.262865267, 0.809017, -1.37638], [-0.262865267, 0.809017, 1.37638], [0.688191733, -0.5, -1.37638], [0.688191733, -0.5, 1.37638], [0.688191733, 0.5, -1.37638], [0.688191733, 0.5, 1.37638]]}, "97": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Orthocupolarotunda", "ShortName": "J32", "vertices": [[4.8e-07, -1.61803, -0.34026], [4.8e-07, 1.61803, -0.34026], [0.42532548, -1.30902, 0.510391], [0.42532548, 1.30902, 0.510391], [0.68819148, -0.5, -0.865991], [0.68819148, -0.5, 1.03612], [0.68819148, 0.5, -0.865991], [0.68819148, 0.5, 1.03612], [-0.85065052, 0.0, -0.865991], [-0.85065052, 0.0, 1.03612], [-1.11351952, -0.809017, 0.510391], [-1.11351952, 0.809017, 0.510391], [-0.26286552, -0.809017, -0.865991], [-0.26286552, -0.809017, 1.03612], [-0.26286552, 0.809017, -0.865991], [-0.26286552, 0.809017, 1.03612], [-0.95105652, -1.30902, -0.34026], [-0.95105652, 1.30902, -0.34026], [0.95105748, -1.30902, -0.34026], [0.95105748, 1.30902, -0.34026], [-1.53883952, -0.5, -0.34026], [-1.53883952, 0.5, -0.34026], [1.53884048, -0.5, -0.34026], [1.53884048, 0.5, -0.34026], [1.37638048, 0.0, 0.510391]]}, "98": {"AlternativeName": null, "Comment": null, "Name": "Pentagonal Rotunda", "ShortName": "J06", "vertices": [[5.5e-07, -1.61803, -0.55675775], [5.5e-07, 1.61803, -0.55675775], [-0.95105645, -1.30902, -0.55675775], [-0.95105645, 1.30902, -0.55675775], [0.95105755, -1.30902, -0.55675775], [0.95105755, 1.30902, -0.55675775], [-1.53883945, -0.5, -0.55675775], [-1.53883945, 0.5, -0.55675775], [1.53884055, -0.5, -0.55675775], [1.53884055, 0.5, -0.55675775], [1.37638055, 0.0, 0.29389325], [0.42532555, -1.30902, 0.29389325], [0.42532555, 1.30902, 0.29389325], [-1.11351945, -0.809017, 0.29389325], [-1.11351945, 0.809017, 0.29389325], [-0.85065045, 0.0, 0.81962225], [-0.26286545, -0.809017, 0.81962225], [-0.26286545, 0.809017, 0.81962225], [0.68819155, -0.5, 0.81962225], [0.68819155, 0.5, 0.81962225]]}, "99": {"AlternativeName": null, "Comment": null, "Name": "Pentakis Dodecahedron", "ShortName": "C06", "vertices": [[-0.958957, 0.0, 1.25529], [1.37638, 0.0, 0.688191], [-0.0, 0.0, 1.53884], [-1.11352, -0.809017, 0.688191], [-1.11352, 0.809017, 0.688191], [-0.0, 0.0, -1.53884], [0.425325, -1.30902, 0.688191], [0.425325, 1.30902, 0.688191], [-0.296334, -0.912023, 1.25529], [-0.296334, 0.912023, 1.25529], [0.775813, 0.563661, 1.25529], [0.775813, -0.563661, 1.25529], [0.958957, 0.0, -1.25529], [0.296334, -0.912023, -1.25529], [0.296334, 0.912023, -1.25529], [-0.775813, 0.563661, -1.25529], [-0.775813, -0.563661, -1.25529], [1.25529, -0.912023, 0.296334], [-0.479479, -1.47568, 0.296334], [1.25529, 0.912023, 0.296334], [-0.479479, 1.47568, 0.296334], [-1.55163, 0.0, 0.296334], [1.55163, 0.0, -0.296334], [0.479479, -1.47568, -0.296334], [0.479479, 1.47568, -0.296334], [-1.25529, -0.912023, -0.296334], [-1.25529, 0.912023, -0.296334], [1.11352, -0.809017, -0.688191], [-1.37638, 0.0, -0.688191], [1.11352, 0.809017, -0.688191], [-0.425325, -1.30902, -0.688191], [-0.425325, 1.30902, -0.688191]]}, "100": {"AlternativeName": null, "Comment": "edge length=1", "Name": "Triangular Prism", "ShortName": "O01", "vertices": [[-0.288675, -0.5, -0.5], [-0.288675, -0.5, 0.5], [-0.288675, 0.5, -0.5], [-0.288675, 0.5, 0.5], [0.57735, 0.0, -0.5], [0.57735, 0.0, 0.5]]}, "101": {"AlternativeName": null, "Comment": "edge length=2", "Name": "Pentagonal Prism", "ShortName": "O02", "vertices": [[0.850650808, 0.0, -0.5], [0.850650808, 0.0, 0.5], [-0.68819096, -0.5, -0.5], [-0.68819096, -0.5, 0.5], [-0.68819096, 0.5, -0.5], [-0.68819096, 0.5, 0.5], [0.262865556, -0.809016994, -0.5], [0.262865556, -0.809016994, 0.5], [0.262865556, 0.809016994, -0.5], [0.262865556, 0.809016994, 0.5]]}, "102": {"AlternativeName": null, "Comment": "edge length=3", "Name": "Hexagonal Prism", "ShortName": "O03", "vertices": [[-1.0, 0.0, -0.5], [-1.0, 0.0, 0.5], [-0.5, -0.866025404, -0.5], [-0.5, -0.866025404, 0.5], [-0.5, 0.866025404, -0.5], [-0.5, 0.866025404, 0.5], [0.5, -0.866025404, -0.5], [0.5, -0.866025404, 0.5], [0.5, 0.866025404, -0.5], [0.5, 0.866025404, 0.5], [1.0, 0.0, -0.5], [1.0, 0.0, 0.5]]}, "103": {"AlternativeName": null, "Comment": "edge length=4", "Name": "Heptagonal Prism", "ShortName": "O04", "vertices": [[-1.038260698, -0.5, -0.5], [-1.038260698, -0.5, 0.5], [-1.038260698, 0.5, -0.5], [-1.038260698, 0.5, 0.5], [1.152382435, 0.0, -0.5], [1.152382435, 0.0, 0.5], [-0.256429216, -1.123489802, -0.5], [-0.256429216, -1.123489802, 0.5], [-0.256429216, 1.123489802, -0.5], [-0.256429216, 1.123489802, 0.5], [0.718498696, -0.900968868, -0.5], [0.718498696, -0.900968868, 0.5], [0.718498696, 0.900968868, -0.5], [0.718498696, 0.900968868, 0.5]]}, "104": {"AlternativeName": null, "Comment": "edge length=5", "Name": "Octagonal Prism", "ShortName": "O05", "vertices": [[-0.0, -1.306562965, -0.5], [-0.0, -1.306562965, 0.5], [-0.0, 1.306562965, -0.5], [-0.0, 1.306562965, 0.5], [-1.306562965, 0.0, -0.5], [-1.306562965, 0.0, 0.5], [1.306562965, 0.0, -0.5], [1.306562965, 0.0, 0.5], [-0.923879533, -0.923879533, -0.5], [-0.923879533, -0.923879533, 0.5], [-0.923879533, 0.923879533, -0.5], [-0.923879533, 0.923879533, 0.5], [0.923879533, -0.923879533, -0.5], [0.923879533, -0.923879533, 0.5], [0.923879533, 0.923879533, -0.5], [0.923879533, 0.923879533, 0.5]]}, "105": {"AlternativeName": null, "Comment": "edge length=6", "Name": "Nonagonal Prism", "ShortName": "O06", "vertices": [[-1.37373871, -0.5, -0.5], [-1.37373871, -0.5, 0.5], [-1.37373871, 0.5, -0.5], [-1.37373871, 0.5, 0.5], [-0.7309511, -1.266044443, -0.5], [-0.7309511, -1.266044443, 0.5], [-0.7309511, 1.266044443, -0.5], [-0.7309511, 1.266044443, 0.5], [1.4619022, 0.0, -0.5], [1.4619022, 0.0, 0.5], [1.119882057, -0.939692621, -0.5], [1.119882057, -0.939692621, 0.5], [1.119882057, 0.939692621, -0.5], [1.119882057, 0.939692621, 0.5], [0.253856653, -1.439692621, -0.5], [0.253856653, -1.439692621, 0.5], [0.253856653, 1.439692621, -0.5], [0.253856653, 1.439692621, 0.5]]}, "106": {"AlternativeName": null, "Comment": "edge length=7", "Name": "Decagonal Prism", "ShortName": "O07", "vertices": [[-1.618033989, 0.0, -0.5], [-1.618033989, 0.0, 0.5], [1.618033989, 0.0, -0.5], [1.618033989, 0.0, 0.5], [-1.309016994, -0.951056516, -0.5], [-1.309016994, -0.951056516, 0.5], [-1.309016994, 0.951056516, -0.5], [-1.309016994, 0.951056516, 0.5], [-0.5, -1.538841769, -0.5], [-0.5, -1.538841769, 0.5], [-0.5, 1.538841769, -0.5], [-0.5, 1.538841769, 0.5], [0.5, -1.538841769, -0.5], [0.5, -1.538841769, 0.5], [0.5, 1.538841769, -0.5], [0.5, 1.538841769, 0.5], [1.309016994, -0.951056516, -0.5], [1.309016994, -0.951056516, 0.5], [1.309016994, 0.951056516, -0.5], [1.309016994, 0.951056516, 0.5]]}, "107": {"AlternativeName": "Square Pyramid", "Comment": null, "Name": "Pyramid 4", "ShortName": "J01", "vertices": [[0.0, 0.0, 0.565685425], [0.0, -0.707106781, -0.141421356], [0.0, 0.707106781, -0.141421356], [-0.707106781, 0.0, -0.141421356], [0.707106781, 0.0, -0.141421356]]}, "108": {"AlternativeName": null, "Comment": null, "Name": "Pyramid 5", "ShortName": "J02", "vertices": [[0.0, 0.0, 0.43810926], [0.850650808, 0.0, -0.087621852], [0.262865556, -0.809016994, -0.087621852], [0.262865556, 0.809016994, -0.087621852], [-0.68819096, -0.5, -0.087621852], [-0.68819096, 0.5, -0.087621852]]}, "109": {"AlternativeName": null, "Comment": null, "Name": "Rhombic Dodecahedron", "ShortName": "C01", "vertices": [[-0.816497, -0.816497, 0.0], [-0.816497, 0.0, -0.57735], [-0.816497, 0.0, 0.57735], [-0.816497, 0.816497, 0.0], [0.0, -0.816497, -0.57735], [0.0, -0.816497, 0.57735], [0.0, 0.0, -1.1547], [0.0, 0.0, 1.1547], [0.0, 0.816497, -0.57735], [0.0, 0.816497, 0.57735], [0.816497, -0.816497, 0.0], [0.816497, 0.0, -0.57735], [0.816497, 0.0, 0.57735], [0.816497, 0.816497, 0.0]]}, "110": {"AlternativeName": null, "Comment": null, "Name": "Rhombic Enneacontahedron", "ShortName": "O17", "vertices": [[-2.41202, 0.0, -0.460654738], [-2.41202, 0.0, 0.539345262], [-2.2847, -0.934172, -0.127321738], [-2.2847, -0.934172, 0.872678262], [-2.2847, 0.934172, -0.127321738], [-2.2847, 0.934172, 0.872678262], [-2.15738, 0.0, 1.206011262], [-2.07869, -0.57735, -1.206013738], [-2.07869, 0.57735, -1.206013738], [-1.95137, -1.51152, -0.872673738], [-1.95137, -1.51152, 0.127322262], [-1.95137, 1.51152, -0.872673738], [-1.95137, 1.51152, 0.127322262], [-1.74536, 0.0, -1.951363738], [-1.61803, -0.934172, -1.618033738], [-1.61803, -0.934172, 1.618036262], [-1.61803, 0.934172, -1.618033738], [-1.61803, 0.934172, 1.618036262], [-1.53934, -1.51152, 1.206011262], [-1.53934, 1.51152, 1.206011262], [-1.49071, 0.0, 1.951366262], [-1.20601, -2.08887, -0.539344738], [-1.20601, -2.08887, 0.460655562], [-1.20601, 2.08887, -0.539344738], [-1.20601, 2.08887, 0.460655562], [-1.07869, -1.86834, -1.206013738], [-1.07869, 1.86834, -1.206013738], [-0.872678, -1.51152, 1.951366262], [-0.872678, -0.356822, -2.284703738], [-0.872678, 0.356822, -2.284703738], [-0.872678, 1.51152, 1.951366262], [-0.745356, -1.29099, -1.951363738], [-0.745356, -0.57735, 2.284696262], [-0.745356, 0.57735, 2.284696262], [-0.745356, 1.29099, -1.951363738], [-0.539345, -2.08887, 1.206011262], [-0.539345, 2.08887, 1.206011262], [-0.333333, -2.44569, -0.872673738], [-0.333333, -2.44569, 0.127322262], [-0.333333, 2.44569, -0.872673738], [-0.333333, 2.44569, 0.127322262], [-0.127322, -0.934172, 2.284696262], [-0.127322, 0.934172, 2.284696262], [-0.0, -1.86834, -1.618033738], [-0.0, -1.86834, 1.618036262], [-0.0, 0.0, -2.618033738], [-0.0, 0.0, 2.618036262], [-0.0, 1.86834, -1.618033738], [-0.0, 1.86834, 1.618036262], [0.127322, -0.934172, -2.284703738], [0.127322, 0.934172, -2.284703738], [0.333333, -2.44569, -0.127321738], [0.333333, -2.44569, 0.872678262], [0.333333, 2.44569, -0.127321738], [0.333333, 2.44569, 0.872678262], [0.539345, -2.08887, -1.206013738], [0.539345, 2.08887, -1.206013738], [0.745356, -1.29099, 1.951366262], [0.745356, -0.57735, -2.284703738], [0.745356, 0.57735, -2.284703738], [0.745356, 1.29099, 1.951366262], [0.872678, -1.51152, -1.951363738], [0.872678, -0.356822, 2.284696262], [0.872678, 0.356822, 2.284696262], [0.872678, 1.51152, -1.951363738], [1.07869, -1.86834, 1.206011262], [1.07869, 1.86834, 1.206011262], [1.20601, -2.08887, -0.460654738], [1.20601, -2.08887, 0.539345262], [1.20601, 2.08887, -0.460654738], [1.20601, 2.08887, 0.539345262], [1.49071, 0.0, -1.951363738], [1.53934, -1.51152, -1.206013738], [1.53934, 1.51152, -1.206013738], [1.61803, -0.934172, -1.618033738], [1.61803, -0.934172, 1.618036262], [1.61803, 0.934172, -1.618033738], [1.61803, 0.934172, 1.618036262], [1.74536, 0.0, 1.951366262], [1.95137, -1.51152, -0.127321738], [1.95137, -1.51152, 0.872678262], [1.95137, 1.51152, -0.127321738], [1.95137, 1.51152, 0.872678262], [2.07869, -0.57735, 1.206011262], [2.07869, 0.57735, 1.206011262], [2.15738, 0.0, -1.206013738], [2.2847, -0.934172, -0.872673738], [2.2847, -0.934172, 0.127322262], [2.2847, 0.934172, -0.872673738], [2.2847, 0.934172, 0.127322262], [2.41202, 0.0, -0.539344738], [2.41202, 0.0, 0.460655562]]}, "111": {"AlternativeName": null, "Comment": null, "Name": "Rhombic Icosahedron", "ShortName": "O16", "vertices": [[-1.44721, 0.0, 0.223607], [-1.17082, -0.850651, -0.223607], [-1.17082, 0.850651, -0.223607], [-0.894427, 0.0, -0.67082], [-0.723607, -0.525731, 0.67082], [-0.723607, 0.525731, 0.67082], [-0.447214, -1.37638, 0.223607], [-0.447214, 1.37638, 0.223607], [-0.276393, -0.850651, -0.67082], [-0.276393, 0.850651, -0.67082], [0.0, 0.0, -1.11803], [0.0, 0.0, 1.11803], [0.276393, -0.850651, 0.67082], [0.276393, 0.850651, 0.67082], [0.447214, -1.37638, -0.223607], [0.447214, 1.37638, -0.223607], [0.723607, -0.525731, -0.67082], [0.723607, 0.525731, -0.67082], [0.894427, 0.0, 0.67082], [1.17082, -0.850651, 0.223607], [1.17082, 0.850651, 0.223607], [1.44721, 0.0, -0.223607]]}, "112": {"AlternativeName": null, "Comment": null, "Name": "Rhombic Triacontahedron", "ShortName": "C02", "vertices": [[-0.0, 0.0, -1.61803], [-0.0, 0.0, 1.61803], [0.276393, -0.850651, 1.17082], [0.276393, 0.850651, 1.17082], [0.894427, 0.0, 1.17082], [1.17082, -0.850651, 0.723607], [1.17082, -0.850651, -0.276393], [1.17082, 0.850651, 0.723607], [1.17082, 0.850651, -0.276393], [-0.894427, 0.0, -1.17082], [-0.447214, -1.37638, 0.723607], [-0.447214, -1.37638, -0.276393], [-0.447214, 1.37638, 0.723607], [-0.447214, 1.37638, -0.276393], [0.447214, -1.37638, 0.276393], [0.447214, -1.37638, -0.723607], [0.447214, 1.37638, 0.276393], [0.447214, 1.37638, -0.723607], [-1.44721, 0.0, 0.723607], [-1.44721, 0.0, -0.276393], [-0.723607, -0.525731, 1.17082], [-0.723607, 0.525731, 1.17082], [0.723607, -0.525731, -1.17082], [0.723607, 0.525731, -1.17082], [1.44721, 0.0, 0.276393], [1.44721, 0.0, -0.723607], [-1.17082, -0.850651, 0.276393], [-1.17082, -0.850651, -0.723607], [-1.17082, 0.850651, 0.276393], [-1.17082, 0.850651, -0.723607], [-0.276393, -0.850651, -1.17082], [-0.276393, 0.850651, -1.17082]]}, "113": {"AlternativeName": null, "Comment": null, "Name": "Rhombicosidodecahedron", "ShortName": "A09", "vertices": [[-0.5, -0.5, -2.11803], [-0.5, -0.5, 2.11803], [-0.5, 0.5, -2.11803], [-0.5, 0.5, 2.11803], [-0.5, -2.11803, -0.5], [-0.5, -2.11803, 0.5], [-0.5, 2.11803, -0.5], [-0.5, 2.11803, 0.5], [-0.0, -1.30902, -1.80902], [-0.0, -1.30902, 1.80902], [-0.0, 1.30902, -1.80902], [-0.0, 1.30902, 1.80902], [0.5, -0.5, -2.11803], [0.5, -0.5, 2.11803], [0.5, 0.5, -2.11803], [0.5, 0.5, 2.11803], [0.5, -2.11803, -0.5], [0.5, -2.11803, 0.5], [0.5, 2.11803, -0.5], [0.5, 2.11803, 0.5], [-1.80902, 0.0, -1.30902], [-1.80902, 0.0, 1.30902], [-0.809017, -1.61803, -1.30902], [-0.809017, -1.61803, 1.30902], [-0.809017, 1.61803, -1.30902], [-0.809017, 1.61803, 1.30902], [-1.61803, -1.30902, -0.809017], [-1.61803, -1.30902, 0.809017], [-1.61803, 1.30902, -0.809017], [-1.61803, 1.30902, 0.809017], [-2.11803, -0.5, -0.5], [-2.11803, -0.5, 0.5], [-2.11803, 0.5, -0.5], [-2.11803, 0.5, 0.5], [-1.30902, -1.80902, 0.0], [-1.30902, -0.809017, -1.61803], [-1.30902, -0.809017, 1.61803], [-1.30902, 0.809017, -1.61803], [-1.30902, 0.809017, 1.61803], [-1.30902, 1.80902, 0.0], [0.809017, -1.61803, -1.30902], [0.809017, -1.61803, 1.30902], [0.809017, 1.61803, -1.30902], [0.809017, 1.61803, 1.30902], [1.61803, -1.30902, -0.809017], [1.61803, -1.30902, 0.809017], [1.61803, 1.30902, -0.809017], [1.61803, 1.30902, 0.809017], [2.11803, -0.5, -0.5], [2.11803, -0.5, 0.5], [2.11803, 0.5, -0.5], [2.11803, 0.5, 0.5], [1.30902, -1.80902, 0.0], [1.30902, -0.809017, -1.61803], [1.30902, -0.809017, 1.61803], [1.30902, 0.809017, -1.61803], [1.30902, 0.809017, 1.61803], [1.30902, 1.80902, 0.0], [1.80902, 0.0, -1.30902], [1.80902, 0.0, 1.30902]]}, "114": {"AlternativeName": null, "Comment": null, "Name": "Rhombicuboctahedron", "ShortName": "A08", "vertices": [[-0.5, -0.5, -1.20711], [-0.5, -0.5, 1.20711], [-0.5, 0.5, -1.20711], [-0.5, 0.5, 1.20711], [-0.5, -1.20711, -0.5], [-0.5, -1.20711, 0.5], [-0.5, 1.20711, -0.5], [-0.5, 1.20711, 0.5], [0.5, -0.5, -1.20711], [0.5, -0.5, 1.20711], [0.5, 0.5, -1.20711], [0.5, 0.5, 1.20711], [0.5, -1.20711, -0.5], [0.5, -1.20711, 0.5], [0.5, 1.20711, -0.5], [0.5, 1.20711, 0.5], [-1.20711, -0.5, -0.5], [-1.20711, -0.5, 0.5], [-1.20711, 0.5, -0.5], [-1.20711, 0.5, 0.5], [1.20711, -0.5, -0.5], [1.20711, -0.5, 0.5], [1.20711, 0.5, -0.5], [1.20711, 0.5, 0.5]]}, "115": {"AlternativeName": null, "Comment": null, "Name": "Triakis Octahedron", "ShortName": "C05", "vertices": [[-0.5, -0.5, -0.5], [-0.5, -0.5, 0.5], [-0.5, 0.5, -0.5], [-0.5, 0.5, 0.5], [0.0, 0.0, -1.20711], [0.0, 0.0, 1.20711], [0.0, -1.20711, 0.0], [0.0, 1.20711, 0.0], [0.5, -0.5, -0.5], [0.5, -0.5, 0.5], [0.5, 0.5, -0.5], [0.5, 0.5, 0.5], [-1.20711, 0.0, 0.0], [1.20711, 0.0, 0.0]]}, "116": {"AlternativeName": null, "Comment": null, "Name": "Snub Cuboctahedron", "ShortName": "A12", "vertices": [[-1.14261, -0.337754, -0.621226], [-1.14261, 0.337754, 0.621226], [-1.14261, -0.621226, 0.337754], [-1.14261, 0.621226, -0.337754], [1.14261, -0.337754, 0.621226], [1.14261, 0.337754, -0.621226], [1.14261, -0.621226, -0.337754], [1.14261, 0.621226, 0.337754], [-0.337754, -1.14261, 0.621226], [-0.337754, 1.14261, -0.621226], [-0.337754, -0.621226, -1.14261], [-0.337754, 0.621226, 1.14261], [0.337754, -1.14261, -0.621226], [0.337754, 1.14261, 0.621226], [0.337754, -0.621226, 1.14261], [0.337754, 0.621226, -1.14261], [-0.621226, -1.14261, -0.337754], [-0.621226, 1.14261, 0.337754], [-0.621226, -0.337754, 1.14261], [-0.621226, 0.337754, -1.14261], [0.621226, -1.14261, 0.337754], [0.621226, 1.14261, -0.337754], [0.621226, -0.337754, -1.14261], [0.621226, 0.337754, 1.14261]]}, "117": {"AlternativeName": null, "Comment": null, "Name": "Snub Disphenoid", "ShortName": "J84", "vertices": [[-0.5, 0.0, -0.78393025], [0.0, -0.5, 0.78392975], [0.0, 0.5, 0.78392975], [0.0, -0.644584, -0.20556125], [0.0, 0.644584, -0.20556125], [0.5, 0.0, -0.78393025], [-0.644584, 0.0, 0.20556175], [0.644584, 0.0, 0.20556175]]}, "118": {"AlternativeName": null, "Comment": null, "Name": "Snub Icosidodecahedron", "ShortName": "A13", "vertices": [[-2.05022, -0.64303, 0.175393], [2.05022, -0.64303, -0.175393], [-1.64507, 0.64303, 1.23608], [1.64507, 0.64303, -1.23608], [-2.09275, 0.330921, 0.398127], [2.09275, 0.330921, -0.398127], [-1.33296, 1.64692, -0.398127], [1.33296, 1.64692, 0.398127], [-1.82527, -0.330921, 1.09842], [1.82527, -0.330921, -1.09842], [-0.626047, 1.74619, -1.09842], [0.626047, 1.74619, 1.09842], [-1.06222, 1.45402, 1.18539], [1.06222, 1.45402, -1.18539], [-1.93214, 0.84755, -0.442882], [1.93214, 0.84755, 0.442882], [-1.14487, -0.84755, 1.6182], [1.14487, -0.84755, -1.6182], [-1.58199, -1.45402, -0.175393], [1.58199, -1.45402, 0.175393], [-1.05741, 0.374822, -1.84093], [1.05741, 0.374822, 1.84093], [-0.439138, -0.374822, -2.07709], [0.439138, -0.374822, 2.07709], [-1.56241, -1.2495, 0.803274], [1.56241, -1.2495, -0.803274], [-1.86331, -0.728335, -0.803274], [1.86331, -0.728335, 0.803274], [-1.70007, 1.2495, 0.442882], [1.70007, 1.2495, -0.442882], [-0.728114, -1.64692, 1.18539], [0.728114, -1.64692, -1.18539], [-0.265655, -1.74619, -1.23608], [0.265655, -1.74619, 1.23608], [-0.759791, -1.97784, -0.398127], [0.759791, -1.97784, 0.398127], [-1.19922, -1.41527, -1.09842], [1.19922, -1.41527, 1.09842], [-1.79033, 0.192894, -1.18539], [1.79033, 0.192894, 1.18539], [-1.30644, -0.567715, -1.6182], [1.30644, -0.567715, 1.6182], [-0.853311, 0.728335, 1.84093], [0.853311, 0.728335, -1.84093], [-1.37941, 1.10316, -1.23608], [1.37941, 1.10316, 1.23608], [-0.105036, 0.567715, -2.07709], [0.105036, 0.567715, 2.07709], [-0.468228, 2.09705, -0.175393], [0.468228, 2.09705, 0.175393], [-0.300897, 1.97784, 0.803274], [0.300897, 1.97784, -0.803274], [-0.161563, 1.41527, 1.6182], [0.161563, 1.41527, -1.6182], [-0.544174, -0.192894, 2.07709], [0.544174, -0.192894, -2.07709], [-0.232068, -2.09705, 0.442882], [0.232068, -2.09705, -0.442882], [-0.204101, -1.10316, 1.84093], [0.204101, -1.10316, -1.84093]]}, "119": {"AlternativeName": null, "Comment": null, "Name": "Snub Square Antiprism", "ShortName": "J85", "vertices": [[-1.093170494, 0.431564919, -0.353605944], [-1.005700494, -0.562189081, -0.422882944], [-0.859633494, -0.107437081, 0.455675056], [-0.719877494, 0.882747919, 0.457000056], [-0.363678494, 0.031303219, -0.908244944], [-0.309622494, -0.905009081, 0.207952056], [-0.197089494, 0.873472919, -0.395411944], [-0.149505494, -0.936855081, -0.778631944], [-0.074030594, 0.285999919, 0.933206056], [0.254707506, 1.106599919, 0.465710056], [0.475980506, -0.511572081, 0.685483056], [0.580741506, -0.254465081, -0.745745944], [0.614673506, -1.052490081, -0.144077944], [0.747331506, 0.587704919, -0.232912944], [0.900494506, 0.392447919, 0.735804056], [1.198379506, -0.261824081, 0.040684156]]}, "120": {"AlternativeName": null, "Comment": null, "Name": "Spenocorona", "ShortName": "J86", "vertices": [[0.0, -0.5, -0.6630428], [0.0, 0.5, -0.6630428], [-0.852727, 0.5, -0.1406858], [-0.5, 0.0, 0.6502572], [-0.852727, -0.5, -0.1406858], [0.852727, 0.5, -0.1406858], [0.5, 0.0, 0.6502572], [0.852727, -0.5, -0.1406858], [0.0, 0.789428, 0.2941572], [0.0, -0.789428, 0.2941572]]}, "121": {"AlternativeName": null, "Comment": null, "Name": "Sphenomegacorona", "ShortName": "J88", "vertices": [[-0.708649899, -0.296833667, -0.152242192], [-0.649328899, -0.105651667, 0.827517808], [-0.619595899, 0.646067333, 0.168705808], [-0.612442898, -1.049476667, 0.499124808], [-0.165930899, 0.360258333, -0.675392192], [-0.002052116, -0.990752667, -0.290796192], [0.165479102, 0.470573333, 0.762723808], [0.225204102, -0.506004667, 0.553818808], [0.278442101, 1.017633333, -0.066799092], [0.540667102, -0.333660667, -0.813946192], [0.767923102, 0.151087333, 0.030669008], [0.780285102, 0.636760333, -0.843384192]]}, "122": {"AlternativeName": null, "Comment": null, "Name": "Square Cupola", "ShortName": "J04", "vertices": [[0, -0.707107, 0.707107], [0, 0.707107, 0.707107], [-0.707107, 0, 0.707107], [0.707107, 0, 0.707107], [0.5, 1.20711, 0], [-0.5, 1.20711, 0], [-1.20711, 0.5, 0], [-1.20711, -0.5, 0], [-0.5, -1.20711, 0], [0.5, -1.20711, 0], [1.20711, -0.5, 0], [1.20711, 0.5, 0]]}, "123": {"AlternativeName": null, "Comment": null, "Name": "Square Gyrobicupola", "ShortName": "J29", "vertices": [[-0.5, -0.5, -0.707107], [-0.5, 0.5, -0.707107], [0.0, -0.707107, 0.707107], [0.0, 0.707107, 0.707107], [0.5, -0.5, -0.707107], [0.5, 0.5, -0.707107], [-0.707107, 0.0, 0.707107], [0.707107, 0.0, 0.707107], [0.5, 1.20711, 0.0], [-0.5, 1.20711, 0.0], [-1.20711, 0.5, 0.0], [-1.20711, -0.5, 0.0], [-0.5, -1.20711, 0.0], [0.5, -1.20711, 0.0], [1.20711, -0.5, 0.0], [1.20711, 0.5, 0.0]]}, "124": {"AlternativeName": null, "Comment": null, "Name": "Square Orthobicupola", "ShortName": "J28", "vertices": [[0.0, -0.707107, -0.707107], [0.0, -0.707107, 0.707107], [0.0, 0.707107, -0.707107], [0.0, 0.707107, 0.707107], [-0.707107, 0.0, -0.707107], [-0.707107, 0.0, 0.707107], [0.707107, 0.0, -0.707107], [0.707107, 0.0, 0.707107], [0.5, 1.20711, 0.0], [-0.5, 1.20711, 0.0], [-1.20711, 0.5, 0.0], [-1.20711, -0.5, 0.0], [-0.5, -1.20711, 0.0], [0.5, -1.20711, 0.0], [1.20711, -0.5, 0.0], [1.20711, 0.5, 0.0]]}, "125": {"AlternativeName": null, "Comment": null, "Name": "Squashed Dodecahedron", "ShortName": "O15", "vertices": [[-1.732050357, 3.57e-07, -3.57e-07], [-0.866025357, -0.866024643, -0.866025357], [-0.866025357, 0.866025357, -0.866025357], [-0.866025357, 0.866025357, 0.866024643], [-0.577350357, -1.154699643, 1.154699643], [-0.288675357, -1.443379643, -0.288675357], [-0.288675357, 0.288675357, 1.443379643], [-3.57e-07, 3.57e-07, -1.732050357], [-3.57e-07, 1.732050357, -3.57e-07], [0.866024643, -0.866024643, 0.866024643], [0.866024643, 0.866025357, -0.866025357], [1.154699643, -1.154699643, -0.577350357], [1.154699643, 0.577350357, 1.154699643], [1.443379643, 0.288675357, -0.288675357]]}, "126": {"AlternativeName": null, "Comment": null, "Name": "Tetrahedron", "ShortName": "P01", "vertices": [[0.0, 0.0, 0.612372], [-0.288675, -0.5, -0.204124], [-0.288675, 0.5, -0.204124], [0.57735, 0.0, -0.204124]]}, "127": {"AlternativeName": null, "Comment": null, "Name": "Tetrakis Hexahedron", "ShortName": "C04", "vertices": [[-0.942809, 0.0, -0.666667], [-0.942809, 0.0, 0.666667], [-0.707107, -0.707107, 0.0], [-0.707107, 0.707107, 0.0], [0.0, -0.942809, -0.666667], [0.0, -0.942809, 0.666667], [0.0, 0.0, -1.0], [0.0, 0.0, 1.0], [0.0, 0.942809, -0.666667], [0.0, 0.942809, 0.666667], [0.707107, -0.707107, 0.0], [0.707107, 0.707107, 0.0], [0.942809, 0.0, -0.666667], [0.942809, 0.0, 0.666667]]}, "128": {"AlternativeName": null, "Comment": null, "Name": "Triakis Icosahedron", "ShortName": "C07", "vertices": [[-0.0, 0.0, 1.63925], [0.262866, -0.809017, 1.11352], [0.262866, 0.809017, 1.11352], [0.850651, 0.0, 1.11352], [-0.0, 0.0, -1.63925], [-0.262866, -0.809017, -1.11352], [-0.262866, 0.809017, -1.11352], [0.688191, -0.5, -1.11352], [0.688191, 0.5, -1.11352], [-0.850651, 0.0, -1.11352], [-0.688191, -0.5, 1.11352], [-0.688191, 0.5, 1.11352], [-1.37638, 0.0, -0.262866], [1.11352, -0.809017, -0.262866], [1.11352, 0.809017, -0.262866], [-0.425325, -1.30902, -0.262866], [-0.425325, 1.30902, -0.262866], [1.37638, 0.0, 0.262866], [0.425325, -1.30902, 0.262866], [0.425325, 1.30902, 0.262866], [-1.11352, -0.809017, 0.262866], [-1.11352, 0.809017, 0.262866], [1.18617, -0.861803, 0.733094], [1.18617, 0.861803, 0.733094], [-1.46619, 0.0, 0.733094], [-0.453077, -1.39443, 0.733094], [-0.453077, 1.39443, 0.733094], [1.46619, 0.0, -0.733094], [-1.18617, -0.861803, -0.733094], [-1.18617, 0.861803, -0.733094], [0.453077, -1.39443, -0.733094], [0.453077, 1.39443, -0.733094]]}, "129": {"AlternativeName": null, "Comment": null, "Name": "Triakis Tetrahedron", "ShortName": "C03", "vertices": [[0.0, 0.0, -1.020620125], [0.0, 0.0, 0.612371875], [-0.96225, 0.0, 0.340206875], [-0.288675, -0.5, -0.204124125], [-0.288675, 0.5, -0.204124125], [0.481125, -0.833333, 0.340206875], [0.481125, 0.833333, 0.340206875], [0.57735, 0.0, -0.204124125]]}, "130": {"AlternativeName": null, "Comment": null, "Name": "Triangular Cupola", "ShortName": "J03", "vertices": [[0.0, -1.0, -0.272165667], [0.0, 1.0, -0.272165667], [-0.288675, -0.5, 0.544331333], [-0.288675, 0.5, 0.544331333], [0.57735, 0.0, 0.544331333], [-0.866025, -0.5, -0.272165667], [-0.866025, 0.5, -0.272165667], [0.866025, -0.5, -0.272165667], [0.866025, 0.5, -0.272165667]]}, "131": {"AlternativeName": null, "Comment": null, "Name": "Triangular Hebesphenorotunda", "ShortName": "J92", "vertices": [[-0.960178539, 0.715213333, -0.038936944], [-0.818773539, 0.161289333, -0.859408644], [-0.761159539, 0.332418333, 0.863204356], [-0.547987539, 1.123904333, -0.853225644], [-0.478350539, -0.775429667, -0.777738644], [-0.420736539, -0.604300667, 0.944875356], [-0.279332539, -1.158225667, 0.124403356], [-0.181614539, 1.336044333, 0.052737256], [-0.026950639, 0.121738333, 1.508613356], [0.017403761, 0.953251333, 0.954879356], [0.047182961, 0.439778333, -1.274816644], [0.343251461, -1.328455667, -0.639410644], [0.387606461, -0.496940667, -1.193146644], [0.568220461, -0.562393667, 1.087023356], [0.639988461, 0.783023333, 0.191065356], [0.709625461, -1.116315667, 0.266553356], [0.781393461, 0.229098333, -0.629407644], [0.980411461, -0.153697667, 0.272736356]]}, "132": {"AlternativeName": null, "Comment": null, "Name": "Triangular Orthobicupola", "ShortName": "J27", "vertices": [[0.0, -1.0, 0.0], [0.0, 1.0, 0.0], [-0.288675, -0.5, -0.816497], [-0.288675, -0.5, 0.816497], [-0.288675, 0.5, -0.816497], [-0.288675, 0.5, 0.816497], [0.57735, 0.0, -0.816497], [0.57735, 0.0, 0.816497], [-0.866025, -0.5, 0.0], [-0.866025, 0.5, 0.0], [0.866025, -0.5, 0.0], [0.866025, 0.5, 0.0]]}, "133": {"AlternativeName": null, "Comment": null, "Name": "Triaugmented Dodecahedron", "ShortName": "J61", "vertices": [[-0.026740043, 0.0, 1.371363478], [-0.026740043, 0.0, -1.431156522], [0.151670957, -1.30902, 0.437189478], [0.151670957, 1.30902, 0.437189478], [0.440345957, -0.809017, -1.074336522], [0.440345957, 0.809017, -1.074336522], [1.017699957, -0.809017, 0.437189478], [1.017699957, 0.809017, 0.437189478], [-1.249590043, -0.5, 0.437189478], [-1.249590043, 0.5, 0.437189478], [1.196109957, -0.5, -0.496982522], [1.196109957, 0.5, -0.496982522], [-0.960912043, 0.0, -1.074336522], [-0.493826043, -0.809017, 1.014543478], [-0.493826043, 0.809017, 1.014543478], [0.907431957, 0.0, 1.014543478], [-1.071180043, -0.809017, -0.496982522], [-1.071180043, 0.809017, -0.496982522], [-1.021865043, 0.0, 1.272743478], [-0.205151043, -1.30902, -0.496982522], [-0.205151043, 1.30902, -0.496982522], [0.778332957, 1.39443, -0.337406522], [0.778332957, -1.39443, -0.337406522]]}, "134": {"AlternativeName": null, "Comment": null, "Name": "Triaugmented Hexagonal Prism", "ShortName": "J57", "vertices": [[-1.0, -1.33e-07, -0.5], [-1.0, -1.33e-07, 0.5], [-0.5, -0.866025133, -0.5], [-0.5, -0.866025133, 0.5], [-0.5, 0.866024867, -0.5], [-0.5, 0.866024867, 0.5], [0.0, -1.573130133, 0.0], [0.5, -0.866025133, -0.5], [0.5, -0.866025133, 0.5], [0.5, 0.866024867, -0.5], [0.5, 0.866024867, 0.5], [1.0, -1.33e-07, -0.5], [1.0, -1.33e-07, 0.5], [-1.36237, 0.786565867, 0.0], [1.36237, 0.786565867, 0.0]]}, "135": {"AlternativeName": null, "Comment": null, "Name": "Triaugmented Triangular Prism", "ShortName": "J51", "vertices": [[-0.288675, -0.5, -0.5], [-0.288675, -0.5, 0.5], [-0.288675, 0.5, -0.5], [-0.288675, 0.5, 0.5], [0.57735, 0.0, -0.5], [0.57735, 0.0, 0.5], [-0.995782, 0.0, 0.0], [0.497891, -0.862372, 0.0], [0.497891, 0.862372, 0.0]]}, "136": {"AlternativeName": null, "Comment": null, "Name": "Triaugmented Truncated Dodecahedron", "ShortName": "J71", "vertices": [[-2.829466198, 0.210868048, -0.449141693], [-2.721906198, 0.967199048, 0.196144307], [-2.691866198, -0.771698952, -0.574157693], [-2.556136198, 1.114481048, -0.778961693], [-2.410266198, 1.208401048, 1.115224307], [-2.408266198, 1.896771048, 0.389863307], [-2.361666198, -1.605198952, -0.131152693], [-2.242496198, 2.044051048, -0.585242693], [-2.195906198, -1.457908952, -1.106255693], [-2.013596198, 0.842345048, 1.957034307], [-1.976296198, 1.593991048, -1.437635693], [-1.964986198, -1.971248952, 0.710661307], [-1.740266198, 1.745961048, 1.627214307], [-1.738266198, 2.434321048, 0.901857307], [-1.683396198, 0.008850038, 2.400044307], [-1.653356198, -1.730048952, 1.629744307], [-1.545796198, -0.973716952, 2.275024307], [-1.531016198, -1.585668952, -1.842195693], [-1.470056198, 2.672631048, -0.675896693], [-1.311416198, 1.466231048, -2.173575693], [-1.203856198, 2.222561048, -1.528295693], [-1.158416198, 2.913831048, 0.243181307], [-1.157396198, -2.416258952, 1.097634307], [-0.967824198, 2.374531048, 1.536564307], [-0.953109198, -1.776348952, 2.342134307], [-0.951175198, -1.106158952, -2.500875693], [-0.875795198, -0.436161952, 2.787024307], [-0.815451198, 0.780019048, -2.705675693], [-0.677853198, -0.202547952, -2.830695693], [-0.620976198, -1.939658952, -2.057865693], [-0.533854198, 2.760121048, -1.016295693], [-0.457147198, -2.462568952, 1.810034307], [-0.387981198, 2.854041048, 0.877885307], [-0.283111198, -1.238798952, 2.854134307], [-0.247349198, -2.770248952, 0.881968307], [-0.222219198, 3.001321048, -0.097220693], [0.008694472, 2.487981048, 1.719694307], [0.094591702, 0.426031048, -2.921345693], [0.100723602, -0.322711952, 2.970154307], [0.186620802, -2.384668952, -1.670895693], [0.417534802, -2.898008952, 0.146027307], [0.442664802, 2.873571048, -0.833161693], [0.519371802, -2.349118952, 1.993174307], [0.583296802, -2.750728952, -0.829077693], [0.626931802, -1.592788952, 2.638464307], [0.729169802, -2.656798952, 1.065104307], [0.816290802, 2.042971048, 2.106674307], [0.873167802, 0.305866048, 2.879504307], [1.010766802, -0.676699952, 2.754484307], [1.071113802, 0.539481048, -2.738215693], [1.146493802, 1.209481048, 2.549684307], [1.163143802, -2.271218952, -1.487755693], [1.352703802, 2.519581048, -1.048835693], [1.399163802, -2.119248952, 1.577104307], [1.506723802, -1.362918952, 2.222384307], [1.726333802, 1.688981048, 1.891004307], [1.741113802, 1.077031048, -2.226215693], [-0.056852698, -0.980022952, -2.930145693], [1.848673802, 1.833371048, -1.580935693], [1.878703802, 0.094468448, -2.351235693], [1.935583802, -1.642638952, -1.578405693], [0.715591802, -0.351444952, -3.020805693], [2.160303802, 2.074571048, -0.661854693], [2.171613802, -1.490668952, 1.486444307], [2.208903802, -0.739026952, -1.908225693], [2.391213802, 1.561231048, 1.155064307], [0.273346802, -1.813518952, -2.487145693], [2.556983802, 1.708511048, 0.179959307], [2.605583802, -1.105078952, -1.066415693], [1.523183802, -0.796456952, -2.633825693], [2.751453802, -1.011158952, 0.827769307], [2.887183802, 0.875017048, 0.622965307], [1.249863802, -1.700068952, -2.304005693], [2.917213802, -0.863880952, -0.147336693], [3.024773802, -0.107549952, 0.497948307]]}, "137": {"AlternativeName": null, "Comment": null, "Name": "Tridiminished Icosahedron", "ShortName": "J63", "vertices": [[-0.57735, 0.0, -0.815231333], [-0.288675, -0.5, 0.696290667], [-0.288675, 0.5, 0.696290667], [0.288675, -0.5, -0.815231333], [0.288675, 0.5, -0.815231333], [0.57735, 0.0, 0.696290667], [0.467086, -0.809017, 0.118940667], [0.467086, 0.809017, 0.118940667], [-0.934172, 0.0, 0.118940667]]}, "138": {"AlternativeName": null, "Comment": null, "Name": "Tridiminished Rhombicosidodecahedron", "ShortName": "J83", "vertices": [[-2.079718862, 0.421930022, 0.428667444], [-2.056218862, 0.538149022, -0.564278556], [-2.033028862, -0.557217978, 0.626379444], [-1.995018862, -0.369170978, -0.980239556], [-1.980688862, -1.046144978, -0.244372556], [-1.603218862, 1.293775022, 0.541985444], [-1.588888862, 0.616801022, 1.277854444], [-1.579728862, 1.409995022, -0.450960556], [-1.542208862, -0.362346978, 1.475564444], [-1.419508862, -0.965401978, -1.539965556], [-1.405168862, -1.642374978, -0.804094556], [-0.785559862, 1.725305022, 0.923051444], [-0.771227862, 1.048325022, 1.658914444], [-0.747548862, 1.913345022, -0.683568556], [-0.695694862, -0.535965978, 1.978824444], [-0.549496862, -1.022801978, -2.029645556], [-0.526307862, -2.118164978, -0.838989556], [-0.256723862, 2.108215022, 0.165617444], [-0.219201862, 0.335879022, 2.092144444], [0.002528728, -1.735254978, -1.596425556], [0.060954338, 1.551685022, 1.426314444], [0.122458138, 1.855945022, -1.173255556], [0.183170138, -1.011758978, 1.943924444], [0.282685138, -0.519447978, -2.262255556], [0.320207138, -2.291784978, -0.335730556], [0.589790138, 1.934605022, 0.668875444], [0.612980138, 0.839233022, 1.859534444], [0.613283138, 2.050815022, -0.324069556], [0.659662138, -0.139913978, 2.057244444], [0.697976138, 1.259715022, -1.732975556], [0.758688138, -1.607984978, 1.384204444], [0.759178138, 0.352398022, -2.148935556], [0.811032138, -2.096914978, 0.513454444], [0.834711138, -1.231894978, -1.829035556], [0.849043138, -1.908874978, -1.093165556], [1.468651138, 1.458805022, 0.633980444], [1.482991138, 0.781833022, 1.369844444], [1.492151138, 1.575025022, -0.358965556], [1.529671138, -0.197314978, 1.567564444], [1.544491138, 1.086095022, -1.229715556], [1.590871138, -1.104634978, 1.151594444], [1.605691138, 0.178778022, -1.645675556], [1.643211138, -1.593564978, 0.280846444], [1.652371138, -0.800369978, -1.447965556], [1.666711138, -1.477344978, -0.712099556]]}, "139": {"AlternativeName": null, "Comment": null, "Name": "Trigyrate Rhombicosidodecahedron", "ShortName": "J75", "vertices": [[-2.189130133, -0.439999098, -0.013079717], [-2.113110133, 0.336730902, -0.638307917], [-2.089600133, 0.178000902, 0.766773083], [-2.013580133, 0.954731902, 0.141544083], [-1.813430133, -1.189070098, 0.532579083], [-1.797080133, -1.240750098, -0.465950917], [-1.713910133, -0.571070098, 1.312430083], [-1.695260133, 1.236569902, -0.763573917], [-1.674080133, 0.016025502, -1.477589917], [-1.536510133, 0.377194902, 1.575730083], [-1.478760133, -0.958912098, -1.371069917], [-1.413510133, 1.633969902, 0.564089083], [-1.256230133, 0.915864902, -1.602859917], [-1.118660133, 1.277029902, 1.450460083], [-1.095190133, 1.915809902, -0.341028917], [-1.079550133, -1.866710098, 0.579671083], [-1.063200133, -1.918390098, -0.418858917], [-0.918512133, -0.866768098, 1.841500083], [-0.864176133, 0.115112902, -2.055729917], [-0.744877133, -1.636560098, -1.323979917], [-0.741119133, 0.081497002, 2.104800083], [-0.668857133, -0.859825098, -1.949199917], [-0.548585133, 1.938449902, 0.963084083], [-0.526460133, -1.667520098, 1.388630083], [-0.384815133, 1.396899902, -1.699019917], [-0.323269133, 0.981335902, 1.979530083], [-0.285287133, 2.014899902, -0.919165917], [-0.267805133, -2.214090098, 0.110209083], [-0.230264133, 2.220289902, 0.057966183], [-0.007237073, -0.596146098, 2.151890083], [0.007236797, 0.596145902, -2.151889917], [0.246808867, 1.642749902, 1.492150083], [0.247249867, -1.758070098, -1.354299917], [0.285286867, -2.014900098, 0.919166083], [0.323268867, -0.981336098, -1.979529917], [0.384814867, -1.396900098, 1.699020083], [0.542096867, -2.115000098, -0.467927917], [0.607311867, 1.275389902, -1.729339917], [0.668856867, 0.859823902, 1.949200083], [0.706839867, 1.893389902, -0.949491917], [0.741118867, -0.081497298, -2.104799917], [0.761862867, 2.098779902, 0.027639883], [0.864175867, -0.115113098, 2.055730083], [1.056709867, 1.741839902, 0.914015083], [1.095189867, -1.915810098, 0.341029083], [1.179149867, -1.590370098, -1.032669917], [1.255159867, -0.813642098, -1.657899917], [1.256229867, -0.915865098, 1.602860083], [1.341189867, 0.597742902, -1.682249917], [1.478759867, 0.958911902, 1.371070083], [1.502229867, 1.597689902, -0.420424917], [1.674079867, -0.016025698, 1.477590083], [1.695259867, -1.236570098, 0.763574083], [1.732239867, -1.391180098, -0.223709917], [1.797079867, 1.240749902, 0.465951083], [1.855239867, -0.134402098, -1.235349917], [1.894279867, 0.796936902, -0.873294917], [2.113109867, -0.336731098, 0.638308083], [2.150089867, -0.491339098, -0.348975917], [2.189129867, 0.439998902, 0.013079983]]}, "140": {"AlternativeName": null, "Comment": null, "Name": "Truncated Cube", "ShortName": "A05", "vertices": [[-0.5, 1.20711, 1.20711], [-0.5, 1.20711, -1.20711], [-0.5, -1.20711, 1.20711], [-0.5, -1.20711, -1.20711], [0.5, 1.20711, 1.20711], [0.5, 1.20711, -1.20711], [0.5, -1.20711, 1.20711], [0.5, -1.20711, -1.20711], [1.20711, -0.5, 1.20711], [1.20711, -0.5, -1.20711], [1.20711, 0.5, 1.20711], [1.20711, 0.5, -1.20711], [1.20711, 1.20711, -0.5], [1.20711, 1.20711, 0.5], [1.20711, -1.20711, -0.5], [1.20711, -1.20711, 0.5], [-1.20711, -0.5, 1.20711], [-1.20711, -0.5, -1.20711], [-1.20711, 0.5, 1.20711], [-1.20711, 0.5, -1.20711], [-1.20711, 1.20711, -0.5], [-1.20711, 1.20711, 0.5], [-1.20711, -1.20711, -0.5], [-1.20711, -1.20711, 0.5]]}, "141": {"AlternativeName": null, "Comment": null, "Name": "Truncated Dodecahedron", "ShortName": "A07", "vertices": [[-0.0, -1.61803, 2.4899], [-0.0, -1.61803, -2.4899], [-0.0, 1.61803, 2.4899], [-0.0, 1.61803, -2.4899], [0.425325, -2.92705, 0.262866], [0.425325, 2.92705, 0.262866], [0.688191, -2.11803, 1.96417], [0.688191, 2.11803, 1.96417], [-2.75276, 0.0, -1.11352], [-2.06457, -2.11803, 0.262866], [-2.06457, 2.11803, 0.262866], [-1.37638, -2.61803, -0.262866], [-1.37638, 2.61803, -0.262866], [-0.688191, -2.11803, -1.96417], [-0.688191, 2.11803, -1.96417], [1.37638, -2.61803, 0.262866], [1.37638, 2.61803, 0.262866], [2.75276, 0.0, 1.11352], [1.80171, -1.30902, -1.96417], [1.80171, 1.30902, -1.96417], [2.06457, -2.11803, -0.262866], [2.06457, 2.11803, -0.262866], [2.22703, 0.0, 1.96417], [2.22703, -1.61803, -1.11352], [2.22703, 1.61803, -1.11352], [-2.65236, -1.30902, 0.262866], [-2.65236, 1.30902, 0.262866], [2.65236, -1.30902, -0.262866], [2.65236, 1.30902, -0.262866], [2.91522, -0.5, 0.262866], [2.91522, 0.5, 0.262866], [-2.91522, -0.5, -0.262866], [-2.91522, 0.5, -0.262866], [0.951057, -1.30902, 2.4899], [0.951057, -1.30902, -2.4899], [0.951057, 1.30902, 2.4899], [0.951057, 1.30902, -2.4899], [0.850651, -2.61803, 1.11352], [0.850651, 2.61803, 1.11352], [-0.951057, -1.30902, 2.4899], [-0.951057, -1.30902, -2.4899], [-0.951057, 1.30902, 2.4899], [-0.951057, 1.30902, -2.4899], [-1.53884, -0.5, 2.4899], [-1.53884, -0.5, -2.4899], [-1.53884, 0.5, 2.4899], [-1.53884, 0.5, -2.4899], [1.53884, -0.5, 2.4899], [1.53884, -0.5, -2.4899], [1.53884, 0.5, 2.4899], [1.53884, 0.5, -2.4899], [-2.22703, 0.0, -1.96417], [-2.22703, -1.61803, 1.11352], [-2.22703, 1.61803, 1.11352], [-0.850651, -2.61803, -1.11352], [-0.850651, 2.61803, -1.11352], [-1.80171, -1.30902, 1.96417], [-1.80171, 1.30902, 1.96417], [-0.425325, -2.92705, -0.262866], [-0.425325, 2.92705, -0.262866]]}, "142": {"AlternativeName": null, "Comment": null, "Name": "Truncated Icosahedron", "ShortName": "A06", "vertices": [[-0.16246, -2.11803, 1.27598], [-0.16246, 2.11803, 1.27598], [0.16246, -2.11803, -1.27598], [0.16246, 2.11803, -1.27598], [-0.262866, -0.809017, -2.32744], [-0.262866, -2.42705, -0.425325], [-0.262866, 0.809017, -2.32744], [-0.262866, 2.42705, -0.425325], [0.262866, -0.809017, 2.32744], [0.262866, -2.42705, 0.425325], [0.262866, 0.809017, 2.32744], [0.262866, 2.42705, 0.425325], [0.688191, -0.5, -2.32744], [0.688191, 0.5, -2.32744], [1.21392, -2.11803, 0.425325], [1.21392, 2.11803, 0.425325], [-2.06457, -0.5, 1.27598], [-2.06457, 0.5, 1.27598], [-1.37638, -1.0, 1.80171], [-1.37638, 1.0, 1.80171], [-1.37638, -1.61803, -1.27598], [-1.37638, 1.61803, -1.27598], [-0.688191, -0.5, 2.32744], [-0.688191, 0.5, 2.32744], [1.37638, -1.0, -1.80171], [1.37638, 1.0, -1.80171], [1.37638, -1.61803, 1.27598], [1.37638, 1.61803, 1.27598], [-1.7013, 0.0, -1.80171], [1.7013, 0.0, 1.80171], [-1.21392, -2.11803, -0.425325], [-1.21392, 2.11803, -0.425325], [-1.96417, -0.809017, -1.27598], [-1.96417, 0.809017, -1.27598], [2.06457, -0.5, -1.27598], [2.06457, 0.5, -1.27598], [2.22703, -1.0, -0.425325], [2.22703, 1.0, -0.425325], [2.38949, -0.5, 0.425325], [2.38949, 0.5, 0.425325], [-1.11352, -1.80902, 1.27598], [-1.11352, 1.80902, 1.27598], [1.11352, -1.80902, -1.27598], [1.11352, 1.80902, -1.27598], [-2.38949, -0.5, -0.425325], [-2.38949, 0.5, -0.425325], [-1.63925, -1.80902, 0.425325], [-1.63925, 1.80902, 0.425325], [1.63925, -1.80902, -0.425325], [1.63925, 1.80902, -0.425325], [1.96417, -0.809017, 1.27598], [1.96417, 0.809017, 1.27598], [0.850651, 0.0, 2.32744], [-2.22703, -1.0, 0.425325], [-2.22703, 1.0, 0.425325], [-0.850651, 0.0, -2.32744], [-0.525731, -1.61803, -1.80171], [-0.525731, 1.61803, -1.80171], [0.525731, -1.61803, 1.80171], [0.525731, 1.61803, 1.80171]]}, "143": {"AlternativeName": null, "Comment": null, "Name": "Truncated Octahedron", "ShortName": "A04", "vertices": [[-1.5, -0.5, 0.0], [-1.5, 0.5, 0.0], [-1.0, -1.0, -0.707107], [-1.0, -1.0, 0.707107], [-1.0, 1.0, -0.707107], [-1.0, 1.0, 0.707107], [-0.5, -1.5, 0.0], [-0.5, -0.5, -1.41421], [-0.5, -0.5, 1.41421], [-0.5, 0.5, -1.41421], [-0.5, 0.5, 1.41421], [-0.5, 1.5, 0.0], [0.5, -1.5, 0.0], [0.5, -0.5, -1.41421], [0.5, -0.5, 1.41421], [0.5, 0.5, -1.41421], [0.5, 0.5, 1.41421], [0.5, 1.5, 0.0], [1.0, -1.0, -0.707107], [1.0, -1.0, 0.707107], [1.0, 1.0, -0.707107], [1.0, 1.0, 0.707107], [1.5, -0.5, 0.0], [1.5, 0.5, 0.0]]}, "144": {"AlternativeName": null, "Comment": null, "Name": "Truncated Tetrahedron", "ShortName": "A03", "vertices": [[0, -1, -0.612372], [0, 1, -0.612372], [-0.57735, -1, 0.204124], [-0.57735, 1, 0.204124], [-0.288675, -0.5, 1.02062], [-0.288675, 0.5, 1.02062], [0.57735, 0, 1.02062], [1.1547, 0, 0.204124], [-0.866025, -0.5, -0.612372], [-0.866025, 0.5, -0.612372], [0.866025, -0.5, -0.612372], [0.866025, 0.5, -0.612372]]}}"""
+
 
 def _damasceno_science_2012():
     "Parse the polyhedra shape data from the json data."
     data = json.loads(DAMASCENO_SCIENCE_2012)
     for shape_id in sorted((int(k) for k in data.keys())):
         yield shape_id, data[str(shape_id)]
+
 
 SHAPES = \
     [SPHERE] +\
@@ -32,6 +37,8 @@ def _sanity_check(shapes):
     "Ensure integrity of the shapes data."
     for i, shape in enumerate(shapes):
         assert shape.id == i
+
+
 _sanity_check(SHAPES)
 
 
@@ -60,3 +67,11 @@ def hpmc_shape_param(shape, *args, **kwargs):
         return {'vertices': shape.vertices}
     else:
         return dict()
+
+
+DAMASCENO_FIX = '{"J18": [1, 0, 0, 0], "A08": [1, 0, 0, 0], "O15": [0.7045564242640256, 0.4557687746359653, 0.060002041867676825, 0.540624476015312], "J45": [0.9569403357322088, 0.0, 0.0, 0.29028467725446233], "A12": [1, 0, 0, 0], "A01": [0.9238795325112867, -0.0, -0.0, 0.3826834323650897], "J61": [5.720154115573641e-17, 0.3568228725033438, 2.1849099433689207e-17, 0.9341720599859871], "J90": [0.19591344291270588, 0.6039888920172233, -0.18730394955680352, 0.7494882065024902], "J27": [1, 0, 0, 0], "C04": [0.9238795325112867, -0.0, -0.0, 0.3826834323650897], "C01": [0.9238795325112867, -0.0, -0.0, 0.3826834323650897], "J60": [0.1467739392257901, -0.3946011110748078, -0.7955911548080715, -0.4356398607724401], "J37": [0.9807852804032304, 0.0, 0.0, 0.19509032201612825], "P05": [1, 0, 0, 0], "P04": [1, 0, 0, 0]}'
+
+# This is a json loadable string that has shapes in the proper orientations as
+# well as their invariant quaternions
+
+SYMSHAPES = '[{"Name": "Rhombic Dodecahedron", "ShortName": "C01", "vertices": [[-1.1547011310369448, 2.7755575615628914e-17, 0.0], [-0.5773505655184724, -0.5773499999999998, -0.5773505655184725], [-0.5773505655184724, 0.57735, -0.5773505655184723], [2.7755575615628914e-17, 2.7755575615628914e-16, -1.1547011310369448], [-0.5773505655184724, -0.5773500000000001, 0.5773505655184724], [-0.5773505655184724, 0.5773499999999998, 0.5773505655184725], [0.0, -1.1547, -1.6653345369377348e-16], [0.0, 1.1547, 1.6653345369377348e-16], [0.5773505655184724, -0.5773499999999998, -0.5773505655184725], [0.5773505655184724, 0.5773500000000001, -0.5773505655184724], [-2.7755575615628914e-17, -2.7755575615628914e-16, 1.1547011310369448], [0.5773505655184724, -0.57735, 0.5773505655184723], [0.5773505655184724, 0.5773499999999998, 0.5773505655184725], [1.1547011310369448, -2.7755575615628914e-17, 0.0]], "invariantQ": [[-1.0, 7.07050159149938e-17, 7.07050159149938e-17, 7.07050159149938e-17], [1.0, -7.07050159149938e-17, -7.07050159149938e-17, -7.07050159149938e-17], [6.123233995736766e-17, 0.0, 0.0, 1.0], [-6.123233995736766e-17, 0.0, 0.0, -1.0], [6.123233995736766e-17, 0.0, 1.0, 0.0], [-6.123233995736766e-17, 0.0, -1.0, 0.0], [6.123233995736766e-17, 1.0, 0.0, 0.0], [-6.123233995736766e-17, -1.0, 0.0, 0.0], [0.5000000000000001, 0.5, 0.5, 0.5], [-0.5000000000000001, -0.5, -0.5, -0.5], [0.5000000000000001, -0.5, 0.5, -0.5], [-0.5000000000000001, 0.5, -0.5, 0.5], [0.5000000000000001, 0.5, -0.5, -0.5], [-0.5000000000000001, -0.5, 0.5, 0.5], [0.5000000000000001, -0.5, -0.5, 0.5], [-0.5000000000000001, 0.5, 0.5, -0.5], [0.5000000000000001, -0.5, -0.5, -0.5], [-0.5000000000000001, 0.5, 0.5, 0.5], [0.5000000000000001, -0.5, 0.5, 0.5], [-0.5000000000000001, 0.5, -0.5, -0.5], [0.5000000000000001, 0.5, 0.5, -0.5], [-0.5000000000000001, -0.5, -0.5, 0.5], [0.5000000000000001, 0.5, -0.5, 0.5], [-0.5000000000000001, -0.5, 0.5, -0.5], [6.123233995736766e-17, 0.7071067811865475, 0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, -0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, 0.7071067811865475, 0.0], [0.7071067811865476, -0.0, -0.0, -0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, -0.7071067811865475], [0.7071067811865476, -0.7071067811865475, -0.0, -0.0], [-0.7071067811865476, 0.7071067811865475, 0.0, 0.0], [6.123233995736766e-17, 0.0, 0.7071067811865475, 0.7071067811865475], [-6.123233995736766e-17, 0.0, -0.7071067811865475, -0.7071067811865475], [6.123233995736766e-17, 0.0, 0.7071067811865475, -0.7071067811865475], [-6.123233995736766e-17, 0.0, -0.7071067811865475, 0.7071067811865475], [0.7071067811865476, 0.7071067811865475, 0.0, 0.0], [-0.7071067811865476, -0.7071067811865475, 0.0, 0.0], [0.7071067811865476, 0.0, 0.7071067811865475, 0.0], [-0.7071067811865476, 0.0, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, 0.0, 0.7071067811865475], [-6.123233995736766e-17, -0.7071067811865475, 0.0, -0.7071067811865475], [0.7071067811865476, -0.0, -0.7071067811865475, -0.0], [-0.7071067811865476, 0.0, 0.7071067811865475, 0.0], [6.123233995736766e-17, -0.7071067811865475, 0.0, 0.7071067811865475], [-6.123233995736766e-17, 0.7071067811865475, 0.0, -0.7071067811865475]], "symGroup": "m-3m"}, {"Name": "Rhombicuboctahedron", "ShortName": "A08", "vertices": [[-0.5, -0.5, -1.20711], [-0.5, -0.5, 1.20711], [-0.5, 0.5, -1.20711], [-0.5, 0.5, 1.20711], [-0.5, -1.20711, -0.5], [-0.5, -1.20711, 0.5], [-0.5, 1.20711, -0.5], [-0.5, 1.20711, 0.5], [0.5, -0.5, -1.20711], [0.5, -0.5, 1.20711], [0.5, 0.5, -1.20711], [0.5, 0.5, 1.20711], [0.5, -1.20711, -0.5], [0.5, -1.20711, 0.5], [0.5, 1.20711, -0.5], [0.5, 1.20711, 0.5], [-1.20711, -0.5, -0.5], [-1.20711, -0.5, 0.5], [-1.20711, 0.5, -0.5], [-1.20711, 0.5, 0.5], [1.20711, -0.5, -0.5], [1.20711, -0.5, 0.5], [1.20711, 0.5, -0.5], [1.20711, 0.5, 0.5]], "invariantQ": [[-1.0, 7.07050159149938e-17, 7.07050159149938e-17, 7.07050159149938e-17], [1.0, -7.07050159149938e-17, -7.07050159149938e-17, -7.07050159149938e-17], [6.123233995736766e-17, 0.0, 0.0, 1.0], [-6.123233995736766e-17, 0.0, 0.0, -1.0], [6.123233995736766e-17, 0.0, 1.0, 0.0], [-6.123233995736766e-17, 0.0, -1.0, 0.0], [6.123233995736766e-17, 1.0, 0.0, 0.0], [-6.123233995736766e-17, -1.0, 0.0, 0.0], [0.5000000000000001, 0.5, 0.5, 0.5], [-0.5000000000000001, -0.5, -0.5, -0.5], [0.5000000000000001, -0.5, 0.5, -0.5], [-0.5000000000000001, 0.5, -0.5, 0.5], [0.5000000000000001, 0.5, -0.5, -0.5], [-0.5000000000000001, -0.5, 0.5, 0.5], [0.5000000000000001, -0.5, -0.5, 0.5], [-0.5000000000000001, 0.5, 0.5, -0.5], [0.5000000000000001, -0.5, -0.5, -0.5], [-0.5000000000000001, 0.5, 0.5, 0.5], [0.5000000000000001, -0.5, 0.5, 0.5], [-0.5000000000000001, 0.5, -0.5, -0.5], [0.5000000000000001, 0.5, 0.5, -0.5], [-0.5000000000000001, -0.5, -0.5, 0.5], [0.5000000000000001, 0.5, -0.5, 0.5], [-0.5000000000000001, -0.5, 0.5, -0.5], [6.123233995736766e-17, 0.7071067811865475, 0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, -0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, 0.7071067811865475, 0.0], [0.7071067811865476, -0.0, -0.0, -0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, -0.7071067811865475], [0.7071067811865476, -0.7071067811865475, -0.0, -0.0], [-0.7071067811865476, 0.7071067811865475, 0.0, 0.0], [6.123233995736766e-17, 0.0, 0.7071067811865475, 0.7071067811865475], [-6.123233995736766e-17, 0.0, -0.7071067811865475, -0.7071067811865475], [6.123233995736766e-17, 0.0, 0.7071067811865475, -0.7071067811865475], [-6.123233995736766e-17, 0.0, -0.7071067811865475, 0.7071067811865475], [0.7071067811865476, 0.7071067811865475, 0.0, 0.0], [-0.7071067811865476, -0.7071067811865475, 0.0, 0.0], [0.7071067811865476, 0.0, 0.7071067811865475, 0.0], [-0.7071067811865476, 0.0, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, 0.0, 0.7071067811865475], [-6.123233995736766e-17, -0.7071067811865475, 0.0, -0.7071067811865475], [0.7071067811865476, -0.0, -0.7071067811865475, -0.0], [-0.7071067811865476, 0.0, 0.7071067811865475, 0.0], [6.123233995736766e-17, -0.7071067811865475, 0.0, 0.7071067811865475], [-6.123233995736766e-17, 0.7071067811865475, 0.0, -0.7071067811865475]], "symGroup": "m-3m"}, {"Name": "Cuboctahedron", "ShortName": "A01", "vertices": [[-0.7071067811865475, 1.1102230246251565e-16, -0.7071067811865475], [-0.7071067811865476, -0.7071069999999999, -1.1102230246251565e-16], [-0.7071067811865475, 0.707107, 1.1102230246251565e-16], [-1.434097586361965e-17, -0.7071069999999998, -0.7071067811865476], [0.0, 0.707107, -0.7071067811865474], [-0.7071067811865475, -1.1102230246251565e-16, 0.7071067811865475], [0.7071067811865475, 1.1102230246251565e-16, -0.7071067811865475], [0.0, -0.707107, 0.7071067811865474], [1.434097586361965e-17, 0.7071069999999998, 0.7071067811865476], [0.7071067811865475, -0.707107, -1.1102230246251565e-16], [0.7071067811865476, 0.7071069999999999, 1.1102230246251565e-16], [0.7071067811865475, -1.1102230246251565e-16, 0.7071067811865475]], "invariantQ": [[-1.0, 7.07050159149938e-17, 7.07050159149938e-17, 7.07050159149938e-17], [1.0, -7.07050159149938e-17, -7.07050159149938e-17, -7.07050159149938e-17], [6.123233995736766e-17, 0.0, 0.0, 1.0], [-6.123233995736766e-17, 0.0, 0.0, -1.0], [6.123233995736766e-17, 0.0, 1.0, 0.0], [-6.123233995736766e-17, 0.0, -1.0, 0.0], [6.123233995736766e-17, 1.0, 0.0, 0.0], [-6.123233995736766e-17, -1.0, 0.0, 0.0], [0.5000000000000001, 0.5, 0.5, 0.5], [-0.5000000000000001, -0.5, -0.5, -0.5], [0.5000000000000001, -0.5, 0.5, -0.5], [-0.5000000000000001, 0.5, -0.5, 0.5], [0.5000000000000001, 0.5, -0.5, -0.5], [-0.5000000000000001, -0.5, 0.5, 0.5], [0.5000000000000001, -0.5, -0.5, 0.5], [-0.5000000000000001, 0.5, 0.5, -0.5], [0.5000000000000001, -0.5, -0.5, -0.5], [-0.5000000000000001, 0.5, 0.5, 0.5], [0.5000000000000001, -0.5, 0.5, 0.5], [-0.5000000000000001, 0.5, -0.5, -0.5], [0.5000000000000001, 0.5, 0.5, -0.5], [-0.5000000000000001, -0.5, -0.5, 0.5], [0.5000000000000001, 0.5, -0.5, 0.5], [-0.5000000000000001, -0.5, 0.5, -0.5], [6.123233995736766e-17, 0.7071067811865475, 0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, -0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, 0.7071067811865475, 0.0], [0.7071067811865476, -0.0, -0.0, -0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, -0.7071067811865475], [0.7071067811865476, -0.7071067811865475, -0.0, -0.0], [-0.7071067811865476, 0.7071067811865475, 0.0, 0.0], [6.123233995736766e-17, 0.0, 0.7071067811865475, 0.7071067811865475], [-6.123233995736766e-17, 0.0, -0.7071067811865475, -0.7071067811865475], [6.123233995736766e-17, 0.0, 0.7071067811865475, -0.7071067811865475], [-6.123233995736766e-17, 0.0, -0.7071067811865475, 0.7071067811865475], [0.7071067811865476, 0.7071067811865475, 0.0, 0.0], [-0.7071067811865476, -0.7071067811865475, 0.0, 0.0], [0.7071067811865476, 0.0, 0.7071067811865475, 0.0], [-0.7071067811865476, 0.0, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, 0.0, 0.7071067811865475], [-6.123233995736766e-17, -0.7071067811865475, 0.0, -0.7071067811865475], [0.7071067811865476, -0.0, -0.7071067811865475, -0.0], [-0.7071067811865476, 0.0, 0.7071067811865475, 0.0], [6.123233995736766e-17, -0.7071067811865475, 0.0, 0.7071067811865475], [-6.123233995736766e-17, 0.7071067811865475, 0.0, -0.7071067811865475]], "symGroup": "m-3m"}, {"Name": "Snub Cuboctahedron", "ShortName": "A12", "vertices": [[-1.14261, -0.337754, -0.621226], [-1.14261, 0.337754, 0.621226], [-1.14261, -0.621226, 0.337754], [-1.14261, 0.621226, -0.337754], [1.14261, -0.337754, 0.621226], [1.14261, 0.337754, -0.621226], [1.14261, -0.621226, -0.337754], [1.14261, 0.621226, 0.337754], [-0.337754, -1.14261, 0.621226], [-0.337754, 1.14261, -0.621226], [-0.337754, -0.621226, -1.14261], [-0.337754, 0.621226, 1.14261], [0.337754, -1.14261, -0.621226], [0.337754, 1.14261, 0.621226], [0.337754, -0.621226, 1.14261], [0.337754, 0.621226, -1.14261], [-0.621226, -1.14261, -0.337754], [-0.621226, 1.14261, 0.337754], [-0.621226, -0.337754, 1.14261], [-0.621226, 0.337754, -1.14261], [0.621226, -1.14261, 0.337754], [0.621226, 1.14261, -0.337754], [0.621226, -0.337754, -1.14261], [0.621226, 0.337754, 1.14261]], "invariantQ": [[-1.0, 7.07050159149938e-17, 7.07050159149938e-17, 7.07050159149938e-17], [1.0, -7.07050159149938e-17, -7.07050159149938e-17, -7.07050159149938e-17], [6.123233995736766e-17, 0.0, 0.0, 1.0], [-6.123233995736766e-17, 0.0, 0.0, -1.0], [6.123233995736766e-17, 0.0, 1.0, 0.0], [-6.123233995736766e-17, 0.0, -1.0, 0.0], [6.123233995736766e-17, 1.0, 0.0, 0.0], [-6.123233995736766e-17, -1.0, 0.0, 0.0], [0.5000000000000001, 0.5, 0.5, 0.5], [-0.5000000000000001, -0.5, -0.5, -0.5], [0.5000000000000001, -0.5, 0.5, -0.5], [-0.5000000000000001, 0.5, -0.5, 0.5], [0.5000000000000001, 0.5, -0.5, -0.5], [-0.5000000000000001, -0.5, 0.5, 0.5], [0.5000000000000001, -0.5, -0.5, 0.5], [-0.5000000000000001, 0.5, 0.5, -0.5], [0.5000000000000001, -0.5, -0.5, -0.5], [-0.5000000000000001, 0.5, 0.5, 0.5], [0.5000000000000001, -0.5, 0.5, 0.5], [-0.5000000000000001, 0.5, -0.5, -0.5], [0.5000000000000001, 0.5, 0.5, -0.5], [-0.5000000000000001, -0.5, -0.5, 0.5], [0.5000000000000001, 0.5, -0.5, 0.5], [-0.5000000000000001, -0.5, 0.5, -0.5], [6.123233995736766e-17, 0.7071067811865475, 0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, -0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, 0.7071067811865475, 0.0], [0.7071067811865476, -0.0, -0.0, -0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, -0.7071067811865475], [0.7071067811865476, -0.7071067811865475, -0.0, -0.0], [-0.7071067811865476, 0.7071067811865475, 0.0, 0.0], [6.123233995736766e-17, 0.0, 0.7071067811865475, 0.7071067811865475], [-6.123233995736766e-17, 0.0, -0.7071067811865475, -0.7071067811865475], [6.123233995736766e-17, 0.0, 0.7071067811865475, -0.7071067811865475], [-6.123233995736766e-17, 0.0, -0.7071067811865475, 0.7071067811865475], [0.7071067811865476, 0.7071067811865475, 0.0, 0.0], [-0.7071067811865476, -0.7071067811865475, 0.0, 0.0], [0.7071067811865476, 0.0, 0.7071067811865475, 0.0], [-0.7071067811865476, 0.0, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, 0.0, 0.7071067811865475], [-6.123233995736766e-17, -0.7071067811865475, 0.0, -0.7071067811865475], [0.7071067811865476, -0.0, -0.7071067811865475, -0.0], [-0.7071067811865476, 0.0, 0.7071067811865475, 0.0], [6.123233995736766e-17, -0.7071067811865475, 0.0, 0.7071067811865475], [-6.123233995736766e-17, 0.7071067811865475, 0.0, -0.7071067811865475]], "symGroup": "432"}, {"Name": "Tetrakis Hexahedron", "ShortName": "C04", "vertices": [[-0.6666666372637077, -0.6666669999999999, -0.6666666372637078], [-0.6666666372637077, 0.6666670000000001, -0.6666666372637076], [-1.000000309448952, 8.326672684688674e-17, 0.0], [-2.7755575615628914e-17, 2.220446049250313e-16, -1.0000003094489522], [-0.6666666372637076, -0.666667, 0.6666666372637075], [-0.6666666372637076, 0.6666669999999999, 0.6666666372637078], [0.0, -1.0, -2.220446049250313e-16], [0.0, 1.0, 2.220446049250313e-16], [0.6666666372637076, -0.6666669999999999, -0.6666666372637078], [0.6666666372637076, 0.666667, -0.6666666372637075], [2.7755575615628914e-17, -2.220446049250313e-16, 1.0000003094489522], [1.000000309448952, -8.326672684688674e-17, 0.0], [0.6666666372637077, -0.6666670000000001, 0.6666666372637076], [0.6666666372637077, 0.6666669999999999, 0.6666666372637078]], "invariantQ": [[-1.0, 7.07050159149938e-17, 7.07050159149938e-17, 7.07050159149938e-17], [1.0, -7.07050159149938e-17, -7.07050159149938e-17, -7.07050159149938e-17], [6.123233995736766e-17, 0.0, 0.0, 1.0], [-6.123233995736766e-17, 0.0, 0.0, -1.0], [6.123233995736766e-17, 0.0, 1.0, 0.0], [-6.123233995736766e-17, 0.0, -1.0, 0.0], [6.123233995736766e-17, 1.0, 0.0, 0.0], [-6.123233995736766e-17, -1.0, 0.0, 0.0], [0.5000000000000001, 0.5, 0.5, 0.5], [-0.5000000000000001, -0.5, -0.5, -0.5], [0.5000000000000001, -0.5, 0.5, -0.5], [-0.5000000000000001, 0.5, -0.5, 0.5], [0.5000000000000001, 0.5, -0.5, -0.5], [-0.5000000000000001, -0.5, 0.5, 0.5], [0.5000000000000001, -0.5, -0.5, 0.5], [-0.5000000000000001, 0.5, 0.5, -0.5], [0.5000000000000001, -0.5, -0.5, -0.5], [-0.5000000000000001, 0.5, 0.5, 0.5], [0.5000000000000001, -0.5, 0.5, 0.5], [-0.5000000000000001, 0.5, -0.5, -0.5], [0.5000000000000001, 0.5, 0.5, -0.5], [-0.5000000000000001, -0.5, -0.5, 0.5], [0.5000000000000001, 0.5, -0.5, 0.5], [-0.5000000000000001, -0.5, 0.5, -0.5], [6.123233995736766e-17, 0.7071067811865475, 0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, -0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, 0.7071067811865475, 0.0], [0.7071067811865476, -0.0, -0.0, -0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, -0.7071067811865475], [0.7071067811865476, -0.7071067811865475, -0.0, -0.0], [-0.7071067811865476, 0.7071067811865475, 0.0, 0.0], [6.123233995736766e-17, 0.0, 0.7071067811865475, 0.7071067811865475], [-6.123233995736766e-17, 0.0, -0.7071067811865475, -0.7071067811865475], [6.123233995736766e-17, 0.0, 0.7071067811865475, -0.7071067811865475], [-6.123233995736766e-17, 0.0, -0.7071067811865475, 0.7071067811865475], [0.7071067811865476, 0.7071067811865475, 0.0, 0.0], [-0.7071067811865476, -0.7071067811865475, 0.0, 0.0], [0.7071067811865476, 0.0, 0.7071067811865475, 0.0], [-0.7071067811865476, 0.0, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, 0.0, 0.7071067811865475], [-6.123233995736766e-17, -0.7071067811865475, 0.0, -0.7071067811865475], [0.7071067811865476, -0.0, -0.7071067811865475, -0.0], [-0.7071067811865476, 0.0, 0.7071067811865475, 0.0], [6.123233995736766e-17, -0.7071067811865475, 0.0, 0.7071067811865475], [-6.123233995736766e-17, 0.7071067811865475, 0.0, -0.7071067811865475]], "symGroup": "m-3m"}, {"Name": "Elongated Square Gyrobicupola", "ShortName": "J37", "vertices": [[-0.6532814824381883, 0.27059805007309845, -1.2071099999999997], [0.27059805007309845, 0.6532814824381883, -1.2071099999999997], [-0.6532816845954584, -0.2705981338093816, 1.2071099999999997], [0.6532816845954584, 0.2705981338093816, 1.2071099999999997], [-0.27059805007309845, -0.6532814824381883, -1.2071099999999997], [0.6532814824381883, -0.27059805007309845, -1.2071099999999997], [-0.2705981338093816, 0.6532816845954584, 1.2071099999999997], [0.2705981338093816, -0.6532816845954584, 1.2071099999999997], [1.3065659386722441, 1.2317865800692829e-06, -0.5], [1.3065659386722441, 1.2317865800692829e-06, 0.5], [0.9238825063071543, 0.9238807642978668, -0.5], [0.9238825063071543, 0.9238807642978668, 0.5], [-1.2317865800692829e-06, 1.3065659386722441, -0.5], [-1.2317865800692829e-06, 1.3065659386722441, 0.5], [-0.9238807642978668, 0.9238825063071543, -0.5], [-0.9238807642978668, 0.9238825063071543, 0.5], [-1.3065659386722441, -1.2317865800692829e-06, -0.5], [-1.3065659386722441, -1.2317865800692829e-06, 0.5], [-0.9238825063071543, -0.9238807642978668, -0.5], [-0.9238825063071543, -0.9238807642978668, 0.5], [1.2317865800692829e-06, -1.3065659386722441, -0.5], [1.2317865800692829e-06, -1.3065659386722441, 0.5], [0.9238807642978668, -0.9238825063071543, -0.5], [0.9238807642978668, -0.9238825063071543, 0.5]], "invariantQ": [[-1.0, 1.2246467991473532e-16, 0.0, 0.0], [1.0, -1.2246467991473532e-16, 0.0, 0.0], [6.123233995736766e-17, 0.0, 0.0, 1.0], [-6.123233995736766e-17, 0.0, 0.0, -1.0], [0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, -0.7071067811865475], [0.7071067811865476, -0.0, -0.0, -0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [6.123233995736766e-17, 0.0, 1.0, 0.0], [-6.123233995736766e-17, 0.0, -1.0, 0.0], [6.123233995736766e-17, 1.0, 0.0, 0.0], [-6.123233995736766e-17, -1.0, 0.0, 0.0], [6.123233995736766e-17, 0.7071067811865475, 0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, -0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, 0.7071067811865475, 0.0]], "symGroup": "4_mmm"}, {"Name": "Gyroelongated Square Bicupola", "ShortName": "J45", "vertices": [[-0.39284760075979186, -0.5879379831464158, -1.13725], [0.39284760075979186, 0.5879379831464158, -1.13725], [-0.7258858436540914, -1.0863649366500132, 0.4301479999999999], [0.7258858436540914, 1.0863649366500132, 0.4301479999999999], [-0.5879379831464158, 0.39284760075979186, -1.13725], [0.5879379831464158, -0.39284760075979186, -1.13725], [-0.39284710587998073, 0.5879374915471228, 1.13725], [0.39284710587998073, -0.5879374915471228, 1.13725], [1.2814604002163266, -0.2548995778290194, -0.4301479999999999], [-1.0863649366500132, 0.7258858436540914, 0.4301479999999999], [1.0863649366500132, -0.7258858436540914, 0.4301479999999999], [-1.2814583722962256, -0.2548979185319254, 0.4301479999999999], [-0.2548979185319254, 1.2814583722962256, 0.4301479999999999], [0.2548979185319254, -1.2814583722962256, 0.4301479999999999], [1.2814583722962256, 0.2548979185319254, 0.4301479999999999], [-1.0863691901315646, -0.7258901671967242, -0.4301479999999999], [1.0863691901315646, 0.7258901671967242, -0.4301479999999999], [-0.2548995778290194, -1.2814604002163266, -0.4301479999999999], [0.2548995778290194, 1.2814604002163266, -0.4301479999999999], [0.7258901671967242, -1.0863691901315646, -0.4301479999999999], [-0.7258901671967242, 1.0863691901315646, -0.4301479999999999], [-1.2814604002163266, 0.2548995778290194, -0.4301479999999999], [-0.5879374915471228, -0.39284710587998073, 1.13725], [0.5879374915471228, 0.39284710587998073, 1.13725]], "invariantQ": [[-1.0, 1.2246467991473532e-16, 0.0, 0.0], [1.0, -1.2246467991473532e-16, 0.0, 0.0], [6.123233995736766e-17, 0.0, 0.0, 1.0], [-6.123233995736766e-17, 0.0, 0.0, -1.0], [0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, -0.7071067811865475], [0.7071067811865476, -0.0, -0.0, -0.7071067811865475], [-0.7071067811865476, 0.0, 0.0, 0.7071067811865475], [6.123233995736766e-17, 0.0, 1.0, 0.0], [-6.123233995736766e-17, 0.0, -1.0, 0.0], [6.123233995736766e-17, 1.0, 0.0, 0.0], [-6.123233995736766e-17, -1.0, 0.0, 0.0], [6.123233995736766e-17, 0.7071067811865475, 0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, -0.7071067811865475, 0.0], [6.123233995736766e-17, 0.7071067811865475, -0.7071067811865475, 0.0], [-6.123233995736766e-17, -0.7071067811865475, 0.7071067811865475, 0.0]], "symGroup": "422"}, {"Name": "Triangular Orthobicupola", "ShortName": "J27", "vertices": [[0.0, -1.0, 0.0], [0.0, 1.0, 0.0], [-0.288675, -0.5, -0.816497], [-0.288675, -0.5, 0.816497], [-0.288675, 0.5, -0.816497], [-0.288675, 0.5, 0.816497], [0.57735, 0.0, -0.816497], [0.57735, 0.0, 0.816497], [-0.866025, -0.5, 0.0], [-0.866025, 0.5, 0.0], [0.866025, -0.5, 0.0], [0.866025, 0.5, 0.0]], "invariantQ": [[-1.0, 1.2246467991473532e-16, 0.0, 0.0], [1.0, -1.2246467991473532e-16, 0.0, 0.0], [0.5000000000000001, 0.0, 0.0, 0.8660254037844386], [-0.5000000000000001, 0.0, 0.0, -0.8660254037844386], [0.5000000000000001, -0.0, -0.0, -0.8660254037844386], [-0.5000000000000001, 0.0, 0.0, 0.8660254037844386], [6.123233995736766e-17, 1.0, 0.0, 0.0], [-6.123233995736766e-17, -1.0, 0.0, 0.0], [6.123233995736766e-17, 0.5000000000000001, 0.8660254037844386, 0.0], [-6.123233995736766e-17, -0.5000000000000001, -0.8660254037844386, 0.0], [6.123233995736766e-17, -0.5000000000000001, 0.8660254037844386, 0.0], [-6.123233995736766e-17, 0.5000000000000001, -0.8660254037844386, 0.0]], "symGroup": "-6m2"}, {"Name": "Disphenocingulum", "ShortName": "J90", "vertices": [[0.35355601067474907, 0.3535516520557829, -1.104432684890909], [0.7965444780135226, 0.7965417080286229, -0.3249983748890818], [-0.18888844101392216, 0.8959966828939351, -0.46294680405955757], [-0.35355134445029374, -0.35355166835769564, -1.1044398333351966], [0.8959974782871691, -0.1888909639845683, -0.4629443493338358], [-0.8959956111502988, 0.1888914945377682, -0.4629502689498131], [0.18888787050894587, 0.8959953189264307, 0.462947544045344], [-0.796544205629287, 0.7965441058194818, 0.325001085416781], [0.18888920475525983, -0.8959957190578602, -0.4629473194673348], [0.8959945563859366, 0.18888938635067853, 0.4629486551073039], [-0.7965441383314684, -0.7965421279138671, -0.3250052836393526], [0.7965421061481694, -0.7965443197479932, 0.3250017962980476], [-0.8959971279738894, -0.18889007893733817, 0.4629451258318339], [-0.35355308668715957, 0.35355417273804174, 1.1044361680318455], [-0.18889083571971202, -0.8959954651213813, 0.46294528478183583], [0.3535530866871597, -0.353554172738042, 1.104439257793885]], "invariantQ": [[-1.0, 1.2246467991473532e-16, 0.0, 0.0], [1.0, -1.2246467991473532e-16, 0.0, 0.0], [6.123233995736766e-17, 0.0, 0.0, 1.0], [-6.123233995736766e-17, 0.0, 0.0, -1.0], [6.123233995736766e-17, 0.0, 1.0, 0.0], [-6.123233995736766e-17, 0.0, -1.0, 0.0], [6.123233995736766e-17, 1.0, 0.0, 0.0], [-6.123233995736766e-17, -1.0, 0.0, 0.0]], "symGroup": "-42m"}, {"Name": "Triaugmented Dodecahedron", "ShortName": "J61", "vertices": [[0.9341748529796053, 1.144034243545423e-16, 1.004325725425985], [-0.9341753141647604, -1.1440348083343469e-16, -1.0845462197472033], [0.17841151082603823, 1.30902, 0.42697546964196753], [0.17841151082603857, -1.30902, 0.42697546964196753], [-1.044439695772282, 0.8090169999999998, -0.5071974430578344], [-1.044439695772282, -0.8090170000000002, -0.5071974430578344], [-0.46708742648980284, 0.809017, 1.0043292180190952], [-0.4670874264898026, -0.809017, 1.0043292180190952], [1.222848228767508, 0.5000000000000002, -0.5072002805981312], [1.222848228767508, -0.4999999999999999, -0.5072002805981312], [-1.222848689952663, 0.4999999999999999, 0.42697978627691274], [-1.222848689952663, -0.5000000000000002, 0.42697978627691274], [-5.213895438560545e-06, -6.385180359922157e-22, -1.441371193294186], [1.0444392345871267, 0.8090170000000002, 0.42697694873661596], [1.0444392345871267, -0.8090169999999998, 0.42697694873661596], [4.75271028338442e-06, 5.820391435821441e-22, 1.3611506989729676], [0.46708696530464744, 0.809017, -1.0845497123403136], [0.46708696530464766, -0.809017, -1.0845497123403136], [1.6101493334317312, 1.971864227336414e-16, 0.26740091799819954], [-0.17841197201119366, 1.30902, -0.5071959639631859], [-0.17841197201119333, -1.30902, -0.5071959639631859], [-0.8050723668895453, -1.3944299999999998, 0.2674020140376014], [-0.8050723668895458, 1.3944299999999998, 0.2674020140376014]], "invariantQ": [[-1.0, 1.2246467991473532e-16, 0.0, 0.0], [1.0, -1.2246467991473532e-16, 0.0, 0.0], [0.5000000000000001, 0.0, 0.0, 0.8660254037844386], [-0.5000000000000001, 0.0, 0.0, -0.8660254037844386], [0.5000000000000001, -0.0, -0.0, -0.8660254037844386], [-0.5000000000000001, 0.0, 0.0, 0.8660254037844386]], "symGroup": "3m"}, {"Name": "Metabiaugmented Dodecahedron", "ShortName": "J60", "vertices": [[0.809020263100338, 0.809014770920135, -0.7306723621609981], [-0.809020381049436, -0.8090148846798465, 0.8873643672980063], [-0.500001872529516, -1.1375971167848498e-07, -1.2306732345547478], [0.8090211392595555, 0.8090154795571065, 0.887365000694301], [-1.3090206159641702, -0.499997989532927, 0.0783484999109888], [-0.5000017545804181, -5.551115123125783e-17, 1.387365122244093], [-0.8090169682739694, 0.8090217652662582, -0.7306707358516563], [1.893109782657998e-06, 1.3090197547991853, 0.5783458864814476], [0.8090156739702739, -0.8090220595484701, -0.7306709298512174], [1.3090168243686908, -0.5000063098590772, 0.07834582896417461], [-1.3090169423177889, 0.5000061960993655, 0.07834617617283365], [-0.8090157919193716, 0.8090219457887583, 0.8873629349882258], [-6.505952075075427e-06, -1.309018595078542, 0.5783487083557945], [0.50000163663132, -1.1375971159521825e-07, -1.2306731171070844], [1.3090204980150724, 0.4999978757732155, 0.0783435052260198], [6.3880029770313485e-06, 1.3090184813188306, -0.42165670321878623], [-2.011058880757588e-06, -1.3090198685588967, -0.42165388134443926], [0.8090168503248716, -0.8090218790259699, 0.8873627409886647], [1.3944297656490652, -2.4180643696247728e-06, -0.7834603558552322], [-0.8090212572086535, -0.8090155933168179, -0.7306729955572926], [0.5000017545804181, 5.551115123125783e-17, 1.3873652396917564], [-1.3944285825945202, 3.5500535754962037e-06, -0.7834596906309649]], "invariantQ": [[-1.0, 1.2246467991473532e-16, 0.0, 0.0], [1.0, -1.2246467991473532e-16, 0.0, 0.0], [6.123233995736766e-17, 0.0, 0.0, 1.0], [-6.123233995736766e-17, 0.0, 0.0, -1.0]], "symGroup": "mm2"}, {"Name": "Squashed Dodecahedron", "ShortName": "O15", "vertices": [[-0.7071091731492744, 1.2247443003541598, -0.999998227589546], [-1.414214263818451, -3.3306690738754696e-16, -0.4999965362176715], [-4.082480096722562e-06, -1.0072534643921875e-16, -1.4999999189614197], [0.707105090667494, 1.2247443003541598, -0.5000023097140125], [-0.7071037298447225, 1.2247443003541598, 1.0000008399028295], [-1.4142156246412232, -3.3306690738754696e-16, 0.5000058842909731], [0.7071098535690792, 1.224747835888066, 0.5000001107779655], [-0.7071091731492739, -1.2247443003541603, -0.9999982275895456], [1.4142101813366716, 3.3306690738754696e-16, -1.0000040010858866], [4.082476730845547e-06, 1.3692414647307853e-16, 1.4999986822771434], [0.7071050906674945, -1.2247443003541596, -0.5000023097140124], [-0.7071037298447219, -1.22474430035416, 1.0000008399028297], [1.4142156246412232, 3.3306690738754696e-16, 0.9999950664064888], [0.7071098535690793, -1.2247478358880657, 0.5000001107779657]], "invariantQ": [[-1.0, 1.2246467991473532e-16, 0.0, 0.0], [1.0, -1.2246467991473532e-16, 0.0, 0.0], [0.5000000000000001, 0.0, 0.0, 0.8660254037844386], [-0.5000000000000001, 0.0, 0.0, -0.8660254037844386], [0.5000000000000001, -0.0, -0.0, -0.8660254037844386], [-0.5000000000000001, 0.0, 0.0, 0.8660254037844386], [6.123233995736766e-17, 1.0, 0.0, 0.0], [-6.123233995736766e-17, -1.0, 0.0, 0.0], [6.123233995736766e-17, 0.5000000000000001, 0.8660254037844386, 0.0], [-6.123233995736766e-17, -0.5000000000000001, -0.8660254037844386, 0.0], [6.123233995736766e-17, -0.5000000000000001, 0.8660254037844386, 0.0], [-6.123233995736766e-17, 0.5000000000000001, -0.8660254037844386, 0.0]], "symGroup": "-6m2"}, {"Name": "Elongated Triangular Cupola", "ShortName": "J18", "vertices": [[0.0, -1.0, -0.7632994], [0.0, -1.0, 0.2367006], [0.0, 1.0, -0.7632994], [0.0, 1.0, 0.2367006], [-0.288675, -0.5, 1.0531976], [-0.288675, 0.5, 1.0531976], [0.57735, 0.0, 1.0531976], [-0.866025, -0.5, -0.7632994], [-0.866025, -0.5, 0.2367006], [-0.866025, 0.5, -0.7632994], [-0.866025, 0.5, 0.2367006], [0.866025, -0.5, -0.7632994], [0.866025, -0.5, 0.2367006], [0.866025, 0.5, -0.7632994], [0.866025, 0.5, 0.2367006]], "invariantQ": [[-1.0, 1.2246467991473532e-16, 0.0, 0.0], [1.0, -1.2246467991473532e-16, 0.0, 0.0], [0.5000000000000001, 0.0, 0.0, 0.8660254037844386], [-0.5000000000000001, 0.0, 0.0, -0.8660254037844386], [0.5000000000000001, -0.0, -0.0, -0.8660254037844386], [-0.5000000000000001, 0.0, 0.0, 0.8660254037844386]], "symGroup": "3m"}]'
